Aggregator function

These functions are used internally and none are exported.

They may depend on the particulars of the simulation code (e.g. internal units).

GalaxyInspector.integrateQty — Method

integrateQty(data_dict::Dict, quantity::Symbol; <keyword arguments>)::Number

Compute quantity for the whole system of cell/particles in data_dict.

Arguments

data_dict::Dict: Data dictionary (see makeDataDict for the canonical description).

quantity::Symbol: Target quantity. See plotParams for possibilities.

agg_function::Union{Function,Symbol}=:default: If quantity is one the the listed symbols in DERIVED_QTY, SFM_STELLAR_QTY or SFM_GAS_QTY, you can pass an agg_function to accumulate the values given by scatterQty. If agg_function is left as :default integrateQty will try to compute the most reasonable global value for quantity.

Returns

The value of quantity for the whole system of cell/particles in data_dict.

GalaxyInspector.scatterQty — Method

scatterQty(data_dict::Dict, quantity::Symbol)::Vector{<:Number}

Compute quantity for each cell/particle in data_dict.

Arguments

data_dict::Dict: Data dictionary (see makeDataDict for the canonical description).

quantity::Symbol: Target quantity. See plotParams for possibilities; only quantities well defined for each cell/particle individually are possible.

Returns

The values of quantity for every cell/particle.