Computation of energy related quantities

These functions are used internally and none are exported.

They may depend on the particulars of the simulation code (e.g. internal units).

GalaxyInspector.computeKineticEnergy — Method

computeKineticEnergy(data_dict::Dict, component::Symbol)::Vector{<:Unitful.Energy}

Compute the kinetic energy.

Arguments

data_dict::Dict: Data dictionary (see makeDataDict for the canonical description). This function requires the following blocks to be present, depending on the value of component:

If component ∈ [:stellar, :darkmatter, :blackhole, :gas]:

component => ["VEL ", "MASS"]

If component == :Z_stellar

:stellar => ["VEL ", "MASS", "GZ2 "]

If component ∈ [:hydrogen, :helium]:

:gas => ["VEL ", "MASS"]

If component == :Z_gas:

:gas => ["VEL ", "MASS", "GZ "]

If component ∈ [:ionized, :neutral]:

:gas => ["VEL ", "MASS", "NH ", "NHP "]

If component ∈ [:bratomic, :brmolecular]:

:gas => ["VEL ", "MASS", "NH ", "NHP ", "PRES"]

If component ∈ [:odeionized, :odeatomic, :odemetals, :odedust, :ode_neutral]:

:gas => ["VEL ", "MASS", "NH ", "NHP ", "FRAC", "RHO ", "GZ "]

If component ∈ [:odemolecular, :odestellar]:

:gas => ["VEL ", "MASS", "FRAC", "RHO "]

component::Symbol: Target component. It can only be one of the elements of COMPONENTS.

Returns

The kinetic energy of each cell/particle.

GalaxyInspector.computeKineticEnergy — Method

computeKineticEnergy(
    masses::Vector{<:Unitful.Mass},
    velocities::Matrix{<:Unitful.Velocity},
)::Vector{<:Unitful.Energy}

Compute the kinetic energy.

Arguments

masses::Vector{<:Unitful.Mass}: Mass of every cell/particle.

velocities::Matrix{<:Unitful.Velocity}: Velocities of the cells/particles. Each column is a cell/particle and each row a dimension.

Returns

The kinetic energy of each cell/particle.

GalaxyInspector.computePotentialEnergy — Method

computePotentialEnergy(data_dict::Dict, component::Symbol)::Vector{<:Unitful.Energy}

Compute the gravitational potencial energy.

Arguments

data_dict::Dict: Data dictionary (see makeDataDict for the canonical description). This function requires the following blocks to be present, depending on the value of component:

If component ∈ [:stellar, :darkmatter, :blackhole, :gas]:

component => ["POT ", "MASS"]

If component == :Z_stellar

:stellar => ["POT ", "MASS", "GZ2 "]

If component ∈ [:hydrogen, :helium]:

:gas => ["POT ", "MASS"]

If component == :Z_gas:

:gas => ["POT ", "MASS", "GZ "]

If component ∈ [:ionized, :neutral]:

:gas => ["POT ", "MASS", "NH ", "NHP "]

If component ∈ [:bratomic, :brmolecular]:

:gas => ["POT ", "MASS", "NH ", "NHP ", "PRES"]

If component ∈ [:odeionized, :odeatomic, :odemetals, :odedust, :ode_neutral]:

:gas => ["POT ", "MASS", "NH ", "NHP ", "FRAC", "RHO ", "GZ "]

If component ∈ [:odemolecular, :odestellar]:

:gas => ["POT ", "MASS", "FRAC", "RHO "]

component::Symbol: Target component. It can only be one of the elements of COMPONENTS.

Returns

The gravitational potencial energy of each cell/particle.

GalaxyInspector.computePotentialEnergy — Method

computePotentialEnergy(
    masses::Vector{<:Unitful.Mass},
    potential::Vector{<:SpecificEnergy},
)::Vector{<:Unitful.Energy}

Compute the gravitational potencial energy.

Arguments

masses::Vector{<:Unitful.Mass}: Mass of every cell/particle.

potential::Vector{<:SpecificEnergy}: Specific potential energy of every cell/particle.

Returns

The gravitational potencial energy of each cell/particle.

GalaxyInspector.computeTemperature — Method

computeTemperature(
    internal_energy::Vector{<:SpecificEnergy},
    electron_fraction::Vector{Float32},
)::Vector{<:Unitful.Temperature}

Compute the gas temperature.

Arguments

internal_energy::Vector{<:SpecificEnergy}: Specific internal energy of every gas cell.

electron_fraction::Vector{Float32}: Number fraction of electrons in every gas cell.

Returns

The temperature of each gas cell.

GalaxyInspector.computeTotalEnergy — Method

computeTotalEnergy(data_dict::Dict, component::Symbol)::Vector{<:Unitful.Energy}

Compute the total energy (kinetic + potential).

Arguments

data_dict::Dict: Data dictionary (see makeDataDict for the canonical description). This function requires the following blocks to be present, depending on the value of component:

If component ∈ [:stellar, :darkmatter, :blackhole, :gas]:

component => ["VEL ", "POT ", "MASS"]

If component == :Z_stellar

:stellar => ["VEL ", "POT ", "MASS", "GZ2 "]

If component ∈ [:hydrogen, :helium]:

:gas => ["VEL ", "POT ", "MASS"]

If component == :Z_gas:

:gas => ["VEL ", "POT ", "MASS", "GZ "]

If component ∈ [:ionized, :neutral]:

:gas => ["VEL ", "POT ", "MASS", "NH ", "NHP "]

If component ∈ [:bratomic, :brmolecular]:

:gas => ["VEL ", "POT ", "MASS", "NH ", "NHP ", "PRES"]

If component ∈ [:odeionized, :odeatomic, :odemetals, :odedust, :ode_neutral]:

:gas => ["VEL ", "POT ", "MASS", "NH ", "NHP ", "FRAC", "RHO ", "GZ "]

If component ∈ [:odemolecular, :odestellar]:

:gas => ["VEL ", "POT ", "MASS", "FRAC", "RHO "]

component::Symbol: Target component. It can only be one of the elements of COMPONENTS.

Returns

The total energy of each cell/particle.