Computation of characteristic masses and mass related quantities

These functions are used internally and none are exported.

They may depend on the particulars of the simulation code (e.g. internal units).

GalaxyInspector._compute_fraction — Method

_compute_fraction(data_dict::Dict, component::Symbol)::Vector{Float64}

Compute the fraction of a given non :ode component in each cell/particle.

Arguments

data_dict::Dict: Data dictionary (see makeDataDict for the canonical description). This function requires the following blocks to be present, depending on the value of component:

If component == :Z_stellar

:stellar => ["GZ2 "]

If component ∈ [:gas, :hydrogen, :helium]:

:gas => ["MASS"]

If component == :Z_gas:

:gas => ["MASS", "GZ "]

If component ∈ [:ionized, :neutral]:

:gas => ["MASS", "NH ", "NHP "]

If component ∈ [:bratomic, :brmolecular]:

:gas => ["MASS", "NH ", "NHP ", "PRES"]

component::Symbol: Target component. It can only be one of the non :ode elements of COMPONENTS (except :stellar, :darkmatter, and :blackhole).

Returns

The fraction of component in each cell/particle.

References

L. Blitz et al. (2006). The Role of Pressure in GMC Formation II: The H2-Pressure Relation. The Astrophysical Journal, 650(2), 933. doi:10.1086/505417

GalaxyInspector._compute_fraction — Method

_compute_fraction(
    data_dict::Dict,
    component::Symbol,
    ode_ic::F,
)::Vector{Float64} where {F<:Function}

Compute the fraction of a given :ode component in each cell/particle.

Arguments

data_dict::Dict: Data dictionary (see makeDataDict for the canonical description). This function requires the following blocks to be present, depending on the value of component:

If component ∈ [:odeionized, :odeatomic, :odemetals, :odedust, :odeneutral, :odecold]:

:gas => ["MASS", "NH ", "NHP ", "FRAC", "RHO ", "GZ "]

If component ∈ [:odemolecular, :odestellar, :odemolecularstellar]:

:gas => ["MASS", "FRAC", "RHO "]

component::Symbol: Target component. It can only be one of the :ode elements of COMPONENTS.

ode_ic::F: Fuction that gives the initial value of component for the i-th cell/particle. It must have the signature ode_ic(i::Int)::Float64.

Returns

The fraction of component in each cell/particle.

GalaxyInspector.computeAbundance — Method

computeAbundance(
    data_dict::Dict,
    type::Symbol,
    element::Symbol;
    <keyword arguments>
)::Vector{Float64}

Compute the abundance of a given element in each cell/particle. The abundance is defined as $n_X / n_H$ where $n_X$ is the number of atoms of element $X$ and $n_H$ the number of hydrogen atoms.

Arguments

data_dict::Dict: Data dictionary (see makeDataDict for the canonical description). This function requires the following blocks to be present, depending on the value of type:

If type == :stellar:

:stellar => ["MASS", "GME2"]

If type == :gas:

:gas => ["MASS", "GMET"]

type::Symbol: For which cell/particle type the abundance will be calculated. The possibilities are :stellar and :gas.

element::Symbol: Target element. The possibilities are the keys of ELEMENT_INDEX.

solar::Bool=false: If the result will be normalized to the solar abundance.

Returns

The abundance of element in each cell/particle.

GalaxyInspector.computeClumpingFactor — Method

computeClumpingFactor(data_dict::Dict, component::Symbol)::Float64

Compute the clumping factor,

\[C_\rho = \frac{\langle \rho^2 \rangle}{\langle \rho \rangle^2} \, .\]

Arguments

data_dict::Dict: Data dictionary (see makeDataDict for the canonical description). This function requires the following blocks to be present, depending on the value of component:

If component ∈ [:gas, :hydrogen, :helium]:

:gas => ["MASS", "RHO "]

If component == :Z_gas:

:gas => ["MASS", "GZ ", "RHO "]

If component ∈ [:ionized, :neutral]:

:gas => ["MASS", "NH ", "NHP ", "RHO "]

If component ∈ [:bratomic, :brmolecular]:

:gas => ["MASS", "NH ", "NHP ", "PRES", "RHO "]

If component ∈ [:odeionized, :odeatomic, :odemetals, :odedust, :odeneutral, :odecold]:

:gas => ["MASS", "NH ", "NHP ", "FRAC", "RHO ", "GZ "]

If component ∈ [:odemolecular, :odestellar, :odemolecularstellar]:

:gas => ["MASS", "FRAC", "RHO "]

component::Symbol: Target component. It can only be one of the gas elements of COMPONENTS.

Returns

The clumping factor.

GalaxyInspector.computeClumpingFactor — Method

computeClumpingFactor(density::Vector{<:Number})::Float64

Compute the clumping factor,

\[C_\rho = \frac{\langle \rho^2 \rangle}{\langle \rho \rangle^2} \, .\]

Arguments

density::Vector{<:Number}: The density of the cells/particles.

Returns

The clumping factor.

GalaxyInspector.computeDiskAccretion — Method

computeDiskAccretion(
    present_dd::Dict,
    past_dd::Dict;
    <keyword arguments>
)::NTuple{3,Unitful.Mass}

Compute the inflow, outflow, or net gain of mass for a given galactic disk, between two snapshots.

Note

It is assumed that the center of the disk is the origin.

Arguments

present_dd::Dict: Data dictionary, for the present snapshot (see makeDataDict for the canonical description). This function requires the following blocks to be present for every cell/particle that you want to be taken into account:

cell/particle type => ["POS ", "ID ", "MASS"]

:group => ["GRCrit200", "GNsubs", "GPos"]

:subhalo => ["S_Pos"]

:tracer => ["PAID", "TRID"]

past_dd::Dict: Data dictionary, for the past snapshot (see makeDataDict for the canonical description). This function requires the following blocks to be present for every cell/particle that you want to be taken into account:

cell/particle type => ["POS ", "ID ", "MASS"]

:group => ["GRCrit200", "GNsubs", "GPos"]

:subhalo => ["S_Pos"]

:tracer => ["PAID", "TRID"]

component::Symbol: Component to compute the accreted mass for. The options are:

:dark_matter -> Dark matter.

:black_hole -> Black holes.

:gas -> Gas.

:stellar -> Stars.

max_r::Unitful.Length=DISK_R: Radius of the disk.

max_z::Unitful.Length=5.0u"kpc": Half height of the disk.

tracers::Bool=false: Whether to compute the accretion using tracer particles (true) or the actual component particles (false).

Returns

A tuple with three elements:

The net increase in mass.

The inflow mass.

The outflow mass.

GalaxyInspector.computeEfficiencyFF — Method

computeEfficiencyFF(data_dict::Dict, component::Symbol)::Vector{Float64}

Compute the star formation efficiency per free-fall time, according to the definition in eq. 1 of Krumholz et al. (2012),

\[\epsilon_\mathrm{ff} = \frac{t_\mathrm{ff}}{t_\mathrm{dep}} \, .\]

where

\[t_\mathrm{ff} = \sqrt{\frac{3 \, \pi}{32 \, G \, \rho}} \, ,\]

is the free-fall time, and

\[t_\mathrm{dep} = \frac{M}{\dot{M}_\star} \, ,\]

is the depletion time. $M$ and $\rho$ are the mass and density of the target gas phase, and $\dot{M}_\star$ is the SFR.

Arguments

data_dict::Dict: Data dictionary (see makeDataDict for the canonical description). This function requires the following blocks to be present, depending on the value of component:

If `component == :stellar

:stellar => ["RHOC", "GMAS", "GSFR"]

If component ∈ [:gas, :hydrogen, :helium]:

:gas => ["SFR ", "MASS", "RHO "]

If component == :Z_gas:

:gas => ["SFR ", "MASS", "GZ ", "RHO "]

If component ∈ [:ionized, :neutral]:

:gas => ["SFR ", "MASS", "NH ", "NHP ", "RHO "]

If component ∈ [:bratomic, :brmolecular]:

:gas => ["SFR ", "MASS", "NH ", "NHP ", "PRES", "RHO "]

If component ∈ [:odeionized, :odeatomic, :odemetals, :odedust, :odeneutral, :odecold]:

:gas => ["SFR ", "MASS", "NH ", "NHP ", "FRAC", "RHO ", "GZ "]

If component ∈ [:odemolecular, :odestellar, :odemolecularstellar]:

:gas => ["SFR ", "MASS", "FRAC", "RHO "]

component::Symbol: Target component. It can only be one of the gas elements of COMPONENTS.

Returns

The star formation efficiency per free-fall time.

References

M. R. Krumholz et al. (2012). A UNIVERSAL, LOCAL STAR FORMATION LAW IN GALACTIC CLOUDS, NEARBY GALAXIES, HIGH-REDSHIFT DISKS, AND STARBURSTS. The Astrophysical Journal, 745(1), 69. doi:10.1088/0004-637X/745/1/69

GalaxyInspector.computeEfficiencyFF — Method

computeEfficiencyFF(
    densities::Vector{<:Unitful.Density},
    masses::Vector{<:Unitful.Mass},
    sfrs::Vector{<:Unitful.MassFlow},
)::Vector{Float64}

Compute the star formation efficiency per free-fall time, according to the definition in eq. 1 of Krumholz et al. (2012),

\[\epsilon_\mathrm{ff} = \frac{t_\mathrm{ff}}{t_\mathrm{dep}} \, .\]

where

\[t_\mathrm{ff} = \sqrt{\frac{3 \, \pi}{32 \, G \, \rho}} \, ,\]

is the free-fall time, and

\[t_\mathrm{dep} = \frac{M}{\dot{M}_\star} \, ,\]

is the depletion time. $M$ and $\rho$ are the mass and density of the target gas phase, and $\dot{M}_\star$ is the SFR.

Arguments

densities::Vector{<:Unitful.Density}: Densities of the cells.

masses::Vector{<:Unitful.Mass}: Masses of the cells.

sfrs::Vector{<:Unitful.MassFlow}: The SFR associated to each cell.

Returns

The star formation efficiency per free-fall time.

References

M. R. Krumholz et al. (2012). A UNIVERSAL, LOCAL STAR FORMATION LAW IN GALACTIC CLOUDS, NEARBY GALAXIES, HIGH-REDSHIFT DISKS, AND STARBURSTS. The Astrophysical Journal, 745(1), 69. doi:10.1088/0004-637X/745/1/69

GalaxyInspector.computeElementMass — Method

computeElementMass(
    data_dict::Dict,
    type::Symbol,
    element::Symbol,
)::Vector{<:Unitful.Mass}

Compute the total mass of element in each cell/particle.

Arguments

data_dict::Dict: Data dictionary (see makeDataDict for the canonical description). This function requires the following blocks to be present, depending on the value of type:

If type == :stellar:

:stellar => ["MASS", "GME2"]

If type == :gas:

:gas => ["MASS", "GMET"]

type::Symbol: For which cell/particle type the mass will be calculated. The possibilities are :stellar and :gas.

element::Symbol: Target element. The possibilities are the keys of ELEMENT_INDEX.

Returns

The total mass of element in each cell/particle.

GalaxyInspector.computeFraction — Method

computeFraction(data_dict::Dict, component::Symbol)::Vector{Float64}

Compute the fraction of a given component in each cell/particle.

Arguments

data_dict::Dict: Data dictionary (see makeDataDict for the canonical description). This function requires the following blocks to be present, depending on the value of component:

If component == :Z_stellar

:stellar => ["GZ2 "]

If component ∈ [:gas, :hydrogen, :helium]:

:gas => ["MASS"]

If component == :Z_gas:

:gas => ["MASS", "GZ "]

If component ∈ [:ionized, :neutral]:

:gas => ["MASS", "NH ", "NHP "]

If component ∈ [:bratomic, :brmolecular]:

:gas => ["MASS", "NH ", "NHP ", "PRES"]

If component ∈ [:odeionized, :odeatomic, :odemetals, :odedust, :odeneutral, :odecold]:

:gas => ["MASS", "NH ", "NHP ", "FRAC", "RHO ", "GZ "]

If component ∈ [:odemolecular, :odestellar, :odemolecularstellar]:

:gas => ["MASS", "FRAC", "RHO "]

component::Symbol: Target component. It can only be one of the elements of COMPONENTS (except :stellar, :darkmatter, and :blackhole).

Returns

The fraction of component in each cell/particle.

References

L. Blitz et al. (2006). The Role of Pressure in GMC Formation II: The H2-Pressure Relation. The Astrophysical Journal, 650(2), 933. doi:10.1086/505417

GalaxyInspector.computeFractionWithin — Method

computeFractionWithin(
    quantity::Vector{<:Number},
    masses::Vector{<:Unitful.Mass},
    qty_limits::Tuple{<:Number,<:Number},
)::Float64

Compute the fraction of the total mass "contained" within a given range of quantity.

Arguments

quantity::Vector{<:Number}: Target quantity.

masses::Vector{<:Unitful.Mass}: Masses of the cells/particles.

qty_limits::Tuple{<:Number,<:Number}: Range of the target quantity.

Returns

The fraction of the total mass "contained" within a given range of quantity.

GalaxyInspector.computeGlobalAbundance — Method

computeGlobalAbundance(
    data_dict::Dict,
    type::Symbol,
    element::Symbol;
    <keyword arguments>
)::Float64

Compute the total abundance of a given element, as $n_X / n_H$ where $n_X$ is the number of atoms of element $X$ and $n_H$ the number of hydrogen atoms.

Arguments

data_dict::Dict: Data dictionary (see makeDataDict for the canonical description). This function requires the following blocks to be present, depending on the value of type:

If type == :stellar:

:stellar => ["MASS", "GME2"]

If type == :gas:

:gas => ["MASS", "GMET"]

type::Symbol: For which cell/particle type the abundance will be calculated. The possibilities are :stellar and :gas.

element::Symbol: Target element. The possibilities are the keys of ELEMENT_INDEX.

solar::Bool=false: If the result will be normalized to the solar abundance or not.

Returns

The total abundance of element.

GalaxyInspector.computeMass — Method

computeMass(data_dict::Dict, component::Symbol)::Vector{<:Unitful.Mass}

Compute the mass in each cell/particle of a given component.

Arguments

data_dict::Dict: Data dictionary (see makeDataDict for the canonical description). This function requires the following blocks to be present, depending on the value of component:

If component ∈ [:stellar, :darkmatter, :blackhole, :gas]:

component => ["MASS"]

If component == :Z_stellar

:stellar => ["MASS", "GZ2 "]

If component ∈ [:hydrogen, :helium]:

:gas => ["MASS"]

If component == :Z_gas:

:gas => ["MASS", "GZ "]

If component ∈ [:ionized, :neutral]:

:gas => ["MASS", "NH ", "NHP "]

If component ∈ [:bratomic, :brmolecular]:

:gas => ["MASS", "NH ", "NHP ", "PRES"]

If component ∈ [:odeionized, :odeatomic, :odemetals, :odedust, :odeneutral, :odecold]:

:gas => ["MASS", "NH ", "NHP ", "FRAC", "RHO ", "GZ "]

If component ∈ [:odemolecular, :odestellar, :odemolecularstellar]:

:gas => ["MASS", "FRAC", "RHO "]

component::Symbol: Target component. It can only be one of the elements of COMPONENTS.

Returns

The mass of component in each cell/particle.

References

L. Blitz et al. (2006). The Role of Pressure in GMC Formation II: The H2-Pressure Relation. The Astrophysical Journal, 650(2), 933. doi:10.1086/505417

GalaxyInspector.computeMassDensity — Method

computeMassDensity(data_dict::Dict, component::Symbol)::Vector{<:Unitful.Density}

Compute the mass density of a given gas component for each cell.

Arguments

data_dict::Dict: Data dictionary (see makeDataDict for the canonical description). This function requires the following blocks to be present, depending on the value of component:

If component ∈ [:gas, :hydrogen, :helium]:

:gas => ["MASS", "RHO "]

If component == :Z_gas:

:gas => ["MASS", "GZ ", "RHO "]

If component ∈ [:ionized, :neutral]:

:gas => ["MASS", "NH ", "NHP ", "RHO "]

If component ∈ [:bratomic, :brmolecular]:

:gas => ["MASS", "NH ", "NHP ", "PRES", "RHO "]

If component ∈ [:odeionized, :odeatomic, :odemetals, :odedust, :odeneutral, :odecold]:

:gas => ["MASS", "NH ", "NHP ", "FRAC", "RHO ", "GZ "]

If component ∈ [:odemolecular, :odestellar, :odemolecularstellar:

:gas => ["MASS", "FRAC", "RHO "]

component::Symbol: Target component. It can only be one of the gas elements of COMPONENTS.

Returns

The mass density of component for each cell.

References

L. Blitz et al. (2006). The Role of Pressure in GMC Formation II: The H2-Pressure Relation. The Astrophysical Journal, 650(2), 933. doi:10.1086/505417

GalaxyInspector.computeMassHeight — Method

computeMassHeight(
    positions::Matrix{<:Unitful.Length},
    masses::Vector{<:Unitful.Mass};
    <keyword arguments>
)::Unitful.Length

Compute the total height of a cylinder, of infinite radius, containing percent% of the total mass.

Arguments

positions::Matrix{<:Unitful.Length}: Positions of the cells/particles. Each column is a cell/particle and each row a dimension.

masses::Vector{<:Unitful.Mass}: Masses of the cells/particles.

percent::Float64=90.0: Target percentage of the total mass.

Returns

The height containing percent% of the total mass.

GalaxyInspector.computeMassQty — Method

computeMassQty(
    quantity::Vector{<:Number},
    masses::Vector{<:Unitful.Mass};
    <keyword arguments>
)::Number

Compute the minimum value of quantity that "contains" percent% of the total mass.

Arguments

quantity::Vector{<:Number}: Target quantity.

masses::Vector{<:Unitful.Mass}: Masses of the cells/particles.

percent::Float64=90.0: Target percentage of the total mass.

Returns

The minimum value of quantity that "contains" percent% of the total mass.

GalaxyInspector.computeMassRadius — Method

computeMassRadius(
    positions::Matrix{<:Unitful.Length},
    masses::Vector{<:Unitful.Mass};
    <keyword arguments>
)::Unitful.Length

Compute the radius containing percent% of the total mass.

Arguments

positions::Matrix{<:Unitful.Length}: Positions of the cells/particles. Each column is a cell/particle and each row a dimension.

masses::Vector{<:Unitful.Mass}: Masses of the cells/particles.

percent::Float64=90.0: Target percentage of the total mass.

Returns

The radius containing percent% of the total mass.

GalaxyInspector.computeNumber — Method

computeNumber(data_dict::Dict, component::Symbol)::Vector{Float64}

Compute the number of a given component.

Note

What number means changes depending on the component, see the comments in the code.

Arguments

data_dict::Dict: Data dictionary (see makeDataDict for the canonical description). This function requires the following blocks to be present, depending on the value of component:

If component ∈ [:stellar, :darkmatter, :blackhole, :gas]:

component => ["MASS"]

If component ∈ [:Z_stellar]:

:stellar => ["MASS"]

If component ∈ [:hydrogen, :helium]:

:gas => ["MASS"]

If component == :Z_gas:

:gas => ["MASS", "GZ "]

If component ∈ [:ionized, :neutral]:

:gas => ["MASS", "NH ", "NHP "]

If component ∈ [:bratomic, :brmolecular]:

:gas => ["MASS", "NH ", "NHP ", "PRES"]

If component ∈ [:odeionized, :odeatomic, :odemetals, :odedust, :odeneutral, :odecold]:

:gas => ["MASS", "NH ", "NHP ", "FRAC", "RHO ", "GZ "]

If component ∈ [:odemolecular, :odestellar, :odemolecularstellar]:

:gas => ["MASS", "FRAC", "RHO "]

component::Symbol: Target component. It can only be one of the elements of COMPONENTS.

Returns

The number of component.

References

L. Blitz et al. (2006). The Role of Pressure in GMC Formation II: The H2-Pressure Relation. The Astrophysical Journal, 650(2), 933. doi:10.1086/505417

GalaxyInspector.computeNumberDensity — Method

computeNumberDensity(data_dict::Dict, component::Symbol)::Vector{<:NumberDensity}

Compute the number density of a given gas component for each cell.

Note

What number density means changes depending on the component, see the comments in the code.

Arguments

data_dict::Dict: Data dictionary (see makeDataDict for the canonical description). This function requires the following blocks to be present, depending on the value of component:

If component ∈ [:gas, :hydrogen, :helium]:

:gas => ["MASS", "RHO "]

If component == :Z_gas:

:gas => ["MASS", "GZ ", "RHO "]

If component ∈ [:ionized, :neutral]:

:gas => ["MASS", "NH ", "NHP ", "RHO "]

If component ∈ [:bratomic, :brmolecular]:

:gas => ["MASS", "NH ", "NHP ", "PRES", "RHO "]

If component ∈ [:odeionized, :odeatomic, :odemetals, :odedust, :odeneutral, :odecold]:

:gas => ["MASS", "NH ", "NHP ", "FRAC", "RHO ", "GZ "]

If component ∈ [:odemolecular, :odestellar, :odemolecularstellar]:

:gas => ["MASS", "FRAC", "RHO "]

component::Symbol: Target component. It can only be one of the gas elements of COMPONENTS.

Returns

The number density of component for each cell.

References

L. Blitz et al. (2006). The Role of Pressure in GMC Formation II: The H2-Pressure Relation. The Astrophysical Journal, 650(2), 933. doi:10.1086/505417

GalaxyInspector.computeVirialAccretion — Method

computeVirialAccretion(
    present_dd::Dict,
    past_dd::Dict;
    <keyword arguments>
)::NTuple{3,Unitful.Mass}

Compute the inflow, outflow, or net gain of mass for a given halo virial radius ($R_{200}$), between two snapshots.

Arguments

present_dd::Dict: Data dictionary, for the present snapshot (see makeDataDict for the canonical description). This function requires the following blocks to be present for every cell/particle that you want to be taken into account:

cell/particle type => ["POS ", "ID ", "MASS"]

:group => ["GRCrit200", "GNsubs", "GPos"]

:subhalo => ["S_Pos"]

:tracer => ["PAID", "TRID"]

past_dd::Dict: Data dictionary, for the past snapshot (see makeDataDict for the canonical description). This function requires the following blocks to be present for every cell/particle that you want to be taken into account:

cell/particle type => ["POS ", "ID ", "MASS"]

:group => ["GRCrit200", "GNsubs", "GPos"]

:subhalo => ["S_Pos"]

:tracer => ["PAID", "TRID"]

component::Symbol: Component to compute the accreted mass for. The options are:

:dark_matter -> Dark matter.

:black_hole -> Black holes.

:gas -> Gas.

:stellar -> Stars.

halo_idx::Int=1: Index of the target halo (FoF group). Starts at 1.

tracers::Bool=false: Whether to compute the accretion using tracer particles (true) or the actual component particles (false).

Returns

A tuple with three elements:

The net increase in mass.

The inflow mass.

The outflow mass.

GalaxyInspector.initialConditionFunction — Method

initialConditionFunction(data_dict::Dict, component::Symbol)::Union{Function,Nothing}

Generate a function that gives the initial condition of component for the i-th cell, according to our star formation model.

Arguments

data_dict::Dict: Data dictionary (see makeDataDict for the canonical description). This function requires the following blocks to be present, depending on the value of component:

If component ∈ [:odeionized, :odeatomic, :odemetals, :odedust, :ode_neutral]:

:gas => ["NH ", "NHP ", "GZ "]

If component ∈ [:ode_cold]:

:gas => ["GZ "]

If component ∈ [:odemolecular, :odestellar, :odemolecularstellar, :Zstellar, :gas, :hydrogen, :helium, :Zgas, :ionized, :neutral, :bratomic, :brmolecular]:

No blocks are required.

component::Symbol: Target component. It can only be one of the elements of COMPONENTS.

Returns

A function with signature ode_ic(i::Int)::Float64, that gives the initial condition of component for the i-th cell, according to our star formation model.

GalaxyInspector.quantity3DProjection — Method

quantity3DProjection(
    data_dict::Dict,
    grid::CubicGrid,
    quantity::Symbol,
    field_type::Symbol;
    <keyword arguments>
)::Tuple{Array{Float64,3},Vector{Int}}

Project a quantity field into a given regular 3D grid.

Note

If the source of the field are particles, a simple 3D histogram is used. If they are Voronoi cells instead, the value of the closest cell to the voxel is used.

Arguments

data_dict::Dict: Data dictionary (see makeDataDict for the canonical description).

grid::CubicGrid: Cubic grid.

quantity::Symbol: Target quantity. It can be any quantity valid for scatterQty, as long as it has a well defined cell/particle type (see plotParams).

field_type::Symbol: If the field is made up of :particles or Voronoi :cells.

scale_by_volume::Bool=true: If quantity is extensive (e.g. a mass), this should be set to true, so the values are correctly scaled to the voxel volume.

empty_nan::Bool=true: If NaN will be put into empty bins, 0 is used otherwise.

density::Union{Unitful.Units,Nothing}=nothing: Target length unit for the density. Set it to nothing if you want the values of quantity instead of the volume densities of quantity.

log::Bool=true: If the returned projection will have $\log_{10}$ applied to it.

filter_function::Function=filterNothing: Filter function to be applied to data_dict before any other computation. See the required signature and examples in ./src/analysis/filters.jl.

Returns

A tuple with two elements:

A 3D tensor with the target value at each bin of the 3D grid.

A vector with the index of the closest cell to each voxel (the result of nn() from NearestNeighbors).

GalaxyInspector.quantity3DProjection — Method

quantity3DProjection(
    data_dict::Dict,
    grid::CubicGrid,
    quantity::Symbol,
    nn_idxs::Vector{Int};
    <keyword arguments>
)::Tuple{Array{Float64,3},Vector{Int}}

Project a quantity field into a given regular 3D grid.

Note

If the source of the field are particles, a simple 3D histogram is used. If they are Voronoi cells instead, the value of the closest cell to the voxel is used.

Arguments

data_dict::Dict: Data dictionary (see makeDataDict for the canonical description).

grid::CubicGrid: Cubic grid.

quantity::Symbol: Target quantity. It can be any quantity valid for scatterQty, as long as it has a well defined cell/particle type (see plotParams).

nn_idxs::Vector{Int}: A vector with the index of the closest cell to each voxel (the result of nn() from NearestNeighbors). If set to an empty vector, it is assumed that the field is made up of particles.

scale_by_volume::Bool=true: If quantity is extensive (e.g. a mass), this should be set to true, so the values are correctly scaled to the voxel volume.

empty_nan::Bool=true: If NaN will be put into empty bins, 0 is used otherwise.

density::Union{Unitful.Units,Nothing}=nothing: Target length unit for the density. Set it to nothing if you want the values of quantity instead of the volume densities of quantity.

log::Bool=true: If the returned projection will have $\log_{10}$ applied to it.

filter_function::Function=filterNothing: Filter function to be applied to data_dict before any other computation. See the required signature and examples in ./src/analysis/filters.jl.

Returns

A tuple with two elements:

A 3D tensor with the target value at each bin of the 3D grid.

A vector with the index of the closest cell to each voxel (the result of nn() from NearestNeighbors).