Computation of derived quantities

These functions are used internally and none are exported.

These functions depend on the particulars of the simulation code (e.g. units).

GalaxyInspector.computeClumpingFactor — Method

computeClumpingFactor(density::Vector{<:Number})::Float64

Compute the clumping factor,

\[C_\rho = \frac{\langle \rho^2 \rangle}{\langle \rho \rangle^2} \, .\]

Arguments

density::Vector{<:Number}: The density of the cells/particles.

Returns

The clumping factor.

GalaxyInspector.computeDepletionTime — Method

computeDepletionTime(
    mass::Vector{<:Unitful.Mass},
    sfr::Vector{<:Unitful.MassFlow},
)::Vector{<:Unitful.Time}

Compute the depletion time,

\[t_\mathrm{ff} = \frac{M_\mathrm{gas}}{\dot{M}_\star} \, .\]

Arguments

mass::Vector{<:Unitful.Mass}: The gas mass of the cells/particles.

sfr::Vector{<:Unitful.MassFlow}: The SFR associated to each cell/particle.

Returns

The depletion time.

GalaxyInspector.computeEfficiencyFF — Method

computeEfficiencyFF(
    density::Vector{<:Unitful.Density},
    mass::Vector{<:Unitful.Mass},
    sfr::Vector{<:Unitful.MassFlow},
)::Vector{Float64}

Compute the star formation efficiency per free-fall time, according to the definition in equation 1 of Krumholz (2012),

\[\epsilon_\mathrm{ff} = \frac{t_\mathrm{ff}}{t_\mathrm{dep}} \, .\]

where

\[t_\mathrm{ff} = \sqrt{\frac{3 \, \pi}{32 \, G \, \rho}} \, ,\]

is the free-fall time, and

\[t_\mathrm{dep} = \frac{M_\mathrm{H_2}}{\dot{M}_\star} \, ,\]

is the depletion time.

Arguments

density::Vector{<:Unitful.Density}: The molecular hydrogen ($\\mathrm{H_2}$) density of the cells/particles.

mass::Vector{<:Unitful.Mass}: The gas mass of the cells/particles.

sfr::Vector{<:Unitful.MassFlow}: The SFR associated to each cell/particle.

Returns

The star formation efficiency per free-fall time.

References

M. R. Krumholz et al. (2012). A UNIVERSAL, LOCAL STAR FORMATION LAW IN GALACTIC CLOUDS, NEARBY GALAXIES, HIGH-REDSHIFT DISKS, AND STARBURSTS. The Astrophysical Journal, 745(1), 69. doi:10.1088/0004-637X/745/1/69

GalaxyInspector.computeKineticEnergy — Method

computeKineticEnergy(
    masses::Vector{<:Unitful.Mass},
    velocities::Matrix{<:Unitful.Velocity},
)::Vector{<:Unitful.Energy}

Compute the kinetic energy.

Arguments

masses::Vector{<:Unitful.Mass}: Mass of every cell/particle.

velocities::Matrix{<:Unitful.Velocity}: Velocities of the cells/particles. Each column is a cell/particle and each row a dimension.

Returns

The kinetic energy of each cell/particle.

GalaxyInspector.computePotentialEnergy — Method

computePotentialEnergy(
    potential::Vector{<:SpecificEnergy},
    masses::Vector{<:Unitful.Mass},
)::Vector{<:Unitful.Energy}

Compute the gravitational potencial energy.

Arguments

potential::Vector{<:SpecificEnergy}: Specific potential energy of every cell/particle.

masses::Vector{<:Unitful.Mass}: Mass of every cell/particle.

Returns

The gravitational potencial energy of each cell/particle.

GalaxyInspector.computeSFR — Method

computeSFR(
    data_dict::Dict;
    <keyword arguments>
)::Vector{<:Unitful.MassFlow}

Compute the star formation rate of each stellar particle.

For stellar particles younger than age_resol, the SFR is its mass divided by age_resol. It is defined as 0 for older particles.

Arguments

data_dict::Dict: A dictionary with the following shape:

:sim_data -> ::Simulation (see Simulation).

:snap_data -> ::Snapshot (see Snapshot).

:gc_data -> ::GroupCatalog (see GroupCatalog).

cell/particle type -> (block -> data of block, block -> data of block, ...).

cell/particle type -> (block -> data of block, block -> data of block, ...).

cell/particle type -> (block -> data of block, block -> data of block, ...).

...

groupcat type -> (block -> data of block, block -> data of block, ...).

groupcat type -> (block -> data of block, block -> data of block, ...).

groupcat type -> (block -> data of block, block -> data of block, ...).

...

age_resol::Unitful.Time=AGE_RESOLUTION: Age resolution for the SFR.

Returns

The star formation rate of each stellar particle.

GalaxyInspector.computeStellarAge — Method

computeStellarAge(data_dict::Dict)::Vector{<:Unitful.Time}

Compute the age of the stars.

Arguments

data_dict::Dict: A dictionary with the following shape:

:sim_data -> ::Simulation (see Simulation).

:snap_data -> ::Snapshot (see Snapshot).

:gc_data -> ::GroupCatalog (see GroupCatalog).

cell/particle type -> (block -> data of block, block -> data of block, ...).

cell/particle type -> (block -> data of block, block -> data of block, ...).

cell/particle type -> (block -> data of block, block -> data of block, ...).

...

groupcat type -> (block -> data of block, block -> data of block, ...).

groupcat type -> (block -> data of block, block -> data of block, ...).

groupcat type -> (block -> data of block, block -> data of block, ...).

...

Returns

The stellar ages.

GalaxyInspector.computeTemperature — Method

computeTemperature(
    internal_energy::Vector{<:SpecificEnergy},
    electron_fraction::Vector{Float32},
)::Vector{<:Unitful.Temperature}

Compute the gas temperature.

Arguments

internal_energy::Vector{<:SpecificEnergy}: Specific internal energy of every gas cell/particle.

electron_fraction::Vector{Float32}: Number fraction of electrons in every gas cell/particle.

Returns

The temperature of each gas cell/particle.

GalaxyInspector.computeTime — Method

computeTime(a::Real, header::SnapshotHeader; <keyword arguments>)::Unitful.Time

Compute the physical time corresponding to the scale factor a.

To get the physical time $t$ from the scale factor a, one does the integral:

\[t = \frac{1}{H_0} \int_0^a \frac{\mathrm{d}a'}{a' \, \sqrt{\mathcal{E}(a')}} \, ,\]

where

\[\mathcal{E}(a) = \Omega_\Lambda + \Omega_m \, a^{-3} + \Omega_r \, a^{-4} + \Omega_K \, a^{-2} \, .\]

Arguments

a::Real: Scale factor.

header::SnapshotHeader: A header of the simulation, containing the cosmological parameters.

a0::Float64=0.0: Initial scale factor.

Returns

The physical time.

GalaxyInspector.computeTime — Method

computeTime(
    scale_factors::Vector{<:Real},
    header::SnapshotHeader;
    <keyword arguments>
)::Vector{<:Unitful.Time}

Compute the physical time corresponding to each of the scale_factors.

To get the physical time $t$ from the scale factor a, one does the integral:

\[t = \frac{1}{H_0} \int_0^a \frac{\mathrm{d}a'}{a' \, \sqrt{\mathcal{E}(a')}} \, ,\]

where

\[\mathcal{E}(a) = \Omega_\Lambda + \Omega_m \, a^{-3} + \Omega_r \, a^{-4} + \Omega_K \, a^{-2} \, .\]

Arguments

scale_factors::Vector{<:Real}: Scale factors.

header::SnapshotHeader: A header of the simulation, containing the cosmological parameters.

a0::Float64=0.0: Initial scale factor.

Returns

A vector with the physical times.

GalaxyInspector.computeTimeTicks — Method

computeTimeTicks(
    paths::Vector{<:Union{Missing,String}},
)::Tuple{Vector{Float64},Vector{Float64},Vector{<:Unitful.Time},Vector{<:Unitful.Time}}

Compute the different times stamps associated with each snapshot in paths.

Arguments

paths::Vector{<:Union{Missing,String}}: Paths to the snapshots.

Returns

A tuple with four elements:

A vector with the scale factors.

A vector with the redshifts.

A vector with the physical times (physical time since the Big Bang).

A vector with the lookback times (physical time left to reach the last snapshot).

GalaxyInspector.computeTotalEnergy — Method

computeTotalEnergy(
    potential::Vector{<:SpecificEnergy},
    masses::Vector{<:Unitful.Mass},
    velocities::Matrix{<:Unitful.Velocity},
)::Vector{<:Unitful.Energy}

Compute the total energy (kinetic + potential).

Arguments

potential::Vector{<:SpecificEnergy}: Specific potential energy of every cell/particle.

masses::Vector{<:Unitful.Mass}: Mass of every cell/particle.

velocities::Matrix{<:Unitful.Velocity}: Velocities of the cells/particles. Each column is a cell/particle and each row a dimension.

Returns

The total energy of each cell/particle.