Computation of characteristic positions

These functions are used internally and none are exported.

They may depend on the particulars of the simulation code (e.g. internal units).

GalaxyInspector.computeCenter — Method

computeCenter(data_dict::Dict, subhalo_abs_idx::Int)::Vector{<:Unitful.Length}

Return the 3D position of the potential minimum for a given subhalo.

Arguments

data_dict::Dict: Data dictionary (see makeDataDict for the canonical description). This function requires the following blocks to be present:

:subhalo => ["S_Pos"]

subhalo_abs_idx::Int: Absolute index of the target subhalo (subfind). Starts at 1.

Returns

The 3D position of the potential minimum.

GalaxyInspector.computeCenter — Method

computeCenter(data_dict::Dict, cm_type::Symbol)::Vector{<:Unitful.Length}

Compute a characteristic center of mass for the system.

Arguments

data_dict::Dict: Data dictionary (see makeDataDict for the canonical description). This function requires the following blocks to be present, depending on the value of cm_type:

If cm_type == :global_cm:

["POS ", "MASS"] for every cell/particle type in the snapshot (see computeGlobalCenterOfMass).

If cm_type ∈ keys(PARTICLE_INDEX):

cm_type => ["POS ", "MASS"].

If cm_type == :zero:

No blocks are required.

cm_type::Symbol: It can be:

:global_cm -> Center of mass of the whole system.

:{component} -> Center of mass of the given component (e.g. :stellar, :gas, :darkmatter, etc). It can be any of the keys of [`PARTICLEINDEX`](@ref).

:zero -> Origin.

Returns

The center of mass.

GalaxyInspector.computeCenter — Method

computeCenter(data_dict::Dict, subfind_idx::NTuple{2,Int})::Vector{<:Unitful.Length}

Return the 3D position of the potential minimum for a halo or subhalo.

Arguments

data_dict::Dict: Data dictionary (see makeDataDict for the canonical description). This function requires the following blocks to be present:

:group => ["G_Nsubs", "G_Pos"]

:subhalo => ["S_Pos"]

subfind_idx::NTuple{2,Int}: Tuple with two elements:

Index of the target halo (FoF group). Starts at 1.

Index of the target subhalo (subfind), relative to the target halo. Starts at 1. If set to 0, the halo potential minimum is returned.

Returns

The 3D position of the potential minimum.

GalaxyInspector.computeCenterOfMass — Method

computeCenterOfMass(
    positions::Matrix{<:Unitful.Length},
    masses::Vector{<:Unitful.Mass},
)::Vector{<:Unitful.Length}

Compute the center of mass of a group of cells/particles.

Arguments

positions::Matrix{<:Unitful.Length}: Positions of the cells/particles. Each column is a cell/particle and each row a dimension.

masses::Vector{<:Unitful.Mass}: Masses of the cells/particles.

Returns

The center of mass.

GalaxyInspector.computeDistance — Method

computeDistance(
    positions::Matrix{<:Number};
    <keyword arguments>
)::Vector{<:Number}

Compute the distance of a group of cells/particles to a center.

Arguments

positions::Matrix{<:Unitful.Length}: Positions of the cells/particles. Each column is a cell/particle and each row a dimension.

center::Union{Vector{<:Number},Nothing}=nothing: Origin used to compute the distances. If set to nothing, 0 is used.

Returns

The distance of every point to center.

GalaxyInspector.computeGlobalCenterOfMass — Method

computeGlobalCenterOfMass(data_dict::Dict)::Vector{<:Unitful.Length}

Compute the center of mass of the whole system.

Arguments

data_dict::Dict: Data dictionary (see makeDataDict for the canonical description). This function requires the following blocks to be present for every cell/particle that you want to be taken into account:

cell/particle type => ["POS ", "MASS"]

Returns

The center of mass.

GalaxyInspector.findStellarHaloSubhalo — Method

findStellarHaloSubhalo(
    data_dict::Dict,
    star_idxs::Vector{Int},
    real_stars_idxs::Vector{Bool},
)::NTuple{2,Vector{Int}}

Find in which halo and subhalo of data_dict each star in star_idxs is located.

For stars with no halo or subhalo, an index of -1 is given. The subhalo index is relative to the corresponding halo.

Arguments

data_dict::Dict: Data dictionary (see makeDataDict for the canonical description). This function requires the following blocks to be present:

:group => ["G_Nsubs", "G_LenType"]

:subhalo => ["S_LenType"]

star_idxs::Vector{Int}: Indices of the target stars in data_dict. The indices are relative to the stars in data_dict[:stellar].

real_stars_idxs::Vector{Bool}: Boolean list of stellar particles. True for real stars and false for wind particles. The indices are relative to the list of stars/wind particles in the snapshot (i.e. the output of findRealStars).

Returns

A tuple with two elements:

A vector with the halo (index starting at 1) of each star (in the order of star_idxs).

A vector with the subhalo (index starting at 1) of each star (in the order of star_idxs).

GalaxyInspector.locateStellarBirthPlace — Method

locateStellarBirthPlace(data_dict::Dict)::NTuple{2,Vector{Int}}

Find in which halo and subhalo each star in data_dict was born.

For stars with no halo or subhalo, an index of -1 is given. The subhalo index is relative to the corresponding halo.

Arguments

data_dict::Dict: Data dictionary (see makeDataDict for the canonical description). This function requires the following blocks to be present:

:stellar => ["GAGE", "ID "]

Returns

A tuple with two elements:

A vector with the birth halo (index starting at 1) of each star (in the order of data_dict).

A vector with the birth subhalo (index starting at 1) of each star (in the order of data_dict).