Compute derived code quantities for our star formation model

These functions are used internally and none are exported.

They may depend on the particulars of the simulation code (e.g. internal units).

GalaxyInspector.computeDerivedSFMQty — Method

computeDerivedSFMQty(data_dict::Dict, quantity::Symbol)::Vector{<:Number}

Compute a derived code quantity for our star formation model.

Arguments

data_dict::Dict: Data dictionary (see makeDataDict for the canonical description). This function requires the following blocks to be present, depending on the value of component and magnitude:

If magnitude == :tau_star

type => ["RHOC"]

If magnitude == :tau_rec

type => ["RHOC", "FRAC"]

If magnitude == :tau_cond

type => ["RHOC", "FRAC"]

If magnitude == :tau_dg

type => ["RHOC", "FRAC"]

If magnitude == :tau_dc

type => ["RHOC", "FRAC"]

If magnitude == :tau_ion

type => ["RHOC", "PARH", "FRAC"]

If magnitude == :tau_diss

type => ["RHOC", "PARH", "FRAC"]

If magnitude == :S_d

type => ["RHOC", "PARH", "FRAC"]

If magnitude == :S_H2

type => ["RHOC", "PARH", "FRAC"]

If magnitude == :equivalent_size

:gas => ["RHOC", "MASS"] or :stellar => ["RHOC", "GMAS"]

where

If component == :odegas

type = :gas

If component == :odestellar

type = :stellar

quantity::Symbol: Target quantity. It can only be one of the elements of SFM_DERIVED_QTY.

Returns

quantity for each cell/particle.