Computation of characteristic velocities and momentums

These functions are used internally and none are exported.

They may depend on the particulars of the simulation code (e.g. internal units).

GalaxyInspector.computeAngularMomentum — Method

computeAngularMomentum(
    data_dict::Dict,
    component::Symbol,
)::Vector{<:AngularMomentum}

Compute the angular momentum in the z direction with respect to the origin, for each cell/particle of the given component.

Arguments

data_dict::Dict: Data dictionary (see makeDataDict for the canonical description). This function requires the following blocks to be present for the cell/particle type corresponding to component:

cell/particle type => ["POS ", "VEL ", "MASS"].

component::Symbol: Target component. It can only be one of the elements of COMPONENTS.

Returns

The angular momentum of each cell/particle.

GalaxyInspector.computeAngularMomentum — Method

computeAngularMomentum(
    positions::Matrix{<:Unitful.Length},
    velocities::Matrix{<:Unitful.Velocity},
    masses::Vector{<:Unitful.Mass},
)::Vector{<:AngularMomentum}

Compute the angular momentum in the z direction with respect to the origin.

Arguments

positions::Matrix{<:Unitful.Length}: Positions of the cells/particles. Each column is a cell/particle and each row a dimension.

velocities::Matrix{<:Unitful.Velocity}: Velocities of the cells/particles. Each column is a cell/particle and each row a dimension.

masses::Vector{<:Unitful.Mass}: Mass of every cell/particle.

Returns

The angular momentum in the z direction of each cell/particle.

GalaxyInspector.computeCircularity — Method

computeCircularity(data_dict::Dict, component::Symbol)::Vector{Float64}

Compute the circularity of each cell/particle, with respect to the origin and the $z$ direction.

The circularity of a cell/particle is,

\[\epsilon = j_z / j_\mathrm{circ} \, ,\]

where $j_z$ is the $z$ component of its specific angular momentum, and $j_\mathrm{circ}$ is the specific angular momentum of a circular orbit,

\[j_\mathrm{circ} = r \, v_\mathrm{circ} = \sqrt{\mathrm{G} \, r \, M(r)} \, ,\]

where $r$ is the radial distance of the particle, and $M(r)$ is the total mass within a sphere of radius $r$.

Arguments

data_dict::Dict: Data dictionary (see makeDataDict for the canonical description). This function requires the following blocks to be present for every cell/particle that you want to be taken into account:

cell/particle type => ["VEL ", "POS ", "MASS"].

component::Symbol: Target component. It can only be one of the elements of COMPONENTS.

Returns

The circularity of each cell/particle.

GalaxyInspector.computeCircularity — Method

computeCircularity(
    jzs::Vector{<:Unitful.KinematicViscosity},
    rs::Vector{<:Unitful.Length},
    vcircs::Vector{<:Unitful.Velocity},
)::Vector{Float64}

Compute the circularity with respect to the origin and the $z$ direction.

The circularity of a cell/particle is,

\[\epsilon = j_z / j_\mathrm{circ} \, ,\]

where $j_z$ is the $z$ component of its specific angular momentum, and $j_\mathrm{circ}$ is the specific angular momentum of a circular orbit,

\[j_\mathrm{circ} = r \, v_\mathrm{circ} = \sqrt{\mathrm{G} \, r \, M(r)} \, ,\]

where $r$ is the radial distance of the particle, and $M(r)$ is the total mass within a sphere of radius $r$.

Arguments

jzs::Vector{<:Unitful.KinematicViscosity}: Specific angular momentum in the z direction of each cell/particle.

rs::Vector{<:Unitful.Length}: Radial distances of each cell/particle to the origin.

vcircs::Vector{<:Unitful.Velocity}: Circular velocity of each cell/particle.

Returns

The circularity of each cell/particle.

GalaxyInspector.computeGlobalAngularMomentum — Method

computeGlobalAngularMomentum(data_dict::Dict; <keyword arguments>)::Vector{<:Number}

Compute the total angular momentum of the system, with respect to the origin

Arguments

data_dict::Dict: Data dictionary (see makeDataDict for the canonical description). This function requires the following blocks to be present for every cell/particle that you want to be taken into account:

cell/particle type => ["POS ", "VEL ", "MASS"].

normal::Bool=true: If the result will be normalized.

Returns

The angular momentum.

GalaxyInspector.computeGlobalSpinParameter — Method

computeGlobalSpinParameter(data_dict::Dict; <keyword arguments>)::Float64

Compute the spin parameter of the whole system.

The spin parameter was originally defined by Peebles (1969) as,

\[\lambda = \frac{J \, \sqrt{E}}{G \, M^{5/2}} \, ,\]

where $J$ is the norm of the total angular momentum, $M$ the total mass, $G$ the gravitational constant, and

\[E = |E_P + E_k| \, ,\]

where $E_P$ is the total potential energy and $E_k$ is the total kinetic energy (including thermal energy of the gas).

Due to the computational complexity of calculating $E_P$ for a large group of particles, Bullock et al. (2001) proposed an alternative definition of the spin parameter,

\[\lambda = \frac{J}{\sqrt{2} \, M \, R \, V} \, ,\]

where $J$ is the norm of the total angular momentum inside a sphere of radius $R$ containing mass $M$, and

\[V = \sqrt{\frac{G \, M}{R}} \, ,\]

is the circular velocity.

Arguments

data_dict::Dict: Data dictionary (see makeDataDict for the canonical description). This function requires the following blocks to be present for every cell/particle that you want to be taken into account:

cell/particle type => ["POS ", "VEL ", "MASS"].

component::Symbol: Target component. It can only be one of the elements of COMPONENTS.

R::Unitful.Length=DISK_R: Characteristic radius.

Returns

The spin parameter of the cell/particles within radius R.

References

P. J. E. Peebles (1969). Origin of the Angular Momentum of Galaxies. Astrophysical Journal, 155, 393. doi:10.1086/149876

J. S. Bullock et al. (2001). A Universal Angular Momentum Profile for Galactic Halos. The Astrophysical Journal, 555(1), 240. doi:10.1086/321477

J. Zjupa et al. (2017). Angular momentum properties of haloes and their baryon content in the Illustris simulation. Monthly Notices of the Royal Astronomical Society, 466(2), 1625–1647. doi:10.1093/mnras/stw2945

GalaxyInspector.computeSpecificAngularMomentum — Method

computeSpecificAngularMomentum(
    data_dict::Dict,
    component::Symbol,
)::Vector{<:Unitful.KinematicViscosity}

Compute the specific angular momentum in the z direction with respect to the origin, for each cell/particle of the given component.

Arguments

data_dict::Dict: Data dictionary (see makeDataDict for the canonical description). This function requires the following blocks to be present for the cell/particle type corresponding to component:

cell/particle type => ["POS ", "VEL "].

component::Symbol: Target component. It can only be one of the elements of COMPONENTS.

Returns

The specific angular momentum of each cell/particle.

GalaxyInspector.computeSpecificAngularMomentum — Method

computeSpecificAngularMomentum(
    positions::Matrix{<:Unitful.Length},
    velocities::Matrix{<:Unitful.Velocity},
)::Vector{<:Unitful.KinematicViscosity}

Compute the specific angular momentum in the z direction with respect to the origin.

Arguments

positions::Matrix{<:Unitful.Length}: Positions of the cells/particles. Each column is a cell/particle and each row a dimension.

velocities::Matrix{<:Unitful.Velocity}: Velocities of the cells/particles. Each column is a cell/particle and each row a dimension.

Returns

The specific angular momentum in the z direction of each cell/particle.

GalaxyInspector.computeSpinParameter — Method

computeSpinParameter(data_dict::Dict, component::Symbol; <keyword arguments>)::Float64

Compute the spin parameter of the given component, with respect to the origin

The spin parameter was originally defined by Peebles (1969) as,

\[\lambda = \frac{J \, \sqrt{E}}{G \, M^{5/2}} \, ,\]

where $J$ is the norm of the total angular momentum, $M$ the total mass, $G$ the gravitational constant, and

\[E = |E_P + E_k| \, ,\]

where $E_P$ is the total potential energy and $E_k$ is the total kinetic energy (including thermal energy of the gas).

Due to the computational complexity of calculating $E_P$ for a large group of particles, Bullock et al. (2001) proposed an alternative definition of the spin parameter,

\[\lambda = \frac{J}{\sqrt{2} \, M \, R \, V} \, ,\]

where $J$ is the norm of the total angular momentum inside a sphere of radius $R$ containing mass $M$, and

\[V = \sqrt{\frac{G \, M}{R}} \, ,\]

is the circular velocity.

Arguments

data_dict::Dict: Data dictionary (see makeDataDict for the canonical description). This function requires the following blocks to be present for the cell/particle type corresponding to component:

cell/particle type => ["POS ", "VEL ", "MASS"].

component::Symbol: Target component. It can only be one of the elements of COMPONENTS.

R::Unitful.Length=DISK_R: Characteristic radius.

Returns

The spin parameter of the cell/particles within radius R.

References

P. J. E. Peebles (1969). Origin of the Angular Momentum of Galaxies. Astrophysical Journal, 155, 393. doi:10.1086/149876

J. S. Bullock et al. (2001). A Universal Angular Momentum Profile for Galactic Halos. The Astrophysical Journal, 555(1), 240. doi:10.1086/321477

J. Zjupa et al. (2017). Angular momentum properties of haloes and their baryon content in the Illustris simulation. Monthly Notices of the Royal Astronomical Society, 466(2), 1625–1647. doi:10.1093/mnras/stw2945

GalaxyInspector.computeSpinParameter — Method

computeSpinParameter(
    positions::Matrix{<:Unitful.Length},
    velocities::Matrix{<:Unitful.Velocity},
    masses::Vector{<:Unitful.Mass};
    <keyword arguments>
)::Float64

Compute the spin parameter with respect to the origin.

The spin parameter was originally defined by Peebles (1969) as,

\[\lambda = \frac{J \, \sqrt{E}}{G \, M^{5/2}} \, ,\]

where $J$ is the norm of the total angular momentum, $M$ the total mass, $G$ the gravitational constant, and

\[E = |E_P + E_k| \, ,\]

where $E_P$ is the total potential energy and $E_k$ is the total kinetic energy (including thermal energy of the gas).

Due to the computational complexity of calculating $E_P$ for a large group of particles, Bullock et al. (2001) proposed an alternative definition of the spin parameter,

\[\lambda = \frac{J}{\sqrt{2} \, M \, R \, V} \, ,\]

where $J$ is the norm of the total angular momentum inside a sphere of radius $R$ containing mass $M$, and

\[V = \sqrt{\frac{G \, M}{R}} \, ,\]

is the circular velocity.

Arguments

positions::Matrix{<:Unitful.Length}: Positions of the cells/particles. Each column is a cell/particle and each row a dimension.

velocities::Matrix{<:Unitful.Velocity}: Velocities of the cells/particles. Each column is a cell/particle and each row a dimension.

masses::Vector{<:Unitful.Mass}: Mass of every cell/particle.

R::Unitful.Length=DISK_R: Characteristic radius.

Returns

The spin parameter of the cell/particles within radius R.

References

P. J. E. Peebles (1969). Origin of the Angular Momentum of Galaxies. Astrophysical Journal, 155, 393. doi:10.1086/149876

J. S. Bullock et al. (2001). A Universal Angular Momentum Profile for Galactic Halos. The Astrophysical Journal, 555(1), 240. doi:10.1086/321477

J. Zjupa et al. (2017). Angular momentum properties of haloes and their baryon content in the Illustris simulation. Monthly Notices of the Royal Astronomical Society, 466(2), 1625–1647. doi:10.1093/mnras/stw2945

GalaxyInspector.computeTotalAngularMomentum — Method

computeTotalAngularMomentum(
    positions::Matrix{<:Unitful.Length},
    velocities::Matrix{<:Unitful.Velocity},
    masses::Vector{<:Unitful.Mass};
    <keyword arguments>
)::Vector{<:Number}

Compute the total angular momentum vector with respect to the origin.

Arguments

positions::Matrix{<:Unitful.Length}: Positions of the cells/particles. Each column is a cell/particle and each row a dimension.

velocities::Matrix{<:Unitful.Velocity}: Velocities of the cells/particles. Each column is a cell/particle and each row a dimension.

masses::Vector{<:Unitful.Mass}: Mass of every cell/particle.

normal::Bool=true: If the result will be normalized.

Returns

The total angular momentum vector.

GalaxyInspector.computeVcirc — Method

computeVcirc(
    data_dict::Dict,
    component::Symbol,
)::Tuple{Vector{<:Unitful.Length},Vector{<:Unitful.Velocity}}

Compute the circular velocity of each cell/particle of the given component, with respect to the origin.

The circular velocity of a cell/particle is,

\[v_\mathrm{circ} = \sqrt{\frac{\mathrm{G} \, M(r)}{r}} \, ,\]

where $r$ is the radial distance of the cell/particle, and $M(r)$ is the total mass within a sphere of radius $r$.

Arguments

data_dict::Dict: Data dictionary (see makeDataDict for the canonical description). This function requires the following blocks to be present for every cell/particle that you want to be taken into account:

cell/particle type => ["POS ", "MASS"].

component::Symbol: Target component. It can only be one of the elements of COMPONENTS.

Returns

A tuple with two elements:

A vector with the radial distance of the target cells/particles to the origin.

A vector with the circular velocity of the target cells/particles.

GalaxyInspector.computeVcirc — Method

computeVcirc(
    distances::Vector{<:Unitful.Length},
    masses::Vector{<:Unitful.Mass},
    rs::Vector{<:Unitful.Length},
)::Tuple{Vector{<:Unitful.Length},Vector{<:Unitful.Velocity}}

Compute the circular velocity.

The circular velocity of a cell/particle is,

\[v_\mathrm{circ} = \sqrt{\frac{\mathrm{G} \, M(r)}{r}} \, ,\]

where $r$ is the radial distance of the cell/particle, and $M(r)$ is the total mass within a sphere of radius $r$.

Arguments

distances::Vector{<:Unitful.Length}: Radial distances of all the cells/particles in the system.

masses::Vector{<:Unitful.Mass}: Mass of every cell/particle in the system.

rs::Vector{<:Unitful.Length}: Radial distances of the target cells/particles.

Returns

A tuple with two elements:

A vector with the radial distance of the target cells/particles to the origin.

A vector with the circular velocity of the target cells/particles.

GalaxyInspector.computeVcm — Method

computeVcm(
    data_dict::Dict,
    halo_idx::Int,
    subhalo_rel_idx::Int,
)::Vector{<:Unitful.Velocity}

Return the velocity of the center of mass of a given halo or subhalo.

Arguments

data_dict::Dict: Data dictionary (see makeDataDict for the canonical description). This function requires the following blocks to be present:

:group => ["G_Nsubs", "G_Vel"]

:subhalo => ["S_Vel"]

halo_idx::Int: Index of the target halo (FoF group). Starts at 1.

subhalo_rel_idx::Int: Index of the target subhalo (subfind), relative to the target halo. Starts at 1. If set to 0, the halo potential minimum is returned.

Returns

The velocity of the center of mass.

GalaxyInspector.computeVcm — Method

computeVcm(data_dict::Dict, subhalo_abs_idx::Int)::Vector{<:Unitful.Velocity}

Return the velocity of the center of mass of a given subhalo.

Arguments

data_dict::Dict: Data dictionary (see makeDataDict for the canonical description). This function requires the following blocks to be present:

:subhalo => ["S_Vel"]

subhalo_abs_idx::Int: Absolute index of the target subhalo (subfind). Starts at 1.

Returns

The velocity of the center of mass.

GalaxyInspector.computeVcm — Method

computeVcm(data_dict::Dict, cm_type::Symbol)::Vector{<:Unitful.Velocity}

Compute a characteristic velocity for the system.

Arguments

data_dict::Dict: Data dictionary (see makeDataDict for the canonical description). This function requires the following blocks to be present, depending on the value of cm_type:

If cm_type == :all:

["VEL ", "MASS"] for every cell/particle type in the snapshot.

If haskey(PARTICLE_INDEX, cm_type):

cm_type => ["VEL ", "MASS"].

If cm_type == :zero:

No blocks are required

cm_type::Symbol: It can be:

:all -> Velocity of the center of mass of the whole system.

:{component} -> Velocity of the center of mass of the given component (e.g. :stellar, :gas, :darkmatter, etc). It can be any of the keys of [`PARTICLEINDEX`](@ref).

:zero -> 0 velocity.

Returns

The velocity.

GalaxyInspector.computeVcm — Method

computeVcm(
    velocities::Matrix{<:Unitful.Velocity},
    masses::Vector{<:Unitful.Mass},
)::Vector{<:Unitful.Velocity}

Compute the velocity of the center of mass.

Arguments

velocities::Matrix{<:Unitful.Velocity}: Velocities of the cells/particles. Each column is a cell/particle and each row a dimension.

masses::Vector{<:Unitful.Mass}: Mass of every cell/particle.

Returns

The velocity of the center of mass of the cells/particles.

GalaxyInspector.computeVpolar — Method

computeVpolar(
    data_dict::Dict,
    component::Symbol,
    vel_type::Symbol,
)::Vector{<:Unitful.Velocity}

Compute the cylindrical components of the velocity, $\mathbf{\vec{v}} = v_r \, \mathbf{e_r} + v_\theta \, \mathbf{e_\theta} + v_z \, \mathbf{e_z}$, for each cell/particle of the given component.

The speed in the radial direction expressed in Cartesian coordinates is

\[v_r = \frac{x \, v_x + y \, v_y}{\sqrt(x^2 + y^2)} \, ,\]

in the tangential direction is

\[v_\tau = \frac{x \, v_y - y \, v_x}{\sqrt(x^2 + y^2)} \, ,\]

and the speed in the z direction will be computes as

\[v^*_z = v_z \, \mathrm{sign}(z) \, ,\]

in order to distinguish between inflows ($v^*_z < 0$) and outflows ($v^*_z > 0$).

Arguments

data_dict::Dict: Data dictionary (see makeDataDict for the canonical description). This function requires the following blocks to be present for the cell/particle type corresponding to component:

cell/particle type => ["POS ", "VEL "].

vel_type::Symbol: Which velocity will be calculated. The options are:

:radial -> Stellar radial speed ($v_r$).

:tangential -> Stellar tangential speed ($v_\theta$).

:zstar -> Stellar speed in the z direction, computed as $v_z \, \mathrm{sign}(z)$.

component::Symbol: Target component. It can only be one of the elements of COMPONENTS.

Returns

The chosen cylindrical velocity.

GalaxyInspector.computeVpolar — Method

computeVpolar(
    positions::Matrix{<:Unitful.Length},
    velocities::Matrix{<:Unitful.Velocity},
    vel_type::Symbol,
)::Vector{<:Unitful.Velocity}

Compute the cylindrical components of the velocity, $\mathbf{\vec{v}} = v_r \, \mathbf{e_r} + v_\theta \, \mathbf{e_\theta} + v_z \, \mathbf{e_z}$.

The speed in the radial direction expressed in Cartesian coordinates is

\[v_r = \frac{x \, v_x + y \, v_y}{\sqrt(x^2 + y^2)} \, ,\]

in the tangential direction is

\[v_\tau = \frac{x \, v_y - y \, v_x}{\sqrt(x^2 + y^2)} \, ,\]

and the speed in the z direction will be computes as

\[v^*_z = v_z \, \mathrm{sign}(z) \, ,\]

in order to distinguish between inflows ($v^*_z < 0$) and outflows ($v^*_z > 0$).

Arguments

positions::Matrix{<:Unitful.Length}: Positions of the cells/particles. Each column is a cell/particle and each row a dimension.

velocities::Matrix{<:Unitful.Velocity}: Velocities of the cells/particles. Each column is a cell/particle and each row a dimension.

vel_type::Symbol: Which velocity will be calculated. The options are:

:radial -> Radial speed ($v_r$).

:tangential -> Tangential speed ($v_\theta$).

:zstar -> Speed in the z direction, computed as $v_z \, \mathrm{sign}(z)$.

Returns

The chosen cylindrical velocity of each cell/particle.