Data analysis functions

These functions are used internally and none are exported.

They may depend on the particulars of the simulation code (e.g. internal units).

These functions read the data generated by the data acquisition functions, and produce the values that will be plotted.

Signature for the plotSnapshot function

  da_function(data_dict, args...; kwargs...) -> (processed_data, ...)

where:

data_dict::Dict

processed_data::Union{VecOrMat{<:Number},Nothing}

A data analysis functions for the pipeline plotSnapshot must take a dictionary with the following shape:

:sim_data -> ::Simulation (see Simulation).

:snap_data -> ::Snapshot (see Snapshot).

:gc_data -> ::GroupCatalog (see GroupCatalog).

cell/particle type -> (block name -> data for block name, block name -> data of block name, ...).

cell/particle type -> (block name -> data for block name, block name -> data of block name, ...).

cell/particle type -> (block name -> data for block name, block name -> data of block name, ...).

...

groupcat type -> (block name -> data for block name, block name -> data of block name, ...).

groupcat type -> (block name -> data for block name, block name -> data of block name, ...).

groupcat type -> (block name -> data for block name, block name -> data of block name, ...).

...

and return one or more vectors or matrices with the processed data (see plotSnapshot for specifics). It should return nothing if the input data has some problem that prevents computation (e.g. it is empty).

Signature for the plotTimeSeries function

  da_function(sim_data, args...; kw_args...) -> (processed_data_x, processed_data_y)

where:

sim_data::Simulation, see Simulation

processed_data_x::Vector{<:Number}

processed_data_y::Vector{<:Number}

A data analysis functions for the pipeline plotTimeSeries must take a Simulation struct, and return two vectors. It should return nothing if the input data has some problem that prevents computation (e.g. is empty).

GalaxyInspector.daBandProfile — Method

daBandProfile(
    data_dict::Dict,
    quantity::Symbol,
    grid::CircularGrid;
    <keyword arguments>
)::Union{
    Tuple{Vector{<:Unitful.Length},Vector{<:Number},Vector{<:Number},Vector{<:Number}},
    Tuple{Vector{<:Unitful.Length},Vector{<:Number},Vector{<:Number}},
    Nothing,
}

Compute a profile with uncertainty bands.

Arguments

data_dict::Dict: Data dictionary (see makeDataDict for the canonical description).

quantity::Symbol: Target quantity. It can be any of the quantities valid for scatterQty.

grid::CircularGrid: Circular grid.

flat::Bool=true: If the profile will be 2D (rings), or 3D (spherical shells).

density::Bool=false: If the profile will be of the density of quantity.

ylog::Bool=false: If true, returns the profile of $\log_{10}$(quantity).

error_bar::Bool=false: If the returned values will be compatible with band! (default) or with errorbars!.

filter_function::Function=filterNothing: Filter function to be applied to data_dict before any other computation. See the required signature and examples in ./src/analysis/filters.jl.

Returns

A tuple with three elements if error_bar = false:

A vector with the position of each ring or spherical shells.

A vector with the "low" value of quantity in each each ring or spherical shells.

A vector with the "high" value of quantity in each each ring or spherical shells.

Or a tuple with four elements if error_bar = true:

A vector with the position of each ring or spherical shells.

A vector with the central value of quantity in each each ring or spherical shells.

A vector with the "low" error of quantity in each each ring or spherical shells.

A vector with the "high" error of quantity in each each ring or spherical shells.

It returns nothing if any of the necessary quantities are missing.

GalaxyInspector.daBarGasFractions — Method

daBarGasFractions(
    data_dict::Dict,
    quantity::Symbol,
    grid::LinearGrid;
    <keyword arguments>
)::Union{NTuple{2,Vector{<:Number}},Nothing}

Compute a bar plot, where the bins are for a given gas quantity and the heights are the corresponding gas fractions.

Arguments

data_dict::Dict: Data dictionary (see makeDataDict for the canonical description).

quantity::Symbol: Target quantity. It can be any of the quantities valid for scatterQty with the cell/particle type :gas.

grid::LinearGrid: Linear grid.

components::Vector{Symbol}=[:ode_ionized, :ode_atomic, :ode_molecular]: List of gas components to be considered. The fractions will be normalized to this list of components. See COMPONENTS for options.

filter_function::Function=filterNothing: Filter function to be applied to data_dict before any other computation. See the required signature and examples in ./src/analysis/filters.jl.

Returns

A tuple with the two elements required by barplot! from Makie:

A vector with the index of each bar.

A vector with the height of each bar.

GalaxyInspector.daCPUtxt — Method

daCPUtxt(
    sim_data::Simulation,
    target::String,
    x_quantity::Symbol,
    y_quantity::Symbol;
    <keyword arguments>
)::NTuple{2,Vector{<:Number}}

Compute the evolution of a measured quantity in the cpu.txt file, for a given target process.

Arguments

sim_data::Simulation: The Simulation struct for the target simulation.

target::String: Target process.

x_quantity::Symbol: Quantity for the x axis. The options are:

:time_step -> Time step.

:physical_time -> Physical time since the Big Bang.

:clock_time_s -> Clock time duration of the time step in seconds.

:clock_time_percent -> Clock time duration of the time step as a percentage.

:tot_clock_time_s -> Total clock time in seconds.

:tot_clock_time_percent -> Total clock time as a percentage.

y_quantity::Symbol: Quantity for the y axis. The options are:

:time_step -> Time step.

:physical_time -> Physical time since the Big Bang.

:clock_time_s -> Clock time duration of the time step in seconds.

:clock_time_percent -> Clock time duration of the time step as a percentage.

:tot_clock_time_s -> Total clock time in seconds.

:tot_clock_time_percent -> Total clock time as a percentage.

y_log::Union{Unitful.Units,Nothing}=nothing: Target unit for y_quantity, if you want to apply $\log_{10}$ to the y_quantity. If set to nothing, the data from cpu.txt is left as is.

smooth::Int=0: The result will be smoothed out using smooth bins. Set it to 0 if you want no smoothing.

Returns

A Tuple with two elements:

A Vector with the time series of x_quantity.

A Vector with the time series of y_quantity.

GalaxyInspector.daClumpingFactor — Method

daClumpingFactor(
    data_dict::Dict,
    component::Symbol;
    <keyword arguments>
)::Tuple{Vector{<:Number},Vector{Float64}}

Compute the clumping factor ($C_\rho$), for the number density of component, at different volume scales.

\[C_\rho = \frac{\langle n^2 \rangle}{\langle n \rangle^2} \, ,\]

Arguments

data_dict::Dict: Data dictionary (see makeDataDict for the canonical description).

component::Symbol: Target component. It can only be one of the elements of COMPONENTS with cell/partcile type :gas.

n_neighbors::Int=32: Number of neighbors.

filter_function::Function=filterNothing: Filter function to be applied to data_dict before any other computation. See the required signature and examples in ./src/analysis/filters.jl.

x_log::Union{Unitful.Units,Nothing}=nothing: Target unit for the volume, if you want to apply $\log_{10}$ to the volume. If set to nothing, the volume is left as is.

y_log::Union{Unitful.Units,Nothing}=nothing: Target unit for the clumping factor, if you want to apply $\log_{10}$ to the clumping factor. If set tonothing`, the clumping factor is left as is.

Returns

A tuple with two elements:

A vector with the the volumes.

A vector with the clumping factors.

GalaxyInspector.daClumpingFactorProfile — Method

daClumpingFactorProfile(
    data_dict::Dict,
    component::Symbol,
    grid::CircularGrid;
    <keyword arguments>
)::Tuple{Vector{<:Unitful.Length},Vector{Float64}}

Compute a clumping factor ($C_\rho$) profile, for the number density of component.

\[C_\rho = \frac{\langle n^2 \rangle}{\langle n \rangle^2} \, ,\]

Arguments

data_dict::Dict: Data dictionary (see makeDataDict for the canonical description).

component::Symbol: Target component. It can only be one of the elements of COMPONENTS with cell/partcile type :gas.

grid::CircularGrid: Circular grid.

filter_function::Function=filterNothing: Filter function to be applied to data_dict before any other computation. See the required signature and examples in ./src/analysis/filters.jl.

Returns

A tuple with two elements:

A vector with the central position of each bin.

A vector with the clumping factors.

GalaxyInspector.daDensity2DProjection — Method

daDensity2DProjection(
    data_dict::Dict,
    grid::CubicGrid,
    component::Symbol,
    field_type::Symbol;
    <keyword arguments>
)::Tuple{Vector{<:Unitful.Length},Vector{<:Unitful.Length},VecOrMat{Float64}}

Project a 3D density field into a given plane.

Note

If the source of the field are particles, a simple 2D histogram is used. If they are Voronoi cells, the density of the cells that cross the line of sight of each pixel are added up.

Arguments

data_dict::Dict: Data dictionary (see makeDataDict for the canonical description).

grid::CubicGrid: Cubic grid.

component::Symbol: Target component. It can only be one of the elements of COMPONENTS.

field_type::Symbol: If the field is made up of :particles or Voronoi :cells.

projection_plane::Symbol=:xy: Projection plane. The options are :xy, :xz, and :yz.

reduce_grid::Symbol=:square: Type of 2D grid to do the final projection. The options are:

:square -> The density distribution will be projected into a regular square grid, with a resolution reduce_factor times lower than grid. This emulates the way the surface densities are measured in observations. reduce_factor = 1 means no reduction in resolution.

:circular -> The density distribution will be projected into a flat circular grid, formed by a series of reduce_factor concentric rings. This emulates the traditional way the Kennicutt-Schmidt law is measured in some simulations. reduce_factor = 1 means that the result will be a single point. Note that this behaves the opposite way than the reduce_grid = :square case.

reduce_factor::Int=1: Factor by which the resolution of the result will be reduced. This will be applied after the density projection. If reduce_grid = :square, the new values will be computed averaging the values of neighboring pixels. reduce_factor has to divide the size of grid exactly. If reduce_grid = :circular, the new values will be computed averaging the values of the pixels the fall within each of the reduce_factor concentric rings.

m_unit::Unitful.Units=u"Msun": Mass unit.

l_unit::Unitful.Units=u"pc": Length unit.

filter_function::Function=filterNothing: Filter function to be applied to data_dict before any other computation. See the required signature and examples in ./src/analysis/filters.jl.

Returns

A tuple with three elements:

A vector with the x coordinates of the grid.

A vector with the y coordinates of the grid.

A matrix or vector with the $\log_{10}$ of the density at each bin of the 2D grid.

GalaxyInspector.daDiskAccretion — Method

daDiskAccretion(
    sim_data::Simulation;
    <keyword arguments>
)::NTuple{2,Vector{<:Number}}

Compute the evolution of the accreted mass into a given galactic disc.

Arguments

sim_data::Simulation: The Simulation struct for the target simulation.

component::Symbol: Component to compute the accreted mass for. The options are:

:dark_matter -> Dark matter.

:black_hole -> Black holes.

:gas -> Gas.

:stellar -> Stars.

:all -> All the matter.

flux_direction::Symbol=:net: What flux direction will be plotted. The options are:

:net_mass -> Net accreted mass.

:inflow_mass -> Inflow mass only.

:outflow_mass -> Outflow mass only.

max_r::Unitful.Length=DISK_R: Radius of the disk.

max_z::Unitful.Length=5.0u"kpc": Half height of the disk.

trans_mode::Union{Symbol,Tuple{TranslationType,RotationType,Dict{Symbol,Vector{String}}}}=:all_box: How to translate and rotate the cells/particles, before filtering with filter_mode. For options see selectTransformation.

tracers::Bool=false: If tracers will be use to compute the mass accretion.

y_log::Union{Unitful.Units,Nothing}=nothing: Target unit for integrated mass flux, if you want to apply $\log_{10}$ to it. If set to nothing, the data from computeDiskAccretion is left as is.

smooth::Int=0: The time series will be smoothed out using smooth bins. Set it to 0 if you want no smoothing.

show_progress::Bool=true: If a progress bar will be shown.

Returns

A Tuple with two elements:

A Vector with the physical times.

A Vector with the accreted mass at each time.

GalaxyInspector.daEvolution — Method

daEvolution(
    sim_data::Simulation,
    x_quantity::Symbol,
    y_quantity::Symbol;
    <keyword arguments>
)::NTuple{2,Vector{<:Number}}

Compute the time series of two quantities.

Arguments

sim_data::Simulation: The Simulation struct for the target simulation.

x_quantity::Symbol: Quantity for the x axis.

y_quantity::Symbol: Quantity for the y axis.

integration_functions::NTuple{2,Function}=(integrateQty, integrateQty): Functions to compute the integral value of x_quantity and y_quantity at a given time. The functions must have the signature:

integration_functions(data_dict::Dict, quantity::Symbol)::Number

where

data_dict::Dict: Data dictionary (see makeDataDict for the canonical description).

quantity::Symbol: The quantity to be integrated.

trans_mode::Union{Symbol,Tuple{TranslationType,RotationType,Dict{Symbol,Vector{String}}}}=:all_box: How to translate and rotate the cells/particles, before filtering with filter_mode. For options see selectTransformation.

filter_mode::Union{Symbol,Tuple{Function,Dict{Symbol,Vector{String}}}}=:all: Which cells/particles will be selected. For options see selectFilter.

extra_filter::Function=filterNothing: Filter function to be applied after trans_mode and filter_mode are applied. See the required signature and examples in ./src/analysis/filters.jl.

ff_request::Dict{Symbol,Vector{String}}=Dict{Symbol,Vector{String}}(): Request dictionary for extra_filter.

x_log::Union{Unitful.Units,Nothing}=nothing: Target unit for x_quantity, if you want to apply $\log_{10}$ to the x_quantity. If set to nothing, the data from scatterQty is left as is.

y_log::Union{Unitful.Units,Nothing}=nothing: Target unit for y_quantity, if you want to apply $\log_{10}$ to the y_quantity. If set to nothing, the data from scatterQty is left as is.

smooth::Int=0: The result of integrateQty will be smoothed out using smooth bins. Set it to 0 if you want no smoothing.

cumulative::Bool=false: If the y_quantity will be accumulated or not.

show_progress::Bool=true: If a progress bar will be shown.

Returns

A Tuple with two elements:

A Vector with the time series of x_quantity.

A Vector with the time series of y_quantity.

GalaxyInspector.daGasSFR2DProjection — Method

daGasSFR2DProjection(
    data_dict::Dict,
    grid::CubicGrid,
    field_type::Symbol;
    <keyword arguments>
)::Tuple{Vector{<:Unitful.Length},Vector{<:Unitful.Length},VecOrMat{Float64}}

Project the 3D gas SFR field into a given plane.

Note

If the source of the field are particles, a simple 2D histogram is used. If they are Voronoi cells instead, the SFR of the cells that cross the line of sight of each pixel are added up.

Arguments

data_dict::Dict: Data dictionary (see makeDataDict for the canonical description).

grid::CubicGrid: Cubic grid.

field_type::Symbol: If the field is made up of :particles or Voronoi :cells.

projection_plane::Symbol=:xy: Projection plane. The options are :xy, :xz, and :yz.

reduce_grid::Symbol=:square: Type of 2D grid to do the final projection. The options are:

:square -> The density distribution will be projected into a regular square grid, with a resolution reduce_factor times lower than grid. This emulates the way the surface densities are measured in observations. reduce_factor = 1 means no reduction in resolution.

:circular -> The density distribution will be projected into a flat circular grid, formed by a series of reduce_factor concentric rings. This emulates the traditional way the Kennicutt-Schmidt law is measured in some simulations. reduce_factor = 1 means that the result will be a single point. Note that this behaves the opposite way than the reduce_grid = :square case.

reduce_factor::Int=1: Factor by which the resolution of the result will be reduced. This will be applied after the density projection. If reduce_grid = :square, the new values will be computed averaging the values of neighboring pixels. reduce_factor has to divide the size of grid exactly. If reduce_grid = :circular, the new values will be computed averaging the values of the pixels the fall within each of the reduce_factor concentric rings.

m_unit::Unitful.Units=u"Msun": Mass unit.

t_unit::Unitful.Units=u"yr": Time unit.

filter_function::Function=filterNothing: Filter function to be applied to data_dict before any other computation. See the required signature and examples in ./src/analysis/filters.jl.

Returns

A tuple with three elements:

A vector with the x coordinates of the grid.

A vector with the y coordinates of the grid.

A matrix or vector with the $\log_{10}$ of the gas SFR at each bin of the 2D grid.

GalaxyInspector.daHistogram — Method

daHistogram(
    data_dict::Dict,
    quantity::Symbol,
    grid::LinearGrid;
    <keyword arguments>
)::Union{Tuple{Vector{<:Number},Vector{<:Number}},Nothing}

Compute a 1D histogram of a given quantity.

Arguments

data_dict::Dict: Data dictionary (see makeDataDict for the canonical description).

quantity::Symbol: Target quantity. It can be any of the quantities valid for scatterQty.

grid::LinearGrid: Linear grid.

norm::Int=0: Number of counts that will be use to normalize the histogram. If left as 0, the histogram will be normalize with the maximum bin count.

filter_function::Function=filterNothing: Filter function to be applied to data_dict before any other computation. See the required signature and examples in ./src/analysis/filters.jl.

Returns

A tuple with two elements:

A vector with the value of quantity corresponding to each bin.

A vector with the counts.

GalaxyInspector.daHistogram — Method

daHistogram(
    data_dict::Dict,
    quantity::Symbol,
    n_bins::Int,
    log::Bool;
    <keyword arguments>
)::Union{Tuple{Vector{<:Number},Vector{<:Number}},Nothing}

Compute a 1D histogram of a given quantity.

Note

This method uses the extrema of the values of quantity as the range of the histogram.

Arguments

data_dict::Dict: Data dictionary (see makeDataDict for the canonical description).

quantity::Symbol: Target quantity. It can be any of the quantities valid for scatterQty.

n_bins::Int: Number of bins.

log::Bool: If the histogram bins will be logarithmic.

norm::Int=0: Number of count that will be use to normalize the histogram. If left as 0, the histogram will be normalize with the maximum bin count.

filter_function::Function=filterNothing: Filter function to be applied to data_dict before any other computation. See the required signature and examples in ./src/analysis/filters.jl.

Returns

A tuple with two elements:

A vector with the value of quantity corresponding to each bin.

A vector with the counts.

GalaxyInspector.daIntegrateGalaxy — Method

daIntegrateGalaxy(
    data_dict::Dict,
    x_quantity::Symbol,
    y_quantity::Symbol;
    <keyword arguments>
)::NTuple{2,Vector{<:Number}}

Compute two global quantities for the simulation.

Arguments

data_dict::Dict: Data dictionary (see makeDataDict for the canonical description).

x_quantity::Symbol: Quantity for the x axis. It can be any of the quantities valid for integrateQty.

y_quantity::Symbol: Quantity for the y axis. It can be any of the quantities valid for integrateQty.

filter_function::Function=filterNothing: Filter function to be applied to data_dict before any other computation. See the required signature and examples in ./src/analysis/filters.jl.

Returns

A tuple with two elements:

A single element vector with the value of x_quantity.

A single element vector with the value of y_quantity.

GalaxyInspector.daMetallicity2DProjection — Method

daMetallicity2DProjection(
    data_dict::Dict,
    grid::CubicGrid,
    component::Symbol,
    field_type::Symbol;
    <keyword arguments>
)::Tuple{Vector{<:Unitful.Length},Vector{<:Unitful.Length},VecOrMat{Float64}}

Project the 3D metallicity field to a given plane.

Note

If if element = :all, the total metallicity (in solar units) is computed. If element = :X, the abundance of element X is computed, ABUNDANCE_SHIFT + ``\log_{10}(X/H). In both cases the total value of the column given by the line of sight of each pixel is computed.

Arguments

data_dict::Dict: Data dictionary (see makeDataDict for the canonical description).

grid::CubicGrid: Cubic grid.

component::Symbol: Target component. It can be either :stellar or :gas.

field_type::Symbol: If the field is made up of :particles or Voronoi :cells.

element::Symbol=:all: Target element. The possibilities are the keys of ELEMENT_INDEX. Set it to :all if you want the total metallicity.

projection_plane::Symbol=:xy: Projection plane. The options are :xy, :xz, and :yz.

reduce_grid::Symbol=:square: Type of 2D grid to do the final projection. The options are:

:square -> The density distribution will be projected into a regular square grid, with a resolution reduce_factor times lower than grid. This emulates the way the surface densities are measured in observations. reduce_factor = 1 means no reduction in resolution.

:circular -> The density distribution will be projected into a flat circular grid, formed by a series of reduce_factor concentric rings. This emulates the traditional way the Kennicutt-Schmidt law is measured in some simulations. reduce_factor = 1 means that the result will be a single point. Note that this behaves the opposite way than the reduce_grid = :square case.

reduce_factor::Int=1: Factor by which the resolution of the result will be reduced. This will be applied after the density projection. If reduce_grid = :square, the new values will be computed averaging the values of neighboring pixels. reduce_factor has to divide the size of grid exactly. If reduce_grid = :circular, the new values will be computed averaging the values of the pixels the fall within each of the reduce_factor concentric rings.

filter_function::Function=filterNothing: Filter function to be applied to data_dict before any other computation. See the required signature and examples in ./src/analysis/filters.jl.

Returns

A tuple with three elements:

A vector with the x coordinates of the grid.

A vector with the y coordinates of the grid.

A matrix or vector with the $\log_{10}$ of the metallicity at each bin of the 2D grid.

GalaxyInspector.daMolla2015 — Method

daMolla2015(
    data_dict::Dict,
    grid::CircularGrid,
    quantity::Symbol;
    <keyword arguments>
)::Union{
    Tuple{
        Vector{<:Unitful.Length},
        <:Union{Vector{<:SurfaceDensity},Vector{<:MassFlowDensity},Vector{Float64}}
    },
    Nothing,
}

Compute a profile for the Milky Way, compatible with the experimental data in Mollá et al. (2015).

Arguments

data_dict::Dict: Data dictionary (see makeDataDict for the canonical description).

grid::CircularGrid: Circular grid.

quantity::Symbol: Quantity for the y axis. The options are:

:stellar_area_density -> Stellar mass surface density.

:sfr_area_density -> Star formation rate surface density.

:ode_molecular_area_density -> Molecular mass surface density.

:br_molecular_area_density -> Molecular mass surface density, computed using the pressure relation in Blitz et al. (2006).

:ode_atomic_area_density -> Atomic mass surface density.

:br_atomic_area_density -> Atomic mass surface density, computed using the pressure relation in Blitz et al. (2006).

:O_stellar_abundance -> Stellar abundance of oxygen, as $12 + \log_{10}(\mathrm{O \, / \, H})$.

:N_stellar_abundance -> Stellar abundance of nitrogen, as $12 + \log_{10}(\mathrm{N \, / \, H})$.

:C_stellar_abundance -> Stellar abundance of carbon, as $12 + \log_{10}(\mathrm{C \, / \, H})$.

y_unit::Unitful.Units=Unitful.NoUnits: Unit for quantity.

filter_function::Function=filterNothing: Filter function to be applied to data_dict before any other computation. See the required signature and examples in ./src/analysis/filters.jl.

Returns

A tuple with two elements:

A vector with the position of each ring.

A vector with the value quantity in each ring.

It returns nothing if any of the necessary quantities are missing.

References

L. Blitz et al. (2006). The Role of Pressure in GMC Formation II: The H2-Pressure Relation. The Astrophysical Journal, 650(2), 933. doi:10.1086/505417

M. Mollá et al. (2015). Galactic chemical evolution: stellar yields and the initial mass function. Monthly Notices of the Royal Astronomical Society 451(4), 3693–3708. doi:10.1093/mnras/stv1102

GalaxyInspector.daProfile — Method

daProfile(
    data_dict::Dict,
    quantity::Symbol,
    grid::CircularGrid;
    <keyword arguments>
)::Union{Tuple{Vector{<:Unitful.Length},Vector{<:Number}},Nothing}

Compute a profile.

Arguments

data_dict::Dict: Data dictionary (see makeDataDict for the canonical description).

quantity::Symbol: Target quantity. It can be any of the quantities valid for scatterQty.

grid::CircularGrid: Circular grid.

norm::Union{Symbol,Nothing}=nothing: The value of quantity in each bin will be divided by the corresponding value of norm. It can be any of the quantities valid for scatterQty. If set to nothing, no operation is applied.

y_log::Union{Unitful.Units,Nothing}=nothing: Target unit for quantity, if you want to apply $\log_{10}$ to quantity. If set to nothing, the data is left as is.

flat::Bool=true: If the profile will be 2D (rings), or 3D (spherical shells).

total::Bool=true: If the sum (default) or the mean of quantity will be computed for each bin. This affects the values of norm too.

cumulative::Bool=false: If the profile will be accumulated (after dividing by norm).

density::Bool=false: If the profile will be of the density of quantity (after dividing by norm).

filter_function::Function=filterNothing: Filter function to be applied to data_dict before any other computation. See the required signature and examples in ./src/analysis/filters.jl.

Returns

A tuple with two elements:

A vector with the position of each ring or spherical shells.

A vector with the value of quantity in each each ring or spherical shells.

It returns nothing if any of the necessary quantities are missing.

GalaxyInspector.daRotationCurve — Method

daRotationCurve(
    data_dict::Dict;
    <keyword arguments>
)::Tuple{Vector{<:Unitful.Length},Vector{<:Unitful.Velocity}}

Compute a rotation curve.

Arguments

data_dict::Dict: Data dictionary (see makeDataDict for the canonical description).

R::Unitful.Length=DISK_R: Maximum radial distance for the rotation curve.

filter_function::Function=filterNothing: Filter function to be applied to data_dict before any other computation. See the required signature and examples in ./src/analysis/filters.jl.

Returns

A tuple with two elements:

A vector with the distances to each star.

A vector with the circular velocity of each star.

GalaxyInspector.daSFRtxt — Method

daSFRtxt(
    sim_data::Simulation,
    x_quantity::Symbol,
    y_quantity::Symbol;
    <keyword arguments>
)::NTuple{2,Vector{<:Number}}

Compute the stellar mass or SFR evolution using the data in the sfr.txt file.

Arguments

sim_data::Simulation: The Simulation struct for the target simulation.

x_quantity::Symbol: Quantity for the x axis. The options are:

:scale_factor -> Scale factor.

:redshift -> Redshift.

:physical_time -> Physical time since the Big Bang.

:lookback_time -> Physical time left to reach the last snapshot.

y_quantity::Symbol: Quantity for the y axis. The options are:

:stellar_mass -> Cumulative stellar mass.

:sfr -> Star formation rate.

smooth::Int=0: The result will be smoothed out using smooth bins. Set it to 0 if you want no smoothing.

Returns

A Tuple with two elements:

A Vector with the time series of x_quantity.

A Vector with the time series of y_quantity.

GalaxyInspector.daScatterDensity — Method

daScatterDensity(
    data_dict::Dict,
    x_quantity::Symbol,
    y_quantity::Symbol;
    <keyword arguments>
)::Tuple{Vector{<:Number},Vector{<:Number},Matrix{Float64}}

Compute two quantities for every cell/particle using a 2D histogram.

Arguments

data_dict::Dict: Data dictionary (see makeDataDict for the canonical description).

x_quantity::Symbol: Quantity for the x axis. It can be any of the quantities valid for scatterQty.

y_quantity::Symbol: Quantity for the y axis. It can be any of the quantities valid for scatterQty.

x_range::Union{NTuple{2,<:Number},Nothing}=nothing: x axis range for the histogram grid. If set to nothing, the extrema of the values will be used.

y_range::Union{NTuple{2,<:Number},Nothing}=nothing: y axis range for the histogram grid. If set to nothing, the extrema of the values will be used.

x_log::Union{Unitful.Units,Nothing}=nothing: Target unit for x_quantity, if you want to apply $\log_{10}$ to the x_quantity. If set to nothing, the data from scatterQty is left as is.

y_log::Union{Unitful.Units,Nothing}=nothing: Target unit for y_quantity, if you want to apply $\log_{10}$ to the y_quantity. If set to nothing, the data from scatterQty is left as is.

n_bins::Int=100: Number of bins per side of the grid.

filter_function::Function=filterNothing: Filter function to be applied to data_dict before any other computation. See the required signature and examples in ./src/analysis/filters.jl.

Returns

A tuple with three elements:

A vector with the x coordinates of the grid.

A vector with the y coordinates of the grid.

A matrix with the the $\log_{10}$ of the counts in each bin.

GalaxyInspector.daScatterGalaxy — Method

daScatterGalaxy(
    data_dict::Dict,
    x_quantity::Symbol,
    y_quantity::Symbol;
    <keyword arguments>
)::NTuple{2,Vector{<:Number}}

Compute two quantities for every cell/particle in the simulation using scatterQty.

Arguments

data_dict::Dict: Data dictionary (see makeDataDict for the canonical description).

x_quantity::Symbol: Quantity for the x axis. It can be any of the quantities valid for scatterQty.

y_quantity::Symbol: Quantity for the y axis. It can be any of the quantities valid for scatterQty.

x_log::Union{Unitful.Units,Nothing}=nothing: Target unit for x_quantity, if you want to apply $\log_{10}$ to the x_quantity. If set to nothing, the data from scatterQty is left as is.

y_log::Union{Unitful.Units,Nothing}=nothing: Target unit for y_quantity, if you want to apply $\log_{10}$ to the y_quantity. If set to nothing, the data from scatterQty is left as is.

filter_function::Function=filterNothing: Filter function to be applied to data_dict before any other computation. See the required signature and examples in ./src/analysis/filters.jl.

Returns

A tuple with two elements:

A vector with the values of x_quantity.

A vector with the values of y_quantity.

GalaxyInspector.daScatterWeightedDensity — Method

daScatterWeightedDensity(
    data_dict::Dict,
    x_quantity::Symbol,
    y_quantity::Symbol,
    z_quantity::Symbol;
    <keyword arguments>
)::Tuple{Vector{<:Number},Vector{<:Number},Matrix{Float64}}

Compute two quantities for every cell/particle using a 2D histogram, weighted by z_quantity.

Arguments

data_dict::Dict: Data dictionary (see makeDataDict for the canonical description).

x_quantity::Symbol: Quantity for the x axis. It can be any of the quantities valid for scatterQty.

y_quantity::Symbol: Quantity for the y axis. It can be any of the quantities valid for scatterQty.

z_quantity::Symbol: Quantity for the weights. It can be any of the quantities valid for scatterQty.

x_range::Union{NTuple{2,<:Number},Nothing}=nothing: x axis range. If set to nothing, the extrema of the values will be used.

y_range::Union{NTuple{2,<:Number},Nothing}=nothing: y axis range. If set to nothing, the extrema of the values will be used.

x_log::Union{Unitful.Units,Nothing}=nothing: Target unit for x_quantity, if you want to apply $\log_{10}$ to the x_quantity. If set to nothing, the data from scatterQty is left as is.

y_log::Union{Unitful.Units,Nothing}=nothing: Target unit for y_quantity, if you want to apply $\log_{10}$ to the y_quantity. If set to nothing, the data from scatterQty is left as is.

total::Bool=true: If the sum (default) or the mean of z_quantity will be used as the value of each bin.

n_bins::Int=100: Number of bins per side of the grid.

filter_function::Function=filterNothing: Filter function to be applied to data_dict before any other computation. See the required signature and examples in ./src/analysis/filters.jl.

Returns

A tuple with three elements:

A vector with the x coordinates of the grid.

A vector with the y coordinates of the grid.

A matrix with the $\log_{10}$ of z_quantity for each bin.

GalaxyInspector.daStellarHistory — Method

daStellarHistory(
    data_dict::Dict;
    <keyword arguments>
)::Union{Tuple{Vector{<:Unitful.Time},Vector{<:Number}},Nothing}

Compute the evolution of a given stellar quantity using the stellar ages at a given instant in time.

Arguments

data_dict::Dict: Data dictionary (see makeDataDict for the canonical description).

quantity::Symbol=:sfr: Target quantity. The options are:

:sfr -> Star formation rate.

:ssfr -> The specific star formation rate.

:stellar_mass -> Stellar mass.

:stellar_metallicity -> Mass fraction of all elements above He in the stars (solar units).

n_bins::Int=100: Number of bins (time intervals).

ylog::Bool=false: If true, returns $\log_{10}$(quantity).

filter_function::Function=filterNothing: Filter function to be applied to data_dict before any other computation. See the required signature and examples in ./src/analysis/filters.jl.

Returns

A tuple with two elements:

A vector with the physical times.

A vector with the values of quantity at each time.

GalaxyInspector.daTemperature2DProjection — Method

daTemperature2DProjection(
    data_dict::Dict,
    grid::CubicGrid,
    field_type::Symbol;
    <keyword arguments>
)::Tuple{Vector{<:Unitful.Length},Vector{<:Unitful.Length},VecOrMat{Float64}}

Project the 3D temperature field to a given plane.

Note

The gas temperature is the mean value in the column given by the line of sight of each pixel.

Arguments

data_dict::Dict: Data dictionary (see makeDataDict for the canonical description).

grid::CubicGrid: Cubic grid.

field_type::Symbol: If the field is made up of :particles or Voronoi :cells.

projection_plane::Symbol=:xy: Projection plane. The options are :xy, :xz, and :yz.

reduce_grid::Symbol=:square: Type of 2D grid to do the final projection. The options are:

:square -> The density distribution will be projected into a regular square grid, with a resolution reduce_factor times lower than grid. This emulates the way the surface densities are measured in observations. reduce_factor = 1 means no reduction in resolution.

:circular -> The density distribution will be projected into a flat circular grid, formed by a series of reduce_factor concentric rings. This emulates the traditional way the Kennicutt-Schmidt law is measured in some simulations. reduce_factor = 1 means that the result will be a single point. Note that this behaves the opposite way than the reduce_grid = :square case.

reduce_factor::Int=1: Factor by which the resolution of the result will be reduced. This will be applied after the density projection. If reduce_grid = :square, the new values will be computed averaging the values of neighboring pixels. reduce_factor has to divide the size of grid exactly. If reduce_grid = :circular, the new values will be computed averaging the values of the pixels the fall within each of the reduce_factor concentric rings.

filter_function::Function=filterNothing: Filter function to be applied to data_dict before any other computation. See the required signature and examples in ./src/analysis/filters.jl.

Returns

A tuple with three elements:

A vector with the x coordinates of the grid.

A vector with the y coordinates of the grid.

A matrix or vector with the $\log_{10}$ of the temperature at each bin of the 2D grid.

GalaxyInspector.daTrajectory — Method

daTrajectory(
    simulation_path::String,
    target_ids::Vector{<:Unsigned},
    cp_type::Symbol;
    <keyword arguments>
)::Dict{UInt64,Matrix{Quantity}}

Compute the trajectory of a set of cells/particles, given their IDs.

Arguments

simulation_path::String: Path to the simulation directory, set in the code variable OutputDir.

target_ids::Vector{<:Unsigned}: IDs of the cells/particles whose trajectories will be computed.

cp_type::Symbol: Target type of cell/particle. The possibilities are the keys of PARTICLE_INDEX.

trans_mode::Union{Symbol,Tuple{TranslationType,RotationType,Dict{Symbol,Vector{String}}}}=:all_box: How to translate and rotate the cells/particles, before filtering with filter_mode. For options see selectTransformation.

filter_mode::Union{Symbol,Tuple{Function,Dict{Symbol,Vector{String}}}}=:all: Which cells/particles will be selected. For options see selectFilter.

Returns

A dictionary with the following shape:

target_id::UInt64 -> A matrix with the trajectory of the cell/particle with ID target_id. The matrix has 7 rows, where:

Row 1: Physical time.

Rows 2-4: Position [x, y, z].

Rows 5-7: Velocity [vx, vy, vz].

If the target ID is not found at a given physical time the position and velocity are NaN.

GalaxyInspector.daVSFLaw — Method

daVSFLaw(
    data_dict::Dict,
    grid::CubicGrid,
    component::Symbol;
    <keyword arguments>
)::Union{NTuple{2,Vector{<:Float64}},Nothing}

Compute the gas density and the SFR density, used in the volumetric star formation (VSF) law.

Arguments

data_dict::Dict: Data dictionary (see makeDataDict for the canonical description).

grid::CubicGrid: Cubic grid.

component::Symbol: Target component. It can only be one of the elements of COMPONENTS.

field_type::Symbol=:cells: If the gas surface density will be calculated assuming the gas is in :particles or in Voronoi :cells.

stellar_ff::Function=filterNothing: Filter function for the stars. See the required signature and examples in ./src/analysis/filters.jl.

gas_ff::Function=filterNothing: Filter function for the gas. See the required signature and examples in ./src/analysis/filters.jl.

Returns

A tuple with two elements:

A vector with $\log_{10}(\rho_\mathrm{H} \, / \, \mathrm{M_\odot \, pc^{-3}})$

A vector with $\log_{10}(\rho_\mathrm{SFR} \, / \, \mathrm{M_\odot \, yr^{-1} \, kpc^{-3}})$

It returns nothing if any of the necessary quantities are missing.

GalaxyInspector.daVelocityField — Method

daVelocityField(
    data_dict::Dict,
    grid::SquareGrid,
    component::Symbol;
    <keyword arguments>
)::Tuple{Vector{<:Unitful.Length},Vector{<:Unitful.Length},Matrix{Float64},Matrix{Float64}}

Project a 3D velocity field into a given plane.

Arguments

data_dict::Dict: Data dictionary (see makeDataDict for the canonical description).

grid::SquareGrid: Square grid.

component::Symbol: Target component. It can only be one of the elements of COMPONENTS.

projection_plane::Symbol=:xy: Projection plane. The options are :xy, :xz, and :yz.

v_unit::Unitful.Units=u"km * s^-1",: Velocity unit

filter_function::Function=filterNothing: Filter function to be applied to data_dict before any other computation. See the required signature and examples in ./src/analysis/filters.jl.

Returns

A tuple with four elements:

A vector with the x coordinates of the grid.

A vector with the y coordinates of the grid.

A matrix with the mean velocity in the x direction at each grid point.

A matrix with the mean velocity in the y direction at each grid point.

GalaxyInspector.daVirialAccretion — Method

daVirialAccretion(
    sim_data::Simulation;
    <keyword arguments>
)::NTuple{2,Vector{<:Number}}

Compute the evolution of the accreted mass into a sphere with the virial radius.

Arguments

sim_data::Simulation: The Simulation struct for the target simulation.

component::Symbol: Component to compute the accreted mass for. The options are:

:dark_matter -> Dark matter.

:black_hole -> Black holes.

:gas -> Gas.

:stellar -> Stars.

:all -> All the matter.

flux_direction::Symbol=:net: What flux direction will be plotted. The options are:

:net_mass -> Net accreted mass.

:inflow_mass -> Inflow mass only.

:outflow_mass -> Outflow mass only.

halo_idx::Int=1: Index of the target halo (FoF group). Starts at 1.

tracers::Bool=false: If tracers will be use to compute the mass accretion.

y_log::Union{Unitful.Units,Nothing}=nothing: Target unit for integrated mass flux, if you want to apply $\log_{10}$ to it. If set to nothing, the data from computeVirialAccretion is left as is.

smooth::Int=0: The time series will be smoothed out using smooth bins. Set it to 0 if you want no smoothing.

show_progress::Bool=true: If a progress bar will be shown.

Returns

A Tuple with two elements:

A Vector with the physical times.

A Vector with the accreted mass at each time.