Tracer functions

These functions are used internally and none are exported.

They may depend on the particulars of the simulation code (e.g. internal units).

GalaxyInspector.filterExistIDs! — Method

filterExistIDs!(target_ids::Vector{UInt}, component::Symbol, data_dict::Dict)::Nothing

Filter out IDs from target_ids that do not exist for component.

Arguments

data_dict::Dict: Data dictionary (see makeDataDict for the canonical description).

component::Symbol: Target component. It can only be one of the elements of COMPONENTS.

target_ids::Vector{UInt}: IDs to be filtered.

Returns

A vector with only the IDs of target_ids that exist for component.

GalaxyInspector.idMass! — Method

idMass!(
    data_dict::Dict,
    component::Symbol,
    target_ids::Vector{UInt};
    <keyword arguments>
)::Unitful.Mass

Compute the total mass of the cells/particles with IDs given by target_ids.

Arguments

data_dict::Dict: Data dictionary (see makeDataDict for the canonical description).

component::Symbol: Target component. It can only be one of the elements of COMPONENTS.

target_ids::Vector{UInt}: Target IDs.

tracers::Bool=false: If the mass will be computed using tracers or the proper mass of each cell/particle.

Returns

The total mass.

GalaxyInspector.idToIndex — Method

idToIndex(data_dict::Dict, target_ids::Vector{<:Unsigned})::Dict{Symbol,Vector{Int}}

Find the indices of the cells/particles given their IDs.

Arguments

data_dict::Dict: Data dictionary (see makeDataDict for the canonical description).

target_ids::Vector{<:Unsigned}: List of IDs.

Returns

A dictionary with the following shape:

cell/particle type::Symbol -> indices::Int

cell/particle type::Symbol -> indices::Int

cell/particle type::Symbol -> indices::Int

...

GalaxyInspector.idWithinDisk — Method

idWithinDisk(
    data_dict::Dict,
    component::Symbol,
    max_r::Unitful.Length,
    max_z::Unitful.Length,
    origin...,
)::Vector{UInt}

Find the IDs the the cell/particles of component that are within the given galactic disk.

Arguments

data_dict::Dict: Data dictionary (see makeDataDict for the canonical description).

component::Symbol: Target component. It can only be one of the elements of COMPONENTS.

max_r::Unitful.Length: Radius of the disk.

max_z::Unitful.Length: Half height of the disk.

origin...: Center of the disk. It can be any number and type of argument compatible with the second to last arguments of a computeCenter method.

Returns

A vector with the IDs.

GalaxyInspector.idWithinR200 — Method

idWithinR200(data_dict::Dict, component::Symbol; <keyword arguments>)::Vector{UInt}

Find the IDs the the cell/particles of component that are within the virial radius of halo halo_idx.

Note

This function assumes that no translation of the coordinate system has been done. It uses as the center of the halo the position given by the group catalog files, which are unaffected by coordinate transformation of the data.

Arguments

data_dict::Dict: Data dictionary (see makeDataDict for the canonical description).

component::Symbol: Target component. It can only be one of the elements of COMPONENTS.

halo_idx::Int=1: Index of the target halo (FoF group). Starts at 1.

Returns

A vector with the IDs.

GalaxyInspector.parentToTracerID — Method

parentToTracerID(data_dict::Dict, target_ids::Vector{<:Unsigned})::Vector{UInt}

Find the tracer IDs given a list of parent IDs.

Arguments

data_dict::Dict: Data dictionary (see makeDataDict for the canonical description).

target_ids::Vector{<:Unsigned}: List of parent IDs.

Returns

A vector with the IDs of the tracers. There is no guarantee length or order for this vector.

GalaxyInspector.parentToTracerMass — Method

parentToTracerMass(
    data_dict::Dict,
    target_ids::Vector{<:Unsigned},
)::Unitful.Mass

Find the mass of tracers given a list of parent IDs, ignoring parents without tracers

Arguments

data_dict::Dict: Data dictionary (see makeDataDict for the canonical description).

target_ids::Vector{<:Unsigned}: List of parent IDs.

Returns

The total mass of tracers.

GalaxyInspector.tracerToParentID — Method

tracerToParentID(data_dict::Dict, target_ids::Vector{<:Unsigned})::Vector{UInt}

Find the parent IDs given a list of tracer IDs.

Arguments

data_dict::Dict: Data dictionary (see makeDataDict for the canonical description).

target_ids::Vector{<:Unsigned}: List of tracer IDs.

Returns

A vector with the IDs of the parents. There is no guarantee length or order for this vector.

GalaxyInspector.tracersToParentMass — Method

tracersToParentMass(
    data_dict::Dict,
    target_ids::Vector{<:Unsigned},
)::Dict{Symbol,Vector{<:Unitful.Mass}}

Find the masses of the parents given a list of tracer IDs.

Arguments

data_dict::Dict: Data dictionary (see makeDataDict for the canonical description).

target_ids::Vector{<:Unsigned}: List of tracer IDs.

Returns

A dictionary with the following shape:

cell/particle type -> masses::Vector{<:Unitful.Mass}

cell/particle type -> masses::Vector{<:Unitful.Mass}

cell/particle type -> masses::Vector{<:Unitful.Mass}

...