Arepo utilities

These functions are only used internally, and depend on the particulars of the Arepo code (e.g. units).

Here you can find several auxiliary functions to interact with Arepo data.

None of these functions are exported.

GalaxyInspector.addRequest — Method

addRequest(
    request::Dict{Symbol,Vector{String}},
    addition::Dict{Symbol,Vector{String}},
)::Dict{Symbol,Vector{String}}

Add the blocks in addition to request, only for the types already present in request.

Arguments

request::Dict{Symbol,Vector{String}}: The request dictionary for readSnapshot.

addition::Dict{Symbol,Vector{String}}: Request dictionary with the blocks to be added, only for the types already present in request.

Returns

A new dictionary with all the requests.

GalaxyInspector.barPlotLabelFormater — Method

barPlotLabelFormater(x::LaTeXString)::LaTeXString

Format a number to be a barplot label.

Method for compatibility with the barplot! function of Makie.

Arguments

x::Number: Value to be formated.

Returns

The bar label.

GalaxyInspector.barPlotLabelFormater — Method

barPlotLabelFormater(x::Number)::LaTeXString

Format a number to be a barplot label.

For values between 0 and 0.01 the label will be "< 0.01", otherwise it will be the value itself with 2 digits.

Arguments

x::Number: Value to be formated.

Returns

The bar label.

GalaxyInspector.formatError — Method

formatError(q_mean::Number, q_error::Number)::NTuple{2,<:Number}

Nicely format a magnitude with uncertainty.

It follows the traditional rules for error presentation: the error has only one significant digit, unless such digit is a one, in which case two significant digits are used. The mean will have as many digits as to match the last significant position of the error. An error equal to 0 will leave the mean unchanged.

Arguments

q_mean::Number: Mean value.

q_error::Number: Error value. It must be positive.

Returns

A tuple with the formatted mean and error values.

Examples

julia> formatError(69.42069, 0.038796)
(69.42, 0.04)

julia> formatError(69.42069, 0.018796)
(69.421, 0.019)

julia> formatError(15.42, 0.00004)
(15.42, 4.0e-5)

julia> formatError(69.42069, 0.0)
(69.42069, 0.0)

julia> formatError(69.42069, 93.4)
(70.0, 90.0)

julia> formatError(69.42069, 123.4)
(70.0, 120.0)

julia> formatError(15.42069, 16.4)
(15.0, 16.0)

GalaxyInspector.getLabel — Method

getLabel(
    label::AbstractString,
    factor::Int,
    unit::Unitful.Units;
    <keyword arguments>
)::AbstractString

Construct an axis label.

Arguments

label::AbstractString: Variable name.

factor::Int: Exponential factor to scale down the units. If different from 0, a term of the form 10^factor will be added to the label.

unit::Unitful.Units: Unit of the axis.

latex::Bool=true: If the output will be a LaTeXString, or a plain String.

Returns

The LaTeXString or String: "label [10^factor unit]". If label is "", an empty string is returned. The factor term only appears if factor != 0, the unit term only appears if unit != Unitful.NoUnits, and the brackets only appears if there are a factor and/or a unit term.

GalaxyInspector.getUnitLabel — Method

getUnitLabel(factor::Int, unit::Unitful.Units; <keyword arguments>)::AbstractString

Construct the unit part of an axis label.

Arguments

factor::Int: Exponential factor to scale down the units. If different from 0, a term of the form 10^factor will be added to the label.

unit::Unitful.Units: Unit of the axis.

latex::Bool=true: If the output will be a LaTeXString, or a plain String.

Returns

The LaTeXString or String: "10^factor unit". The factor term only appears if factor != 0, the unit term only appears if unit != Unitful.NoUnits.

GalaxyInspector.groupcatTypes — Method

groupcatTypes(data_dict::Dict)::Vector{Symbol}

Find which group catalog data types are part of the keys of data_dict.

Arguments

data_dict::Dict: A dictionary.

Returns

A vector with the group catalog data types.

GalaxyInspector.groupcatTypes — Method

groupcatTypes(path::String)::Vector{Symbol}

Find which group catalog data types are part of the snapshot in path.

Note

If each snapshot is made of multiple files, I'll check the first chunck.

Arguments

path::String: Path to the snapshot file or folder.

Returns

A vector with the group catalog data types.

GalaxyInspector.internalUnits — Method

internalUnits(quantity::String, path::String)::Union{Unitful.Quantity,Unitful.Units}

Get the factor to convert a plain number into a Unitful quantity, using the correct internal code units.

Arguments

quantity::String: Target quantity. The options are the keys of QUANTITIES.

path::String: Path to the snapshot file or folder.

Returns

A Unitful quantity or unit.

GalaxyInspector.isCosmological — Method

isCosmological(path::String)::Bool

Check if the snapshot in path comes from a cosmological simulation.

Note

If each snapshot is made of multiple files, I'll read the first chunck to check if the simulation is cosmological.

Arguments

path::String: Path to the snapshot file or folder.

Returns

If the simulation is cosmological

false -> Newtonian simulation (ComovingIntegrationOn = 0, Redshift = 0.0).

true -> Cosmological simulation (ComovingIntegrationOn = 1, Redshift != 0.0).

GalaxyInspector.isSubfindActive — Method

isSubfindActive(path::String)::Bool

Check if there is information about the halos and subhalos in the group catalog file.

Arguments

path::String: Path to the group catalog file or folder.

Returns

If there are halo and subhalo information in the group catalog file.

GalaxyInspector.mergeRequests — Method

mergeRequests(requests::Dict{Symbol,Vector{String}}...)::Dict{Symbol,Vector{String}}

Merge several request dictionaries, ignoring duplicates.

Arguments

requests: The request dictionaries for readSnapshot.

Returns

A new dictionary with all the requests.

GalaxyInspector.plotParams — Method

plotParams(quantity::Symbol)::PlotParams

Select the plotting parameters for a given quantity.

Arguments

quantity::Symbol: The options are:

:stellar_mass -> Stellar mass.

:gas_mass -> Gas mass.

:hydrogen_mass -> Hydrogen mass.

:dm_mass -> Dark matter mass.

:bh_mass -> Black hole mass.

:molecular_mass -> Molecular hydrogen ($\mathrm{H_2}$) mass.

:atomic_mass -> Atomic hydrogen ($\mathrm{HI}$) mass.

:ionized_mass -> Ionized hydrogen ($\mathrm{HII}$) mass.

:neutral_mass -> Neutral hydrogen ($\mathrm{HI + H_2}$) mass.

:generic_mass -> Parameters for plots with several diferent masses.

:stellar_number -> Number of stellar particles.

:gas_number -> Number of gas cells.

:dm_number -> Number of dark matter particles.

:bh_number -> Number of black hole particles.

:molecular_fraction -> Gas mass fraction of molecular hydrogen.

:atomic_fraction -> Gas mass fraction of atomic hydrogen.

:ionized_fraction -> Gas mass fraction of ionized hydrogen.

:neutral_fraction -> Gas mass fraction of neutral hydrogen.

:molecular_neutral_fraction -> Fraction of molecular hydrogen in the neutral gas.

:generic_fraction -> Parameters for plots with several diferent fraction.

:gas_mass_density -> Gas mass density.

:hydrogen_mass_density -> Hydrogen mass density.

:gas_number_density -> Gas number density.

:molecular_number_density -> Molecular hydrogen number density.

:atomic_number_density -> Atomic hydrogen number density.

:ionized_number_density -> Ionized hydrogen number density.

:neutral_number_density -> Neutral hydrogen number density.

:stellar_area_density -> Stellar area mass density, for a radius of FILTER_R.

:gas_area_density -> Gas area mass density, for a radius of FILTER_R.

:molecular_area_density -> Molecular hydrogen area mass density, for a radius of FILTER_R.

:atomic_area_density -> Atomic hydrogen area mass density, for a radius of FILTER_R.

:ionized_area_density -> Ionized hydrogen area mass density, for a radius of FILTER_R.

:neutral_area_density -> Neutral hydrogen area mass density, for a radius of FILTER_R.

:sfr_area_density -> Star formation rate area density, for the last AGE_RESOLUTION_ρ and a radius of FILTER_R.

:generic_area_density -> Parameters for plots with several diferent area densities.

:gas_metallicity -> Mass fraction of all elements above He in the gas (solar units).

:stellar_metallicity -> Mass fraction of all elements above He in the stars (solar units).

:X_gas_abundance -> Gas abundance of element $\mathrm{X}$, as $12 + \log_{10}(\mathrm{X \, / \, H})$. The possibilities are the keys of ELEMENT_INDEX.

:X_stellar_abundance -> Stellar abundance of element $\mathrm{X}$, as $12 + \log_{10}(\mathrm{X \, / \, H})$. The possibilities are the keys of ELEMENT_INDEX.

:stellar_radial_distance -> Distance of every stellar particle to the origin.

:gas_radial_distance -> Distance of every gas cell to the origin.

:dm_radial_distance -> Distance of every dark matter particle to the origin.

:stellar_xy_distance -> Projected distance of every stellar particle to the origin.

:gas_xy_distance -> Projected distance of every gas cell to the origin.

:dm_xy_distance -> Projected distance of every dark matter particle to the origin.

:mass_accretion -> Gas accretion rate. Positive values mean gas infall into the virial radius $R_{200}$, and negative values mean outflow.

:stellar_specific_am -> Norm of the stellar specific angular momentum.

:gas_specific_am -> Norm of the gas specific angular momentum.

:dm_specific_am -> Norm of the dark matter specific angular momentum.

:stellar_circularity -> Stellar circularity.

:stellar_vcirc -> Stellar circular velocity.

:stellar_vradial -> Stellar radial speed.

:stellar_vtangential -> Stellar tangential speed.

:stellar_vzstar -> Stellar speed in the z direction, computed as $v_z \, \mathrm{sign}(z)$.

:stellar_age -> Stellar age.

:sfr -> The star formation rate.

:ssfr -> The specific star formation rate.

:observational_sfr -> The star formation rate of the last AGE_RESOLUTION.

:observational_ssfr -> The specific star formation rate of the last AGE_RESOLUTION.

:temperature -> Gas temperature, as $\log_{10}(T \, / \, \mathrm{K})$.

:scale_factor -> Scale factor.

:redshift -> Redshift.

:physical_time -> Physical time since the Big Bang.

:lookback_time -> Physical time left to reach the last snapshot.

Returns

A PlotParams object, with entries:

request::Dict{Symbol,Vector{String}} -> Data request for readSnapshot.

var_name::AbstractString -> Name of the quantity for the plot axis.

exp_factor::Int -> Numerical exponent to scale down the axis.

unit::Unitful.Units -> Target unit for the axis.

axis_label::AbstractString -> Label for the axis.

GalaxyInspector.snapshotTypes — Method

snapshotTypes(data_dict::Dict)::Vector{Symbol}

Find which cell/particle types are part of the keys of data_dict.

Arguments

data_dict::Dict: A dictionary.

Returns

A vector with the cell/particle types.

GalaxyInspector.snapshotTypes — Method

snapshotTypes(path::String)::Vector{Symbol}

Find which cell/particle types are part of the snapshot in path.

Note

If each snapshot is made of multiple files, I'll check the first chunck.

Arguments

path::String: Path to the snapshot file or folder.

Returns

A vector with the cell/particle types.