Coordinate transformations.

These functions are only used internally, and depend on the particulars of the Arepo code (e.g. units).

Here you can find the lowest level functions that transform the coordinates (translations, rotations, etc.).

None of these functions are exported.

GalaxyInspector.rotateData! — Method

rotateData!(data_dict::Dict, axis_type::Int)::Nothing

Rotate the positions and velocities of the cells/particles in data_dict.

Arguments

data_dict::Dict: A dictionary with the following shape:

:sim_data -> ::Simulation (see Simulation).

:snap_data -> ::Snapshot (see Snapshot).

:gc_data -> ::GroupCatalog (see GroupCatalog).

cell/particle type -> (block -> data of block, block -> data of block, ...).

cell/particle type -> (block -> data of block, block -> data of block, ...).

cell/particle type -> (block -> data of block, block -> data of block, ...).

...

groupcat type -> (block -> data of block, block -> data of block, ...).

groupcat type -> (block -> data of block, block -> data of block, ...).

groupcat type -> (block -> data of block, block -> data of block, ...).

...

rotation::Int: Target subhalo absolute index, starting at 1. Sets the principal axis of the stars in the subhalo as the new coordinate system.

GalaxyInspector.rotateData! — Method

rotateData!(data_dict::Dict, axis_type::Symbol)::Nothing

Rotate the positions and velocities of the cells/particles in data_dict.

Arguments

data_dict::Dict: A dictionary with the following shape:

:sim_data -> ::Simulation (see Simulation).

:snap_data -> ::Snapshot (see Snapshot).

:gc_data -> ::GroupCatalog (see GroupCatalog).

cell/particle type -> (block -> data of block, block -> data of block, ...).

cell/particle type -> (block -> data of block, block -> data of block, ...).

cell/particle type -> (block -> data of block, block -> data of block, ...).

...

groupcat type -> (block -> data of block, block -> data of block, ...).

groupcat type -> (block -> data of block, block -> data of block, ...).

groupcat type -> (block -> data of block, block -> data of block, ...).

...

rotation::Symbol: Type of rotation. The options are:

:zero -> No rotation is appplied.

:global_am -> Sets the angular momentum of the whole system as the new z axis.

:stellar_am -> Sets the stellar angular momentum as the new z axis.

:stellar_pa -> Sets the stellar principal axis as the new coordinate system.

:stellar_subhalo_pa -> Sets the principal axis of the stars in the main subhalo as the new coordinate system.

GalaxyInspector.rotateData! — Method

rotateData!(data_dict::Dict, axis_type::NTuple{2,Int})::Nothing

Rotate the positions and velocities of the cells/particles in data_dict.

Arguments

data_dict::Dict: A dictionary with the following shape:

:sim_data -> ::Simulation (see Simulation).

:snap_data -> ::Snapshot (see Snapshot).

:gc_data -> ::GroupCatalog (see GroupCatalog).

cell/particle type -> (block -> data of block, block -> data of block, ...).

cell/particle type -> (block -> data of block, block -> data of block, ...).

cell/particle type -> (block -> data of block, block -> data of block, ...).

...

groupcat type -> (block -> data of block, block -> data of block, ...).

groupcat type -> (block -> data of block, block -> data of block, ...).

groupcat type -> (block -> data of block, block -> data of block, ...).

...

rotation::NTuple{2,Int}: Type of rotation. The options are:

(halo_idx, subhalo_rel_idx) -> Sets the principal axis of the stars in subhalo_rel_idx::Int subhalo (of the halo_idx::Int halo) as the new coordinate system.

(halo_idx, 0) -> Sets the principal axis of the stars in the halo_idx::Int halo as the new coordinate system.

GalaxyInspector.translateData! — Method

translateData!(data_dict::Dict, translation::Union{Symbol,NTuple{2,Int},Int})::Nothing

Translate the positions of the cells/particles in data_dict.

Note

The velocities will be boosted to the stellar center of mass of the system. If there are no stars, no transformation in applied to the velocities.

Arguments

data_dict::Dict: A dictionary with the following shape:

:sim_data -> ::Simulation (see Simulation).

:snap_data -> ::Snapshot (see Snapshot).

:gc_data -> ::GroupCatalog (see GroupCatalog).

cell/particle type -> (block -> data of block, block -> data of block, ...).

cell/particle type -> (block -> data of block, block -> data of block, ...).

cell/particle type -> (block -> data of block, block -> data of block, ...).

...

groupcat type -> (block -> data of block, block -> data of block, ...).

groupcat type -> (block -> data of block, block -> data of block, ...).

groupcat type -> (block -> data of block, block -> data of block, ...).

...

translation::Union{Symbol,NTuple{2,Int},Int}=:zero: Type of translation. The options are:

:zero -> No translation is applied.

:global_cm -> Sets the center of mass of the whole system as the new origin.

:stellar_cm -> Sets the stellar center of mass as the new origin.

(halo_idx, subhalo_rel_idx) -> Sets the position of the potencial minimum for the subhalo_rel_idx::Int subhalo (of the halo_idx::Int halo) as the new origin.

(halo_idx, 0) -> Sets the center of mass of the halo_idx::Int halo as the new origin.

subhalo_abs_idx -> Sets the center of mass of the subhalo_abs_idx::Int as the new origin.

GalaxyInspector.translatePoints — Method

translatePoints(
    positions::Matrix{<:Number},
    new_origin::Vector{<:Number},
)::Matrix{<:Number}

Translate a system of points, moving new_origin to the origin.

Arguments

positions::Matrix{<:Number}: Points to be translated. Each column is a point and each row a dimension.

new_origin::Vector{<:Number}: Target origin.

Returns

Matrix with the translated points.