Auxiliary functions

These functions are used internally and none are exported.

This auxiliary functions do not depend on the specifics of the simulation code.

Base.intersect — Method

New methods for Base.intersect to use with the Colon type.

Base.intersect — Method

New methods for Base.intersect to use with the Vector{Bool} type.

Base.isempty — Method

New method for Base.isempty to check for empty LaTeXStrings.

Base.iszero — Method

New method for Base.iszero to compare IndexType with 0 as an interger.

Base.union — Method

New methods for Base.union to use with the Vector{Bool} type.

GalaxyInspector.absCoor — Method

absCoor(
    plot::Union{Makie.FigureAxisPlot,Makie.Axis,Makie.Figure},
    r_x::Real,
    r_y::Real,
)::NTuple{2,Float64}

Compute the absolute x and y coordinates of a plot, from the relative ones.

Arguments

plot::Union{Makie.FigureAxisPlot,Makie.Axis,Makie.Figure}: Plot, axis, or figure for which the absolute coordinates will be calculated. In the case of a figure, it will use the limits from the current axis object.

r_x::Real: Relative x coordinate.

r_y::Real: Relative y coordinate.

Returns

A tuple with the absolute coordinates, (x, y).

Examples

julia> absCoor(lines(rand(100)), 0.5, 0.5)
(50.50000071525574, 0.48792968317866325)

GalaxyInspector.area — Method

Area of a circle with radius r.

GalaxyInspector.cleanPlot! — Method

Default function to end cleanPlot! recursion if an unknown type is encountered.

GalaxyInspector.cleanPlot! — Method

cleanPlot!(ax::Makie.Axis)::Bool

Empty an axis.

Arguments

ax::Makie.Axis: Axis to be emptied.

Returns

Flag to indicate that an axis has been emptied.

GalaxyInspector.cleanPlot! — Method

cleanPlot!(figure::Makie.Figure)::Nothing

Delete all the legends of a figure and empty all its axes.

Arguments

figure::Makie.Figure: Figure to be cleaned.

GalaxyInspector.cleanPlot! — Method

cleanPlot!(legend::Union{Makie.Legend,Makie.Colorbar})::Bool

Delete a legend or colorbar.

Arguments

legend::Union{Makie.Legend,Makie.Colorbar}: Legend or colorbar to be deleted.

Returns

Flag to indicate that a legend or colorbar has been deleted.

GalaxyInspector.cubicSplineKernel — Method

cubicSplineKernel(q::Real, h::Number)::Number

2D cubic spline kernel.

Arguments

q::Real: Relative distance to the neighbor, $|r - r'| / h$.

h::Number: Smoothing length.

Returns

The kernel function evaluated at a separation q * h, and with a smoothing length h.

References

PySPH documentation

J. J. Monaghan (1992). Smoothed Particle Hydrodynamics. Annual Review of Astronomy and Astrophysics, 30, 543-574. doi:10.1146/annurev.aa.30.090192.002551

M.B. Liu et al. (2010). Smoothed Particle Hydrodynamics (SPH): an Overview and Recent Developments. Archives of Computational Methods in Engineering, 17, 25–76. doi:10.1007/s11831-010-9040-7

GalaxyInspector.deltas — Method

deltas(data::Vector{<:Number})::Vector{<:Number}

Compute the difference between each consecutive pair of elements in data.

Arguments

data::Vector{<:Number}: Data vector. It has to have at least 2 elements.

Returns

A vector with the difference between each consecutive pair of elements in data, the first element is 0 by convention.

GalaxyInspector.evaluateNormal — Method

evaluateNormal(data::Vector{<:Number})::Vector{<:Number}

Evaluate a normal distribution at the values in data.

The mean and standard deviation of the distribution are the ones from the data itself.

Arguments

data::Vector{<:Number}: Data vector used to compute the mean and standard deviation of the normal distribution.

Returns

The normal distribution evaluated at the values in data.

GalaxyInspector.getEmpty — Method

Always returns an empty vector, for any type and number of arguments.

GalaxyInspector.getNothing — Method

Always returns nothing, for any type and number of arguments.

GalaxyInspector.histogram1D — Method

histogram1D(
    positions::Vector{<:Number},
    grid::Union{LinearGrid,CircularGrid},
)::Vector{Int}

Compute a 1D histogram of positions.

Arguments

positions::Vector{<:Number}: Values for which the histogram will be constructed.

grid::Union{LinearGrid,CircularGrid}: A linear or circular grid.

Returns

A vector with the counts.

GalaxyInspector.histogram1D — Method

histogram1D(
    positions::Vector{<:Number},
    values::Vector{<:Number},
    grid::Union{LinearGrid,CircularGrid};
    <keyword arguments>
)::Vector{<:Number}

Compute a 1D histogram of values.

Arguments

positions::Vector{<:Number}: Positions of the values within a 1D axis. This determines to which bin each value will be added.

values::Vector{<:Number}: The values that will be added up in each bin, according to their positions.

grid::Union{LinearGrid,CircularGrid}: A linear or circular grid.

total::Bool=true: If the sum (default) or the mean of values will be computed for each bin.

empty_nan::Bool=true: If NaN will be put into empty bins, 0 is used otherwise.

Returns

A vector with the histogram values.

GalaxyInspector.histogram1D — Method

histogram1D(
    positions::Vector{<:Number},
    values::Vector{<:Number},
    edges::Vector{<:Number};
    <keyword arguments>
)::Vector{<:Number}

Compute a 1D histogram of values.

Arguments

positions::Vector{<:Number}: Positions of the values within a 1D axis. This determines to which bin each value will be added.

values::Vector{<:Number}: The values that will be added up in each bin, according to their positions.

edges::Vector{<:Number}: A sorted list of bin edges.

total::Bool=true: If the sum (default) or the mean of values will be computed for each bin.

empty_nan::Bool=true: If NaN will be put into empty bins, 0 is used otherwise.

Returns

A vector with the histogram values.

GalaxyInspector.histogram1D — Method

histogram1D(positions::Vector{<:Number}, edges::Vector{<:Number})::Vector{Int}

Compute a 1D histogram of positions.

Arguments

positions::Vector{<:Number}: Values for which the histogram will be constructed.

edges::Vector{<:Number}: A sorted list of bin edges.

Returns

A vector with the counts.

GalaxyInspector.histogram2D — Method

histogram2D(positions::Matrix{<:Number}, grid::SquareGrid)::Matrix{Int}

Compute a 2D histogram of positions.

Arguments

positions::Matrix{<:Number}: Values for which the histogram will be constructed.

grid::SquareGrid: A square grid.

Returns

A matrix with the counts.

GalaxyInspector.histogram2D — Method

histogram2D(
    positions::Matrix{<:Number},
    values::Vector{<:Number},
    grid::SquareGrid;
    <keyword arguments>
)::Matrix{<:Number}

Compute a 2D histogram of values.

Arguments

positions::Matrix{<:Number}: Positions of the values in the grid. Each column correspond to a value and each row is a dimension. This determines to which bin each value will be added.

values::Vector{<:Number}: The values that will be added up in each square bin, according to their positions.

grid::SquareGrid: A square grid.

total::Bool=true: If the sum (default) or the mean of values will be computed in each bin.

empty_nan::Bool=true: If NaN will be put into empty bins, 0 is used otherwise.

Returns

A matrix with the histogram values.

GalaxyInspector.histogram2D — Method

histogram2D(
    positions::Matrix{<:Number},
    values::Vector{<:Number},
    x_edges::Vector{<:Number},
    y_edges::Vector{<:Number};
    <keyword arguments>
)::Matrix{<:Number}

Compute a 2D histogram of values.

Arguments

positions::Matrix{<:Number}: Positions of the values in the grid. Each column correspond to a value and each row is a dimension. This determines to which bin each value will be added.

values::Vector{<:Number}: The values that will be added up in each square bin, according to their positions.

x_edges::Vector{<:Number}: A sorted list of bin edges for the x axis.

y_edges::Vector{<:Number}: A sorted list of bin edges for the y axis.

total::Bool=true: If the sum (default) or the mean of values will be computed in each bin.

empty_nan::Bool=true: If NaN will be put into empty bins, 0 is used otherwise.

Returns

A matrix with the histogram values.

GalaxyInspector.histogram2D — Method

histogram2D(
    positions::Matrix{<:Number},
    x_edges::Vector{<:Number},
    y_edges::Vector{<:Number},
)::Matrix{Int}

Compute a 2D histogram of positions.

Arguments

positions::Matrix{<:Number}: Values for which the histogram will be constructed.

x_edges::Vector{<:Number}: A sorted list of bin edges for the x axis.

y_edges::Vector{<:Number}: A sorted list of bin edges for the y axis.

Returns

A matrix with the counts.

GalaxyInspector.histogram3D — Method

histogram3D(positions::Matrix{<:Number}, grid::CubicGrid)::Array{Int,3}

Compute a 3D histogram of positions.

Arguments

positions::Matrix{<:Number}: Values for which the histogram will be constructed.

grid::CubicGrid: A cubic grid.

Returns

A 3D tensor with the counts.

GalaxyInspector.histogram3D — Method

histogram3D(
    positions::Matrix{<:Number},
    values::Vector{<:Number},
    grid::CubicGrid;
    <keyword arguments>
)::Array{<:Number,3}

Compute a 3D histogram of values.

Arguments

positions::Matrix{<:Number}: Positions of the values in the grid. Each column correspond to a value and each row is a dimension. This determines to which bin each value will be added.

values::Vector{<:Number}: The values that will be added up in each square bin, according to their positions.

grid::CubicGrid: A cubic grid.

total::Bool=true: If the sum (default) or the mean of values will be computed in each bin.

empty_nan::Bool=true: If NaN will be put into empty bins, 0 is used otherwise.

Returns

A 3D tensor with the histogram values.

GalaxyInspector.hvcatImages — Method

hvcatImages(
    blocks_per_row::Int,
    paths::Vector{String};
    <keyword arguments>
)::Nothing

Join several images vertically and horizontally

The elements will fill the rows and columns starting at the top left, going from left to right and from top to bottom.

Arguments

blocks_per_row::Int: Number of columns.

paths::Vector{String}: Paths to the images.

output_path::String="./joined_images.png": Path to the output image.

GalaxyInspector.isPositive — Method

Test for strict positivity.

GalaxyInspector.listHistogram1D — Method

listHistogram1D(
    positions::Vector{<:Number},
    values::Vector{<:Number},
    grid::Union{LinearGrid,CircularGrid},
)::Vector{Vector{<:Number}}

Compute a 1D histogram of values, returning the full list of values within each bin.

Arguments

positions::Vector{<:Number}: Positions of the values within a 1D axis. This determines to which bin each value will be added.

values::Vector{<:Number}: The values that will be added up in each bin, according to their positions.

grid::Union{LinearGrid,CircularGrid}: A linear or circular grid.

Returns

- A vector with the lists of `values` within each bin.

GalaxyInspector.listHistogram1D — Method

listHistogram1D(
    positions::Vector{<:Number},
    values::Vector{<:Number},
    edges::Vector{<:Number},
)::Vector{Vector{<:Number}}

Compute a 1D histogram of values, returning the full list of values within each bin.

Arguments

positions::Vector{<:Number}: Positions of the values within a 1D axis. This determines to which bin each value will be added.

values::Vector{<:Number}: The values that will be added up in each bin, according to their positions.

edges::Vector{<:Number}: A sorted list of bin edges.

Returns

A vector with the lists of values within each bin.

GalaxyInspector.listHistogram3D — Method

listHistogram3D(positions::Matrix{<:Number}, grid::CubicGrid)::Array{Vector{Int},3}

Compute a 3D histogram of positions, returning the full list of indices within each bin.

Arguments

positions::Matrix{<:Number}: Positions of the points in the grid. Each column correspond to a point and each row is a dimension. This determines to which bin the index of each point will be added.

grid::CubicGrid: A cubic grid.

Returns

- A tensor with the indices of the points within each bin.

GalaxyInspector.parserWS — Method

parserWS(data::AbstractString)::Union{Float64,Missing}

Parse a string as a Float64, ignoring white spaces. If the string is empty return missing.

Arguments

data::AbstractString: String to be parsed.

Returns

Number in the string as a Float64.

GalaxyInspector.pointData — Method

Extract all the data points in a Makie plot, axis, or figure. In the case of a figure, it will only take the data from the current axis object. It only works for scatter, line and scatterline plots.

GalaxyInspector.projectIntoCircularGrid — Method

projectIntoCircularGrid(
    image::Matrix{<:Number},
    n_bins::Int;
    <keyword arguments>
)::Vector{<:Number}

Project image into a circular grid, averaging the values in each concentric ring.

Arguments

image::Matrix{<:Number}: Original matrix. It has to be a square matrix.

n_bins::Int: Number of bins for the circular grid.

inscribed::Bool=true: If the circular grid will be inscribed in image when doing the projection. If set to false, the matrix will be inscribed into the circular grid instead.

Returns

A vector with the averages of the values in each concentric ring.

GalaxyInspector.rangeCut! — Method

rangeCut!(
    raw_values::Vector{<:Number},
    range::Tuple{<:Number,<:Number};
    <keyword arguments>
)::Bool

Delete every element in raw_values that is outside the given range.

Arguments

raw_values::Vector{<:Number}: Dataset that will be pruned.

range::Tuple{<:Number,<:Number}: The range in question.

keep_edges::Bool=true: If the edges of the range will be kept.

min_left::Int=1: Minimum number of values that need to be left after pruning to proceed with the transformation.

Returns

If a transformation was performed.

GalaxyInspector.rangeCut! — Method

rangeCut!(
    m_data::Vector{<:Number},
    s_data::Vector,
    range::Tuple{<:Number,<:Number};
    <keyword arguments>
)::Bool

Delete every element in m_data that is outside the given range.

Every corresponding element in s_data (i.e. with the same index) will be deleted too.

Arguments

m_data::Vector{<:Number}: Master dataset that will be pruned.

s_data::Vector: Slave dataset that will be pruned according to which values of m_data are outside range.

range::Tuple{<:Number,<:Number}: The range in question.

keep_edges::Bool=true: If the edges of the range will be kept.

min_left::Int=1: Minimum number of values that need to be left in the master dataset after pruning to proceed with the transformation.

Returns

If a transformation was performed.

GalaxyInspector.reduceResolution — Method

reduceResolution(hr_matrix::Matrix{<:Number}, factor::Int)::Matrix{<:Number}

Reduce the number of rows and columns of hr_matrix by factor, averaging its values.

Arguments

hr_matrix::Matrix{<:Number}: Original "high resolution" matrix. It has to be a square matrix.

factor::Int: Factor by wich the number of rows and columns will be reduced. It has to divede the size of hr_matrix exactly.

Returns

The new smaller matrix, with the average values.

GalaxyInspector.reduceTicks — Method

reduceTicks(hr_ticks::Vector{<:Number}, factor::Int)::Vector{<:Number}

Reduce the number of ticks in hr_ticks by factor keeping the total length of the axis the same and assuming hr_ticks are regularly spaced.

Arguments

hr_ticks::Vector{<:Number}: Original "high resolution" list of ticks.

factor::Int: Factor by wich the number of ticks will be reduced. It has to divede the size of hr_ticks exactly.

Returns

The new shorter tick list.

GalaxyInspector.ring — Method

ring(vec::Vector, index::Integer)

Make the indexing operation vec[index] work using modular arithmetic for the indices.

Arguments

vec::Vector: Vector.

index::Integer: Index.

Returns

vec[mod1(index, length(vec))]

Examples

julia> ring([1, 2, 3], 11)
2

julia> ring([1, 2, 3], 3)
3

julia> ring([1, 2, 3], -5)
1

GalaxyInspector.safeSelect — Method

safeSelect(vec::Vector, index::IndexType; warnings::Bool=true)

Make the indexing operation vec[index] ignore indices that are out of bounds.

Arguments

vec::Vector: Vector.

index::IndexType: Indices. It can be an integer (a single element), a vector of integers (several elements), an UnitRange (e.g. 5:13), an StepRange (e.g. 5:2:13) or (:) (every element).

warnings::Bool=true: If a warning will be given when index has out of bounds indices.

Returns

vec[index (minus out of bounds indices)]

Examples

julia> safeSelect([1, 2, 3], 11; warnings=false)
Int[]

julia> safeSelect([1, 2, 3], 1:5; warnings=false)
3-element Vector{Int}:
 1
 2
 3

julia> safeSelect([1, 2, 3], 1:3:10; warnings=false)
1

julia> safeSelect([1, 2, 3], [1, 2, 5, 9]; warnings=false)
2-element Vector{Int}:
 1
 2

julia> safeSelect([1, 2, 3], (:); warnings=false)
3-element Vector{Int}:
 1
 2
 3

GalaxyInspector.sanitizeData! — Method

sanitizeData!(
    x_data::Vector{<:Number},
    y_data::Vector{<:Number};
    <keyword arguments>
)::NTuple{4,Bool}

Do the following transformations over x_data and y_data, in order:

Trim them to fit within the domain of the functions func_domain[1] and func_domain[2], respectively.

Trim them to fit within range[1] and range[2], respectively.

Scale them down by a factor 10^exp_factor[1] and 10^exp_factor[2], respectively.

By default, no transformation is done.

Note

The datasets must have the same length, and any operation that deletes an element, will delete the corresponding element (i.e. with the same index) in the other dataset, so that the dataset will remain of equal length.

Arguments

x_data::Vector{<:Number}: First dataset to be sanitized.

y_data::Vector{<:Number}: Second dataset to be sanitized.

func_domain::NTuple{2,Function}=(identity, identity): x_data will be trimmed to fit within the domain of the function func_domain[1], and y_data will be trimmed to fit within the domain of the function func_domain[2]. The options are the scaling functions accepted by Makie: log10, log2, log, sqrt, Makie.logit, Makie.Symlog10, Makie.pseudolog10, and identity.

range::Tuple{Tuple{<:Number,<:Number},Tuple{<:Number,<:Number}}=((-Inf, Inf), (-Inf, Inf)): Every element in x_data that falls outside of range[1] will be deleted, and every element in y_data that falls outside of range[2] will be deleted.

keep_edges::NTuple{2,Bool}=(true, true): If the edges of each corresponding range will be kept.

min_left::Int=1: Minimum number of values that need to be left in each dataset after any of the transformations to procced with them.

exp_factor::NTuple{2,Int}=(0, 0): Every element in x_data will be divided by 10^exp_factor[1], and every element in y_data will be divided by 10^exp_factor[2].

warnings::Bool=true: If a warning will be given when any of the datasets is a vector of Integers, which may cause wrong results when dividing by 10^exp_factor.

Returns

A tuple with four flags:

If x_data was successfully modified to fit within the domain of func_domain[1].

If y_data was successfully modified to fit within the domain of func_domain[2].

If x_data was successfully modified to fit within range[1].

If y_data was successfully modified to fit within range[2].

GalaxyInspector.sanitizeData! — Method

sanitizeData!(
    raw_values::Vector{<:Number};
    <keyword arguments>
)::NTuple{2,Bool}

Do the following transformations over raw_values, in order:

Trim it to fit within the domain of the function func_domain.

Trim it to fit within range.

Scale it down by a factor of 10^exp_factor.

By default, no transformation is done.

Arguments

raw_values::Vector{<:Number}: Dataset to be sanitized.

func_domain::Function=identity: raw_values will be trimmed to fit within the domain of the function func_domain. The options are the scaling functions accepted by Makie: log10, log2, log, sqrt, Makie.logit, Makie.Symlog10, Makie.pseudolog10, and identity.

range::Tuple{<:Number,<:Number}=(-Inf, Inf): Every element in raw_values that falls outside of range will be deleted.

keep_edges::Bool=true: If the edges of range will be kept.

min_left::Int=1: Minimum number of values that need to be left after each transformation to procced with it.

exp_factor::Int=0: Every element in raw_values will be divided by 10^exp_factor.

warnings::Bool=true: If a warning will be given when raw_values is a vector of Integers, which may cause wrong results when dividing by 10^exp_factor.

Returns

A tuple with two flags:

If raw_values was mutated to fit within the domain of func_domain.

If raw_values was mutated to fit within range.

GalaxyInspector.scaledBins — Method

scaledBins(
    values::Vector{<:Number},
    n_bins::Int;
    <keyword arguments>
)::Vector{Float64}

Compute a set of bin edges, to encompass a given list of values.

Arguments

values::Vector{<:Number}: Values to be binned.

n_bins::Int: Number of bins.

scaling::Function=identity: Scaling function. The options are the scaling functions accepted by Makie: log10, log2, log, sqrt, Makie.logit, Makie.Symlog10, Makie.pseudolog10, and identity.

limits::Tuple{<:Number,<:Number}=(-Inf, Inf): Set it to a value different than (-Inf, Inf) if you want to fix the limits of the bins.

Returns

A sorted list of bin edges.

GalaxyInspector.setPositive — Method

Create a copy of list with every negative value set to 0.

GalaxyInspector.smoothWindow — Method

smoothWindow(
    x_data::Vector{<:Number},
    y_data::Vector{<:Number},
    n_bins::Int;
    <keyword arguments>
)::NTuple{2,Vector{<:Number}}

Separate the values of x_data in n_bins bins and compute the mean value of x_data and y_data within each one.

Arguments

x_data::Vector{<:Number}: x-axis data.

y_data::Vector{<:Number}: y-axis data.

n_bins::Int: Number of bins.

scaling::Function=identity: Scaling function for the x axis. The options are the scaling functions accepted by Makie: log10, log2, log, sqrt, Makie.logit, Makie.Symlog10, Makie.pseudolog10, and identity. All the values of x_data must be in the domain of scaling.

Returns

A tuple with two vectors, containing the smoothed-out x and y values.

GalaxyInspector.volume — Method

Volume of a sphere with radius r.

GalaxyInspector.xlimits! — Method

Extract the limits of the x axis, from a Makie plot, axis, or figure. In the case of a figure, it will take the limits from the current axis object.

GalaxyInspector.xscale — Method

Extract the scale function of the x axis, from a Makie plot, axis, or figure. In the case of a figure, it will take the scale from the current axis object.

GalaxyInspector.ylimits! — Method

Extract the limits of the y axis, from a Makie plot, axis, or figure. In the case of a figure, it will take the limits from the current axis object.

GalaxyInspector.yscale — Method

Extract the scale function of the y axis, from a Makie plot, axis, or figure. In the case of a figure, it will take the scale from the current axis object.