Grid utilities

These functions are used internally and none are exported.

They may depend on the particulars of the simulation code (e.g. internal units).

GalaxyInspector.equalAreaBins — Method

equalAreaBins(radius::Number, n_bins::Int; shift::Number=zero(radius))::Vector{<:Number}

Compute the bin edges that make each bin have equal area.

Arguments

radius::Number: Radius of the grid (equal to the last bin edge).

n_bins::Int: Number of bins.

shift::Number=zero(radius): Distance of the first bin edge to the center.

Returns

The bin edges

GalaxyInspector.equalVolumeBins — Method

equalVolumeBins(radius::Number, n_bins::Int; shift::Number=zero(radius))::Vector{<:Number}

Compute the bin edges that make each bin have equal volume.

Arguments

radius::Number: Radius of the grid (equal to the last bin edge).

n_bins::Int: Number of bins.

shift::Number=zero(radius): Distance of the first bin edge to the center.

Returns

The bin edges

GalaxyInspector.flattenGrid — Method

flattenGrid(cubic_grid::CubicGrid)::SquareGrid

From a CubicGrid construct a SquareGrid with the same center, number of bins, and physical side length.

Arguments

cubic_grid::CubicGrid: Cubic grid.

Returns

A square grid.

GalaxyInspector.projectIntoCircularGrid — Method

projectIntoCircularGrid(
    image::Matrix{<:Number},
    n_bins::Int;
    <keyword arguments>
)::Vector{<:Number}

Project a given matrix into a circular grid, averaging or adding up the values in each concentric ring.

Arguments

image::Matrix{<:Number}: Original matrix. It has to be a square matrix.

n_bins::Int: Number of bins for the circular grid.

inscribed::Bool=true: If the circular grid will be inscribed in image when doing the projection. If set to false, the matrix will be inscribed into the circular grid instead.

total::Bool=false: If the sum (total = true) or the mean (total = false) of the values in image, that fall within each ring, will be used.

Returns

A vector with the average or sum of the values that fall within each concentric ring.

GalaxyInspector.reduceMatrix — Method

reduceMatrix(
    hr_matrix::Matrix{<:Number},
    factor::Int;
    <keyword arguments>
)::Matrix{<:Number}

Reduce the number of rows and columns of a given square matrix by factor, averaging or adding up its values.

Arguments

hr_matrix::Matrix{<:Number}: Original "high resolution" matrix. It has to be a square matrix.

factor::Int: Factor by which the number of rows and columns will be reduced. It has to divide the size of hr_matrix exactly.

total::Bool=false: If the sum (total = true) or the mean (total = false) of the high resolution values will be used for the new matrix.

Returns

The new smaller matrix.

GalaxyInspector.scaledBins — Method

scaledBins(
    values::Vector{<:Number},
    n_bins::Int;
    <keyword arguments>
)::Vector{Float64}

Compute a set of bin edges, that will encompass a given list of values.

Arguments

values::Vector{<:Number}: Values to be binned.

n_bins::Int: Number of bins.

scaling::Function=identity: Scaling function. The options are the scaling functions accepted by Makie: log10, log2, log, sqrt, Makie.logit, Makie.Symlog10, Makie.pseudolog10, and identity.

limits::Tuple{<:Number,<:Number}=(-Inf, Inf): Set it to a value different than (-Inf, Inf) if you want to fix the limits of the bins.

Returns

A sorted list of bin edges.