Histogram utilities

These functions are used internally and none are exported.

They may depend on the particulars of the simulation code (e.g. internal units).

GalaxyInspector.histogram1D — Method

histogram1D(positions::Vector{<:Number}, grid::Union{LinearGrid,CircularGrid})::Vector{Int}

Compute a 1D histogram of positions.

Arguments

positions::Vector{<:Number}: Values for which the histogram will be constructed.

grid::Union{LinearGrid,CircularGrid}: A linear or circular grid.

Returns

A vector with the counts.

GalaxyInspector.histogram1D — Method

histogram1D(
    positions::Vector{<:Number},
    values::Vector{<:Number},
    grid::Union{LinearGrid,CircularGrid};
    <keyword arguments>
)::Vector{<:Number}

Compute a 1D histogram of values.

Arguments

positions::Vector{<:Number}: Positions of the values within a 1D axis.

values::Vector{<:Number}: The values that will be added up to each bin, according to their positions.

grid::Union{LinearGrid,CircularGrid}: A linear or circular grid.

total::Bool=true: If the sum (total = true) or the mean (total = false) of values will be computed for each bin.

empty_nan::Bool=true: If NaN will be put into empty bins, 0 is used otherwise.

Returns

A vector with the histogram values.

GalaxyInspector.histogram1D — Method

histogram1D(
    positions::Vector{<:Number},
    values::Vector{<:Number},
    edges::Vector{<:Number};
    <keyword arguments>
)::Vector{<:Number}

Compute a 1D histogram of values.

Arguments

positions::Vector{<:Number}: Positions of the values within a 1D axis.

values::Vector{<:Number}: The values that will be added up in each bin, according to their positions.

edges::Vector{<:Number}: A list of bin edges.

total::Bool=true: If the sum (total = true) or the mean (total = false) of values will be computed for each bin.

empty_nan::Bool=true: If NaN will be put into empty bins, 0 is used otherwise.

Returns

A vector with the histogram values.

GalaxyInspector.histogram1D — Method

histogram1D(positions::Vector{<:Number}, edges::Vector{<:Number})::Vector{Int}

Compute a 1D histogram of positions.

Arguments

positions::Vector{<:Number}: Values for which the histogram will be constructed.

edges::Vector{<:Number}: A list of bin edges.

Returns

A vector with the counts.

GalaxyInspector.histogram2D — Method

histogram2D(
    positions::Matrix{<:Number},
    values::Vector{<:Number},
    grid::SquareGrid;
    <keyword arguments>
)::Matrix{<:Number}

Compute a 2D histogram of values.

Arguments

positions::Matrix{<:Number}: Positions of the values in the grid. Each column corresponds to a value and each row is a dimension.

values::Vector{<:Number}: The values that will be added up in each square bin, according to their positions.

grid::SquareGrid: A square grid.

total::Bool=true: If the sum (total = true) or the mean (total = false) of values will be computed for each bin.

empty_nan::Bool=true: If NaN will be put into empty bins, 0 is used otherwise.

Returns

A matrix with the histogram values.

GalaxyInspector.histogram2D — Method

histogram2D(
    positions::Matrix{<:Number},
    values::Vector{<:Number},
    x_edges::Vector{<:Number},
    y_edges::Vector{<:Number};
    <keyword arguments>
)::Matrix{<:Number}

Compute a 2D histogram of values.

Arguments

positions::Matrix{<:Number}: Positions of the values in the grid. Each column corresponds to a value and each row is a dimension.

values::Vector{<:Number}: The values that will be added up in each square bin, according to their positions.

x_edges::Vector{<:Number}: A list of bin edges for the x axis.

y_edges::Vector{<:Number}: A list of bin edges for the y axis.

total::Bool=true: If the sum (total = true) or the mean (total = false) of values will be computed for each bin.

empty_nan::Bool=true: If NaN will be put into empty bins, 0 is used otherwise.

Returns

A matrix with the histogram values.

GalaxyInspector.histogram2D — Method

histogram2D(
    positions::Matrix{<:Number},
    x_edges::Vector{<:Number},
    y_edges::Vector{<:Number},
)::Matrix{Int}

Compute a 2D histogram of positions.

Arguments

positions::Matrix{<:Number}: Values for which the histogram will be constructed.

x_edges::Vector{<:Number}: A list of bin edges for the x axis.

y_edges::Vector{<:Number}: A list of bin edges for the y axis.

Returns

A matrix with the counts.

GalaxyInspector.histogram3D — Method

histogram3D(
    positions::Matrix{<:Number},
    values::Vector{<:Number},
    grid::CubicGrid;
    <keyword arguments>
)::Array{<:Number,3}

Compute a 3D histogram of values.

Arguments

positions::Matrix{<:Number}: Positions of the values in the grid. Each column corresponds to a value and each row is a dimension.

values::Vector{<:Number}: The values that will be added up in each square bin, according to their positions.

grid::CubicGrid: A cubic grid.

total::Bool=true: If the sum (total = true) or the mean (total = false) of values will be computed for each bin.

empty_nan::Bool=true: If NaN will be put into empty bins, 0 is used otherwise.

Returns

A 3D tensor with the histogram values.

GalaxyInspector.listHistogram1D — Method

listHistogram1D(
    positions::Vector{<:Number},
    values::Vector{<:Number},
    grid::Union{LinearGrid,CircularGrid},
)::Vector{Vector{<:Number}}

Compute a 1D histogram of values, returning the full list of values within each bin.

Arguments

positions::Vector{<:Number}: The positions of values within a 1D axis.

values::Vector{<:Number}: The values that will be sorted, according to their positions.

grid::Union{LinearGrid,CircularGrid}: A linear or circular grid.

Returns

- A vector with the lists of `values` within each bin.

GalaxyInspector.listHistogram3D — Method

listHistogram3D(positions::Matrix{<:Number}, grid::CubicGrid)::Array{Vector{Int},3}

Compute a 3D histogram of positions, returning the list of indices that fall within each bin.

Arguments

positions::Matrix{<:Number}: Positions of the points. Each column corresponds to a point and each row is a dimension.

grid::CubicGrid: A cubic grid.

Returns

- A tensor with the indices of the points within each bin.