Auxiliary functions

These functions are used internally and none are exported.

They may depend on the particulars of the simulation code (e.g. internal units).

Base.intersect — Method

New methods for Base.intersect to use with the Colon type.

Base.intersect — Method

New methods for Base.intersect to use with the Vector{Bool} type.

Base.isempty — Method

New method for Base.isempty to check for empty LaTeXStrings.

Base.iszero — Method

New method for Base.iszero to compare IndexType with 0.

Base.iterate — Method

New method for Base.iterate to manage symbols.

Base.round — Method

New method for Base.round to use with the Unitful.Quantity type.

Base.union — Method

New methods for Base.union to use with the Vector{Bool} type.

GalaxyInspector.allocateArray — Method

allocateArray(T::Type, dims::NTuple{N, Integer}; <keyword arguments>) where {N}

Allocate an array of type T and dimensions dims, using memory-mapping if the size exceeds a certain threshold.

Arguments

T::Type: The element type of the array.

dims::NTuple{N, Integer}: The dimensions of the array.

mmap::Bool=false: Whether to use memory-mapping. If false, the method will decide based on the size of the array and the available physical memory.

fill=nothing: If not nothing, the value with which to fill the array after allocation.

Returns

An array of type T and dimensions dims, allocated in RAM or using memory-mapping depending on the size.

GalaxyInspector.area — Method

Area of a circle with radius r.

GalaxyInspector.copyArray — Method

copyArray(A::AbstractArray; <keyword arguments>)::AbstractArray

Copy an array, using memory-mapping if the size exceeds a certain threshold.

Arguments

A::AbstractArray: The array to be copied.

mmap::Bool=false: Whether to use memory-mapping. If false, the method will decide based on the size of the array and the available physical memory.

Returns

- A copy of `A`, allocated in RAM or using memory-mapping depending on the size.

GalaxyInspector.copyDataDict — Method

copyDataDict(data_dict::Dict)::Dict

Copy a data dictionary.

Arguments

data_dict::Dict: Data dictionary to be copied (see makeDataDict for the canonical description).

Returns

The copied dictionary.

GalaxyInspector.evaluateNormal — Method

evaluateNormal(values::Vector{<:Number})::Vector{<:Number}

Evaluate a normal distribution at values.

The median and standard deviation for the distribution are the ones of values.

Arguments

values::Vector{<:Number}: Vector used to compute the mean and standard deviation of the normal distribution.

Returns

The normal distribution evaluated at values.

GalaxyInspector.findFiles — Method

findFiles(base_path::String, pattern_string::AbstractString)::Vector{String}

Find every file that matches pattern_string within base_path.

Note

This method does a recursive search to any depth within base_path.

Arguments

base_path::String: High level path from when the search starts.

pattern_string::AbstractString: Glob like pattern.

Returns

A vector with the full path to every file that matched.

GalaxyInspector.getEmpty — Method

Always returns an empty vector, for any type and number of arguments.

GalaxyInspector.getNothing — Method

Always returns nothing, for any type and number of arguments.

GalaxyInspector.griMagnitudeInterpolation — Method

griMagnitudeInterpolation(path::String)::NTuple{3,Interpolations.Extrapolation}

Compute three linear interpolation functions for the stellar magnitudes in the g, r, and i filters of SDSS.

Arguments

path::String: Path to the magnitude table. It has to have the columns:

:Z -> Stellar metallicity.

:log_age -> log10(stellar age [yr]).

:g -> Stellar magnitude for the g filter of SDSS (λeff = 4627 Å -> "blue" channel).

:r -> Stellar magnitude for the r filter of SDSS (λeff = 6140 Å -> "green" channel).

:i -> Stellar magnitude for the i filter of SDSS (λeff = 7467 Å -> "red" channel).

Returns

A tuple with the three interpolation functions:

g(log10(age [yr]), Z*)

r(log10(age [yr]), Z*)

i(log10(age [yr]), Z*)

GalaxyInspector.isPositive — Method

Test for strict positivity.

GalaxyInspector.loadConfig! — Function

loadConfig!(path::Union{Nothing,String}=nothing)::Nothing

Load a TOML configuration file and apply the resulting runtime settings.

Arguments

path::Union{Nothing,String}=nothing: The path to the TOML configuration file. If nothing, the file is assumed to be config.toml in the current working directory.

GalaxyInspector.memoryThreshold — Function

memoryThreshold(mmap_memory_fraction::Float64=MMAP_MEMORY_FRACTION[])::Int

Compute the threshold for memory-mapping as a fraction of the free physical memory.

Arguments

mmap_memory_fraction::Float64=MMAP_MEMORY_FRACTION[]: Fraction of the free physical memory that will be used as the threshold for memory-mapping. It has to be a number between 0 and 1.

Returns

The threshold for memory-mapping in bytes.

GalaxyInspector.metaFormatter — Method

metaFormatter(level::LogLevel, _module, group, id, file, line)

Formatter for loggers.

See the documentation of ConsoleLogger

GalaxyInspector.parseString — Method

parseString(s::AbstractString; <keyword arguments>)::Union{Float64,Missing}

Parse a string as a Float64, ignoring white spaces.

Arguments

s::AbstractString: String to be parsed.

empty::Symbol=:missing: Return value if the string is empty. It can be either :missing or :nan.

Returns

Number in the string as a Float64.

GalaxyInspector.ring — Method

ring(vec::Vector, index::Integer)::Vector

Make the indexing operation vec[index] work with modular arithmetic for the indices.

Arguments

vec::Vector: Vector.

index::Integer: Index.

Returns

vec[mod1(index, length(vec))]

Examples

julia> ring([1, 2, 3], 11)
2

julia> ring([1, 2, 3], 3)
3

julia> ring([1, 2, 3], -5)
1

GalaxyInspector.runtimeType — Method

Get the most specific common type of the elements in an array.

GalaxyInspector.safeSelect — Method

safeSelect(vec::Vector, index::IndexType)

Do the indexing operation vec[index] while ignoring indices that are out of bounds.

Arguments

vec::Vector: Vector.

index::IndexType: Indices. It can be an integer (a single element), a vector of integers (several elements), an UnitRange (e.g. 5:13), an StepRange (e.g. 5:2:13) or (:) (every element).

Returns

vec[index (minus out of bounds indices)]

Examples

julia> safeSelect([1, 2, 3], 11)
Int[]

julia> safeSelect([1, 2, 3], 1:5)
3-element Vector{Int}:
 1
 2
 3

julia> safeSelect([1, 2, 3], 1:3:10)
1

julia> safeSelect([1, 2, 3], [1, 2, 5, 9])
2-element Vector{Int}:
 1
 2

julia> safeSelect([1, 2, 3], (:))
3-element Vector{Int}:
 1
 2
 3

GalaxyInspector.safeSlice — Method

safeSlice(vec::Vector, slice::IndexType)::Vector

Do the slicing operation vec[slice] while ignoring indices that are out of bounds.

Note

This method is like safeSelect, but it always return a vector (a slice).

Arguments

vec::Vector: Vector.

slice::IndexType: Indices. It can be an integer (a single element), a vector of integers (several elements), an UnitRange (e.g. 5:13), an StepRange (e.g. 5:2:13) or (:) (every element).

Returns

[vec[slice (minus out of bounds indices)]...]

Examples

julia> safeSlice([1, 2, 3], 11)
Int[]

julia> safeSlice([1, 2, 3], 1:5)
3-element Vector{Int}:
 1
 2
 3

julia> safeSlice([1, 2, 3], 1:3:10)
1-element Vector{Int64}:
 1

julia> safeSlice([1, 2, 3], [1, 2, 5, 9])
2-element Vector{Int}:
 1
 2

julia> safeSlice([1, 2, 3], (:))
3-element Vector{Int}:
 1
 2
 3

GalaxyInspector.setLogging — Method

setLogging(log_file::String)::IO

Set if logging messages will be printed out. By default no logs are printed.

Arguments

log_file::String: The path to the log file.

Returns

The IO stream to which the logger is writing.

GalaxyInspector.setPositive — Method

Return a copy of list with every negative value set to 0.

GalaxyInspector.volume — Method

Volume of a sphere with radius r.

Unitful.showrep — Method

New methods for Unitful.showrep to use with AU derived units.

Unitful.showrep — Method

New methods for Unitful.showrep to use with light year derived units.

Unitful.showrep — Method

New methods for Unitful.showrep to use with parsec derived units.