Plotting utilities

These functions are used internally and none are exported.

They may depend on the particulars of the simulation code (e.g. internal units).

GalaxyInspector.KSLawBigiel2008 — Method

KSLawBigiel2008(
    ΣH::Vector{<:SurfaceDensity};
    <keyword arguments>
)::Vector{<:Number}

Evaluate the Kennicutt-Schmidt law for the molecular or neutral gas, taken from Bigiel et al. (2008).

From Bigiel et al. (2008) (Section 3.1), we have

\[\Sigma_\mathrm{SFR} = a \left( \frac{\Sigma_\mathrm{HI, H_2, gas}}{10 \, \mathrm{M_\odot \, pc^{-2}}} \right)^{\!N} \, ,\]

where N is the power-law index, and $A = \log_{10}(a)$, where $a$ is $\Sigma_\mathrm{SFR}$ at the fiducial gas surface density of $10 \, \mathrm{M_\odot \, pc^{-2}}$.

Arguments

ΣH::Vector{<:SurfaceDensity}: Values of the molecular or neutral gas surface density.

molecular::Bool=true: If ΣH is the surface density of molecular hydrogen, or of neutral hydrogen.

log_output::Bool=true: If the output will the $\log_{10} \, \Sigma_\text{SFR}$, or $\Sigma_\text{SFR}$. If log_output = true, the implied units are $\mathrm{M_\odot \, yr^{-1} \, kpc^{-2}}$

Returns

The SFR surface density.

References

F. Bigiel et al. (2008). THE STAR FORMATION LAW IN NEARBY GALAXIES ON SUB-KPC SCALES. The Astrophysical Journal, 136(6), 2846. doi:10.1088/0004-6256/136/6/2846

GalaxyInspector.KSLawCologni2026 — Method

KSLawCologni2026(ΣH2::Vector{<:SurfaceDensity}; <keyword arguments>)::Vector{<:Number}

Evaluate the Kennicutt-Schmidt law for the molecular gas, taken from Cologni et al. (2026).

From Cologni et al. (2026) (Section 4.3.2), we have

\[\Sigma_\mathrm{SFR} = a \left( \frac{\Sigma_\mathrm{H_2}}{1.0 \, \mathrm{M_\odot \, kpc^{-2}}} \right)^{\!N} \, ,\]

where N is the power-law index, and $A = \log_{10}(a)$, where $a$ is $\Sigma_\mathrm{SFR}$ at the fiducial gas surface density of $1.0 \, \mathrm{M_\odot \, kpc^{-2}}$.

Arguments

ΣH2::Vector{<:SurfaceDensity}: Values of the molecular gas surface density.

log_output::Bool=true: If the output will the $\log_{10} \, \Sigma_\text{SFR}$, or $\Sigma_\text{SFR}$. If log_output = true, the implied units are $\mathrm{M_\odot \, yr^{-1} \, kpc^{-2}}$

Returns

The SFR surface density.

References

R. Cologni et al. (2026). Resolved molecular gas and star formation in massive unquenched spirals. Astronomy and Astrophysics, 709, A148. doi:10.1051/0004-6361/202558486

GalaxyInspector.KSLawKennicutt1998 — Method

KSLawKennicutt1998(Σgas::Vector{<:SurfaceDensity}; <keyword arguments>)::Vector{<:Number}

Evaluate the Kennicutt-Schmidt law, taken from Kennicutt (1998).

From Kennicutt (1998) (Section 4, Eq. 4), we have

\[\Sigma_\mathrm{SFR} = a \left( \frac{\Sigma_\mathrm{gas}}{1 \, \mathrm{M_\odot \, pc^{-2}}} \right)^{\!N} \mathrm{M_\odot \, yr^{-1] \, kpc^{-2}} \, ,\]

where N is the power-law index and $a$ is $\Sigma_\mathrm{SFR}$ at the fiducial gas surface density of $1 \, \mathrm{M_\odot \, pc^{-2}}$.

Arguments

Σgas::Vector{<:SurfaceDensity}: Values of the gas mass surface density.

log_output::Bool=true: If the output will the $\log_{10} \, \Sigma_\text{SFR}$, or $\Sigma_\text{SFR}$. If log_output = true, the implied units are $\mathrm{M_\odot \, yr^{-1} \, kpc^{-2}}$

Returns

The SFR surface density.

References

R. C. Kennicutt (1998). The Global Schmidt Law in Star-forming Galaxies. The Astrophysical Journal, 498(2), 541-552. doi:10.1086/305588

GalaxyInspector.KSLawLin2019 — Method

KSLawLin2019(ΣH2::Vector{<:SurfaceDensity}; <keyword arguments>)::Vector{<:Number}

Evaluate the Kennicutt-Schmidt law for the molecular gas, taken from Lin et al. (2019).

From Lin et al. (2019) (Section 3.1, Table 1), we have

\[\Sigma_\mathrm{SFR} = a \left( \frac{\Sigma_\mathrm{H_2}}{1.0 \, \mathrm{M_\odot \, kpc^{-2}}} \right)^{\!N} \, ,\]

where N is the power-law index, and $A = \log_{10}(a)$, where $a$ is $\Sigma_\mathrm{SFR}$ at the fiducial gas surface density of $1.0 \, \mathrm{M_\odot \, kpc^{-2}}$.

Arguments

ΣH2::Vector{<:SurfaceDensity}: Values of the molecular gas surface density.

log_output::Bool=true: If the output will the $\log_{10} \, \Sigma_\text{SFR}$, or $\Sigma_\text{SFR}$. If log_output = true, the implied units are $\mathrm{M_\odot \, yr^{-1} \, kpc^{-2}}$

Returns

The SFR surface density.

References

L. Lin et al. (2019). The ALMaQUEST Survey: The Molecular Gas Main Sequence and the Origin of the Star-forming Main Sequence. The Astrophysical Journal Letters, 884(2), L33. doi:10.3847/2041-8213/ab4815

GalaxyInspector.KSLawQuerejeta2021 — Method

KSLawQuerejeta2021(ΣH2::Vector{<:SurfaceDensity}; <keyword arguments>)::Vector{<:Number}

Evaluate the Kennicutt-Schmidt law for the molecular gas, taken from Querejeta et al. (2021).

From Querejeta et al. (2021) (Section 4.3, Table 4, row All, mean of the columns), we have

\[\Sigma_\mathrm{SFR} = a \left( \frac{\Sigma_\mathrm{H_2}}{1.0 \, \mathrm{M_\odot \, pc^{-2}}} \right)^{\!N} \, ,\]

where N is the power-law index, and $A = \log_{10}(a)$, where $a$ is $\Sigma_\mathrm{SFR}$ at the fiducial gas surface density of $1.0 \, \mathrm{M_\odot \, pc^{-2}}$.

Arguments

ΣH2::Vector{<:SurfaceDensity}: Values of the molecular gas surface density.

log_output::Bool=true: If the output will the $\log_{10} \, \Sigma_\text{SFR}$, or $\Sigma_\text{SFR}$. If log_output = true, the implied units are $\mathrm{M_\odot \, yr^{-1} \, kpc^{-2}}$

Returns

The SFR surface density.

References

M. Querejeta et al. (2021). Stellar structures, molecular gas, and star formation across the PHANGS sample of nearby galaxies. Astronomy and Astrophysics, 656, A133. doi:10.1051/0004-6361/202140695

GalaxyInspector.absCoor — Method

absCoor(
    plot::Union{Makie.FigureAxisPlot,Makie.Axis,Makie.Figure},
    r_x::Real,
    r_y::Real,
)::NTuple{2,Float64}

Compute the absolute x and y coordinates of a plot, from the relative ones.

Arguments

plot::Union{Makie.FigureAxisPlot,Makie.Axis,Makie.Figure}: Plot, axis, or figure for which the absolute coordinates will be calculated. In the case of a figure, it will use the limits from the current axis object.

r_x::Real: Relative x coordinate.

r_y::Real: Relative y coordinate.

Returns

A tuple with the absolute coordinates, (x, y).

Examples

julia> absCoor(lines(rand(100)), 0.5, 0.5)
(50.50000071525574, 0.48792968317866325)

GalaxyInspector.barPlotLabelFormater — Method

barPlotLabelFormater(x::LaTeXString)::LaTeXString

Format a number to be a barplot label.

Method for compatibility with the barplot! function of Makie.

Arguments

x::Number: Value to be formatted.

Returns

The bar label.

GalaxyInspector.barPlotLabelFormater — Method

barPlotLabelFormater(x::Number)::LaTeXString

Format a number to be a barplot label.

For values between 0 and 0.01 the label will be "< 0.01", otherwise it will be the value itself with 2 digits.

Arguments

x::Number: Value to be formatted.

Returns

The bar label.

GalaxyInspector.cleanPlot! — Method

Default function to end cleanPlot! recursion if an unknown type is encountered.

GalaxyInspector.cleanPlot! — Method

cleanPlot!(ax::Makie.Axis)::Bool

Empty an axis.

Arguments

ax::Makie.Axis: Axis to be emptied.

Returns

Flag to indicate that an axis has been emptied.

GalaxyInspector.cleanPlot! — Method

cleanPlot!(figure::Makie.Figure)::Nothing

Delete all the legends of a figure and empty all its axes.

Arguments

figure::Makie.Figure: Figure to be cleaned.

GalaxyInspector.cleanPlot! — Method

cleanPlot!(legend::Union{Makie.Legend,Makie.Colorbar})::Bool

Delete a legend or colorbar.

Arguments

legend::Union{Makie.Legend,Makie.Colorbar}: Legend or colorbar to be deleted.

Returns

Flag to indicate that a legend or colorbar has been deleted.

GalaxyInspector.formatError — Method

formatError(q_mean::Number, q_error::Number)::NTuple{2,<:Number}

Nicely format a magnitude with uncertainty.

It follows the traditional rules for error presentation: the error has only one significant digit, unless such digit is a one, in which case two significant digits are used. The mean will have as many digits as to match the last significant position of the error. An error equal to 0 will leave the mean unchanged.

Arguments

q_mean::Number: Mean value.

q_error::Number: Error value. It must be positive.

Returns

A tuple with the formatted mean and error values.

Examples

julia> formatError(69.42069, 0.038796)
(69.42, 0.04)

julia> formatError(69.42069, 0.018796)
(69.421, 0.019)

julia> formatError(15.42, 0.00004)
(15.42, 4.0e-5)

julia> formatError(69.42069, 0.0)
(69.42069, 0.0)

julia> formatError(69.42069, 93.4)
(70.0, 90.0)

julia> formatError(69.42069, 123.4)
(70.0, 120.0)

julia> formatError(15.42069, 16.4)
(15.0, 16.0)

GalaxyInspector.formatSeconds — Method

formatSeconds(sec::Float64)::String

Format given number of seconds as "DDd-HHh:MM':SS''", where D is days, H is hours, M is minutes, and S is seconds.

Arguments

sec::Float64: Number of seconds.

Returns

String with the formatted time.

GalaxyInspector.getLabel — Method

getLabel(
    label::AbstractString,
    factor::Int,
    unit::Unitful.Units;
    <keyword arguments>
)::AbstractString

Construct an axis label.

Arguments

label::AbstractString: Variable name.

factor::Int: Exponential factor to scale down the units. If different from 0, a term of the form 10^factor will be added to the label.

unit::Unitful.Units: Unit of the axis.

latex::Bool=true: If the output will be a LaTeXString or a plain String.

Returns

The LaTeXString or String: "label [10^factor unit]". If label is "", an empty string is returned. The factor term only appears if factor != 0, the unit term only appears if unit != Unitful.NoUnits, and the brackets only appears if there are a factor and/or a unit term.

GalaxyInspector.getLabel — Method

getLabel(quantity::AbstractPlotQuantity; <keyword arguments>)::AbstractString

Construct an axis label for a given AbstractPlotQuantity.

Note

This label will ignore the exp_factor of AbstractPlotQuantity and take it as 0.

Arguments

quantity::AbstractPlotQuantity: Target quantity.

log::Bool=false: If true adds $\log_{10}$ to the label.

latex::Bool=true: If the output will be a LaTeXString or a plain String.

Returns

The label for quantity, according to QTY_REGISTRY.

GalaxyInspector.getLabel — Method

getLabel(quantity::Symbol; <keyword arguments>)::AbstractString

Construct an axis label for a given quantity.

Note

This label will ignore the exp_factor field of BaseQuantity and take it as 0.

Arguments

quantity::Symbol: Target quantity.

log::Bool=false: If true adds $\log_{10}$ to the label.

latex::Bool=true: If the output will be a LaTeXString or a plain String.

Returns

The label for quantity, according to QTY_REGISTRY.

GalaxyInspector.getLabelArgs — Method

getLabelArgs(quantity::AbstractPlotQuantity; <keyword arguments>)::NamedTuple

Construct the label arguments of plotSnapshot for a given AbstractPlotQuantity.

Arguments

quantity::AbstractPlotQuantity: Target quantity.

log::Bool=false: If true adds $\log_{10}$ to the axis label.

Returns

A named tuple with

log_unit::Union{Unitful.Units,Nothing} -> Quantity unit for the functions that use to decide if the axis will be logarithmic or not. It is set to nothing if log = false, and to the corresponding unit if log = true.

unit::Unitful.Units -> Target unit for the axis data.

exp_factor::Int -> Numerical exponent to scale down the axis data, e.g. if exp_factor = 10 the values will be divided by $10^{10}$.

qty_label::AbstractString -> Name of the variable for the axis.

label::AbstractString -> Label for the axis. It can contain the string auto_label, which will be replaced by: qty_label [10^exp_factor unit]. If a LaTeXString with auto_label inside is used, it is recommended that each section around auto_label is delimited with a $ $ pair.

request::Dict{Symbol,Vector{String}} -> Data request for readSnapshot.

GalaxyInspector.getLabelArgs — Method

getLabelArgs(quantity::Symbol; <keyword arguments>)::NamedTuple

Construct the label arguments of plotSnapshot for a given quantity.

Arguments

quantity::Symbol: Target quantity.

log::Bool=false: If true adds $\log_{10}$ to the axis label.

Returns

A named tuple with

log_unit::Union{Unitful.Units,Nothing} -> Quantity unit for the functions that use to decide if the axis will be logarithmic or not. It is set to nothing if log = false, and to the corresponding unit if log = true.

unit::Unitful.Units -> Target unit for the axis data.

exp_factor::Int -> Numerical exponent to scale down the axis data, e.g. if exp_factor = 10 the values will be divided by $10^{10}$.

qty_label::AbstractString -> Name of the variable for the axis.

label::AbstractString -> Label for the axis. It can contain the string auto_label, which will be replaced by: qty_label [10^exp_factor unit]. If a LaTeXString with auto_label inside is used, it is recommended that each section around auto_label is delimited with a $ $ pair.

request::Dict{Symbol,Vector{String}} -> Data request for readSnapshot.

GalaxyInspector.getUnitLabel — Method

getUnitLabel(factor::Int, unit::Unitful.Units; <keyword arguments>)::AbstractString

Construct the unit part of an axis label.

Arguments

factor::Int: Exponential factor to scale down the units. If different from 0, a term of the form 10^factor will be added to the label.

unit::Unitful.Units: Unit of the axis.

latex::Bool=true: If the output will be a LaTeXString or a plain String.

Returns

The LaTeXString or String: "10^factor unit". The factor term only appears if factor != 0, the unit term only appears if unit != Unitful.NoUnits.

GalaxyInspector.hvcatImages — Method

hvcatImages(
    blocks_per_row::Int,
    paths::Vector{String};
    <keyword arguments>
)::Nothing

Join several images vertically and horizontally.

The images in paths will fill the rows and columns starting at the top left, going from left to right and from top to bottom (row-major order).

Arguments

blocks_per_row::Int: Number of columns.

paths::Vector{String}: Paths to the images.

output_path::String="./joined_images.png": Path to the output image.

GalaxyInspector.invKSLawBigiel2008 — Method

invKSLawBigiel2008(
    Σsfr ::Vector{<:MassFlowDensity};
    <keyword arguments>
)::Vector{<:Number}

Evaluate the inverse Kennicutt-Schmidt law for the molecular or neutral gas, taken from Bigiel et al. (2008).

From Bigiel et al. (2008) (Section 3.1, Eq. 2), we have

\[\Sigma_\mathrm{SFR} = a \left( \frac{\Sigma_\mathrm{HI, H_2, gas}}{10 \, \mathrm{M_\odot \, pc^{-2}}} \right)^{\!N} \, ,\]

where N is the power-law index, and $A = \log_{10}(a)$, where $a$ is $\Sigma_\mathrm{SFR}$ at the fiducial gas surface density of $10 \, \mathrm{M_\odot \, pc^{-2}}$.

Arguments

Σsfr ::Vector{<:MassFlowDensity}: Values of the SFR surface density.

molecular::Bool=true: If the output will be the surface density of molecular hydrogen, or of neutral hydrogen.

log_output::Bool=true: If the output will the $\log_{10} \, \Sigma_\text{H2}$, or $\Sigma_\text{HI}$. If log_output = true, the implied units are $10 \, \mathrm{M_\odot \, pc^{-2}}$

Returns

The molecular or neutral gas surface density.

References

F. Bigiel et al. (2008). THE STAR FORMATION LAW IN NEARBY GALAXIES ON SUB-KPC SCALES. The Astrophysical Journal, 136(6), 2846. doi:10.1088/0004-6256/136/6/2846

GalaxyInspector.invKSLawCologni2026 — Method

invKSLawCologni2026(Σsfr ::Vector{<:MassFlowDensity}; <keyword arguments>)::Vector{<:Number}

Evaluate the inverse Kennicutt-Schmidt law for the molecular gas, taken from Cologni et al. (2026).

From Cologni et al. (2026) (Section 4.3.2), we have

\[\Sigma_\mathrm{SFR} = a \left( \frac{\Sigma_\mathrm{H_2}}{1.0 \, \mathrm{M_\odot \, kpc^{-2}}} \right)^{\!N} \, ,\]

where N is the power-law index, and $A = \log_{10}(a)$, where $a$ is $\Sigma_\mathrm{SFR}$ at the fiducial gas surface density of $1.0 \, \mathrm{M_\odot \, kpc^{-2}}$.

Arguments

Σsfr ::Vector{<:MassFlowDensity}: Values of the SFR surface density.

log_output::Bool=true: If the output will the $\log_{10} \, \Sigma_\text{H2}$. If log_output = true, the implied units are $1.0 \, \mathrm{M_\odot \, kpc^{-2}}$

Returns

The molecular gas surface density.

References

R. Cologni et al. (2026). Resolved molecular gas and star formation in massive unquenched spirals. Astronomy and Astrophysics, 709, A148. doi:10.1051/0004-6361/202558486

GalaxyInspector.invKSLawKennicutt1998 — Method

invKSLawKennicutt1998(
    Σsfr::Vector{<:MassFlowDensity};
    <keyword arguments>
)::Vector{<:Number}

Evaluate the inverse Kennicutt-Schmidt law, taken from Kennicutt (1998).

From Kennicutt (1998) (Section 4, Eq. 4), we have

\[\Sigma_\mathrm{SFR} = a \left( \frac{\Sigma_\mathrm{gas}}{1 \, \mathrm{M_\odot \, pc^{-2}}} \right)^{\!N} \mathrm{M_\odot \, yr^{-1] \, kpc^{-2}} \, ,\]

where N is the power-law index and $a$ is $\Sigma_\mathrm{SFR}$ at the fiducial gas surface density of $1 \, \mathrm{M_\odot \, pc^{-2}}$.

Arguments

Σsfr::Vector{<:MassFlowDensity}: Values of the SFR surface density..

log_output::Bool=true: If the output will the $\log_{10} \, \Sigma_\text{gas}$, or $\Sigma_\text{gas}$. If log_output = true, the implied units are $\mathrm{M_\odot \, pc^{-2}}$

Returns

The gas mass surface density.

References

R. C. Kennicutt (1998). The Global Schmidt Law in Star-forming Galaxies. The Astrophysical Journal, 498(2), 541-552. doi:10.1086/305588

GalaxyInspector.invKSLawLin2019 — Method

invKSLawLin2019(Σsfr ::Vector{<:MassFlowDensity}; <keyword arguments>)::Vector{<:Number}

Evaluate the inverse Kennicutt-Schmidt law for the molecular gas, taken from Lin et al. (2019).

From Lin et al. (2019) (Section 3.1, Table 1), we have

\[\Sigma_\mathrm{SFR} = a \left( \frac{\Sigma_\mathrm{H_2}}{1.0 \, \mathrm{M_\odot \, kpc^{-2}}} \right)^{\!N} \, ,\]

where N is the power-law index, and $A = \log_{10}(a)$, where $a$ is $\Sigma_\mathrm{SFR}$ at the fiducial gas surface density of $1.0 \, \mathrm{M_\odot \, kpc^{-2}}$.

Arguments

Σsfr ::Vector{<:MassFlowDensity}: Values of the SFR surface density.

log_output::Bool=true: If the output will the $\log_{10} \, \Sigma_\text{H2}$. If log_output = true, the implied units are $1.0 \, \mathrm{M_\odot \, kpc^{-2}}$

Returns

The molecular gas surface density.

References

L. Lin et al. (2019). The ALMaQUEST Survey: The Molecular Gas Main Sequence and the Origin of the Star-forming Main Sequence. The Astrophysical Journal Letters, 884(2), L33. doi:10.3847/2041-8213/ab4815

GalaxyInspector.invKSLawQuerejeta2021 — Method

invKSLawQuerejeta2021(Σsfr ::Vector{<:MassFlowDensity}; <keyword arguments>)::Vector{<:Number}

Evaluate the inverse Kennicutt-Schmidt law for the molecular gas, taken from Querejeta et al. (2021).

From Querejeta et al. (2021) (Section 4.3, Table 4, row All, mean of the columns), we have

\[\Sigma_\mathrm{SFR} = a \left( \frac{\Sigma_\mathrm{H_2}}{1.0 \, \mathrm{M_\odot \, pc^{-2}}} \right)^{\!N} \, ,\]

where N is the power-law index, and $A = \log_{10}(a)$, where $a$ is $\Sigma_\mathrm{SFR}$ at the fiducial gas surface density of $1.0 \, \mathrm{M_\odot \, pc^{-2}}$.

Arguments

Σsfr ::Vector{<:MassFlowDensity}: Values of the SFR surface density.

log_output::Bool=true: If the output will the $\log_{10} \, \Sigma_\text{H2}$. If log_output = true, the implied units are $1.0 \, \mathrm{M_\odot \, pc^{-2}}$

Returns

The molecular gas surface density.

References

M. Querejeta et al. (2021). Stellar structures, molecular gas, and star formation across the PHANGS sample of nearby galaxies. Astronomy and Astrophysics, 656, A133. doi:10.1051/0004-6361/202140695

GalaxyInspector.pointData — Method

Extract all the data points in a Makie plot, axis, or figure. In the case of a figure, it will only take the data from the current axis object. It only works for scatter, line, and scatterline plots.

GalaxyInspector.rangeCut! — Method

rangeCut!(
    raw_values::Vector{<:Number},
    range::Tuple{<:Number,<:Number};
    <keyword arguments>
)::Bool

Delete every element in raw_values that is outside the given range.

Arguments

raw_values::Vector{<:Number}: Dataset that will be pruned.

range::Tuple{<:Number,<:Number}: The range in question.

keep_edges::Bool=true: If the edges of the range will be kept.

min_left::Int=1: Minimum number of values that need to be left after pruning to proceed with the transformation.

Returns

If a transformation was performed.

GalaxyInspector.rangeCut! — Method

rangeCut!(
    m_data::Vector{<:Number},
    s_data::Vector,
    range::Tuple{<:Number,<:Number};
    <keyword arguments>
)::Bool

Delete every element in m_data that is outside the given range.

Every corresponding element in s_data (i.e. with the same index) will be deleted too.

Arguments

m_data::Vector{<:Number}: Master dataset that will be pruned.

s_data::Vector: Slave dataset that will be pruned according to which values of m_data are outside range.

range::Tuple{<:Number,<:Number}: The range in question.

keep_edges::Bool=true: If the edges of the range will be kept.

min_left::Int=1: Minimum number of values that need to be left in the master dataset after pruning to proceed with the transformation.

Returns

If a transformation was performed.

GalaxyInspector.reduceTicks — Method

reduceTicks(hr_ticks::Vector{T}, factor::Int)::Vector{T} where {T<:FloatOrQuantity}

Reduce the length of a given list of axis ticks, while keeping the total length of the axis the same.

Note

If factor is 1, the original list of ticks will be returned without any checks.

Arguments

hr_ticks::Vector{T}: Original "high resolution" list of ticks. It has to be regularly spaced

factor::Int: Factor by which the number of values will be reduced. It has to divide the size of hr_ticks exactly.

Returns

The new shorter list of ticks.

GalaxyInspector.sanitizeData! — Method

sanitizeData!(
    x_data::Vector{<:Number},
    y_data::Vector{<:Number};
    <keyword arguments>
)::NTuple{4,Bool}

Do the following transformations over x_data and y_data, in order:

Trim them to fit within the domain of the functions func_domain[1] and func_domain[2], respectively.

Trim them to fit within range[1] and range[2], respectively.

Scale them down by a factor 10^exp_factor[1] and 10^exp_factor[2], respectively.

By default, no transformation is done.

Note

The datasets must have the same length, and any operation that deletes an element, will delete the corresponding element (i.e. with the same index) in the other dataset, so that the datasets will remain of equal length.

Arguments

x_data::Vector{<:Number}: First dataset to be sanitized.

y_data::Vector{<:Number}: Second dataset to be sanitized.

func_domain::NTuple{2,Function}=(identity, identity): x_data will be trimmed to fit within the domain of the function func_domain[1], and y_data will be trimmed to fit within the domain of the function func_domain[2]. The options are the scaling functions accepted by Makie: log10, log2, log, sqrt, Makie.logit, Makie.Symlog10, Makie.pseudolog10, and identity.

range::Tuple{Tuple{<:Number,<:Number},Tuple{<:Number,<:Number}}=((-Inf, Inf), (-Inf, Inf)): Every element in x_data that falls outside of range[1] will be deleted, and every element in y_data that falls outside of range[2] will be deleted.

keep_edges::NTuple{2,Bool}=(true, true): If the edges of each corresponding range will be kept.

min_left::Int=1: Minimum number of values that need to be left in each dataset after any of the transformations to proceed with them.

exp_factor::NTuple{2,Int}=(0, 0): Every element in x_data will be divided by 10^exp_factor[1], and every element in y_data will be divided by 10^exp_factor[2].

Returns

A tuple with four flags:

If x_data was mutated to fit within the domain of func_domain[1].

If y_data was mutated to fit within the domain of func_domain[2].

If x_data was mutated to fit within range[1].

If y_data was mutated to fit within range[2].

GalaxyInspector.sanitizeData! — Method

sanitizeData!(
    raw_values::Vector{<:Number};
    <keyword arguments>
)::NTuple{2,Bool}

Do the following transformations over raw_values, in order:

Trim it to fit within the domain of the function func_domain.

Trim it to fit within range.

Scale it down by a factor of 10^exp_factor.

By default, no transformation is done.

Arguments

raw_values::Vector{<:Number}: Dataset to be sanitized.

func_domain::Function=identity: raw_values will be trimmed to fit within the domain of the function func_domain. The options are the scaling functions accepted by Makie: log10, log2, log, sqrt, Makie.logit, Makie.Symlog10, Makie.pseudolog10, and identity.

range::Tuple{<:Number,<:Number}=(-Inf, Inf): Every element in raw_values that falls outside of range will be deleted.

keep_edges::Bool=true: If the edges of range will be kept.

min_left::Int=1: Minimum number of values that need to be left after each transformation to proceed with it.

exp_factor::Int=0: Every element in raw_values will be divided by 10^exp_factor.

Returns

A tuple with two flags:

If raw_values was mutated to fit within the domain of func_domain.

If raw_values was mutated to fit within range.

GalaxyInspector.smoothMovingWindow — Method

smoothMovingWindow(
    x_data::Vector{<:Number},
    y_data::Vector{<:Number};
    <keyword arguments>
)::NTuple{2,Vector{<:Number}}

Smooth data using a central moving window of a fixed number of points, moving one point at a time. The output vectors retain the exact same length as the input vectors.

Arguments

x_data::Vector{<:Number}: x-axis data.

y_data::Vector{<:Number}: y-axis data.

window_radius::Int=0: Number of points to gather to the left and to the right of the center point. A value of 0 bypasses calculations and returns the original data.

agg::Function=mean: The aggregation function applied to the points inside the window (e.g., mean, median).

Returns

A tuple with two vectors, containing the smoothed-out x and y values.

GalaxyInspector.validatePlotData — Method

validatePlotData(
    plot_function::Function,
    plot_data::VecOrMat{<:Number}...;
    <keyword arguments>
)::Tuple{Vector{<:VecOrMat{<:Number}},Bool,Bool}

Adapt plot_data to be plotted with plot_function.

This function strips the units of every argument in plot_data, and it may trims down the first two arguments in plot_data.

Arguments

plot_function::Function: Target plotting function from Makie. The supported functions are:

scatter! -> Scatter plot.

lines! -> Line plot.

scatterlines! -> Scatter plot with lines between the markers.

hist! -> Histogram.

heatmap! -> Heatmap.

arrows2d! -> 2D vector field.

barplot! -> Bar plots.

band! -> Band plots.

errorbars! -> Error bars.

plot_data::VecOrMat{<:Number}: Raw plot data.

x_unit::Unitful.Units=Unitful.NoUnits: Target unit for the x axis data (first arguments in plot_data). The values will be converted accordingly. Use the default value of Unitful.NoUnits for dimensionless quantities.

y_unit::Unitful.Units=Unitful.NoUnits: Target unit for the y axis data (second arguments in plot_data). The values will be converted accordingly. Use the default value of Unitful.NoUnits for dimensionless quantities.

x_exp_factor::Int=0: Numerical exponent to scale down the x axis data (first arguments in plot_data), e.g. if x_exp_factor = 10 the values will be divided by $10^{10}$. The default is no scaling.

y_exp_factor::Int=0: Numerical exponent to scale down the y axis data (second arguments in plot_data), e.g. if y_exp_factor = 10 the values will be divided by $10^{10}$. The default is no scaling.

x_trim::NTuple{2,<:Real}=(-Inf, Inf): The data (first arguments in plot_data) will be trim down so the x coordinates fit within x_trim, in the units given by x_unit.

y_trim::NTuple{2,<:Real}=(-Inf, Inf): The data (second arguments in plot_data) will be trim down so the y coordinates fit within y_trim, in the units given by y_unit. This option does not affect histograms.

x_edges::Bool=false: Set it to true if you want to keep the borders of x_trim.

y_edges::Bool=false: Set it to true if you want to keep the borders of y_trim.

x_scale_func::Function=identity: Scaling function for the x axis (first arguments in plot_data). The options are the scaling functions accepted by Makie: log10, log2, log, sqrt, Makie.logit, Makie.Symlog10, Makie.pseudolog10, and identity. The data will be trimmed down to fit within the domain of x_scale_func.

y_scale_func::Function=identity: Scaling function for the y axis (second arguments in plot_data). The options are the scaling functions accepted by Makie: log10, log2, log, sqrt, Makie.logit, Makie.Symlog10, Makie.pseudolog10, and identity. The data will be trimmed down to fit within the domain of y_scale_func.

Returns

A tuple with three elements:

A vector with the elements of plot_data ready to be plotted by plot_function.

If plot_data[1] was mutated to fit within the domain of x_scale_func.

If plot_data[2] was mutated to fit within the domain of y_scale_func (only relevant if length(plot_data) > 1).

GalaxyInspector.xlimits! — Method

Extract the limits of the x axis, from a Makie plot, axis, or figure. In the case of a figure, it will take the limits from the current axis object.

GalaxyInspector.xscale — Method

Extract the scale function of the x axis, from a Makie plot, axis, or figure. In the case of a figure, it will take the scale from the current axis object.

GalaxyInspector.ylimits! — Method

Extract the limits of the y axis, from a Makie plot, axis, or figure. In the case of a figure, it will take the limits from the current axis object.

GalaxyInspector.yscale — Method

Extract the scale function of the y axis, from a Makie plot, axis, or figure. In the case of a figure, it will take the scale from the current axis object.