Constants

List of constant and structures used to configure GalaxyInspector, and to store information about Arepo (e.g. the available quantities in the snapshot and their units).

GalaxyInspector.AGE_RESOLUTION — Constant

Stellar age limit for the SFR and sSFR calculations.

GalaxyInspector.AGE_RESOLUTION_ρ — Constant

Stellar age limit for the SFR area density calculation.

GalaxyInspector.AtomicWeight — Constant

Standard atomic weights.

References

T. Prohaska et al. (2022). Standard atomic weights of the elements 2021 (IUPAC Technical Report). Pure and Applied Chemistry, 94(5), 573-600. doi:10.1515/pac-2019-0603

GalaxyInspector.BIGIEL2008_DATA_PATH — Constant

Path to the file with the fits for the molecular Kennicutt-Schmidt relation, taken from Bigiel et al. (2008).

References

F. Bigiel et al. (2008). THE STAR FORMATION LAW IN NEARBY GALAXIES ON SUB-KPC SCALES. The Astrophysical Journal, 136(6), 2846. doi:10.1088/0004-6256/136/6/2846

GalaxyInspector.CRITICAL_DENSITY — Constant

Critical density for Newtonian simulations, above which the gas particles enter the star formation routine.

This value corresponds to CritPhysDensity $= 0.318 \, [\mathrm{cm^{-3}}]$ in the param.txt file. Which is converted to internal units within the code using CritPhysDensity * PROTONMASS / HYDROGEN_MASSFRAC / UnitDensity_in_cgs

GalaxyInspector.ColorType — Type

Color type.

GalaxyInspector.DefaultParticleIndex — Constant

Code index for each type of cell/particle.

Note

This index is for simulations with 6 cell/particle types.

References

See for example Gadget2 User's Guide, or Gadget4 documentation.

GalaxyInspector.ElementIndex — Constant

Code index for each tracked element.

GalaxyInspector.FELDMANN2020_DATA_PATH — Constant

Path to the file with the global galaxy properties, taken from Feldmann (2020).

References

R. Feldmann (2020). The link between star formation and gas in nearby galaxies. Communications Physics 3(226). doi:10.1038/s42005-020-00493-0

GalaxyInspector.FILTER_R — Constant

Radius of the sphere used in filterWithin, and for area density calculation.

GalaxyInspector.FullParticleIndex — Constant

Code index for each type of cell/particle.

Note

This index is for simulations with 7 cell/particle types.

References

See for example Gadget2 User's Guide, or Gadget4 documentation.

GalaxyInspector.GC_BASENAME — Constant

Base name of the group catalog files.

GalaxyInspector.GasAbundance — Constant

Symbol list for the gas abundance quantities.

GalaxyInspector.HUBBLE_CONSTANT — Constant

Hubble constant in $\mathrm{Gyr^{-1}}$.

This value corresponds to $H_0 = 0.102201 \, \mathrm{Gyr}^{-1} = 100 \, \mathrm{km} \, \mathrm{s}^{-1} \, \mathrm{Mpc}^{-1}$.

GalaxyInspector.ILLUSTRIS_L_UNIT — Constant

Internal unit of length used in IllustrisTNG, equivalent to $1.0 \, \mathrm{Mpc}$. See the documentation here

GalaxyInspector.ILLUSTRIS_M_UNIT — Constant

Internal unit of mass used in IllustrisTNG, equivalent to $10^{10} \, \mathrm{M_\odot}$. See the documentation here

GalaxyInspector.ILLUSTRIS_V_UNIT — Constant

Internal unit of velocity used in IllustrisTNG, equivalent to $1.0 \, \mathrm{km \, s^{-1}}$. See the documentation here

GalaxyInspector.IndexParticle — Constant

Type of cell/particle corresponding to each code index.

GalaxyInspector.IndexType — Type

Index type.

GalaxyInspector.KS98_INTERCEPT — Constant

Intercept of the Kennicutt-Schmidt law, taken from Kennicutt (1998).

References

R. C. Kennicutt (1998). The Global Schmidt Law in Star-forming Galaxies. The Astrophysical Journal, 498(2), 541-552. doi:10.1086/305588

GalaxyInspector.KS98_RHO_UNIT — Constant

Unit of surface density of the Kennicutt-Schmidt law, taken from Kennicutt (1998).

References

R. C. Kennicutt (1998). The Global Schmidt Law in Star-forming Galaxies. The Astrophysical Journal, 498(2), 541-552. doi:10.1086/305588

GalaxyInspector.KS98_SLOPE — Constant

Slope of the Kennicutt-Schmidt law, taken from Kennicutt (1998).

References

R. C. Kennicutt (1998). The Global Schmidt Law in Star-forming Galaxies. The Astrophysical Journal, 498(2), 541-552. doi:10.1086/305588

GalaxyInspector.LINE_STYLES — Constant

List of line styles.

GalaxyInspector.LineStyleType — Type

Line style type.

GalaxyInspector.MARKERS — Constant

List of marker types.

GalaxyInspector.MOLLA2015_DATA_PATH — Constant

Path to the file with the Milky Way profiles, taken from Mollá et al. (2015).

References

M. Mollá et al. (2015). Galactic chemical evolution: stellar yields and the initial mass function. Monthly Notices of the Royal Astronomical Society 451(4), 3693–3708. doi:10.1093/mnras/stv1102

GalaxyInspector.P0 — Constant

Reference pressure for the molecular fraction-pressure relation, taken from Blitz et al. (2006).

References

L. Blitz et al. (2006). The Role of Pressure in GMC Formation II: The H2-Pressure Relation. The Astrophysical Journal, 650(2), 933. doi:10.1086/505417

GalaxyInspector.PHYSICAL_UNITS — Constant

If physical units will be used throughout, instead of comoving units.

GalaxyInspector.ParticleCodeName — Constant

Internal code name (data group in the HDF5 output) corresponding to each type of cell/particle.

GalaxyInspector.ParticleIndex — Constant

Current cell/particle index in use.

GalaxyInspector.ParticleNames — Constant

Human readable name corresponding to each type of cell/particle.

GalaxyInspector.ParticleType — Constant

Type of cell/particle corresponding to each internal code name (data group in the HDF5 output).

GalaxyInspector.PassAll — Constant

Default filter dictionary that does not exclude any cells/particles.

GalaxyInspector.QUANTITIES — Constant

Dictionary of dimensional properties for the quantities in the code.

GalaxyInspector.SFR_REL_PATH — Constant

Relative path, within the simulation directory, to the sfr.txt file.

GalaxyInspector.SNAP_BASENAME — Constant

Base name of the snapshot files, set in the code variable SnapshotFileBase.

GalaxyInspector.SOLAR_METALLICITY — Constant

Solar metallicity, as used in the code.

References

M. Asplund et al. (2006). The new solar abundances - Part I: the observations. Communications in Asteroseismology, 147. doi:10.1553/cia147s76

GalaxyInspector.SolarAbundance — Constant

Solar abundances.

They are defined as $12 + \log_{10}(N_\mathrm{X} / N_\mathrm{H})$, where $N_\mathrm{X}$ and $N_\mathrm{H}$ are the number densities of element $\mathrm{X}$ and hydrogen respectively.

References

M. Asplund et al. (2009). The Chemical Composition of the Sun. Annual Review of Astronomy and Astrophysics, 47(1), 481–522. doi:10.1146/annurev.astro.46.060407.145222

GalaxyInspector.StellarAbundance — Constant

Symbol list for the stellar abundance quantities.

GalaxyInspector.THEME — Constant

Global plot theme.

GalaxyInspector.CircularGrid — Type

Circular grid (2D or 3D), formed by a series of concentric rings or spherical shells.

Fields

grid::Vector{<:Number}: Vector with the distance of each bin to the center of the grid.

ticks::Vector{<:Number}: Vector with the edges of the bins.

center::Vector{<:Number}: 3D location of the center of the grid. For the 2D grid, the grid is assumed to be in the xy plane.

bin_area::Vector{<:Number}: Area of each ring.

bin_volumes::Vector{<:Number}: Volume of each spherical shell.

log::Bool: If the grid is logarithmic.

GalaxyInspector.GroupCatHeader — Type

Data in the "Header" group of a HDF5 group catalog file.

Fields

box_size::Float64 = NaN: Total size of the simulation box.

h0::Float64 = NaN: Hubble parameter.

n_groups_part::Int32 = -1: Number of halos (FoF groups) in this file chunk.

n_groups_total::Int32 = -1: Total number of halos (FoF groups) in this snapshot.

n_subgroups_part::Int32 = -1: Number of subhalos (subfind) in this file chunk.

n_subgroups_total::Int32 = -1: Total number of subhalos (subfind) in this snapshot.

num_files::Int32 = -1: Number of file chunks per snapshot.

omega_0::Float64 = NaN: The cosmological density parameter for matter.

omega_l::Float64 = NaN: The cosmological density parameter for the cosmological constant.

redshift::Float64 = NaN: The redshift.

time::Float64 = NaN: The physical time/scale factor.

GalaxyInspector.GroupCatalog — Type

Metadata for a group catalog file.

Fields

path::Union{String,Missing}: Full path to the group catalog file.

header::GroupCatHeader: Header of the group catalog.

GalaxyInspector.InternalUnits — Type

Unit conversion struct.

Fields

x_cgs::Unitful.Length: Length, from internal units to $\mathrm{cm}$.

x_cosmo::Unitful.Length: Length, from internal units to $\mathrm{kpc}$.

x_comoving::Unitful.Length: Length, from internal units to $\mathrm{ckpc}$.

v_cgs::Unitful.Velocity: Velocity, from internal units to $\mathrm{cm \, s^{-1}}$.

v_cosmo::Unitful.Velocity: Velocity, from internal units to $\mathrm{km \, s^{-1}}$.

m_cgs::Unitful.Mass: Mass, from internal units to $\mathrm{g}$.

m_cosmo::Unitful.Mass: Mass, from internal units to $\mathrm{M_\odot}$.

t_cgs::Unitful.Time: Time, from internal units to $\mathrm{s}$.

t_cosmo::Unitful.Time: Time, from internal units to $\mathrm{Myr}$.

U_cgs::Unitful.Energy: Specific energy, from internal units to $\mathrm{erg \, g^{-1}}$.

rho_cgs::Unitful.Density: Density, from internal units to $\mathrm{g \, cm^{-3}}$.

P_Pa::Unitful.Pressure: Pressure, from internal units to $\mathrm{Pa}$.

GalaxyInspector.LinearGrid — Type

Linear grid (1D).

Fields

grid::Vector{<:Number}: Vector with the center value of each bin.

ticks::Vector{<:Number}: Vector with the edges of the bins.

bin_widths::Vector{<:Number}: Widths of the bins.

log::Bool: If the grid is logarithmic.

GalaxyInspector.PlotParams — Type

Plotting parameters for a quantity.

Fields

request::Dict{Symbol,Vector{String}} = Dict{Symbol,Vector{String}}(): Data request for readSnapshot. It must have the shape cell/particle type -> [block, block, block, ...].

var_name::AbstractString = "": Name of the quantity for the plot axis. It should not include units or scaling factors.

exp_factor::Int64 = 0: Numerical exponent to scale down the axis, e.g. if x_exp_factor = 10 the values will be divided by $10^{10}$. The default is no scaling.

unit::Unitful.Units = Unitful.NoUnits: Target unit for the axis.

axis_label::AbstractString = "auto_label": Label for the axis. It can contain the string auto_label, which will be replaced by the default label: var_name / 10^exp_factor unit.

GalaxyInspector.Qty — Type

Dimensional information about a physical quantity.

Fields

hdf5_name::String: HDF5 block name.

dimensions::Unitful.Dimensions: Physical dimensions of the quantity, e.g. Unitful.𝐋 / Unitful.𝐓.

unit::Union{Unitful.Units,Symbol}: Units of the quantity within the code. It can be a unit from Unitful or UnitfulAstro, or it can be the symbol :internal which denotes internal code units.

GalaxyInspector.Simulation — Type

Metadata for a simulation.

Fields

path::String: Full path to the simulation directory.

index::Int64: An index associated with the simulation.

slice::IndexType: Slice of the simulation, i.e. which snapshots will be read. It can be an integer (a single snapshot), a vector of integers (several snapshots), an UnitRange (e.g. 5:13), an StepRange (e.g. 5:2:13) or (:) (all snapshots).

cosmological::Bool: If the simulation is cosmological,

false -> Newtonian simulation (ComovingIntegrationOn = 0).

true -> Cosmological simulation (ComovingIntegrationOn = 1).

table::DataFrame: A dataframe where each row is a snapshot, and with 8 colums:

:ids -> Dataframe index of each snapshot, i.e. if there are 10 snapshots in total it runs from 1 to 10.

:numbers -> Number in the file name of each snapshot.

:scale_factors -> Scale factor of each snapshot.

:redshifts -> Redshift of each snapshot.

:physical_times -> Physical time since the Big Bang.

:lookback_times -> Physical time left to reach the last snapshot.

:snapshot_paths -> Full path to the snapshots.

:groupcat_paths -> Full path to the group catalog files.

GalaxyInspector.Snapshot — Type

Metadata for a snapshot.

Fields

path::String: Full path to the snapshot.

global_index::Int64: Index of the snapshot in the context of the whole simulation.

slice_index::Int64: Index of the snapshot in the context of the slice.

physical_time::Unitful.Time: Physical time since the Big Bang.

lookback_time::Unitful.Time: Physical time left to reach the last snapshot.

scale_factor::Float64: Scale factor of the snapshot.

redshift::Float64: Redshift of the snapshot.

header::SnapshotHeader: Header of the snapshot.

GalaxyInspector.SnapshotHeader — Type

Data in the "Header" group of a HDF5 snapshot file.

Fields

box_size::Float64: Total size of the simulation box.

h0::Float64: Hubble parameter.

mass_table::Vector{Float64}: Masses of particle types which have a constant mass.

num_files::Int32: Number of file chunks per snapshot.

num_part::Vector{Int32}: Number of particles (of each type) included in this file chunk.

num_total::Vector{UInt32}: Total number of particles (of each type) for this snapshot.

omega_0::Float64: The cosmological density parameter for matter.

omega_l::Float64: The cosmological density parameter for the cosmological constant.

redshift::Float64: The redshift.

time::Float64: The physical time/scale factor.

l_unit::Unitful.Length: Conversion factor from internal units of length to centimeters.

m_unit::Unitful.Mass: Conversion factor from internal units of mass to grams.

v_unit::Unitful.Velocity: Conversion factor from internal units of velocity to centimeters per second.

GalaxyInspector.SquareGrid — Type

Square grid (2D).

Fields

grid::Matrix{NTuple{2,<:Number}}: Matrix with the physical coordinates of the center of each pixel in the grid.

x_ticks::Vector{<:Number}: Full set of possible values for the x coordinate.

y_ticks::Vector{<:Number}: Full set of possible values for the y coordinate.

size::Number: Side length of the square grid.

n_bins::Int64: Number of bins per side of the grid.

bin_width::Number: Side length of each bin.

bin_area::Number: Area of each bin.