Quantities

This is a list of constant related to code and derived quantities used in GalaxyInspector.

None of these are exported.

GalaxyInspector.COMPONENTS — Constant

List of symbols for the physical components of a simulation.

GalaxyInspector.DERIVED_QTY — Constant

List of symbols for the derived quantities (component + magnitude).

GalaxyInspector.MAGNITUDES — Constant

List of symbols for physical magnitudes.

GalaxyInspector.QTY_GLOBAL_LIST — Constant

List of symbols for almost all the quantities in GalaxyInspector.

GalaxyInspector.QUANTITIES — Constant

Dimensional properties for the quantities in the code.

GalaxyInspector.QUANTITY_SPLITS — Constant

Dictionary mapping each derived quantity to its component and magnitude.

GalaxyInspector.RATIO_QTY — Constant

List of symbols for the ratios between derived quantities.

GalaxyInspector.RATIO_SPLITS — Constant

Dictionary mapping each ratio to its two derived quantities.

GalaxyInspector.SFM_DERIVED_MAGNITUDES — Constant

List of symbols for every derived code magnitude in our SF model.

GalaxyInspector.SFM_DERIVED_QTY — Constant

List of symbols for every derived code quantity in our SF model.

GalaxyInspector.SFM_DERIVED_QTY_SPLITS — Constant

Dictionary mapping each derived code quantity to its component and magnitude.

GalaxyInspector.SFM_GAS_DERIVED_QTY — Constant

List of symbols for every derived code quantity associated with a gas cell in our SF model.

GalaxyInspector.SFM_GAS_QTY — Constant

List of symbols for every code quantity associated with a gas cell in our SF model.

GalaxyInspector.SFM_KEYS — Constant

Dictionary mapping each code quantity in our SF model with its QUANTITIES key.

GalaxyInspector.SFM_QTY — Constant

List of symbols for every code quantity in our SF model.

GalaxyInspector.SFM_QTY_SPLITS — Constant

Dictionary mapping each code quantity in our SF model to its component and magnitude.

GalaxyInspector.SFM_STELLAR_DERIVED_QTY — Constant

List of symbols for every derived code quantity associated with a stellar particle in our SF model.

This values come from the state of the parent gas cell (at the stellar birth time).

GalaxyInspector.SFM_STELLAR_QTY — Constant

List of symbols for every code quantity associated with a stellar particle in our SF model.

This values come from the state of the parent gas cell (at the stellar birth time).