Convenience functions

These functions are pre-made recipes for specific plot types.

They are also examples of how to use the more configurable, but much more complicated, pipeline functions.

All these functions are exported.

GalaxyInspector.compareWithFeldmann2020 — Method

compareWithFeldmann2020(
    simulation_paths::Vector{String},
    x_quantity::Symbol,
    y_quantity::Symbol;
    <keyword arguments>
)::Nothing

Plot a time series plus the corresponding experimental results from Feldmann (2020).

Arguments

simulation_paths::Vector{String}: Paths to the simulation directories, set in the code variable OutputDir.

x_quantity::Symbol: Quantity for the x axis. The possibilities are:

:stellar_mass -> Stellar mass.

:molecular_mass -> Molecular hydrogen ($\mathrm{H_2}$) mass.

:atomic_mass -> Atomic hydrogen ($\mathrm{HI}$) mass.

:sfr -> The star formation rate of the last AGE_RESOLUTION.

y_quantity::Symbol: Quantity for the y axis. The possibilities are:

:stellar_mass -> Stellar mass.

:molecular_mass -> Molecular hydrogen ($\mathrm{H_2}$) mass.

:atomic_mass -> Atomic hydrogen ($\mathrm{HI}$) mass.

:sfr -> The star formation rate of the last AGE_RESOLUTION.

slice::IndexType=(:): Slice of the simulation, i.e. which snapshots will be read. It can be an integer (a single snapshot), a vector of integers (several snapshots), an UnitRange (e.g. 5:13), an StepRange (e.g. 5:2:13) or (:) (all snapshots). It works over the longest possible list of snapshots among the simulations (grouped by the number in the file names). Out of bounds indices are ignored.

output_path::String="./": Path to the output folder.

filter_mode::Symbol=:all: Which cells/particles will be plotted, the options are:

:all -> Plot every cell/particle within the simulation box.

:halo -> Plot only the cells/particles that belong to the main halo.

:subhalo -> Plot only the cells/particles that belong to the main subhalo.

:sphere -> Plot only the cell/particle inside a sphere with radius FILTER_R (see ./src/constants.jl).

:stellar_subhalo -> Plot only the cells/particles that belong to the main subhalo.

sim_labels::Union{Vector{String},Nothing}=basename.(simulation_paths): Labels for the plot legend, one per simulation. Set it to nothing if you don't want a legend.

References

R. Feldmann (2020). The link between star formation and gas in nearby galaxies. Communications Physics 3(226). doi:10.1038/s42005-020-00493-0

GalaxyInspector.compareWithKennicuttBigiel — Method

compareWithKennicuttBigiel(
    simulation_paths::Vector{String},
    slice_n::Int64;
    <keyword arguments>
)::Nothing

Plot the resolved Kennicutt-Schmidt relation plus the results of Kennicutt (1998) or Bigiel et al. (2008), depending on the chosen quantity.

Note

This method plots the KS relation using cylindrical bins at a fix moment in time.

Arguments

simulation_paths::Vector{String}: Paths to the simulation directories, set in the code variable OutputDir.

slice_n::Int64: Selects which snapshot to plot, starts at 1 and is independent of the number in the file name. If every snapshot is present, slice_n = filename_number + 1.

quantity::Symbol=:molecular_area_density: Quantity for the x axis. The possibilities are:

:gas_area_density -> Gas area mass density, for a radius of FILTER_R. This one will be plotted with the results of Kennicutt (1998).

:molecular_area_density -> Molecular hydrogen area mass density, for a radius of FILTER_R. This one will be plotted with the results of Bigiel et al. (2008).

:neutral_area_density -> Neutral hydrogen area mass density, for a radius of FILTER_R. This one will be plotted with the results of Bigiel et al. (2008).

output_path::String="./": Path to the output folder.

filter_mode::Symbol=:all: Which cells/particles will be plotted, the options are:

:all -> Plot every cell/particle within the simulation box.

:halo -> Plot only the cells/particles that belong to the main halo.

:subhalo -> Plot only the cells/particles that belong to the main subhalo.

:sphere -> Plot only the cell/particle inside a sphere with radius FILTER_R (see ./src/constants.jl).

:stellar_subhalo -> Plot only the cells/particles that belong to the main subhalo.

sim_labels::Union{Vector{String},Nothing}=basename.(simulation_paths): Labels for the plot legend, one per simulation. Set it to nothing if you don't want a legend.

References

R. C. Kennicutt (1998). The Global Schmidt Law in Star-forming Galaxies. The Astrophysical Journal, 498(2), 541-552. doi:10.1086/305588

F. Bigiel et al. (2008). THE STAR FORMATION LAW IN NEARBY GALAXIES ON SUB-KPC SCALES. The Astrophysical Journal, 136(6), 2846. doi:10.1088/0004-6256/136/6/2846

GalaxyInspector.compareWithKennicuttBigiel — Method

compareWithKennicuttBigiel(
    simulation_paths::Vector{String},
    slice::Union{Colon,UnitRange{<:Integer},StepRange{<:Integer,<:Integer},Vector{<:Integer}};
    <keyword arguments>
)::Nothing

Plot the integrated Kennicutt-Schmidt relation plus the results of Kennicutt (1998) or Bigiel et al. (2008), depending on the chosen quantity.

Note

This method plots the KS relation for the whole galaxy at different points in time.

Arguments

simulation_paths::Vector{String}: Paths to the simulation directories, set in the code variable OutputDir.

slice::Union{Colon,UnitRange{<:Integer},StepRange{<:Integer,<:Integer},Vector{<:Integer}}: Slice of the simulation, i.e. which snapshots will be read. It can be a vector of integers (several snapshots), an UnitRange (e.g. 5:13), an StepRange (e.g. 5:2:13) or (:) (all snapshots). It works over the longest possible list of snapshots among the simulations (grouped by the number in the file names). Out of bounds indices are ignored.

quantity::Symbol=:molecular_area_density: Quantity for the x axis. The possibilities are:

:gas_area_density -> Gas area mass density, for a radius of FILTER_R. This one will be plotted with the results of Kennicutt (1998).

:molecular_area_density -> Molecular hydrogen area mass density, for a radius of FILTER_R. This one will be plotted with the results of Bigiel et al. (2008).

:neutral_area_density -> Neutral hydrogen area mass density, for a radius of FILTER_R. This one will be plotted with the results of Bigiel et al. (2008).

output_path::String="./": Path to the output folder.

filter_mode::Symbol=:all: Which cells/particles will be plotted, the options are:

:all -> Plot every cell/particle within the simulation box.

:halo -> Plot only the cells/particles that belong to the main halo.

:subhalo -> Plot only the cells/particles that belong to the main subhalo.

:sphere -> Plot only the cell/particle inside a sphere with radius FILTER_R (see ./src/constants.jl).

:stellar_subhalo -> Plot only the cells/particles that belong to the main subhalo.

sim_labels::Union{Vector{String},Nothing}=basename.(simulation_paths): Labels for the plot legend, one per simulation. Set it to nothing if you don't want a legend.

References

R. C. Kennicutt (1998). The Global Schmidt Law in Star-forming Galaxies. The Astrophysical Journal, 498(2), 541-552. doi:10.1086/305588

F. Bigiel et al. (2008). THE STAR FORMATION LAW IN NEARBY GALAXIES ON SUB-KPC SCALES. The Astrophysical Journal, 136(6), 2846. doi:10.1088/0004-6256/136/6/2846

GalaxyInspector.compareWithMolla2015 — Method

compareWithMolla2015(
    simulation_paths::Vector{String},
    slice_n::Int64,
    quantity::Symbol;
    <keyword arguments>
)::Nothing

Plot a Milky Way profile plus the corresponding experimental results from Mollá et al. (2015).

Arguments

simulation_paths::Vector{String}: Paths to the simulation directories, set in the code variable OutputDir.

slice_n::Int64: Selects which snapshot to plot, starts at 1 and is independent of the number in the file name. If every snapshot is present, slice_n = filename_number + 1.

quantity::Symbol: Quantity for the y axis. The options are:

:stellar_area_density -> Stellar area mass density.

:molecular_area_density -> Molecular hydrogen area mass density.

:atomic_area_density -> Atomic hydrogen area mass density.

:sfr_area_density -> Star formation rate area density, for the last AGE_RESOLUTION_ρ.

:O_stellar_abundance -> Stellar abundance of oxygen, as $12 + \log_{10}(\mathrm{O \, / \, H})$.

:N_stellar_abundance -> Stellar abundance of nitrogen, as $12 + \log_{10}(\mathrm{N \, / \, H})$.

:C_stellar_abundance -> Stellar abundance of carbon, as $12 + \log_{10}(\mathrm{C \, / \, H})$.

output_path::String="./": Path to the output folder.

filter_mode::Symbol=:all: Which cells/particles will be plotted, the options are:

:all -> Plot every cell/particle within the simulation box.

:halo -> Plot only the cells/particles that belong to the main halo.

:subhalo -> Plot only the cells/particles that belong to the main subhalo.

:sphere -> Plot only the cell/particle inside a sphere with radius FILTER_R (see ./src/constants.jl).

:stellar_subhalo -> Plot only the cells/particles that belong to the main subhalo.

sim_labels::Union{Vector{String},Nothing}=basename.(simulation_paths): Labels for the plot legend, one per simulation. Set it to nothing if you don't want a legend.

References

M. Mollá et al. (2015). Galactic chemical evolution: stellar yields and the initial mass function. Monthly Notices of the Royal Astronomical Society 451(4), 3693–3708. doi:10.1093/mnras/stv1102

GalaxyInspector.densityMap — Method

densityMap(
    simulation_paths::Vector{String},
    slice_n::Int64;
    <keyword arguments>
)::Nothing

Plot a 2D histogram of the density.

Arguments

simulation_paths::Vector{String}: Paths to the simulation directories, set in the code variable OutputDir.

slice_n::Int64: Selects which snapshot to plot, starts at 1 and is independent of the number in the file name. If set to 0, an animation using every snapshots will be made. If every snapshot is present, slice_n = filename_number + 1.

quantities::Vector{Symbol}=[:gas_mass]: Quantities for which the density will be calculated. The options are:

:stellar_mass -> Stellar mass.

:gas_mass -> Gas mass.

:dm_mass -> Dark matter mass.

:bh_mass -> Black hole mass.

:molecular_mass -> Molecular hydrogen ($\mathrm{H_2}$) mass.

:atomic_mass -> Atomic hydrogen ($\mathrm{HI}$) mass.

:ionized_mass -> Ionized hydrogen ($\mathrm{HII}$) mass.

:neutral_mass -> Neutral hydrogen ($\mathrm{HI + H_2}$) mass.

output_path::String="./": Path to the output folder.

filter_mode::Symbol=:all: Which cells/particles will be plotted, the options are:

:all -> Plot every cell/particle within the simulation box.

:halo -> Plot only the cells/particles that belong to the main halo.

:subhalo -> Plot only the cells/particles that belong to the main subhalo.

:sphere -> Plot only the cell/particle inside a sphere with radius FILTER_R (see ./src/constants.jl).

:stellar_subhalo -> Plot only the cells/particles that belong to the main subhalo.

projection_planes::Vector{Symbol}=[:xy]: Projection planes. The options are :xy, :xz and :yz.

box_size::Unitful.Length=100u"kpc": Physical side length of the plot window.

pixel_length::Unitful.Length=0.1u"kpc": Pixel (bin of the 2D histogram) side length.

smooth::Bool=false: If the results will be smooth out using the kernel function cubicSplineKernel.

GalaxyInspector.rotationCurve — Method

rotationCurve(
    simulation_paths::Vector{String},
    slice_n::Int64;
    <keyword arguments>
)::Nothing

Plot the galaxy rotation curve of a set of simulations.

Arguments

simulation_paths::Vector{String}: Paths to the simulation directories, set in the code variable OutputDir.

slice_n::Int64: Selects which snapshot to plot, starts at 1 and is independent of the number in the file name. If every snapshot is present, slice_n = filename_number + 1.

radius::Unitful.Length=FILTER_R: Maximum radial distance for the rotation curve.

output_path::String="./": Path to the output folder.

filter_mode::Symbol=:all: Which cells/particles will be plotted, the options are:

:all -> Plot every cell/particle within the simulation box.

:halo -> Plot only the cells/particles that belong to the main halo.

:subhalo -> Plot only the cells/particles that belong to the main subhalo.

:sphere -> Plot only the cell/particle inside a sphere with radius FILTER_R (see ./src/constants.jl).

:stellar_subhalo -> Plot only the cells/particles that belong to the main subhalo.

sim_labels::Union{Vector{String},Nothing}=basename.(simulation_paths): Labels for the plot legend, one per simulation. Set it to nothing if you don't want a legend.

GalaxyInspector.scatterDensityMap — Method

scatterDensityMap(
    simulation_paths::Vector{String},
    slice_n::Int64,
    x_quantity::Symbol,
    y_quantity::Symbol,
    x_range::NTuple{2, <:Number},
    y_range::NTuple{2, <:Number},
    n_bins::Int64;
    <keyword arguments>
)::Nothing

Plot two quantities as a density scatter plot (2D histogram).

Arguments

simulation_paths::Vector{String}: Paths to the simulation directories, set in the code variable OutputDir.

slice_n::Int64: Selects which snapshot to plot, starts at 1 and is independent of the number in the file name. If set to 0, an animation using every snapshots will be made. If every snapshot is present, slice_n = filename_number + 1.

x_quantity::Symbol: Quantity for the x axis. The possibilities are:

:stellar_mass -> Stellar mass.

:gas_mass -> Gas mass.

:dm_mass -> Dark matter mass.

:bh_mass -> Black hole mass.

:molecular_mass -> Molecular hydrogen ($\mathrm{H_2}$) mass.

:atomic_mass -> Atomic hydrogen ($\mathrm{HI}$) mass.

:ionized_mass -> Ionized hydrogen ($\mathrm{HII}$) mass.

:neutral_mass -> Neutral hydrogen ($\mathrm{HI + H_2}$) mass.

:molecular_fraction -> Gas mass fraction of molecular hydrogen.

:atomic_fraction -> Gas mass fraction of atomic hydrogen.

:ionized_fraction -> Gas mass fraction of ionized hydrogen.

:neutral_fraction -> Gas mass fraction of neutral hydrogen.

:gas_mass_density -> Gas mass density.

:gas_number_density -> Gas number density.

:molecular_number_density -> Molecular hydrogen number density.

:atomic_number_density -> Atomic hydrogen number density.

:ionized_number_density -> Ionized hydrogen number density.

:neutral_number_density -> Neutral hydrogen number density.

:gas_metallicity -> Mass fraction of all elements above He in the gas (solar units).

:stellar_metallicity -> Mass fraction of all elements above He in the stars (solar units).

:X_gas_abundance -> Gas abundance of element $\mathrm{X}$, as $12 + \log_{10}(\mathrm{X \, / \, H})$. The possibilities are the keys of ElementIndex.

:X_stellar_abundance -> Stellar abundance of element $\mathrm{X}$, as $12 + \log_{10}(\mathrm{X \, / \, H})$. The possibilities are the keys of ElementIndex.

:stellar_radial_distance -> Distance of every stellar particle to the origin.

:gas_radial_distance -> Distance of every gas cell to the origin.

:dm_radial_distance -> Distance of every dark matter particle to the origin.

:stellar_xy_distance -> Projected distance of every stellar particle to the origin.

:gas_xy_distance -> Projected distance of every gas cell to the origin.

:dm_xy_distance -> Projected distance of every dark matter particle to the origin.

:stellar_circularity -> Stellar circularity.

:stellar_vcirc -> Stellar circular velocity.

:stellar_age -> Stellar age.

:sfr -> The star formation rate of the last AGE_RESOLUTION.

:ssfr -> The specific star formation rate of the last AGE_RESOLUTION.

:temperature -> Gas temperature, as $\log_{10}(T \, / \, \mathrm{K})$.

y_quantity::Symbol: Quantity for the y axis. The possibilities are:

:stellar_mass -> Stellar mass.

:gas_mass -> Gas mass.

:dm_mass -> Dark matter mass.

:bh_mass -> Black hole mass.

:molecular_mass -> Molecular hydrogen ($\mathrm{H_2}$) mass.

:atomic_mass -> Atomic hydrogen ($\mathrm{HI}$) mass.

:ionized_mass -> Ionized hydrogen ($\mathrm{HII}$) mass.

:neutral_mass -> Neutral hydrogen ($\mathrm{HI + H_2}$) mass.

:molecular_fraction -> Gas mass fraction of molecular hydrogen.

:atomic_fraction -> Gas mass fraction of atomic hydrogen.

:ionized_fraction -> Gas mass fraction of ionized hydrogen.

:neutral_fraction -> Gas mass fraction of neutral hydrogen.

:gas_mass_density -> Gas mass density.

:gas_number_density -> Gas number density.

:molecular_number_density -> Molecular hydrogen number density.

:atomic_number_density -> Atomic hydrogen number density.

:ionized_number_density -> Ionized hydrogen number density.

:neutral_number_density -> Neutral hydrogen number density.

:gas_metallicity -> Mass fraction of all elements above He in the gas (solar units).

:stellar_metallicity -> Mass fraction of all elements above He in the stars (solar units).

:X_gas_abundance -> Gas abundance of element $\mathrm{X}$, as $12 + \log_{10}(\mathrm{X \, / \, H})$. The possibilities are the keys of ElementIndex.

:X_stellar_abundance -> Stellar abundance of element $\mathrm{X}$, as $12 + \log_{10}(\mathrm{X \, / \, H})$. The possibilities are the keys of ElementIndex.

:stellar_radial_distance -> Distance of every stellar particle to the origin.

:gas_radial_distance -> Distance of every gas cell to the origin.

:dm_radial_distance -> Distance of every dark matter particle to the origin.

:stellar_xy_distance -> Projected distance of every stellar particle to the origin.

:gas_xy_distance -> Projected distance of every gas cell to the origin.

:dm_xy_distance -> Projected distance of every dark matter particle to the origin.

:stellar_circularity -> Stellar circularity.

:stellar_vcirc -> Stellar circular velocity.

:stellar_age -> Stellar age.

:sfr -> The star formation rate of the last AGE_RESOLUTION.

:ssfr -> The specific star formation rate of the last AGE_RESOLUTION.

:temperature -> Gas temperature, as $\log_{10}(T \, / \, \mathrm{K})$.

x_range::NTuple{2,<:Number}: x axis range.

y_range::NTuple{2,<:Number}: y axis range.

n_bins::Int64: Number of bins per side of the square grid.

output_path::String="./": Path to the output folder.

filter_mode::Symbol=:all: Which cells/particles will be plotted, the options are:

:all -> Plot every cell/particle within the simulation box.

:halo -> Plot only the cells/particles that belong to the main halo.

:subhalo -> Plot only the cells/particles that belong to the main subhalo.

:sphere -> Plot only the cell/particle inside a sphere with radius FILTER_R (see ./src/constants.jl).

:stellar_subhalo -> Plot only the cells/particles that belong to the main subhalo

GalaxyInspector.scatterPlot — Method

scatterPlot(
    simulation_paths::Vector{String},
    slice_n::Int64,
    x_quantity::Symbol,
    y_quantity::Symbol;
    <keyword arguments>
)::Nothing

Plot two quantities as a scatter plot, one marker for every cell/particle.

Arguments

simulation_paths::Vector{String}: Paths to the simulation directories, set in the code variable OutputDir.

slice_n::Int64: Selects which snapshot to plot, starts at 1 and is independent of the number in the file name. If every snapshot is present, slice_n = filename_number + 1.

x_quantity::Symbol: Quantity for the x axis. The possibilities are:

:stellar_mass -> Stellar mass.

:gas_mass -> Gas mass.

:dm_mass -> Dark matter mass.

:bh_mass -> Black hole mass.

:molecular_mass -> Molecular hydrogen ($\mathrm{H_2}$) mass.

:atomic_mass -> Atomic hydrogen ($\mathrm{HI}$) mass.

:ionized_mass -> Ionized hydrogen ($\mathrm{HII}$) mass.

:neutral_mass -> Neutral hydrogen ($\mathrm{HI + H_2}$) mass.

:molecular_fraction -> Gas mass fraction of molecular hydrogen.

:atomic_fraction -> Gas mass fraction of atomic hydrogen.

:ionized_fraction -> Gas mass fraction of ionized hydrogen.

:neutral_fraction -> Gas mass fraction of neutral hydrogen.

:gas_mass_density -> Gas mass density.

:gas_number_density -> Gas number density.

:molecular_number_density -> Molecular hydrogen number density.

:atomic_number_density -> Atomic hydrogen number density.

:ionized_number_density -> Ionized hydrogen number density.

:neutral_number_density -> Neutral hydrogen number density.

:gas_metallicity -> Mass fraction of all elements above He in the gas (solar units).

:stellar_metallicity -> Mass fraction of all elements above He in the stars (solar units).

:X_gas_abundance -> Gas abundance of element $\mathrm{X}$, as $12 + \log_{10}(\mathrm{X \, / \, H})$. The possibilities are the keys of ElementIndex.

:X_stellar_abundance -> Stellar abundance of element $\mathrm{X}$, as $12 + \log_{10}(\mathrm{X \, / \, H})$. The possibilities are the keys of ElementIndex.

:stellar_radial_distance -> Distance of every stellar particle to the origin.

:gas_radial_distance -> Distance of every gas cell to the origin.

:dm_radial_distance -> Distance of every dark matter particle to the origin.

:stellar_xy_distance -> Projected distance of every stellar particle to the origin.

:gas_xy_distance -> Projected distance of every gas cell to the origin.

:dm_xy_distance -> Projected distance of every dark matter particle to the origin.

:stellar_circularity -> Stellar circularity.

:stellar_vcirc -> Stellar circular velocity.

:stellar_age -> Stellar age.

:sfr -> The star formation rate of the last AGE_RESOLUTION.

:ssfr -> The specific star formation rate of the last AGE_RESOLUTION.

:temperature -> Gas temperature, as $\log_{10}(T \, / \, \mathrm{K})$.

y_quantity::Symbol: Quantity for the y axis. The possibilities are:

:stellar_mass -> Stellar mass.

:gas_mass -> Gas mass.

:dm_mass -> Dark matter mass.

:bh_mass -> Black hole mass.

:molecular_mass -> Molecular hydrogen ($\mathrm{H_2}$) mass.

:atomic_mass -> Atomic hydrogen ($\mathrm{HI}$) mass.

:ionized_mass -> Ionized hydrogen ($\mathrm{HII}$) mass.

:neutral_mass -> Neutral hydrogen ($\mathrm{HI + H_2}$) mass.

:molecular_fraction -> Gas mass fraction of molecular hydrogen.

:atomic_fraction -> Gas mass fraction of atomic hydrogen.

:ionized_fraction -> Gas mass fraction of ionized hydrogen.

:neutral_fraction -> Gas mass fraction of neutral hydrogen.

:gas_mass_density -> Gas mass density.

:gas_number_density -> Gas number density.

:molecular_number_density -> Molecular hydrogen number density.

:atomic_number_density -> Atomic hydrogen number density.

:ionized_number_density -> Ionized hydrogen number density.

:neutral_number_density -> Neutral hydrogen number density.

:gas_metallicity -> Mass fraction of all elements above He in the gas (solar units).

:stellar_metallicity -> Mass fraction of all elements above He in the stars (solar units).

:X_gas_abundance -> Gas abundance of element $\mathrm{X}$, as $12 + \log_{10}(\mathrm{X \, / \, H})$. The possibilities are the keys of ElementIndex.

:X_stellar_abundance -> Stellar abundance of element $\mathrm{X}$, as $12 + \log_{10}(\mathrm{X \, / \, H})$. The possibilities are the keys of ElementIndex.

:stellar_radial_distance -> Distance of every stellar particle to the origin.

:gas_radial_distance -> Distance of every gas cell to the origin.

:dm_radial_distance -> Distance of every dark matter particle to the origin.

:stellar_xy_distance -> Projected distance of every stellar particle to the origin.

:gas_xy_distance -> Projected distance of every gas cell to the origin.

:dm_xy_distance -> Projected distance of every dark matter particle to the origin.

:stellar_circularity -> Stellar circularity.

:stellar_vcirc -> Stellar circular velocity.

:stellar_age -> Stellar age.

:sfr -> The star formation rate of the last AGE_RESOLUTION.

:ssfr -> The specific star formation rate of the last AGE_RESOLUTION.

:temperature -> Gas temperature, as $\log_{10}(T \, / \, \mathrm{K})$.

output_path::String="./": Path to the output folder.

filter_mode::Symbol=:all: Which cells/particles will be plotted, the options are:

:all -> Plot every cell/particle within the simulation box.

:halo -> Plot only the cells/particles that belong to the main halo.

:subhalo -> Plot only the cells/particles that belong to the main subhalo.

:sphere -> Plot only the cell/particle inside a sphere with radius FILTER_R (see ./src/constants.jl).

:stellar_subhalo -> Plot only the cells/particles that belong to the main subhalo.

GalaxyInspector.sfrTXT — Method

sfrTXT(
    simulation_paths::Vector{String},
    x_quantity::Symbol,
    y_quantity::Symbol;
    <keyword arguments>
)::Nothing

Plot a time series of the data in the sfr.txt file.

Arguments

simulation_paths::Vector{String}: Paths to the simulation directories, set in the code variable OutputDir.

x_quantity::Symbol: Quantity for the x axis. The possibilities are:

:physical_time -> Physical time since the Big Bang.

:lookback_time -> Physical time left to reach the last snapshot.

:scale_factor -> Scale factor.

:redshift -> Redshift.

y_quantity::Symbol: Quantity for the y axis. The possibilities are:

:stellar_mass -> Stellar mass.

:sfr -> The star formation rate.

output_path::String="./": Path to the output folder.

sim_labels::Union{Vector{String},Nothing}=basename.(simulation_paths): Labels for the plot legend, one per simulation. Set it to nothing if you don't want a legend.

GalaxyInspector.simulationReport — Method

simulationReport(
    simulation_paths::Vector{String},
    slice_n::Int64; 
    <keyword arguments>
)::Nothing

Write a text file with information about a given snapshot and simulation.

Arguments

simulation_paths::Vector{String}: Paths to the simulation directories, set in the code variable OutputDir. One text file will be printed for each simulation.

slice_n::Int64: Selects which snapshot to plot, starts at 1 and is independent of the number in the file name. If every snapshot is present, slice_n = filename_number + 1.

output_path::String="./": Path to the output folder.

filter_mode::Symbol=:all: Which cells/particles will be considered in the "filtered" section of the report. The options are:

:all -> Consider every cell/particle within the simulation box.

:halo -> Consider only the cells/particles that belong to the main halo.

:subhalo -> Consider only the cells/particles that belong to the main subhalo.

:sphere -> Consider only the cell/particle inside a sphere with radius FILTER_R (see ./src/constants.jl).

:stellar_subhalo -> Consider only the cells/particles that belong to the main subhalo.

halo_idx::Int64=1: Index of the target halo (FoF group). Starts at 1.

subhalo_rel_idx::Int64=1: Index of the target subhalo (subfind), relative the target halo. Starts at 1.

warnings::Bool=true: If a warning will be given when there is missing files.

GalaxyInspector.stellarCircularity — Method

stellarCircularity(
    simulation_paths::Vector{String},
    slice_n::Int64;
    <keyword arguments>
)::Nothing

Plot a histogram of the stellar circularity.

Arguments

simulation_paths::Vector{String}: Paths to the simulation directories, set in the code variable OutputDir.

slice_n::Int64: Selects which snapshot to plot, starts at 1 and is independent of the number in the file name. If every snapshot is present, slice_n = filename_number + 1.

range::NTuple{2,<:Number}=(-2.0, 2.0): Circularity range.

n_bins::Int64=60: Number of bins.

output_path::String="./": Path to the output folder.

filter_mode::Symbol=:all: Which cells/particles will be plotted, the options are:

:all -> Plot every cell/particle within the simulation box.

:halo -> Plot only the cells/particles that belong to the main halo.

:subhalo -> Plot only the cells/particles that belong to the main subhalo.

:sphere -> Plot only the cell/particle inside a sphere with radius FILTER_R (see ./src/constants.jl).

:stellar_subhalo -> Plot only the cells/particles that belong to the main subhalo.

sim_labels::Union{Vector{String},Nothing}=basename.(simulation_paths): Labels for the plot legend, one per simulation. Set it to nothing if you don't want a legend.

GalaxyInspector.stellarHistory — Method

stellarHistory(
    simulation_paths::Vector{String},
    slice_n::Int64,
    quantity::Symbol;
    <keyword arguments>
)::Nothing

Plot the evolution of a given stellar quantity using the stellar ages at a given instant in time.

Arguments

simulation_paths::Vector{String}: Paths to the simulation directories, set in the code variable OutputDir.

slice_n::Int64: Selects which snapshot to plot, starts at 1 and is independent of the number in the file name. If every snapshot is present, slice_n = filename_number + 1.

quantity::Symbol: Quantity for the y axis. The options are:

:sfr -> The star formation rate.

:ssfr -> The specific star formation rate.

:stellar_mass -> Stellar mass.

n_bins::Int64=20: Number of bins (time intervals).

output_path::String="./": Path to the output folder.

filter_mode::Symbol=:all: Which cells/particles will be plotted, the options are:

:all -> Plot every cell/particle within the simulation box.

:halo -> Plot only the cells/particles that belong to the main halo.

:subhalo -> Plot only the cells/particles that belong to the main subhalo.

:sphere -> Plot only the cell/particle inside a sphere with radius FILTER_R (see ./src/constants.jl).

:stellar_subhalo -> Plot only the cells/particles that belong to the main subhalo.

sim_labels::Union{Vector{String},Nothing}=basename.(simulation_paths): Labels for the plot legend, one per simulation. Set it to nothing if you don't want a legend.

GalaxyInspector.timeSeries — Method

timeSeries(
    simulation_paths::Vector{String},
    x_quantity::Symbol,
    y_quantity::Symbol;
    <keyword arguments>
)::Nothing

Plot a time series.

Arguments

simulation_paths::Vector{String}: Paths to the simulation directories, set in the code variable OutputDir.

x_quantity::Symbol: Quantity for the x axis. The possibilities are:

:stellar_mass -> Stellar mass.

:gas_mass -> Gas mass.

:dm_mass -> Dark matter mass.

:bh_mass -> Black hole mass.

:molecular_mass -> Molecular hydrogen ($\mathrm{H_2}$) mass.

:atomic_mass -> Atomic hydrogen ($\mathrm{HI}$) mass.

:ionized_mass -> Ionized hydrogen ($\mathrm{HII}$) mass.

:neutral_mass -> Neutral hydrogen ($\mathrm{HI + H_2}$) mass.

:molecular_fraction -> Gas mass fraction of molecular hydrogen.

:atomic_fraction -> Gas mass fraction of atomic hydrogen.

:ionized_fraction -> Gas mass fraction of ionized hydrogen.

:neutral_fraction -> Gas mass fraction of neutral hydrogen.

:stellar_area_density -> Stellar area mass density, for a radius of FILTER_R.

:gas_area_density -> Gas area mass density, for a radius of FILTER_R.

:molecular_area_density -> Molecular hydrogen area mass density, for a radius of FILTER_R.

:atomic_area_density -> Atomic hydrogen area mass density, for a radius of FILTER_R.

:ionized_area_density -> Ionized hydrogen area mass density, for a radius of FILTER_R.

:neutral_area_density -> Neutral hydrogen area mass density, for a radius of FILTER_R.

:sfr_area_density -> Star formation rate area density, for the last AGE_RESOLUTION_ρ and a radius of FILTER_R.

:gas_metallicity -> Mass fraction of all elements above He in the gas (solar units).

:stellar_metallicity -> Mass fraction of all elements above He in the stars (solar units).

:X_gas_abundance -> Gas abundance of element $\mathrm{X}$, as $12 + \log_{10}(\mathrm{X \, / \, H})$. The possibilities are the keys of ElementIndex.

:X_stellar_abundance -> Stellar abundance of element $\mathrm{X}$, as $12 + \log_{10}(\mathrm{X \, / \, H})$. The possibilities are the keys of ElementIndex.

:stellar_specific_am -> Norm of the stellar specific angular momentum.

:gas_specific_am -> Norm of the gas specific angular momentum.

:dm_specific_am -> Norm of the dark matter specific angular momentum.

:sfr -> The star formation rate of the last AGE_RESOLUTION.

:ssfr -> The specific star formation rate of the last AGE_RESOLUTION.

:scale_factor -> Scale factor.

:redshift -> Redshift.

:physical_time -> Physical time since the Big Bang.

:lookback_time -> Physical time left to reach the last snapshot.

y_quantity::Symbol: Quantity for the y axis. The possibilities are:

:stellar_mass -> Stellar mass.

:gas_mass -> Gas mass.

:dm_mass -> Dark matter mass.

:bh_mass -> Black hole mass.

:molecular_mass -> Molecular hydrogen ($\mathrm{H_2}$) mass.

:atomic_mass -> Atomic hydrogen ($\mathrm{HI}$) mass.

:ionized_mass -> Ionized hydrogen ($\mathrm{HII}$) mass.

:neutral_mass -> Neutral hydrogen ($\mathrm{HI + H_2}$) mass.

:molecular_fraction -> Gas mass fraction of molecular hydrogen.

:atomic_fraction -> Gas mass fraction of atomic hydrogen.

:ionized_fraction -> Gas mass fraction of ionized hydrogen.

:neutral_fraction -> Gas mass fraction of neutral hydrogen.

:stellar_area_density -> Stellar area mass density, for a radius of FILTER_R.

:gas_area_density -> Gas area mass density, for a radius of FILTER_R.

:molecular_area_density -> Molecular hydrogen area mass density, for a radius of FILTER_R.

:atomic_area_density -> Atomic hydrogen area mass density, for a radius of FILTER_R.

:ionized_area_density -> Ionized hydrogen area mass density, for a radius of FILTER_R.

:neutral_area_density -> Neutral hydrogen area mass density, for a radius of FILTER_R.

:sfr_area_density -> Star formation rate area density, for the last AGE_RESOLUTION_ρ and a radius of FILTER_R.

:gas_metallicity -> Mass fraction of all elements above He in the gas (solar units).

:stellar_metallicity -> Mass fraction of all elements above He in the stars (solar units).

:X_gas_abundance -> Gas abundance of element $\mathrm{X}$, as $12 + \log_{10}(\mathrm{X \, / \, H})$. The possibilities are the keys of ElementIndex.

:X_stellar_abundance -> Stellar abundance of element $\mathrm{X}$, as $12 + \log_{10}(\mathrm{X \, / \, H})$. The possibilities are the keys of ElementIndex.

:stellar_specific_am -> Norm of the stellar specific angular momentum.

:gas_specific_am -> Norm of the gas specific angular momentum.

:dm_specific_am -> Norm of the dark matter specific angular momentum.

:sfr -> The star formation rate of the last AGE_RESOLUTION.

:ssfr -> The specific star formation rate of the last AGE_RESOLUTION.

:scale_factor -> Scale factor.

:redshift -> Redshift.

:physical_time -> Physical time since the Big Bang.

:lookback_time -> Physical time left to reach the last snapshot.

slice::IndexType=(:): Slice of the simulation, i.e. which snapshots will be read. It can be an integer (a single snapshot), a vector of integers (several snapshots), an UnitRange (e.g. 5:13), an StepRange (e.g. 5:2:13) or (:) (all snapshots). It works over the longest possible list of snapshots among the simulations (grouped by the number in the file names). Out of bounds indices are ignored.

filter_mode::Symbol=:all: Which cells/particles will be plotted, the options are:

:all -> Plot every cell/particle within the simulation box.

:halo -> Plot only the cells/particles that belong to the main halo.

:subhalo -> Plot only the cells/particles that belong to the main subhalo.

:sphere -> Plot only the cell/particle inside a sphere with radius FILTER_R (see ./src/constants.jl).

:stellar_subhalo -> Plot only the cells/particles that belong to the main subhalo.

sim_labels::Union{Vector{String},Nothing}=nothing: Labels for the plot legend, one per simulation. Set it to nothing if you don't want a legend.