Data acquisition functions

These function read the different output files from the simulation (snapshots, FoF catalogs, sfr.txt, etc.), and load the data into memory as dictionaries.

Some of these functions are exported.

GalaxyInspector.blockPresent — Method

blockPresent(block::String, group::HDF5.Group)::Bool

Checks if a given data block exist in a HDF5 group.

Arguments

block::String: Target block. The possibilities are the keys of QUANTITIES.

group::HDF5.Group: HDF5 group.

Returns

If block exist in group.

GalaxyInspector.blockPresent — Method

blockPresent(type_symbol::Symbol, block::String, path::String)::Bool

Checks if a given block exist in a snapshot.

Note

If each snapshot is made of multiple files, I'll check only in the first chunck.

Arguments

type_symbol::Symbol: The cell/particle type of the target block. The possibilities are the keys of PARTICLE_INDEX.

block::String: Target block. The possibilities are the keys of QUANTITIES.

path::String: Path to the snapshot file or folder.

Returns

If block exist in the snapshot.

GalaxyInspector.getBlock — Method

getBlock(path::String, type_symbol::Symbol, block::String)::VecOrMat{<:Number}

Convenience function to directly get the data associated with one block.

Arguments

path::String: Path to the snapshot file or folder.

type_symbol::Symbol: Type of cell/particle. The possibilities are the keys of PARTICLE_INDEX.

block::String: Target block. The possibilities are the keys of QUANTITIES.

Returns

The data for block.

GalaxyInspector.getGroupCatPaths — Method

getGroupCatPaths(simulation_path::String; <keyword arguments>)::Dict{Symbol,Vector{String}}

Find the path and number of every group catalog in simulation_path.

Note

If each group catalog is made of multiple files, the :paths field will have paths to folders, each one containing the sub-files of the corresponding group catalog.

Arguments

simulation_path::String: Path to the simulation directory, set in the code variable OutputDir.

warnings::Bool=true: If a warning will be raised when no group catalog file or folders are found.

Returns

A dictionary with the following shape:

:numbers -> The number that characterize each group catalog.

:paths -> The full path to each group catalog.

GalaxyInspector.getSnapshotPaths — Method

getSnapshotPaths(simulation_path::String; <keyword arguments>)::Dict{Symbol,Vector{String}}

Find the path and number of every snapshot in simulation_path.

Note

If each snapshot is made of multiple files, the :paths field will have paths to folders, each one containing the sub-files of the corresponding snapshot.

Arguments

simulation_path::String: Path to the simulation directory, set in the code variable OutputDir.

warnings::Bool=true: If a warning will be raised when no snapshot files or folders are found.

Returns

A dictionary with the following shape:

:numbers -> The number that characterize each snapshot.

:paths -> The full path to each snapshot.

GalaxyInspector.makeSimulationTable — Method

makeSimulationTable(simulation_path::String; <keyword arguments>)::DataFrame

Construct a dataframe with the path, time stamps and number of each snapshot and group catalog file in simulation_path.

Arguments

simulation_path::String: Path to the simulation directory, set in the code variable OutputDir.

warnings::Bool=true: If a warning will be raised when there are missing files.

Returns

A dataframe with 8 colums:

:ids -> Dataframe index of each snapshot, i.e. if there are 10 snapshots in total it runs from 1 to 10.

:numbers -> Number in the file name of each snapshot.

:scale_factors -> Scale factor of each snapshot.

:redshifts -> Redshift of each snapshot.

:physical_times -> Physical time since the Big Bang of each snapshot.

:lookback_times -> Physical time left to reach the last snapshot.

:snapshot_paths -> Full path to the snapshots.

:groupcat_paths -> Full path to the group catalog files.

GalaxyInspector.readCpuFile — Method

readCpuFile(
    file_path::String,
    targets::Vector{String};
    <keyword arguments>
)::Dict{String,Matrix{Float64}}

Read the cpu.txt file.

For each process in targets a matrix with all the CPU usage data is returned.

Arguments

file_path::String: Path to the cpu.txt file.

targets::Vector{String}: Target processes.

step::Int=1: Step used to traverse the rows.

warnings::Bool=true: If a warning will be given when there are missing targets.

Returns

A dictionary with the following shape:

target process -> matrix with columns:

Time step.

Simulation time (scale factor for cosmological simulations and physical time for non cosmological simulations).

Clock time in seconds.

Clock time as a percentage.

Cumulative clock time in seconds.

Cumulative clock time as a percentage.

GalaxyInspector.readGoupCatBlocks — Method

readGoupCatBlocks(
    file_path::String,
    snapshot_path::String,
    request::Dict{Symbol,Vector{String}};
    <keyword arguments>
)::Dict{Symbol,Dict{String,VecOrMat{<:Number}}}

Read the specified blocks from a group catalog file.

Arguments

file_path::String: Path to the group catalog file.

snapshot_path::String: Path to the corresponding snapshot file or folder. This is needed for unit conversion.

request::Dict{Symbol,Vector{String}}: The blocks to be read. It must have the shape group type -> [block, block, block].

warnings::Bool=true: If a warning will be given when there are missing blocks.

Returns

A dictionary with the following shape: group type -> (block -> data of block).

GalaxyInspector.readGroupCatHeader — Method

readGroupCatHeader(path::Union{String,Missing}; <keyword arguments>)::GroupCatHeader

Read the header of a group catalog in the HDF5 format.

Note

If each group catalog is made of multiple files, I'll read the header on the first one.

Arguments

path::Union{String,Missing}: Path to the group catalog file or folder.

warnings::Bool=true: If a warning will be given when path is missing.

Returns

A GroupCatHeader.

GalaxyInspector.readGroupCatalog — Method

readGroupCatalog(
    path::Union{String,Missing},
    snapshot_path::String,
    request::Dict{Symbol,Vector{String}};
    <keyword arguments>
)::Dict{Symbol,Dict{String,VecOrMat{<:Number}}}

Read the specified blocks from a group catalog file or folder.

Arguments

path::Union{String,Missing}: Path to the group catalog file or folder.

snapshot_path::String: Path to the corresponding snapshot file or folder. This is needed for unit conversion.

request::Dict{Symbol,Vector{String}}: Which blocks will be read. It must have the shape group type -> [block, block, block].

warnings::Bool=true: If a warning will be given when there are missing blocks.

Returns

A dictionary with the following shape: group type -> (block -> data of block).

GalaxyInspector.readSfrFile — Method

readSfrFile(
    file_path::String,
    snap_path::String;
    <keyword arguments>
)::Dict{Int32,VecOrMat{<:Number}}

Read the sfr.txt file.

Arguments

file_path::String: Path to the sfr.txt file.

snapshot_path::String: Path to one snapshot file or folder of the simulation. This is needed for unit conversion.

warnings::Bool=true: If a warning will be given when the sfr.txt file does not have the expected structure.

Returns

A dictionary with the following shape:

1 -> Time or scale factor (internal units).

2 -> Total stellar mass to be formed prior to stochastic sampling (internal units).

3 -> Instantaneous star formation rate of all cells ($\mathrm{M_\odot \, yr^{-1}}$).

4 -> Instantaneous star formation rate of active cells ($\mathrm{M_\odot \, yr^{-1}}$).

5 -> Total mass in stars formed after stochastic sampling (internal units).

6 -> Cumulative stellar mass formed (internal units).

GalaxyInspector.readSnapBlocks — Method

readSnapBlocks(
    file_path::String,
    request::Dict{Symbol,Vector{String}};
    <keyword arguments>
)::Dict{Symbol,Dict{String,VecOrMat{<:Number}}}

Read the specified blocks from a snapshot file.

Arguments

file_path::String: Path to the snapshot file.

request::Dict{Symbol,Vector{String}}: The blocks to be read. It must have the shape cell/particle type -> [block, block, block].

warnings::Bool=true: If a warning will be given when there are missing blocks.

Returns

A dictionary with the following shape: cell/particle type -> (block -> data of block).

GalaxyInspector.readSnapHeader — Method

readSnapHeader(path::String)::SnapshotHeader

Read the header of a snapshot in the HDF5 format.

Note

If each snapshot is made of multiple files, I'll read the header on the first chunck.

Arguments

path::String: Path to the snapshot file or folder.

Returns

A SnapshotHeader structure.

GalaxyInspector.readSnapshot — Method

readSnapshot(
    path::Union{String,Missing},
    request::Dict{Symbol,Vector{String}};
    <keyword arguments>
)::Dict{Symbol,Dict{String,VecOrMat{<:Number}}}

Read the specified blocks from a snapshot file or folder.

Arguments

path::Union{String,Missing}: Path to the snapshot file or folder.

request::Dict{Symbol,Vector{String}}: Which blocks will be read. It must have the shape cell/particle type -> [block, block, block].

warnings::Bool=true: If a warning will be given when some/all the data is missing.

Returns

A dictionary with the following shape: cell/particle type -> (block -> data of block).

GalaxyInspector.readTemperature — Method

readTemperature(file_path::String)::Vector{<:Unitful.Temperature}

Compute the temperature of the gas cells in a snapshot.

Arguments

file_path::String: Path to the snapshot file.

Returns

The temperature of the gas cells.

GalaxyInspector.readTime — Method

readTime(path::String)::Float64

Read the "Time" field in the header of a snapshot file.

Note

If each snapshot is made of multiple files, I'll read the header on the first chunck.

Arguments

path::String: Path to the snapshot file or folder.

Returns

The "Time" field in the header (for cosmological simulations is the scale factor).