General utilities

These functions are only used internally, and are general in the sense that do not depend on the specifics of Arepo.

None of these functions are exported.

Base.isempty — Method

Extension of Base.isempty to check for empty LaTeXStrings.

GalaxyInspector.absCoor — Method

absCoor(
    plot::Union{Makie.FigureAxisPlot,Makie.Axis,Makie.Figure},
    r_x::Real,
    r_y::Real,
)::NTuple{2,Float64}

Compute the absolute x and y coordinates of a plot, from the relative ones.

Arguments

plot::Union{Makie.FigureAxisPlot,Makie.Axis,Makie.Figure}: Plot, axis, or figure for which the absolute coordinates will be calculated. In the case of a figure, it will use the limits from the current axis object.

r_x::Real: Relative x coordinate, rx ∈ [0, 1].

r_y::Real: Relative y coordinate, ry ∈ [0, 1].

Returns

A tuple with the absolute coordinates, (ax, ay).

Examples

julia> absCoor(lines(rand(100)), 0.5, 0.5)
(50.50000071525574, 0.48792968317866325)

GalaxyInspector.area — Method

Area of a circle with radius r.

GalaxyInspector.cleanPlot! — Method

Default function to end cleanPlot! recursion if an unknown type is encountered.

GalaxyInspector.cleanPlot! — Method

cleanPlot!(ax::Makie.Axis)::Bool

Empty an axis.

Arguments

ax::Makie.Axis: Axis to be emptied.

Returns

Flag to indicate that an axis has been emptied.

GalaxyInspector.cleanPlot! — Method

cleanPlot!(figure::Makie.Figure)::Nothing

Delete all the legends of a figure and empty all its axes.

Arguments

figure::Makie.Figure: Figure to be cleaned.

GalaxyInspector.cleanPlot! — Method

cleanPlot!(legend::Makie.Legend)::Bool

Delete a legend.

Arguments

legend::Makie.Legend: Legend to be deleted.

Returns

Flag to indicate that a legend has been deleted.

GalaxyInspector.cubicSplineKernel — Method

cubicSplineKernel(q::Real, h::Number)::Number

2D cubic spline kernel.

Arguments

q::Real: Relative distance to the neighbor, $|r - r'| / h$.

h::Number: Smoothing length.

Returns

The kernel function evaluated at a separation q * h, and with a smoothing length h.

References

PySPH documentation

J. J. Monaghan (1992). Smoothed Particle Hydrodynamics. Annual Review of Astronomy and Astrophysics, 30, 543-574. doi:10.1146/annurev.aa.30.090192.002551

M.B. Liu et al. (2010). Smoothed Particle Hydrodynamics (SPH): an Overview and Recent Developments. Archives of Computational Methods in Engineering, 17, 25–76. doi:10.1007/s11831-010-9040-7

GalaxyInspector.getEmpty — Method

Always returns an empty vector, for any type and number of arguments.

GalaxyInspector.getNothing — Method

Always returns nothing, for any type and number of arguments.

GalaxyInspector.histogram1D — Method

histogram1D(
    positions::Vector{<:Number},
    grid::Union{LinearGrid,CircularGrid},
)::Vector{Int64}

Compute a 1D histogram of positions.

Arguments

positions::Vector{<:Number}: Values for which the histogram will be constructed.

grid::Union{LinearGrid,CircularGrid}: A linear or circular grid.

Returns

A vector with the counts.

GalaxyInspector.histogram1D — Method

histogram1D(
    positions::Vector{<:Number},
    values::Vector{<:Number},
    grid::Union{LinearGrid,CircularGrid};
    <keyword arguments>
)::Vector{<:Number}

Compute a 1D histogram of values.

Arguments

positions::Vector{<:Number}: Positions of the values within a 1D axis. This determines to which bin each value will be added.

values::Vector{<:Number}: The values that will be added up in each bin, according to their positions.

grid::Union{LinearGrid,CircularGrid}: A linear or circular grid.

total::Bool=true: If the sum (default) or the mean of values will be computed for each bin.

empty_nan::Bool=true: If NaN will be put into empty bins, 0 is used otherwise.

Returns

A vector with the histogram values.

GalaxyInspector.histogram1D — Method

histogram1D(
    positions::Vector{<:Number},
    values::Vector{<:Number},
    edges::Vector{<:Number};
    <keyword arguments>
)::Vector{<:Number}

Compute a 1D histogram of values.

Arguments

positions::Vector{<:Number}: Positions of the values within a 1D axis. This determines to which bin each value will be added.

values::Vector{<:Number}: The values that will be added up in each bin, according to their positions.

edges::Vector{<:Number}: A sorted list of bin edges.

total::Bool=true: If the sum (default) or the mean of values will be computed for each bin.

empty_nan::Bool=true: If NaN will be put into empty bins, 0 is used otherwise.

Returns

A vector with the histogram values.

GalaxyInspector.histogram1D — Method

histogram1D(positions::Vector{<:Number}, edges::Vector{<:Number})::Vector{Int64}

Compute a 1D histogram of positions.

Arguments

positions::Vector{<:Number}: Values for which the histogram will be constructed.

edges::Vector{<:Number}: A sorted list of bin edges.

Returns

A vector with the counts.

GalaxyInspector.histogram2D — Method

histogram2D(positions::Matrix{<:Number}, grid::SquareGrid)::Matrix{Int64}

Compute a 2D histogram of positions.

Arguments

positions::Matrix{<:Number}: Values for which the histogram will be constructed.

grid::SquareGrid: A square grid.

Returns

A matrix with the counts.

GalaxyInspector.histogram2D — Method

histogram2D(
    positions::Matrix{<:Number},
    values::Vector{<:Number},
    grid::SquareGrid;
    <keyword arguments>
)::Matrix{<:Number}

Compute a 2D histogram of values.

Arguments

positions::Matrix{<:Number}: Positions of the values in the grid. Each column correspond to a value and each row is a dimension. This determines to which bin each value will be added.

values::Vector{<:Number}: The values that will be added up in each square bin, according to their positions.

grid::SquareGrid: A square grid.

total::Bool=true: If the sum (default) or the mean of values will be computed in each bin.

empty_nan::Bool=true: If NaN will be put into empty bins, 0 is used otherwise.

Returns

A matrix with the histogram values.

GalaxyInspector.histogram2D — Method

histogram2D(
    positions::Matrix{<:Number},
    values::Vector{<:Number},
    x_edges::Vector{<:Number},
    y_edges::Vector{<:Number};
    <keyword arguments>
)::Matrix{<:Number}

Compute a 2D histogram of values.

Arguments

positions::Matrix{<:Number}: Positions of the values in the grid. Each column correspond to a value and each row is a dimension. This determines to which bin each value will be added.

values::Vector{<:Number}: The values that will be added up in each square bin, according to their positions.

x_edges::Vector{<:Number}: A sorted list of bin edges for the x axis.

y_edges::Vector{<:Number}: A sorted list of bin edges for the y axis.

total::Bool=true: If the sum (default) or the mean of values will be computed in each bin.

empty_nan::Bool=true: If NaN will be put into empty bins, 0 is used otherwise.

Returns

A matrix with the histogram values.

GalaxyInspector.histogram2D — Method

histogram2D(
    positions::Matrix{<:Number},
    x_edges::Vector{<:Number},
    y_edges::Vector{<:Number},
)::Matrix{Int64}

Compute a 2D histogram of positions.

Arguments

positions::Matrix{<:Number}: Values for which the histogram will be constructed.

grid::SquareGrid: A square grid.

x_edges::Vector{<:Number}: A sorted list of bin edges for the x axis.

y_edges::Vector{<:Number}: A sorted list of bin edges for the y axis.

Returns

A matrix with the counts.

GalaxyInspector.isPositive — Method

Test for strict positivity.

GalaxyInspector.pointData — Method

Extract all the data points in a Makie plot, axis, or figure. In the case of a figure, it will only take the data from the current axis object. It only works for scatter, line and scatterline plots.

GalaxyInspector.rangeCut! — Method

rangeCut!(
    data::Vector{<:Number},  
    range::Tuple{<:Number,<:Number};
    <keyword arguments>
)::Bool

Delete every element in data that is outside the given range.

Arguments

data::Vector{<:Number}: Dataset that will be pruned.

range::Tuple{<:Number,<:Number}: The range in question.

keep_edges::Bool=true: If the edges of the range will be kept.

min_left::Int64=0: Minimum number of values that need to be left after pruning to proceed with the transformation.

Returns

If a transformation was performed.

GalaxyInspector.rangeCut! — Method

rangeCut!(
    m_data::Vector{<:Number}, 
    s_data::Vector, 
    range::Tuple{<:Number,<:Number};
    <keyword arguments>
)::Bool

Delete every element in m_data that is outside the given range.

Every corresponding element in s_data (i.e. with the same index) will be deleted too.

Arguments

m_data::Vector{<:Number}: Master dataset that will be pruned.

s_data::Vector: Slave dataset that will be pruned according to which values of m_data are outside range.

range::Tuple{<:Number,<:Number}: The range in question.

keep_edges::Bool=true: If the edges of the range will be kept.

min_left::Int64=0: Minimum number of values that need to be left in the master dataset after pruning to proceed with the transformation.

Returns

If a transformation was performed.

GalaxyInspector.ring — Method

ring(vec::Vector, index::Integer)

Make the indexing operation vec[index] work using modular arithmetic for the indices.

Arguments

vec::Vector: Vector.

index::Integer: Index.

Returns

vec[mod1(index, length(vec))]

Examples

julia> ring([1, 2, 3], 11)
2

julia> ring([1, 2, 3], 3)
3

julia> ring([1, 2, 3], -5)
1

GalaxyInspector.safeSelect — Method

safeSelect(vec::Vector, index::IndexType; warnings::Bool=true)

Make the indexing operation vec[index] ignore indices that are out of bounds.

Arguments

vec::Vector: Vector.

index::IndexType: Indices. It can be an integer (a single element), a vector of integers (several elements), an UnitRange (e.g. 5:13), an StepRange (e.g. 5:2:13) or (:) (every element).

warnings::Bool=true: If a warning will be given when index has out of bounds indices.

Returns

vec[index (minus out of bounds indices)]

Examples

julia> safeSelect([1, 2, 3], 11; warnings=false)
Int64[]

julia> safeSelect([1, 2, 3], 1:5; warnings=false)
3-element Vector{Int64}:
 1
 2
 3

julia> safeSelect([1, 2, 3], 1:3:10; warnings=false)
1

julia> safeSelect([1, 2, 3], [1, 2, 5, 9]; warnings=false)
2-element Vector{Int64}:
 1
 2

julia> safeSelect([1, 2, 3], (:); warnings=false)
3-element Vector{Int64}:
 1
 2
 3

GalaxyInspector.sanitizeData! — Method

sanitizeData!(
    x_data::Vector{<:Number},
    y_data::Vector{<:Number};
    <keyword arguments> 
)::NTuple{4,Bool}

Do the following transformations over x_data and y_data, in order:

Trim them to fit within the domain of the functions func_domain[1] and func_domain[2], respectively.

Trim them to fit within range[1] and range[2], respectively.

Scale them down by a factor 10^exp_factor[1] and 10^exp_factor[2], respectively.

By default, no transformation is done.

Note

The datasets must have the same length, and any operation that deletes an element, will delete the corresponding element (i.e. with the same index) in the other dataset, so that the dataset will remain of equal length.

Arguments

x_data::Vector{<:Number}: First dataset to be sanitized.

y_data::Vector{<:Number}: Second dataset to be sanitized.

func_domain::NTuple{2,Function}=(identity, identity): x_data will be trimmed to fit within the domain of the function func_domain[1], and y_data will be trimmed to fit within the domain of the function func_domain[2]. The options are the scaling functions accepted by Makie: log10, log2, log, sqrt, Makie.logit, Makie.Symlog10, Makie.pseudolog10, and identity.

range::Tuple{Tuple{<:Number,<:Number},Tuple{<:Number,<:Number}}=((-Inf, Inf), (-Inf, Inf)): Every element in x_data that falls outside of range[1] will be deleted, and every element in y_data that falls outside of range[2] will be deleted.

keep_edges::NTuple{2,Bool}=(true, true): If the edges of each corresponding range will be kept.

min_left::Int64=0: Minimum number of values that need to be left in each dataset after any of the transformations to procced with them.

exp_factor::NTuple{2,Int64}=(0, 0): Every element in x_data will be divided by 10^exp_factor[1], and every element in y_data will be divided by 10^exp_factor[2].

warnings::Bool=true: If a warning will be given when any of the datasets is a vector of Integers, which may cause wrong results when dividing by 10^exp_factor.

Returns

A tuple with four flags:

If x_data was successfully modified to fit within the domain of func_domain[1].

If y_data was successfully modified to fit within the domain of func_domain[2].

If x_data was successfully modified to fit within range[1].

If y_data was successfully modified to fit within range[2].

GalaxyInspector.sanitizeData! — Method

sanitizeData!(
    data::Vector{<:Number};
    <keyword arguments> 
)::NTuple{2,Bool}

Do the following transformations over data, in order:

Trim it to fit within the domain of the function func_domain.

Trim it to fit within range.

Scale it down by a factor of 10^exp_factor.

By default, no transformation is done.

Arguments

data::Vector{<:Number}: Dataset to be sanitized.

func_domain::Function=identity: data will be trimmed to fit within the domain of the function func_domain. The options are the scaling functions accepted by Makie: log10, log2, log, sqrt, Makie.logit, Makie.Symlog10, Makie.pseudolog10, and identity.

range::Tuple{<:Number,<:Number}=(-Inf, Inf): Every element in data that falls outside of range will be deleted.

keep_edges::Bool=true: If the edges of range will be kept.

min_left::Int64=0: Minimum number of values that need to be left after each transformation to procced with it.

exp_factor::Int64=0: Every element in data will be divided by 10^exp_factor.

warnings::Bool=true: If a warning will be given when data is a vector of Integers, which may cause wrong results when dividing by 10^exp_factor.

Returns

A tuple with two flags:

If data was mutated to fit within the domain of func_domain.

If data was mutated to fit within range.

GalaxyInspector.scaledBins — Method

scaledBins(
    values::Vector{<:Number}, 
    n_bins::Int64; 
    <keyword arguments>
)::Vector{Float64}

Compute a set of bin edges, for a given list values.

Arguments

values::Vector{<:Number}: Values to be binned.

n_bins::Int64: Number of bins.

scaling::Function=identity: Scaling function. The options are the scaling functions accepted by Makie: log10, log2, log, sqrt, Makie.logit, Makie.Symlog10, Makie.pseudolog10, and identity.

limits::Tuple{<:Number,<:Number}=(-Inf, Inf): Set it to a value different than Inf if you want to fix the limits of the binning.

Returns

A sorted list of bin edges.

GalaxyInspector.setPositive — Method

Create a copy of list with every negative value set to 0.

GalaxyInspector.smoothWindow — Method

smoothWindow(
    x_data::Vector{<:Number},
    y_data::Vector{<:Number},
    n_bins::Int64;
    <keyword arguments>
)::NTuple{2,Vector{<:Number}}

Separate the values of x_data in n_bins bins and compute the mean value of x_data and y_data within each one.

Arguments

x_data::Vector{<:Number}: x-axis data.

y_data::Vector{<:Number}: y-axis data.

n_bins::Int64: Number of bins.

scaling::Function=identity: Scaling function for the x axis. The options are the scaling functions accepted by Makie: log10, log2, log, sqrt, Makie.logit, Makie.Symlog10, Makie.pseudolog10, and identity. All the values of x_data must be in the domain of scaling.

Returns

A tuple with two vectors, containing the smoothed-out x and y values.

GalaxyInspector.volume — Method

Volume of a sphere with radius r.

GalaxyInspector.xlimits! — Method

Extract the limits of the x axis, from a Makie plot, axis, or figure. In the case of a figure, it will take the limits from the current axis object.

GalaxyInspector.xscale — Method

Extract the scale function of the x axis, from a Makie plot, axis, or figure. In the case of a figure, it will take the scale from the current axis object.

GalaxyInspector.ylimits! — Method

Extract the limits of the y axis, from a Makie plot, axis, or figure. In the case of a figure, it will take the limits from the current axis object.

GalaxyInspector.yscale — Method

Extract the scale function of the y axis, from a Makie plot, axis, or figure. In the case of a figure, it will take the scale from the current axis object.