Configuration constants

This is a list of constant used to configure GalaxyInspector.

None of these are exported.

GalaxyInspector.ABUNDANCE_SHIFT — Constant

Shift for the solar abundance of each element in dex.

GalaxyInspector.AGE_RESOLUTION — Constant

Characteristic stellar age for the SFR and sSFR.

GalaxyInspector.BOX_L — Constant

Characteristic plotting box size.

GalaxyInspector.BULGE_R — Constant

Characteristic bulge radius (for R25 calculation).

GalaxyInspector.CONFIG_SCHEMA — Constant

List of config keys, their corresponding global variables, and parsing functions.

GalaxyInspector.DISK_HEIGHT — Constant

Characteristic disk height.

GalaxyInspector.DISK_R — Constant

Characteristic disk radius.

GalaxyInspector.GAMMA — Constant

Adiabatic index.

GalaxyInspector.GC_BASENAME — Constant

Base name of the group catalog files.

GalaxyInspector.HUBBLE_CONSTANT — Constant

Hubble constant in $\mathrm{Gyr^{-1}}$.

This value corresponds to H_0 = 100 \, \mathrm{km} \, \mathrm{s}^{-1} \, \mathrm{Mpc}^{-1} = 0.102201 \, \mathrm{Gyr}^{-1} = $.

GalaxyInspector.HYDROGEN_MASSFRAC — Constant

Mass fraction of hydrogen.

GalaxyInspector.INTERNAL_L_UNIT — Constant

Internal unit of length

The default value is the internal unit of length used in IllustrisTNG, equal to $1.0 \, \mathrm{kpc}$.

GalaxyInspector.INTERNAL_M_UNIT — Constant

Internal unit of mass

The default value is the internal unit of mass used in IllustrisTNG, equal to $10^{10} \, \mathrm{M_\odot}$.

GalaxyInspector.INTERNAL_V_UNIT — Constant

Internal unit of velocity

The default value is the internal unit of velocity used in IllustrisTNG, equal to $1.0 \, \mathrm{km \, s^{-1}}$.

GalaxyInspector.MMAP_MEMORY_FRACTION — Constant

Maximum fraction of the free physical memory that an array can occupy before we start using memory mapping.

GalaxyInspector.PHYSICAL_UNITS — Constant

If physical lengths units will be used, instead of comoving lengths units.

GalaxyInspector.ROTATION_R — Constant

Characteristic radius for rotations.

GalaxyInspector.SNAP_BASENAME — Constant

Base name of the snapshot files, set in the code variable SnapshotFileBase.

GalaxyInspector.SOLAR_METALLICITY — Constant

Solar metallicity, as used in Arepo.

References

M. Asplund et al. (2006). The new solar abundances - Part I: the observations. Communications in Asteroseismology, 147. doi:10.1553/cia147s76

GalaxyInspector.TRACER_MASS — Constant

Mass of the tracers in internal code units. Its value comes from All.TargetGasMass = All.TargetGasMassFactor * All.ReferenceGasPartMass in the code.

It is only printed in the output files (stdout_n), as All.TargetGasMass=XXXXX

GalaxyInspector.parseAbundanceShift — Method

parseAbundanceShift(d::Dict{String,Any})::Dict{Symbol,Float64}

Parse the abundance shift dictionary from the config file, merging it with the default values.

If a key in d is not in the defaults, it will be ignored.

Arguments

d::Dict{String,Any}: The dictionary from the config file.

Returns

- The result of merging `d` with the default values of [`ABUNDANCE_SHIFT`](@ref).

GalaxyInspector.parseUnitful — Method

parseUnitful(s::String)::Unitful.Quantity

Parse a string of the form "40.0 kpc" into a Unitful quantity.

Note

Works for any unit string that Unitful.uparse accepts (e.g. "kpc", "km/s", "K"). The string must have a space between the value and the unit, otherwise it will not be parsed correctly (e.g. "40.0kpc" will not work, but "40.0 kpc" will). And the unit should have no spaces (e.g. "km/s" is correct, but "km / s" is not). This is because the function splits the string into two parts using whitespace as a delimiter, and then parses the first part as a number and the second part as a unit.

Arguments

s::String: The string to parse, expected to be of the form "<value> <unit>", e.g. "40.0 kpc".

Returns

The parsed quantity, e.g. 40.0u"kpc".