Global constants

This is a list of constant and types used in GalaxyInspector.

None of these are exported.

GalaxyInspector.ABUNDANCE_QUANTITIES — Constant

List of quantities related to chemical abundance.

GalaxyInspector.ALPHA_BLITZ — Constant

Reference exponent for the molecular fraction-pressure relation, from Blitz et al. (2006) (Table 2, "Mean" row, Second column).

We use -α here.

References

L. Blitz et al. (2006). The Role of Pressure in GMC Formation II: The H2-Pressure Relation. The Astrophysical Journal, 650(2), 933. doi:10.1086/505417

GalaxyInspector.AREPO_SFM — Constant

Star formation model parameters

GalaxyInspector.ATOMIC_WEIGHTS — Constant

Standard atomic weights.

References

T. Prohaska et al. (2022). Standard atomic weights of the elements 2021 (IUPAC Technical Report). Pure and Applied Chemistry, 94(5), 573-600. doi:10.1515/pac-2019-0603

GalaxyInspector.A_BIGIEL2008_BF_MOLECULAR — Constant

Kennicutt-Schmidt law best-fit for molecular gas, from Bigiel et al. (2008) (Section 4.3, Equation 3).

Power-law index, N, and $A = \log_{10}(a)$, where $a$ is $\Sigma_\mathrm{SFR}$ at the fiducial gas surface density of $10 \, \mathrm{M_\odot \, pc^{-2}}$ are given.

\[\Sigma_\mathrm{SFR} = a \left( \frac{\Sigma_\mathrm{H_2}}{10 \, \mathrm{M_\odot \, pc^{-2}}} \right)^{\!N} \, ,\]

References

F. Bigiel et al. (2008). THE STAR FORMATION LAW IN NEARBY GALAXIES ON SUB-KPC SCALES. The Astrophysical Journal, 136(6), 2846. doi:10.1088/0004-6256/136/6/2846

GalaxyInspector.A_BIGIEL2008_MOLECULAR — Constant

Kennicutt-Schmidt law fits for molecular and neutral gas, from Bigiel et al. (2008) (Table 2, Average).

Power-law index, N, and $A = \log_{10}(a)$, where $a$ is $\Sigma_\mathrm{SFR}$ at the fiducial gas surface density of $10 \, \mathrm{M_\odot \, pc^{-2}}$ are given.

\[\Sigma_\mathrm{SFR} = a \left( \frac{\Sigma_\mathrm{HI, H_2, gas}}{10 \, \mathrm{M_\odot \, pc^{-2}}} \right)^{\!N} \, ,\]

References

F. Bigiel et al. (2008). THE STAR FORMATION LAW IN NEARBY GALAXIES ON SUB-KPC SCALES. The Astrophysical Journal, 136(6), 2846. doi:10.1088/0004-6256/136/6/2846

GalaxyInspector.A_COLOGNI2026 — Constant

Kennicutt-Schmidt law best-fit for molecular gas, from Cologni et al. (2026) (Section 4.3.2).

Power-law index, N, and $A = \log_{10}(a)$, where $a$ is $\Sigma_\mathrm{SFR}$ at the fiducial molecular gas surface density of $1.0 \mathrm{M_\odot \, kpc^{-2}}$ are given.

\[\Sigma_\mathrm{SFR} = a \left( \frac{\Sigma_\mathrm{H_2}}{1.0 \, \mathrm{M_\odot \, kpc^{-2}}} \right)^{\!N} \, ,\]

References

R. Cologni et al. (2026). Resolved molecular gas and star formation in massive unquenched spirals. Astronomy and Astrophysics, 709, A148. doi:10.1051/0004-6361/202558486

GalaxyInspector.A_GIANNETTI2017 — Constant

Best fit for the gas-to-dust ratio vs. galactocentric distance, from Giannetti et al. (2017) (Section 3, Equation 2).

\[\log_{10} \, \gamma = A \, R_\text{GC} + B \, ,\]

References

A. Giannetti et al. (2017). Galactocentric variation of the gas-to-dust ratio and its relation with metallicity. Astronomy and Astrophysics, 606, L12. doi:10.1051/0004-6361/201731728

GalaxyInspector.A_LIN2019 — Constant

Kennicutt-Schmidt law best-fit for molecular gas, from Lin et al. (2019) (Section 3.1, Table 1).

Power-law index, N, and $A = \log_{10}(a)$, where $a$ is $\Sigma_\mathrm{SFR}$ at the fiducial molecular gas surface density of $1.0 \mathrm{M_\odot \, kpc^{-2}}$ are given.

\[\Sigma_\mathrm{SFR} = a \left( \frac{\Sigma_\mathrm{H_2}}{1.0 \, \mathrm{M_\odot \, kpc^{-2}}} \right)^{\!N} \, ,\]

References

L. Lin et al. (2019). The ALMaQUEST Survey: The Molecular Gas Main Sequence and the Origin of the Star-forming Main Sequence. The Astrophysical Journal Letters, 884(2), L33. doi:10.3847/2041-8213/ab4815

GalaxyInspector.A_QUEREJETA2021 — Constant

Kennicutt-Schmidt law best-fit for molecular gas, from Querejeta et al. (2021) (Section 4.3, Table 4, row All, mean of the columns).

Power-law index, N, and $A = \log_{10}(a)$, where $a$ is $\Sigma_\mathrm{SFR}$ at the fiducial molecular gas surface density of $1.0 \mathrm{M_\odot \, pc^{-2}}$ are given.

\[\Sigma_\mathrm{SFR} = a \left( \frac{\Sigma_\mathrm{H_2}}{1.0 \, \mathrm{M_\odot \, pc^{-2}}} \right)^{\!N} \, ,\]

References

M. Querejeta et al. (2021). Stellar structures, molecular gas, and star formation across the PHANGS sample of nearby galaxies. Astronomy and Astrophysics, 656, A133. doi:10.1051/0004-6361/202140695

GalaxyInspector.AλAV_g — Constant

Relation between the extinction in the g, r, and i bands of SDSS and the extinction in the V band from Table 3 of S. Wang et al. (2019).

\[\left. A_\lambda / A_V \right\vert_g = 1.205 \pm 0.010 \, .\]

\[\left. A_\lambda / A_V \right\vert_r = 0.848 \pm 0.006 \, .\]

\[\left. A_\lambda / A_V \right\vert_i = 0.630 \pm 0.004 \, .\]

S. Wang et al. (2019). The Optical to Mid-infrared Extinction Law Based on the APOGEE, Gaia DR2, Pan-STARRS1, SDSS, APASS, 2MASS, and WISE Surveys. The American Astronomical Society 877(2), 116. doi:10.3847/1538-4357/ab1c61

GalaxyInspector.BIGIEL2010_TABLE_2 — Constant

Path to Table 2 of Bigiel et al. (2010).

References

F. Bigiel et al. (2010). EXTREMELY INEFFICIENT STAR FORMATION IN THE OUTER DISKS OF NEARBY GALAXIES. The Astrophysical Journal, 140(5), 1194. doi:10.1088/0004-6256/140/5/1194

GalaxyInspector.BIGIEL2010_TABLE_3 — Constant

Path to Table 3 of Bigiel et al. (2010).

References

F. Bigiel et al. (2010). EXTREMELY INEFFICIENT STAR FORMATION IN THE OUTER DISKS OF NEARBY GALAXIES. The Astrophysical Journal, 140(5), 1194. doi:10.1088/0004-6256/140/5/1194

GalaxyInspector.COMPONENTS — Constant

List of physical components of a simulation.

GalaxyInspector.COSMO_THRESHOLD_DENSITY — Constant

Cosmological threshold density above which the gas cells/particles can turn into stars.

This value corresponds to CritOverDensity $= 57.7 \, [\mathrm{cm^{-3}}]$ in the param.txt file (used only in cosmological simulations). Which is converted to internal units within the code using OverDensThresh = CritOverDensity * OmegaBaryon * 3 * Hubble * Hubble / (8 * M_PI * G). Then, to go to physical units again one has to do:OverDensThresh*UnitDensityincgs*cf_a3inv*HubbleParam*HubbleParam`.

Using the unit factors,

UnitLength_in_cm = $3.085678 \times 10^{24}$

UnitMass_in_g = $1.989 \times 10^{43}$

UnitVelocity_in_cm_per_s = $100000$

The derived units,

UnitTime_in_s = UnitLength_in_cm * UnitVelocity_in_cm_per_s^-1 = $3.08568 \times 10^{19}$

UnitDensity_in_cgs = UnitMass_in_g * UnitLength_in_cm^-3 = $6.76991 \times 10^{-31}$

The parameters,

OmegaBaryon = $0.048$

HubbleParam = $0.6777$

PROTONMASS = $1.67262178 \times 10^{-24}$

HYDROGEN_MASSFRAC = $0.76$

GRAVITY = $6.6738 \times 10^{-8}$

HUBBLE = $3.2407789 \times 10^{-18}$

And the derived parameters,

Hubble = HUBBLE * UnitTime_in_s = $100$

G = GRAVITY * UnitLength_in_cm^-3 * UnitMass_in_g * UnitTime_in_s^2 = $43.0187$

One gets,

OverDensThresh = 76.8495 [internal units of density]

And, for a cosmological simulation at redshift 0 (cf_a3inv = 1), this result in a physical density threshold of $1.42857 \times 10^{-5} \, [\mathrm{cm^{-3}}]$, or, adding the proton mass, a value of

$\log_{10} \rho \ [\mathrm{M_\odot \, kpc^{-3}}] = 2.548$

GalaxyInspector.CPU_REL_PATH — Constant

Relative path, within the simulation directory, of cpu.txt.

GalaxyInspector.CYCLE — Constant

Default cycler.

GalaxyInspector.ColorType — Type

Type for colors.

GalaxyInspector.DEFAULT_THEME — Constant

Default plot theme.

Regarding the graphic units used, we know that $1 \, \mathrm{mm} = 2.83466 \, \mathrm{pt}$ and $1 \, \mathrm{in} = 25.4 \, \mathrm{mm}$. Then, if we want $1 \, \mathrm{[code\,\,]unit} = 0.1 \, \mathrm{mm}$ in vector graphics, we have to use pt_per_unit = 0.283466.

For pixel images, we control the ppi with px_per_unit. A reasonable high ppi is 600. So, using px_per_unit = $2.3622$ we get $23.622 \, \mathrm{px/mm} \sim 600 \, \mathrm{px/in}$ (remember that $1 \, \mathrm{[code\,\,]unit} = 0.1 \, \mathrm{mm}$).

GalaxyInspector.DELOSREYES2019_DATA_PATH — Constant

Path to Table 4 (corrected) from de los Reyes et al. (2019).

References

M. A. C. de los Reyes et al. (2019). Revisiting the Integrated Star Formation Law. I. Non-starbursting Galaxies. The Astrophysical Journal, 872(1), 16. doi:10.3847/1538-4357/aafa82)

M. A. C. de los Reyes et al. (2019). Erratum: “Revisiting the Integrated Star Formation Law. I. Non-starbursting Galaxies” (2019 ApJ, 872, 16). The Astrophysical Journal, 878(1), 74. doi:10.3847/1538-4357/ab22af)

GalaxyInspector.ELEMENT_INDEX — Constant

Code index for each tracked element.

GalaxyInspector.EVANS2014_ϵff — Constant

Median star formation efficiency per free-fall time, from Evans et al. (2014) (Section 1.1).

N. J. Evans II et al. (2014). STAR FORMATION RELATIONS IN NEARBY MOLECULAR CLOUDS. The Astrophysical Journal, 782(2), 114. doi:10.1088/0004-637X/782/2/114

GalaxyInspector.EXTINCTION_FACTOR_D — Constant

Extinction factor from Nozawa et al. (2013) eq. 12.

\[\text{Extinction factor}^{-1} = \frac{A(V)}{\Sigma_\text{dust}} = (3.7 \pm 0.5) \times 10^{4} \, \mathrm{g^{-1} \, cm^2 \, mag} \, .\]

T. Nozawa et al. (2013). PROPERTIES OF DUST GRAINS PROBED WITH EXTINCTION CURVES. The Astrophysical Journal 770(1), 27. doi:10.1088/0004-637X/770/1/27

GalaxyInspector.EXTINCTION_FACTOR_NH — Constant

Extinction factor from T. Güver et al. (2009).

\[\text{Extinction factor} = \frac{N_\mathrm{H}}{A(V)} = (2.21 \pm 0.09) \times 10^{21} \, \mathrm{atoms \, cm^{-2} \, mag^{-1}} \, .\]

We assume that an atom mass is the proton mass.

T. Güver et al. (2009). The relation between optical extinction and hydrogen column density in the Galaxy. Monthly Notices of the Royal Astronomical Society 400(4), 2050-2053. doi:10.1111/j.1365-2966.2009.15598.x

GalaxyInspector.FELDMANN2020_DATA_PATH — Constant

Path to the file with the global galactic properties from Feldmann (2020).

References

R. Feldmann (2020). The link between star formation and gas in nearby galaxies. Communications Physics 3(226). doi:10.1038/s42005-020-00493-0

GalaxyInspector.FloatOrQuantity — Type

Type for numbers.

GalaxyInspector.HALO_QUANTITIES — Constant

List of halo quantities.

GalaxyInspector.HESTIA_SUBHALOS — Constant

Subhalo numbers for the MW and M31 in Hestia simulations.

GalaxyInspector.INDEX_PARTICLE — Constant

Type of cell/particle corresponding to each code index.

GalaxyInspector.ISOLATED_PARTICLE_NAMES — Constant

Human readable name for each type of cell/particle.

GalaxyInspector.IndexType — Type

Index type.

GalaxyInspector.LEE2016_ϵff — Constant

Median star formation efficiency per free-fall time, from Lee et al. (2016) (Table 3 and eq. 15).

E. J. Lee et al. (2016). OBSERVATIONAL EVIDENCE OF DYNAMIC STAR FORMATION RATE IN MILKY WAY GIANT MOLECULAR CLOUDS. The Astrophysical Journal, 833(2), 229. doi:10.3847/1538-4357/833/2/229

GalaxyInspector.LEROY2008_DATA_PATH — Constant

Path to the file with the profiles from Leroy et al. (2008).

References

A. K. Leroy et al. (2008). THE STAR FORMATION EFFICIENCY IN NEARBY GALAXIES: MEASURING WHERE GAS FORMS STARS EFFECTIVELY. The Astronomical Journal 136(6), 2782–2845. doi:10.1088/0004-6256/136/6/2782

GalaxyInspector.LINE_STYLES — Constant

Default line styles.

GalaxyInspector.LOGGING — Constant

If logging messages will be printed out.

GalaxyInspector.LOG_STREAM — Constant

IO stream for logging messages.

GalaxyInspector.LONG_PARTICLE_INDEX — Constant

Code index for each type of cell/particle, in C zero-based numbering

References

See for example Gadget2 User's Guide, or Gadget4 documentation.

GalaxyInspector.LONG_PARTICLE_NAMES — Constant

Human readable name for each type of cell/particle.

GalaxyInspector.LineStyleType — Type

Type for line styles.

GalaxyInspector.MAGNITUDES — Constant

List of physical magnitudes of a simulation.

GalaxyInspector.MARKERS — Constant

Default markers.

GalaxyInspector.MCMILLAN2011_DATA_PATH — Constant

Path to the file with the fits from McMillan (2011).

References

P. J. McMillan (2011). Mass models of the Milky Way. Monthly Notices of the Royal Astronomical Society 414(3), 2446-2457. doi:10.1111/j.1365-2966.2011.18564.x

GalaxyInspector.METAL_LIST — Constant

List of element above helium.

GalaxyInspector.METAL_LIST_IDX — Constant

List of element indices above helium.

GalaxyInspector.MILLANIRIGOYEN2025_DATA_PATH — Constant

Path to the file with the stellar magnitudes for the g, r, and i filters (SDSS AB system) from Millán-Irigoyen et al. (2025).

The effective wavelengths of each filter are taken from Table 2 (2.5m reference) of Doi et al. (2010):

g band -> λeff = 4627 Å -> "blue" channel r band -> λeff = 6140 Å -> "green" channel i band -> λeff = 7467 Å -> "red" channel

References

I. Millán-Irigoyen et al. (2025). HR-pyPopStar II: high spectral resolution evolutionary synthesis models low metallicity expansion and the properties of the stellar populations of dwarf galaxies. arXiv. doi:10.48550/arxiv.2510.02886

M. Doi et al. (2010). PHOTOMETRIC RESPONSE FUNCTIONS OF THE SLOAN DIGITAL SKY SURVEY IMAGER. The Astronomical Journal 139(4), 1628-1648. doi:10.1088/0004-6256/139/4/1628

GalaxyInspector.MMAP_PATH — Constant

Path to the directory where memory-mapped arrays will be stored.

GalaxyInspector.MOLLA2015_DATA_PATH — Constant

Path to the file with the Milky Way profiles from Mollá et al. (2015).

References

M. Mollá et al. (2015). Galactic chemical evolution: stellar yields and the initial mass function. Monthly Notices of the Royal Astronomical Society 451(4), 3693–3708. doi:10.1093/mnras/stv1102

GalaxyInspector.MORPHOLOGICAL_COMPONENTS — Constant

Human readable name for each morphological component.

GalaxyInspector.N_KS98 — Constant

Slope of the Kennicutt-Schmidt law, taken from Kennicutt (1998) (Section 4, Equation 4).

\[\Sigma_\mathrm{SFR} = a \left( \frac{\Sigma_\mathrm{gas}}{1 \, \mathrm{M_\odot \, pc^{-2}}} \right)^{\!N} \mathrm{M_\odot \, yr^{-1} \, kpc^{-2}} \, ,\]

References

R. C. Kennicutt (1998). The Global Schmidt Law in Star-forming Galaxies. The Astrophysical Journal, 498(2), 541-552. doi:10.1086/305588

GalaxyInspector.P0 — Constant

Reference pressure for the molecular fraction-pressure relation, from Blitz et al. (2006) (Table 2, "Mean" row, Third column).

References

L. Blitz et al. (2006). The Role of Pressure in GMC Formation II: The H2-Pressure Relation. The Astrophysical Journal, 650(2), 933. doi:10.1086/505417

GalaxyInspector.PARTICLE_CODE_NAME — Constant

Internal code name (data group in the HDF5 output) corresponding to each type of cell/particle.

GalaxyInspector.PARTICLE_INDEX — Constant

Current cell/particle index in use.

GalaxyInspector.PARTICLE_NAMES — Constant

Current human readable name of each cell/particle type in use.

GalaxyInspector.PARTICLE_TYPE — Constant

Type of cell/particle corresponding to each internal code name (data group in the HDF5 output).

GalaxyInspector.PASS_ALL — Constant

Filter dictionary that does not exclude any cell/particle.

GalaxyInspector.PASS_NONE — Constant

Filter dictionary that excludes every cell/particle.

GalaxyInspector.PXSIZE_BIGIEL2008 — Constant

Spatial resolution used in Bigiel et al. (2008) (see Section 1).

References

F. Bigiel et al. (2008). THE STAR FORMATION LAW IN NEARBY GALAXIES ON SUB-KPC SCALES. The Astrophysical Journal, 136(6), 2846. doi:10.1088/0004-6256/136/6/2846

GalaxyInspector.PXSIZE_COLOGNI2026 — Constant

Spatial resolution used in Cologni et al. (2026) (see Section 3.2).

References

R. Cologni et al. (2026). Resolved molecular gas and star formation in massive unquenched spirals. Astronomy and Astrophysics, 709, A148. doi:10.1051/0004-6361/202558486

GalaxyInspector.PXSIZE_LIN2019 — Constant

Spatial resolution used in Lin et al. (2019) (see Section 4).

References

L. Lin et al. (2019). The ALMaQUEST Survey: The Molecular Gas Main Sequence and the Origin of the Star-forming Main Sequence. The Astrophysical Journal Letters, 884(2), L33. doi:10.3847/2041-8213/ab4815

GalaxyInspector.PXSIZE_QUEREJETA2021 — Constant

Spatial resolution used in Querejeta et al. (2021) (see Section 2.3).

References

M. Querejeta et al. (2021). Stellar structures, molecular gas, and star formation across the PHANGS sample of nearby galaxies. Astronomy and Astrophysics, 656, A133. doi:10.1051/0004-6361/202140695

GalaxyInspector.PXSIZE_SUN2023 — Constant

Spatial resolution used in Sun et al. (2023).

References

J. Sun et al. (2023). Star Formation Laws and Efficiencies across 80 Nearby Galaxies. The Astrophysical Journal Letters, 945(2), L19. doi:10.3847/2041-8213/acbd9c

GalaxyInspector.QTY_REGISTRY — Constant

List of physical quantities and their properties.

GalaxyInspector.QUANTITIES — Constant

Dimensional properties of the quantities in the snapshots and group catalog files.

GalaxyInspector.RGC_RANGE_GIANNETTI2017 — Constant

Range of values for the galactocentric distances in Figure 4 and Table A.1 of Giannetti et al. (2017).

References

A. Giannetti et al. (2017). Galactocentric variation of the gas-to-dust ratio and its relation with metallicity. Astronomy and Astrophysics, 606, L12. doi:10.1051/0004-6361/201731728

GalaxyInspector.ReducedIndexType — Type

Reduced index type.

GalaxyInspector.RotationType — Type

Type for rotations.

GalaxyInspector.SCHREIBER2015_m0 — Constant

Fitted parameters for the SFR of main-sequence galaxies as a function of stellar mass and redshift, from Schreiber et al. (2015) (Equation 9).

C. Schreiber et al. (2015). The Herschel view of the dominant mode of galaxy growth from z = 4 to the present day. Astronomy & Astrophysics, 575, A74. doi:10.1051/0004-6361/201425017

GalaxyInspector.SCOVILLE2016_α0 — Constant

Fitted parameters for the gas mass fraction as a function of stellar mass, redshift, and sSFR, from Scoville et al. (2016) (Equation 1).

N. Scoville et al. (2016). ISM MASSES AND THE STAR FORMATION LAW AT Z = 1 TO 6: ALMA OBSERVATIONS OF DUST CONTINUUM IN 145 GALAXIES IN THE COSMOS SURVEY FIELD. The Astrophysical Journal, 820(2), 83. doi:10.3847/0004-637X/820/2/83

GalaxyInspector.SFM_COMPONENTS — Constant

List of physical components of our SF model.

GalaxyInspector.SFM_IDX — Constant

ODE index corresponding to each gas phase in our star formation model.

Ionized gas fraction: fi(t) = Mi(t) / MC –> data_dict[:gas]["FRAC"][1, :]

Atomic gas fraction: fa(t) = Ma(t) / MC –> data_dict[:gas]["FRAC"][2, :]

Molecular gas fraction: fm(t) = Mm(t) / MC –> data_dict[:gas]["FRAC"][3, :]

Stellar fraction: fs(t) = Ms(t) / MC –> data_dict[:gas]["FRAC"][4, :]

Metal fraction: fZ(t) = MZ(t) / MC –> data_dict[:gas]["FRAC"][5, :]

Dust fraction: fd(t) = Md(t) / MC –> data_dict[:gas]["FRAC"][6, :]

GalaxyInspector.SFM_MAGNITUDES — Constant

List of physical magnitudes of our SF model.

GalaxyInspector.SFR_RANGE_BIGIEL2008 — Constant

Range of values for

\[\Sigma_\mathrm{SFR} \, [\mathrm{M_\odot \, yr^{-1} \, kpc^{-2}}] \, ,\]

in the seven spirals in Table 1 of Bigiel et al. (2008), with associated molecular data.

The actual values for the SFR density are taken from Table 2 in Bigiel et al. (2010), using only the ones with associated molecular data.

References

F. Bigiel et al. (2008). THE STAR FORMATION LAW IN NEARBY GALAXIES ON SUB-KPC SCALES. The Astrophysical Journal, 136(6), 2846. doi:10.1088/0004-6256/136/6/2846

F. Bigiel et al. (2010). EXTREMELY INEFFICIENT STAR FORMATION IN THE OUTER DISKS OF NEARBY GALAXIES. The Astrophysical Journal, 140(5), 1194. doi:10.1088/0004-6256/140/5/1194

GalaxyInspector.SFR_RANGE_COLOGNI2026 — Constant

Range of values for

\[\Sigma_\mathrm{SFR} \, [\mathrm{M_\odot \, yr^{-1} \, kpc^{-2}}] \, ,\]

in Figure 7 of Cologni et al. (2026), with associated molecular data.

References

R. Cologni et al. (2026). Resolved molecular gas and star formation in massive unquenched spirals. Astronomy and Astrophysics, 709, A148. doi:10.1051/0004-6361/202558486

GalaxyInspector.SFR_RANGE_KS98 — Constant

Range of values for

\[\Sigma_\mathrm{SFR} \, [\mathrm{M_\odot \, yr^{-1} \, kpc^{-2}}] \, ,\]

from the combine data (Table 1 and 2) in Kennicutt (1998).

References

R. C. Kennicutt (1998). The Global Schmidt Law in Star-forming Galaxies. The Astrophysical Journal, 498(2), 541-552. doi:10.1086/305588

GalaxyInspector.SFR_RANGE_LIN2019 — Constant

Range of values for

\[\Sigma_\mathrm{SFR} \, [\mathrm{M_\odot \, yr^{-1} \, kpc^{-2}}] \, ,\]

in Figure 3 of Lin et al. (2019), with associated molecular data.

References

L. Lin et al. (2019). The ALMaQUEST Survey: The Molecular Gas Main Sequence and the Origin of the Star-forming Main Sequence. The Astrophysical Journal Letters, 884(2), L33. doi:10.3847/2041-8213/ab4815

GalaxyInspector.SFR_RANGE_QUEREJETA2021 — Constant

Range of values for

\[\Sigma_\mathrm{SFR} \, [\mathrm{M_\odot \, yr^{-1} \, kpc^{-2}}] \, ,\]

in Figure 6 and 7 of Querejeta et al. (2021), with associated molecular data.

References

M. Querejeta et al. (2021). Stellar structures, molecular gas, and star formation across the PHANGS sample of nearby galaxies. Astronomy and Astrophysics, 656, A133. doi:10.1051/0004-6361/202140695

GalaxyInspector.SFR_REL_PATH — Constant

Relative path, within the simulation directory, of sfr.txt.

GalaxyInspector.SOLAR_ABUNDANCE — Constant

Solar abundances.

These are defined as $12 + \log_{10}(N_\mathrm{X} / N_\mathrm{H})$, where $N_\mathrm{X}$ and $N_\mathrm{H}$ are the number densities of element $\mathrm{X}$ and hydrogen respectively.

References

M. Asplund et al. (2009). The Chemical Composition of the Sun. Annual Review of Astronomy and Astrophysics, 47(1), 481–522. doi:10.1146/annurev.astro.46.060407.145222

GalaxyInspector.SUN2023_DATA_PATH — Constant

Path to Table A1 from Sun et al. (2023).

References

J. Sun et al. (2023). Star Formation Laws and Efficiencies across 80 Nearby Galaxies. The Astrophysical Journal Letters, 945(2), L19. doi:10.3847/2041-8213/acbd9c

GalaxyInspector.THRESHOLD_DENSITY — Constant

Threshold density above which the gas cells/particles can turn into stars.

This value corresponds to CritPhysDensity $= 0.318 \, [\mathrm{cm^{-3}}]$ in the param.txt file (used in cosmological and non-cosmological simulations). Which is converted to internal units within the code using PhysDensThresh = CritPhysDensity * PROTONMASS / HYDROGEN_MASSFRAC / UnitDensity_in_cgs. Then, to go to physical units again one has to do: PhysDensThresh * UnitDensity_in_cgs * cf_a3inv * HubbleParam * HubbleParam.

PhysDensThresh = $1.03378 \times 10^{6}$ [internal units of density]

For a cosmological simulation at redshift 0 (cf_a3inv = 1), this result in a physical density threshold of $0.192 \, [\mathrm{cm^{-3}}]$, or, adding the proton mass, a value of

$\log_{10} \rho \, [\mathrm{M_\odot \, kpc^{-3}}] = 6.677$

GalaxyInspector.TRANSFORM_LIST — Constant

List of symbols for the transformation presets.

The options have the form :{component}_{group} selecting which cells/particle to consider for the center of mass (new origin for the translation) and principal axis (new reference system for the rotation), where component can be:

  + :all         -> Every component present in data_dict
  + :{component} -> One of the keys of PARTICLE_INDEX

and group can be:
  + :box     -> Whole simulation box
  + :halo    -> Main halo
  + :subhalo -> Main subhalo

GalaxyInspector.TRANSFORM_SPLITS — Constant

Dictionary mapping each transformation to its component and group.

GalaxyInspector.TranslationType — Type

Type for translations.

GalaxyInspector.WONG_BLUE — Constant

Default colors.

GalaxyInspector.a_KS98 — Constant

Intercept of the Kennicutt-Schmidt law, taken from Kennicutt (1998) (Section 4, Equation 4).

\[\Sigma_\mathrm{SFR} = a \left( \frac{\Sigma_\mathrm{gas}}{1 \, \mathrm{M_\odot \, pc^{-2}}} \right)^{\!N} \mathrm{M_\odot \, yr^{-1} \, kpc^{-2}} \, ,\]

References

R. C. Kennicutt (1998). The Global Schmidt Law in Star-forming Galaxies. The Astrophysical Journal, 498(2), 541-552. doi:10.1086/305588

GalaxyInspector.AbstractPlotQuantity — Type

Abstract derived quantity type.

GalaxyInspector.BaseQuantity — Type

Derived quantity type.

Fields

id :: Symbol: Symbol representing the quantity.

qty_label :: AbstractString: Name of the quantity for the axis label.

request :: Dict{Symbol,Vector{String}} = Dict{Symbol,Vector{String}}(): Data request for readSnapshot. It must have the shape cell/particle type -> [block, block, ...], where the possible types are the keys of PARTICLE_INDEX, :group, and :subhalo, and the possible blocks are the keys of QUANTITIES.

unit :: Unitful.Units = Unitful.NoUnits: Fiducial unit for the quantity.

exp_factor :: Int = 0: Numerical exponent to scale down the quantity, e.g. if exp_factor = 10 the values will be divided by $10^{10}$.

cp_type :: Union{Symbol,Nothing} = nothing: Cell/particle type corresponding to the quantity. It can be any of the keys of PARTICLE_INDEX.

scatter_func :: Function = getNothing: Function that computes the quantity for each cell/particle given a data dictionary (see makeDataDict for the canonical description). It should return nothing if it does not apply to the quantity.

integrate_func :: Function = getNothing: Function that computes the quantity for all cells/particles as a single value, given a data dictionary (see makeDataDict for the canonical description). It should return nothing if it does not apply to the quantity.

GalaxyInspector.BinaryQuantity — Type

Represents a mathematical operation between two derived quantities.

Fields

op ::Function: Binary operation, e.g. +, -, *, /, etc.

op_label ::String: The binary operation as a string, for the axis label, e.g. "+", "/", etc.

left ::AbstractPlotQuantity: Left quantity.

right ::AbstractPlotQuantity: Right quantity.

GalaxyInspector.CubicGrid — Type

Grid with 3 d.o.f.

Fields

x_axis :: Vector{<:Number}: Absolute position of (the center of) the bins in the x direction.

y_axis :: Vector{<:Number}: Absolute position of (the center of) the bins in the y direction.

z_axis :: Vector{<:Number}: Absolute position of (the center of) the bins in the z direction.

x_edges :: Vector{<:Number}: Absolute position of (the edges of) the bins in the x direction.

y_edges :: Vector{<:Number}: Absolute position of (the edges of) the bins in the y direction.

z_edges :: Vector{<:Number}: Absolute position of (the edges of) the bins in the z direction.

n_bins :: NTuple{3,Int}: Number of bins in each direction.

bin_size_1D :: Vector{<:Number}: Widths of the bins in each direction.

bin_size_2D :: Vector{<:Number}: Areas of the faces normal to each direction.

bin_size_3D :: Number: Volume of the bins.

origin :: Vector{<:Number}: Physical position of the center of the grid.

size :: Vector{<:Number}: Physical side length of the grid in each direction (passthrough argument).

n_voxels :: Int: Total number of bins.

GalaxyInspector.GroupCatHeader — Type

Data in the "Header" group of a HDF5 group catalog file.

The default values are for when there are no group catalog files.

Fields

box_size :: Float64 = NaN: Total size of the simulation box, in internal units of length.

h :: Float64 = NaN: Dimensionless Hubble parameter, "little h".

n_groups_part :: Int32 = -1: Number of halos (FoF groups) in this file chunk.

n_groups_total :: Int32 = -1: Total number of halos (FoF groups) in this snapshot.

n_subgroups_part :: Int32 = -1: Number of subhalos (subfind) in this file chunk.

n_subgroups_total :: Int32 = -1: Total number of subhalos (subfind) in this snapshot.

num_files :: Int32 = -1: Number of file chunks per snapshot.

omega_0 :: Float64 = NaN: The cosmological density parameter for matter.

omega_l :: Float64 = NaN: The cosmological density parameter for the cosmological constant.

redshift :: Float64 = NaN: The redshift.

time :: Float64 = NaN: The physical time or the scale factor, depending on the type of simulation.

GalaxyInspector.GroupCatalog — Type

Metadata for a group catalog file.

Fields

path :: Union{String,Missing}: Full path to the group catalog file.

header :: GroupCatHeader: Header.

GalaxyInspector.InternalUnits — Type

Unit conversion factors.

Fields

x_cgs::Unitful.Length: From internal units of length to $\mathrm{cm}$.

x_cosmo::Unitful.Length: From internal units of length to $\mathrm{kpc}$.

x_comoving::Unitful.Length: From internal units of length to $\mathrm{ckpc}$.

v_cgs::Unitful.Velocity: From internal units of velocity to $\mathrm{cm \, s^{-1}}$.

v_cosmo::Unitful.Velocity: From internal units of velocity to $\mathrm{km \, s^{-1}}$.

m_cgs::Unitful.Mass: From internal units of mass to $\mathrm{g}$.

m_cosmo::Unitful.Mass: From internal units of mass to $\mathrm{M_\odot}$.

t_cgs::Unitful.Time: From internal units of time to $\mathrm{s}$.

t_cosmo::Unitful.Time: From internal units of time to $\mathrm{Myr}$.

t_newton::Unitful.Time: From internal units (non-cosmological simulations) of time to $\mathrm{Myr}$.

U_cgs::Unitful.Energy: From internal units of specific energy to $\mathrm{erg \, g^{-1}}$.

rho_cgs::Unitful.Density: From internal units of density to $\mathrm{g \, cm^{-3}}$.

P_Pa::Unitful.Pressure: From internal units of pressure to $\mathrm{Pa}$.

GalaxyInspector.LinearGrid — Type

Grid with 1 d.o.f.

Embeddings:

1D -> Ticks.
2D -> Concentric rings.
3D -> Spherical shells.

Fields

x_axis :: Vector{<:Number}: Position of (the center of) the bins, relative to the origin.

x_edges :: Vector{<:Number}: Position of (the edges of) the bins, relative to the origin.

n_bins :: Int: Number of bins.

bin_size_1D :: Vector{<:Number}: Widths of the bins in linear space.

bin_size_2D :: Vector{<:Number}: Areas of the bins in linear space (assuming a 2D embedding).

bin_size_3D :: Vector{<:Number}: Volumes of the bins in linear space (assuming a 3D embedding).

origin :: Vector{<:Number}: Physical position of the coordinate origin for the grid.

GalaxyInspector.Simulation — Type

Metadata for a simulation.

Fields

path :: String: Full path to the simulation directory.

index :: Int: An index associated with the simulation.

slice :: IndexType: Slice of the simulation, i.e. which snapshots will be read. It can be an integer (a single snapshot), a vector of integers (several snapshots), an UnitRange (e.g. 5:13), an StepRange (e.g. 5:2:13) or (:) (all snapshots).

cosmological :: Bool: If the simulation is cosmological,

false -> Newtonian simulation (ComovingIntegrationOn = 0).

true -> Cosmological simulation (ComovingIntegrationOn = 1).

simulation_table :: DataFrame: A dataframe where each row is a snapshot, and the colums are

:ids -> Dataframe index, i.e. if there are 10 snapshots in total it runs from 1 to 10.

:numbers -> Number in the file name.

:scale_factors -> Scale factor.

:redshifts -> Redshift.

:physical_times -> Physical time since the Big Bang.

:lookback_times -> Physical time left to reach the last snapshot.

:snapshot_paths -> Full path to each snapshots.

:groupcat_paths -> Full path to each group catalog files.

GalaxyInspector.Snapshot — Type

Metadata for a snapshot.

Fields

path :: String: Full path to the snapshot.

global_index :: Int: Index of the snapshot in the context of the whole simulation.

slice_index :: Int: Index of the snapshot in the context of the simulation slice.

physical_time :: Unitful.Time: Physical time since the Big Bang.

lookback_time :: Unitful.Time: Physical time left to reach the last snapshot.

scale_factor :: Float64: Scale factor.

redshift :: Float64: Redshift.

header :: SnapshotHeader: Header.

GalaxyInspector.SnapshotBlock — Type

Dimensional information about a physical quantity.

Fields

hdf5_name :: String: HDF5 block name.

dimensions :: Unitful.Dimensions: Physical dimensions of the quantity, e.g. Unitful.𝐋 * Unitful.𝐓^-1.

unit :: Union{Unitful.Units,Symbol}: Units of the quantity within the simulation code. It can be a unit from Unitful or UnitfulAstro, or it can be the symbol :internal which denotes internal code units.

GalaxyInspector.SnapshotHeader — Type

Data in the "Header" group of a HDF5 snapshot file.

Fields

box_size :: Float64: Total size of the simulation box, in internal units of length.

h :: Float64: Dimensionless Hubble parameter, "little h".

mass_table :: Vector{Float64}: Masses of particle/cell types which have a constant mass, in internal units of mass.

num_files :: Int32: Number of file chunks per snapshot.

num_part :: Vector{Int32}: Number of particles/cells (of each type) included in this file chunk.

num_total :: Vector{UInt32}: Total number of particles/cells (of each type) for this snapshot.

omega_0 :: Float64: The cosmological density parameter for matter.

omega_l :: Float64: The cosmological density parameter for the cosmological constant.

redshift :: Float64: The redshift.

time :: Float64: The physical time or the scale factor, depending on the type of simulation.

l_unit :: Unitful.Length: Conversion factor from internal units of length to centimeters.

m_unit :: Unitful.Mass: Conversion factor from internal units of mass to grams.

v_unit :: Unitful.Velocity: Conversion factor from internal units of velocity to centimeters per second.

GalaxyInspector.SquareGrid — Type

Grid with 2 d.o.f.

Fields

x_axis :: Vector{<:Number}: Absolute position of (the center of) the bins in the x direction.

y_axis :: Vector{<:Number}: Absolute position of (the center of) the bins in the y direction.

x_edges :: Vector{<:Number}: Absolute position of (the edges of) the bins in the x direction.

y_edges :: Vector{<:Number}: Absolute position of (the edges of) the bins in the y direction.

n_bins :: NTuple{2,Int}: Number of bins in each direction.

bin_size_1D :: Vector{<:Number}: Widths of the bins in each direction.

bin_size_2D :: Number: Area of the bins.

origin :: Vector{<:Number}: Physical position of the center of the grid.

size :: Vector{<:Number}: Physical side length of the grid in each direction (passthrough argument).

n_pixels :: Int: Total number of bins.

GalaxyInspector.LEE2015_PARAMETERS — Method

LEE2015_PARAMETERS(z::Float64)::NTuple{3,Measurements.Measurement{Float64}}

Best-fit parameters for the SFR as a function of stellar mass, from Lee et al. (2015) (Table 1).

Arguments

- `z :: Float64`: Redshift. The parameters are only valid for 0.25 <= z <= 1.3, but the function will return the values for z = 0.25 or z = 1.3 if the input value is outside this range.

Returns

A tuple with three elements (See Section 4.1 of Lee et al. (2015) for the definition of the parameters):

S0 [log10(M⊙ * yr^-1)]: The maximum value of log10(SFR / M⊙ * yr^-1).

M0 [log10(M⊙)]: Turnover mass.

γ [dimensionless]: Power-law slope at low stellar masses.

References

N. Lee et al. (2015). A TURNOVER IN THE GALAXY MAIN SEQUENCE OF STAR FORMATION AT M ~ 10^10 M☉ FOR REDSHIFTS z < 1.3*. The Astrophysical Journal, 801(2), 80. doi:10.1088/0004-637X/801/2/80

GalaxyInspector.fillRegistry! — Method

fillRegistry!(; <keyword arguments>)::Nothing

Fill the given registry with the component-magnitude quantities defined in MAGNITUDES and COMPONENTS, the abundance quantities defined in ABUNDANCE_QUANTITIES, and the SFM quantities defined in SFM_MAGNITUDES and SFM_COMPONENTS.

Arguments

registry::Dict{Symbol, BaseQuantity}: The registry to be filled. By default, it is the global QTY_REGISTRY.

GalaxyInspector.haloQuantity — Method

haloQuantity(magnitude::Symbol, halo_idx::Int)::BaseQuantity

Construct a halo BaseQuantity.

Arguments

magnitude::Symbol: Target halo quantity. The options are the keys of HALO_QUANTITIES.

halo_idx::Int: Index of the target halo (FoF group). Starts at 1.

Returns

The selected BaseQuantity for the target halo.