Type aliases and structs

This is a list of type aliases and structs used in GalaxyInspector.

None of these are exported.

GalaxyInspector.ColorType — Type

Color type.

GalaxyInspector.IndexType — Type

Index type.

GalaxyInspector.LineStyleType — Type

Line style type.

GalaxyInspector.ReducedIndexType — Type

Reduced index type.

GalaxyInspector.RotationType — Type

Rotation type

GalaxyInspector.TranslationType — Type

Translation type

GalaxyInspector.CubicGrid — Type

Grid with 3 d.o.f.

Fields

x_axis :: Vector{<:Number}: Absolute position of (the center of) the bins in the x direction.

y_axis :: Vector{<:Number}: Absolute position of (the center of) the bins in the y direction.

z_axis :: Vector{<:Number}: Absolute position of (the center of) the bins in the z direction.

x_edges :: Vector{<:Number}: Absolute position of (the edges of) the bins in the x direction.

y_edges :: Vector{<:Number}: Absolute position of (the edges of) the bins in the y direction.

z_edges :: Vector{<:Number}: Absolute position of (the edges of) the bins in the z direction.

n_bins :: NTuple{3,Int}: Number of bins in each direction.

bin_size_1D :: Vector{<:Number}: Widths of the bins in each direction.

bin_size_2D :: Vector{<:Number}: Areas of the faces normal to each direction.

bin_size_3D :: Number: Volume of the bins.

origin :: Vector{<:Number}: Physical position of the center of the grid.

size :: Vector{<:Number}: Physical side length of the grid in each direction (passthrough argument).

n_voxels :: Int: Total number of bins.

GalaxyInspector.GroupCatHeader — Type

Data in the "Header" group of a HDF5 group catalog file.

The default values are for when there are no group catalog files.

Fields

box_size :: Float64 = NaN: Total size of the simulation box, in internal units of length.

h :: Float64 = NaN: Dimensionless Hubble parameter, "little h".

n_groups_part :: Int32 = -1: Number of halos (FoF groups) in this file chunk.

n_groups_total :: Int32 = -1: Total number of halos (FoF groups) in this snapshot.

n_subgroups_part :: Int32 = -1: Number of subhalos (subfind) in this file chunk.

n_subgroups_total :: Int32 = -1: Total number of subhalos (subfind) in this snapshot.

num_files :: Int32 = -1: Number of file chunks per snapshot.

omega_0 :: Float64 = NaN: The cosmological density parameter for matter.

omega_l :: Float64 = NaN: The cosmological density parameter for the cosmological constant.

redshift :: Float64 = NaN: The redshift.

time :: Float64 = NaN: The physical time or the scale factor, depending on the type of simulation.

GalaxyInspector.GroupCatalog — Type

Metadata for a group catalog file.

Fields

path :: Union{String,Missing}: Full path to the group catalog file.

header :: GroupCatHeader: Header.

GalaxyInspector.InternalUnits — Type

Unit conversion factors.

Fields

x_cgs::Unitful.Length: From internal units of length to $\mathrm{cm}$.

x_cosmo::Unitful.Length: From internal units of length to $\mathrm{kpc}$.

x_comoving::Unitful.Length: From internal units of length to $\mathrm{ckpc}$.

v_cgs::Unitful.Velocity: From internal units of velocity to $\mathrm{cm \, s^{-1}}$.

v_cosmo::Unitful.Velocity: From internal units of velocity to $\mathrm{km \, s^{-1}}$.

m_cgs::Unitful.Mass: From internal units of mass to $\mathrm{g}$.

m_cosmo::Unitful.Mass: From internal units of mass to $\mathrm{M_\odot}$.

t_cgs::Unitful.Time: From internal units of time to $\mathrm{s}$.

t_cosmo::Unitful.Time: From internal units of time to $\mathrm{Myr}$.

t_newton::Unitful.Time: From internal units (non-cosmological simulations) of time to $\mathrm{Myr}$.

U_cgs::Unitful.Energy: From internal units of specific energy to $\mathrm{erg \, g^{-1}}$.

rho_cgs::Unitful.Density: From internal units of density to $\mathrm{g \, cm^{-3}}$.

P_Pa::Unitful.Pressure: From internal units of pressure to $\mathrm{Pa}$.

GalaxyInspector.LinearGrid — Type

Grid with 1 d.o.f.

Embeddings:

1D -> Ticks.
2D -> Concentric rings.
3D -> Spherical shells.

Fields

x_axis :: Vector{<:Number}: Position of (the center of) the bins, relative to the origin.

x_edges :: Vector{<:Number}: Position of (the edges of) the bins, relative to the origin.

n_bins :: Int: Number of bins.

bin_size_1D :: Vector{<:Number}: Widths of the bins in linear space.

bin_size_2D :: Vector{<:Number}: Areas of the bins in linear space (assuming a 2D embedding).

bin_size_3D :: Vector{<:Number}: Volumes of the bins in linear space (assuming a 3D embedding).

origin :: Vector{<:Number}: Physical position of the coordinate origin for the grid.

GalaxyInspector.PlotParams — Type

Plotting parameters for a quantity.

Fields

request::Dict{Symbol,Vector{String}} = Dict{Symbol,Vector{String}}(): Data request for readSnapshot. It must have the shape cell/particle type -> [block, block, ...], where the possible types are the keys of PARTICLE_INDEX, :group, and :subhalo, and the possible blocks are the keys of QUANTITIES.

var_name::AbstractString = "": Name of the quantity for the axis label. It should not include units or scaling factors.

exp_factor::Int = 0: Numerical exponent to scale down the axis, e.g. if exp_factor = 10 the values will be divided by $10^{10}$. The default is no scaling.

unit::Unitful.Units = Unitful.NoUnits: Target unit for the axis.

axis_label::AbstractString = "auto_label": Label for the axis. It can contain the string auto_label, which will be replaced by the default label: var_name / 10^exp_factor unit.

cp_type::Union{Symbol,Nothing} = nothing: Cell/particle type corresponding to the quantity. It can be any of the keys of PARTICLE_INDEX.

GalaxyInspector.Qty — Type

Dimensional information about a physical quantity.

Fields

hdf5_name :: String: HDF5 block name.

dimensions :: Unitful.Dimensions: Physical dimensions of the quantity, e.g. Unitful.𝐋 * Unitful.𝐓^-1.

unit :: Union{Unitful.Units,Symbol}: Units of the quantity within the simulation code. It can be a unit from Unitful or UnitfulAstro, or it can be the symbol :internal which denotes internal code units.

GalaxyInspector.Simulation — Type

Metadata for a simulation.

Fields

path :: String: Full path to the simulation directory.

index :: Int: An index associated with the simulation.

slice :: IndexType: Slice of the simulation, i.e. which snapshots will be read. It can be an integer (a single snapshot), a vector of integers (several snapshots), an UnitRange (e.g. 5:13), an StepRange (e.g. 5:2:13) or (:) (all snapshots).

cosmological :: Bool: If the simulation is cosmological,

false -> Newtonian simulation (ComovingIntegrationOn = 0).

true -> Cosmological simulation (ComovingIntegrationOn = 1).

simulation_table :: DataFrame: A dataframe where each row is a snapshot, and the colums are

:ids -> Dataframe index, i.e. if there are 10 snapshots in total it runs from 1 to 10.

:numbers -> Number in the file name.

:scale_factors -> Scale factor.

:redshifts -> Redshift.

:physical_times -> Physical time since the Big Bang.

:lookback_times -> Physical time left to reach the last snapshot.

:snapshot_paths -> Full path to each snapshots.

:groupcat_paths -> Full path to each group catalog files.

GalaxyInspector.Snapshot — Type

Metadata for a snapshot.

Fields

path :: String: Full path to the snapshot.

global_index :: Int: Index of the snapshot in the context of the whole simulation.

slice_index :: Int: Index of the snapshot in the context of the simulation slice.

physical_time :: Unitful.Time: Physical time since the Big Bang.

lookback_time :: Unitful.Time: Physical time left to reach the last snapshot.

scale_factor :: Float64: Scale factor.

redshift :: Float64: Redshift.

header :: SnapshotHeader: Header.

GalaxyInspector.SnapshotHeader — Type

Data in the "Header" group of a HDF5 snapshot file.

Fields

box_size :: Float64: Total size of the simulation box, in internal units of length.

h :: Float64: Dimensionless Hubble parameter, "little h".

mass_table :: Vector{Float64}: Masses of particle/cell types which have a constant mass, in internal units of mass.

num_files :: Int32: Number of file chunks per snapshot.

num_part :: Vector{Int32}: Number of particles/cells (of each type) included in this file chunk.

num_total :: Vector{UInt32}: Total number of particles/cells (of each type) for this snapshot.

omega_0 :: Float64: The cosmological density parameter for matter.

omega_l :: Float64: The cosmological density parameter for the cosmological constant.

redshift :: Float64: The redshift.

time :: Float64: The physical time or the scale factor, depending on the type of simulation.

l_unit :: Unitful.Length: Conversion factor from internal units of length to centimeters.

m_unit :: Unitful.Mass: Conversion factor from internal units of mass to grams.

v_unit :: Unitful.Velocity: Conversion factor from internal units of velocity to centimeters per second.

GalaxyInspector.SquareGrid — Type

Grid with 2 d.o.f.

Fields

x_axis :: Vector{<:Number}: Absolute position of (the center of) the bins in the x direction.

y_axis :: Vector{<:Number}: Absolute position of (the center of) the bins in the y direction.

x_edges :: Vector{<:Number}: Absolute position of (the edges of) the bins in the x direction.

y_edges :: Vector{<:Number}: Absolute position of (the edges of) the bins in the y direction.

n_bins :: NTuple{2,Int}: Number of bins in each direction.

bin_size_1D :: Vector{<:Number}: Widths of the bins in each direction.

bin_size_2D :: Number: Area of the bins.

origin :: Vector{<:Number}: Physical position of the center of the grid.

size :: Vector{<:Number}: Physical side length of the grid in each direction (passthrough argument).

n_pixels :: Int: Total number of bins.