Post processing functions

A post-processing function must take a Makie figure, add something to it, and return how to label the additions (or nothing when no new labels should be drawn).

None of these functions are exported.

Expected signature:

  post_processing(figure, args...; kwargs...) -> ([marker, ...], [label, ...])

where:

figure::Makie.Figure

marker::LegendElement

label::String

GalaxyInspector.ppAnnotation! — Method

ppAnnotation!(figure::Makie.Figure, text::AbstractString; <keyword arguments>)::Nothing

Add an annotation to the plot.

Arguments

figure::Makie.Figure: Makie figure to be drawn over.

text::AbstractString: Text to be written.

position::Tuple{<:Real,<:Real}=(0, 1): Relative position of the top left corner of the text box.

GalaxyInspector.ppBarPlotLabels — Method

ppBarPlotLabels(
    ::Makie.Figure,
    include_stars::Bool;
    <keyword arguments>
)::Tuple{Vector{<:LegendElement},Vector{AbstractString}}

Return the legend elements for the plot made by gasFractionsBarPlot.

Arguments

::Makie.Figure: Makie figure to be drawn over.

include_stars::Bool=false: If the stars will be included as one of the gas phases.

colors=Makie.wong_colors(): Colors for the bars.

Returns

A tuple with the elements for the legend:

A PolyElements to be used in the legend.

The label strings.

GalaxyInspector.ppBigiel2008! — Method

ppBigiel2008!(
    figure::Makie.Figure,
    quantity::Symbol;
    <keyword arguments>
)::Tuple{Vector{<:LegendElement},Vector{AbstractString}}

Draw a line plot with the fit for the KS relation in Bigiel et al. (2008).

Arguments

figure::Makie.Figure: Makie figure to be drawn over.

quantity::Symbol: Quantity for the x axis. The options are:

:molecular_area_density -> Molecular hydrogen area mass density, for a radius of FILTER_R.

:neutral_area_density -> Neutral hydrogen area mass density, for a radius of FILTER_R.

x_unit::Unitful.Units=u"Msun * pc^-2": Unit for the area density of gas used in figure.

y_unit::Unitful.Units=u"Msun * yr^-1 * kpc^-2": Unit for the area density of star formation rate used in figure.

x_log::Bool=true: If the x axis is $\log_{10}(\Sigma_\mathrm{H})$ (x_log = true) or just $\Sigma_\mathrm{H}$ (x_log = false).

y_log::Bool=true: If the y axis is $\log_{10}(\Sigma_\mathrm{SFR})$ (y_log = true) or just $\Sigma_\mathrm{SFR}$ (y_log = false).

color::ColorType=:red: Color of the line.

linestyle::LineStyleType=nothing: Style of the line. nothing will produce a solid line.

Returns

A tuple with the elements for the legend:

A LineElement to be used as the marker.

The label string.

References

F. Bigiel et al. (2008). THE STAR FORMATION LAW IN NEARBY GALAXIES ON SUB-KPC SCALES. The Astrophysical Journal, 136(6), 2846. doi:10.1088/0004-6256/136/6/2846

GalaxyInspector.ppCross! — Method

ppCross!(
    figure::Makie.Figure,
    cross_point::Tuple{<:Real,<:Real};
    <keyword arguments>
)::Nothing

Draw two lines, one horizontal and one vertical.

Arguments

figure::Makie.Figure: Makie figure to be drawn over.

cross_point::Tuple{<:Real,<:Real}: Crossing point of the lines.

color::ColorType=:red: Color of the lines.

linestyle::LineStyleType=nothing: Style of the lines. nothing will produce a solid line.

warnings::Bool=true: If a warning will be raised when at least one of the lines is outside the plot range.

GalaxyInspector.ppFeldmann2020! — Method

ppFeldmann2020!(
    figure::Makie.Figure,
    x_quantity::Symbol,
    y_quantity::Symbol;
    <keyword arguments>
)::Tuple{Vector{<:LegendElement},Vector{AbstractString}}

Draw a line, or scatter, plot using the experimental data from the xGASS and xCOLD GASS collaborations, processed by Feldmann (2020).

Arguments

figure::Makie.Figure: Makie figure to be drawn over.

x_quantity::Symbol: Quantity for the x axis. The options are:

:stellar_mass -> Stellar mass.

:molecular_mass -> Molecular hydrogen ($\mathrm{H_2}$) mass.

:atomic_mass -> Atomic hydrogen ($\mathrm{HI}$) mass.

:observational_sfr -> The star formation rate of the last AGE_RESOLUTION.

y_quantity::Symbol: Quantity for the y axis. The options are:

:stellar_mass -> Stellar mass.

:molecular_mass -> Molecular hydrogen ($\mathrm{H_2}$) mass.

:atomic_mass -> Atomic hydrogen ($\mathrm{HI}$) mass.

:observational_sfr -> The star formation rate of the last AGE_RESOLUTION.

scatter::Bool=false: If the data will be presented as a line plot with error bands (default), or alternatively, a scatter plot.

Returns

A tuple with the elements for the legend:

A MarkerElement to be used as the marker.

The label string.

References

R. Feldmann (2020). The link between star formation and gas in nearby galaxies. Communications Physics 3(226). doi:10.1038/s42005-020-00493-0

GalaxyInspector.ppFitLine! — Method

ppFitLine!(
    figure::Makie.Figure;
    <keyword arguments>
)::Union{Tuple{Vector{<:LegendElement},Vector{AbstractString}},Nothing}

Draw a linear fit for the data in figure.

An annotation with the equation $y = a \, x + b$, and the fitted values for $a$ and $b$, will be positioned in the upper right corner of the plot.

Arguments

figure::Makie.Figure: Makie figure to be drawn over.

error_formating::Symbol=:std_error: Error format for the annotation. The options are:

:std_error -> mean ± standard_error.

:conf_interval -> mean ± max(upper$_{95\%}$ - mean, mean - lower$_{95\%}$).

color::ColorType=:red: Color of the line.

linestyle::LineStyleType=nothing: Style of the line. nothing will produce a solid line.

warnings::Bool=true: If a warning will be raised when there are no points to fit.

Returns

A tuple with the elements for the legend:

A LineElement to be used as the marker.

The label string.

GalaxyInspector.ppHorizontalFlags! — Method

ppHorizontalFlags!(
    figure::Makie.Figure,
    positions::Vector{<:Real};
    <keyword arguments>
)::Nothing

Draw horizontal lines.

Arguments

figure::Makie.Figure: Makie figure to be drawn over.

positions::Vector{<:Real}: The y coordinates of the lines.

colors::Vector{<:ColorType}=[:red]: Colors of the lines.

line_styles::Vector{<:LineStyleType}=[nothing]: Styles of the lines. nothing will produce a solid line.

warnings::Bool=true: If a warning will be raised when all the lines are outside the plot range.

GalaxyInspector.ppKennicutt1998! — Method

ppKennicutt1998!(
    figure::Makie.Figure;
    <keyword arguments>
)::Union{Tuple{Vector{<:LegendElement},Vector{AbstractString}},Nothing}

Draw a line plot with the fit for the KS relation in Kennicutt (1998).

Arguments

figure::Makie.Figure: Makie figure to be drawn over.

x_unit::Unitful.Units=u"Msun * pc^-2": Unit for the area density of gas used in figure.

y_unit::Unitful.Units=u"Msun * yr^-1 * kpc^-2": Unit for the area density of star formation rate used in figure.

x_log::Bool=true: If the x axis is $\log_{10}(\Sigma_\mathrm{gas})$ (x_log = true) or just $\Sigma_\mathrm{gas}$ (x_log = false).

y_log::Bool=true: If the y axis is $\log_{10}(\Sigma_\mathrm{SFR})$ (y_log = true) or just $\Sigma_\mathrm{SFR}$ (y_log = false).

color::ColorType=:red: Color of the line.

linestyle::LineStyleType=nothing: Style of the line. nothing will produce a solid line.

Returns

A tuple with the elements for the legend:

A LineElement to be used as the marker.

The label string.

References

R. C. Kennicutt (1998). The Global Schmidt Law in Star-forming Galaxies. The Astrophysical Journal, 498(2), 541-552. doi:10.1086/305588

GalaxyInspector.ppMolla2015! — Method

ppMolla2015!(
    figure::Makie.Figure,
    quantity::Symbol,
    <keyword arguments>
)::Tuple{Vector{<:LegendElement},Vector{AbstractString}}

Draw a profile for the Milky Way using the data compiled by Mollá et al. (2015).

Arguments

figure::Makie.Figure: Makie figure to be drawn over.

quantity::Symbol: Quantity for the y axis. The options are:

:stellar_area_density -> Stellar area mass density.

:molecular_area_density -> Molecular hydrogen area mass density.

:atomic_area_density -> Atomic hydrogen area mass density.

:sfr_area_density -> Star formation rate area density.

:X_stellar_abundance -> Stellar abundance of element $\mathrm{X}$, as $12 + \log_{10}(\mathrm{X \, / \, H})$. $\mathrm{X}$ can be O (oxygen), N (nitrogen), or C (carbon).

color::ColorType=:red: Color of the line.

linestyle::LineStyleType=nothing: Style of the line. nothing will produce a solid line.

error_bars::Bool=true: If the error bars will be plotted.

Returns

A tuple with the elements for the legend:

A MarkerElement to be used as the marker.

The label string.

References

M. Mollá et al. (2015). Galactic chemical evolution: stellar yields and the initial mass function. Monthly Notices of the Royal Astronomical Society 451(4), 3693–3708. doi:10.1093/mnras/stv1102

GalaxyInspector.ppVerticalFlags! — Method

ppVerticalFlags!(
    figure::Makie.Figure,
    positions::Vector{<:Real};
    <keyword arguments>
)::Nothing

Draw vertical lines.

Arguments

figure::Makie.Figure: Makie figure to be drawn over.

positions::Vector{<:Real}: The x coordinates of the lines.

colors::Vector{<:ColorType}=[:red]: Colors of the lines.

line_styles::Vector{<:LineStyleType}=[nothing]: Styles of the lines. nothing will produce a solid line.

warnings::Bool=true: If a warning will be raised when all the lines are outside the plot range.