Constants

GalaxyInspector.CRITICAL_DENSITY — Constant

Critical density for Newtonian simulations, above which the gas particles enter the star formation routine.

This value correspond to $\mathrm{CritPhysDensity} = 0.318 \, \mathrm{cm}^{-3}$ in param.txt. Which is converted to internal units withing GADGET3 with

CritPhysDensity * PROTONMASS / HYDROGEN_MASSFRAC / UnitDensity_in_cgs

GalaxyInspector.ColorInput — Type

Union of color types.

GalaxyInspector.HDF5Names — Constant

Database of the HDF5 names for the blocks.

GalaxyInspector.HUBBLE_CONST — Constant

$H_0 = 0.102201 \, \mathrm{Gyr}^{-1} = 100 \, \mathrm{km} \, \mathrm{s}^{-1} \, \mathrm{Mpc}^{-1}$

GalaxyInspector.IndexType — Type

Union of index types.

GalaxyInspector.KS98_INTERCEPT — Constant

Intercept for the Kennicutt-Schmidt law.

R. C. Kennicutt (1998). The Global Schmidt Law in Star-forming Galaxies. The Astrophysical Journal, 498(2), 541-552. https://doi.org/10.1086/305588

GalaxyInspector.KS98_RHO_UNIT — Constant

Unit of area density for the Kennicutt-Schmidt law.

R. C. Kennicutt (1998). The Global Schmidt Law in Star-forming Galaxies. The Astrophysical Journal, 498(2), 541-552. https://doi.org/10.1086/305588

GalaxyInspector.KS98_SLOPE — Constant

The slope for the Kennicutt-Schmidt law.

R. C. Kennicutt (1998). The Global Schmidt Law in Star-forming Galaxies. The Astrophysical Journal, 498(2), 541-552. https://doi.org/10.1086/305588

GalaxyInspector.LINE_STYLES — Constant

List of line styles for line plots.

GalaxyInspector.LineStyleInput — Type

Union of line style types.

GalaxyInspector.MARKERS — Constant

List of markers for scatter plots.

GalaxyInspector.ParticleType — Constant

GADGET index for the types of particles that can be simulated.

See GADGET2 User's Guide.

Note: In most simulations only :gas, :halo (dark matter) and :stars are used.

GalaxyInspector.QUANTITIES — Constant

Database of dimensional properties for the quantities in the simulations.

GalaxyInspector.RealOrQty — Type

Union of real numbers and physical quantities types.

GalaxyInspector.SOLAR_METALLICITY — Constant

Solar metallicity.

M. Asplund et al. (2009). The Chemical Composition of the Sun. Annual Review of Astronomy and Astrophysics, 47(1), 481–522. https://doi.org/10.1146/annurev.astro.46.060407.145222

GalaxyInspector.theme — Constant

Global plot theme.

GalaxyInspector.Qty — Type

Structure with the dimensional information of a GADGET quantity.

Fields

dimension::Unitful.Dimensions: Physical dimensions of the quantity, e.g. Unitful.𝐋 / Unitful.𝐓.

unit::Union{Symbol,Unitful.Units}: Original units of the quantity. It can be a unit object from Unitful or UnitfulAstro, or it can be the symbol :internal which denotes that GADGET internal units are being used.

GalaxyInspector.SimData — Type

Structure with information about a GADGET simulation.

Fields

path::String: Full path to the directory containing the snapshot files.

sim_index::Int32: Index associated with the simulation.

idx::IndexType: Indexing of the simulation, i.e. which snapshots will be read. It can be an integer (a single snapshot), a vector of integers (several snapshots), an UnitRange (e.g. 5:13) or : (every snapshot will be read).

base_name::String: Base name of the snapshot files, set in the GADGET variable SnapshotFileBase.

header::GIO.SnapshotHeader: Header of the snapshot.

filter_function::Union{Function,Nothing} = nothing: A function with the signature:

foo(file_path::String)::Vector{Int64}

It indicates which particles will be read, taking the file path to a snapshot and returning the list of indices of the selected particles. If set to nothing, then no particles are filtered. See the GadgetIO.jl documentation for examples.

a0::Float64: Scale factor of the initial snapshot (only relevant for cosmological simulations).

GalaxyInspector.SnapData — Type

Structure with information about a GADGET snapshot.

Fields

path::String: Full path to the snapshot.

global_index::Int32: Index of the snapshot in the context of the whole simulation.

local_index::Int32: Index of the snapshot in the context of the current running loop.

time_stamp::Unitful.Time: Clock time associated with the snapshot.