Data Acquisition Functions

GadgetInspector.getCpuFile — Method

getCpuFile(
    source_path::String, 
    target::String; 
    <keyword arguments>
)::Dict{String,Matrix{Float64}}

Get the CPU consumption of the target process, from the cpu.txt file.

Arguments

source_path::String: Path to the directory containing the cpu.txt file.

target::String: Target process.

step::Int64 = 1: Step used to traverse the CPU cycles, i.e. one every step cycles is returned.

warnings::Bool = true: If a warning will be given when the data is not as expected, but the function can still run using sane defaults.

Returns

A dictionary with one entry:

process ⟹ Matrix with CPU cycles in its first column, and CPU usage (as a percentage) in its second column.

GadgetInspector.getCpuFile — Method

getCpuFile(
    source_path::String, 
    targets::Vector{String}; 
    <keyword arguments>
)::Dict{String,Matrix{Float64}}

Get the CPU utilization of several processes from the cpu.txt file.

For each process in targets a matrix with all the CPU usage data (as percentages of total CPU time) is returned.

Arguments

source_path::String: Path to the directory containing the cpu.txt file.

targets::Vector{String}: Target processes.

step::Int64 = 1: Step used to traverse the CPU cycles, i.e. one every step cycles is returned.

warnings::Bool = true: If a warning will be given when the data is not as expected, but the function can still run using sane defaults.

Returns

A dictionary with one entry per target process:

process ⟹ Matrix with CPU cycles in its first column, and CPU usage (as a percentage) in its second column.

GadgetInspector.getMollá2015 — Method

getMollá2015(
    source_path::String,
)::Tuple{DataFrame,Dict{String,Tuple{Unitful.Units,Bool}}}

Get the experimental data compiled by Mollá et al. (2015), from the data.csv file. The quantities in the file are

"R" ⟶ Distance from the galactic center.

"ΣHI" ⟶ Atomic Hydrogen area density.

"ΣHII" ⟶ Molecular Hydrogen area density.

"logΣ*" ⟶ Logarithm of the stellar mass area density.

"logΣsfr" ⟶ Logarithm of the SFR area density.

"C/H" ⟶ 12 + log(C/H), where C and H are the concentration of Carbon and Hydrogen respectively.

"N/H" ⟶ 12 + log(N/H), where N and H are the concentration of Nitrogen and Hydrogen respectively.

"O/H" ⟶ 12 + log(O/H), where O and H are the concentration of Oxygen and Hydrogen respectively.

Every quantity but "R" has its error and is measured as a 2D profile.

Source

M. Mollá et al. (2015). Galactic chemical evolution: stellar yields and the initial mass function. Monthly Notices of the Royal Astronomical Society, 451(4), 3693–3708. https://doi.org/10.1093/mnras/stv1102

Arguments

source_path::String: Path to the directory containing the data.csv file.

Returns

A Tuple with two elements:

A dataframe with all the data.

A dictionary where the keys are the quantity, and the values are the units and if the unit is

inside a logarithm.

GadgetInspector.getRawData — Method

getRawData(
    file_path::String,
    type::Symbol,
    block::String; 
    <keyword arguments>
)::Dict{Symbol,Dict{String,VecOrMat{<:RealOrQty}}}

Extract the raw data from a snapshot file, for a given type of particle and data block.

Arguments

file_path::String: Path to the snapshot file or folder (for multiple-file snapshots).

type::Symbol: Particle type, for which the data will be extracted. The posibilities are given by ParticleType in src/constants.jl.

block::String: Data block to be extracted.

sim_cosmo::Bool = false: If the simulation is cosmological,

false ⟶ Newtonian simulation (ComovingIntegrationOn = 0).

true ⟶ Cosmological simulation (ComovingIntegrationOn = 1).

filter_function::Union{Function,Nothing} = nothing: A function with the signature:

foo(file_path::String)::Vector{Int64}

It indicates which particles will be read, taking the file path to a snapshot and returning the list of indices of the selected particles. If set to nothing, then no particles are filtered. See the GadgetIO.jl documentation for examples.

warnings::Bool = true: If a warning will be given when the data is not as expected, but the function can still run using sane defaults.

Returns

A vector or matrix with the raw data.

GadgetInspector.getSfrFile — Method

getSfrFile(
    source_path::String,
    header::SnapshotHeader; 
    <keyword arguments>
)::Dict{Int32,VecOrMat{<:RealOrQty}}

Get the data from the sfr.txt file.

Arguments

source_path::String: Path to the directory containing the sfr.txt file.

header::SnapshotHeader: Header of one snapshot from the same simulation of the sfr.txt file.

sim_cosmo::Bool = false: If the simulation is cosmological,

false ⟶ Newtonian simulation (ComovingIntegrationOn = 0).

true ⟶ Cosmological simulation (ComovingIntegrationOn = 1).

Returns

A dictionary with as many entries as columns has the sfr.txt file. In the version produced by GADGET3, there are six columns:

1 ⟹ Time.

2 ⟹ Total mass (probability).

3 ⟹ SFR (original GADGET).

4 ⟹ SFR (probability).

5 ⟹ Real total mass.

6 ⟹ Real SFR.

GadgetInspector.getSnapshotData — Method

getSnapshotData(
    file_path::String,
    type_blocks::Dict{Symbol,<:Union{String,Vector{String}}}; 
    <keyword arguments>
)::Dict{Symbol,Dict{String,VecOrMat{<:RealOrQty}}}

Extract the raw data from a snapshot file.

Use this method when you want one or more blocks from more than one type of particle, not necessarily the same from each one.

Arguments

file_path::String: Path to the snapshot file or folder (for multiple-file snapshots).

type_blocks::Dict{Symbol,<:Union{String,Vector{String}}}: Dictionary where the keys are the particle types (the posibilities are given by ParticleType in src/constants.jl), and the values are the data blocks (or lists of data blocks) to be extracted, for each type.

sim_cosmo::Bool = false: If the simulation is cosmological,

false ⟶ Newtonian simulation (ComovingIntegrationOn = 0).

true ⟶ Cosmological simulation (ComovingIntegrationOn = 1).

filter_function::Union{Function,Nothing} = nothing: A function with the signature:

foo(file_path::String)::Vector{Int64}

It indicates which particles will be read, taking the file path to a snapshot and returning the list of indices of the selected particles. If set to nothing, then no particles are filtered. See the GadgetIO.jl documentation for examples.

warnings::Bool = true: If a warning will be given when the data is not as expected, but the function can still run using sane defaults.

Returns

A dictionary with the following structue:

type => (block ⟹ data, block ⟹ data, block ⟹ data, ...), where data is a vector or matrix with the raw data

GadgetInspector.getSnapshotData — Method

getSnapshotData(
    file_path::String,
    type::Symbol,
    block::String; 
    <keyword arguments>
)::Dict{Symbol,Dict{String,VecOrMat{<:RealOrQty}}}

Extract the raw data from a snapshot file.

Use this method when you want one block from one type of particle.

Arguments

file_path::String: Path to the snapshot file or folder (for multiple-file snapshots).

type::Symbol: Particle type, for which the data will be extracted. The posibilities are given by ParticleType in src/constants.jl.

block::String: Data block to be extracted.

sim_cosmo::Bool = false: If the simulation is cosmological,

false ⟶ Newtonian simulation (ComovingIntegrationOn = 0).

true ⟶ Cosmological simulation (ComovingIntegrationOn = 1).

filter_function::Union{Function,Nothing} = nothing: A function with the signature:

foo(file_path::String)::Vector{Int64}

It indicates which particles will be read, taking the file path to a snapshot and returning the list of indices of the selected particles. If set to nothing, then no particles are filtered. See the GadgetIO.jl documentation for examples.

warnings::Bool = true: If a warning will be given when the data is not as expected, but the function can still run using sane defaults.

Returns

A dictionary with the following structue:

type => (block ⟹ Vector or matrix with the raw data).

GadgetInspector.getSnapshotData — Method

getSnapshotData(
    file_path::String,
    type::Symbol,
    blocks::Vector{String}; 
    <keyword arguments>
)::Dict{Symbol,Dict{String,VecOrMat{<:RealOrQty}}}

Extract the raw data from a snapshot file.

Use this method when you want many blocks from the same type of particle.

Arguments

file_path::String: Path to the snapshot file or folder (for multiple-file snapshots).

type::Symbol: Particle type, for which the data will be extracted. The posibilities are given by ParticleType in src/constants.jl.

blocks::Vector{String}: Data blocks to be extracted.

sim_cosmo::Bool = false: If the simulation is cosmological,

false ⟶ Newtonian simulation (ComovingIntegrationOn = 0).

true ⟶ Cosmological simulation (ComovingIntegrationOn = 1).

filter_function::Union{Function,Nothing} = nothing: A function with the signature:

foo(file_path::String)::Vector{Int64}

It indicates which particles will be read, taking the file path to a snapshot and returning the list of indices of the selected particles. If set to nothing, then no particles are filtered. See the GadgetIO.jl documentation for examples.

warnings::Bool = true: If a warning will be given when the data is not as expected, but the function can still run using sane defaults.

Returns

A dictionary with the following structue:

type => (block ⟹ data, block ⟹ data, block ⟹ data, ...), where data is a vector or matrix with the raw data

GadgetInspector.getSnapshotData — Method

getSnapshotData(
    file_path::String,
    types::Vector{Symbol},
    blocks::Union{String,Vector{String}}; 
    <keyword arguments>
)::Dict{Symbol,Dict{String,VecOrMat{<:RealOrQty}}}

Extract the raw data from a snapshot file.

Use this method when you want the same block (or list of blocks) from many types of particles.

Arguments

file_path::String: Path to the snapshot file or folder (for multiple-file snapshots).

types::Vector{Symbol}: Particle types, for which the data will be extracted. The posibilities are given by ParticleType in src/constants.jl.

blocks::Union{String,Vector{String}}: Data block (or list of data blocks) to be extracted.

sim_cosmo::Bool = false: If the simulation is cosmological,

false ⟶ Newtonian simulation (ComovingIntegrationOn = 0).

true ⟶ Cosmological simulation (ComovingIntegrationOn = 1).

filter_function::Union{Function,Nothing} = nothing: A function with the signature:

foo(file_path::String)::Vector{Int64}

It indicates which particles will be read, taking the file path to a snapshot and returning the list of indices of the selected particles. If set to nothing, then no particles are filtered. See the GadgetIO.jl documentation for examples.

warnings::Bool = true: If a warning will be given when the data is not as expected, but the function can still run using sane defaults.

Returns

A dictionary with the following structue:

type => (block ⟹ data, block ⟹ data, block ⟹ data, ...), where data is a vector or matrix with the raw data

GadgetInspector.getSnapshotPaths — Method

function getSnapshotPaths(
    base_name::String,
    source_path::String,
)::Dict{String,Vector{String}}

Find the paths to the GADGET output files and group them by snapshot.

Arguments

base_name::String: Base name of the snapshot files, set in the GADGET variable SnapshotFileBase.

source_path::String: Path to the directory containing the snapshot files, set in the GADGET variable OutputDir.

Returns

A dictionary with two entries:

"snap_numbers" ⟹ The numbers that characterize each snapshot.

"snap_paths" ⟹ The full paths to the snapshot files.

GadgetInspector.getTemperature — Method

getTemperature(
    file_path::String; 
    <keyword arguments>
)::Vector{<:RealOrQty}

Get the temperature of the gas particles in a given snapshot.

Arguments

file_path::String: Path to the snapshot file or folder (for multiple-file snapshots).

sim_cosmo::Bool = false: If the simulation is cosmological,

false ⟶ Newtonian simulation (ComovingIntegrationOn = 0).

true ⟶ Cosmological simulation (ComovingIntegrationOn = 1).

filter_function::Union{Function,Nothing} = nothing: A function with the signature:

foo(file_path::String)::Vector{Int64}

It indicates which particles will be read, taking the file path to a snapshot and returning the list of indices of the selected particles. If set to nothing, then no particles are filtered. See the GadgetIO.jl documentation for examples.

warnings::Bool = true: If a warning will be given when the data is not as expected, but the function can still run using sane defaults.

Returns

The temperature of the gas particles.

GadgetInspector.makeSourceTable — Function

makeSourceTable(
    source::Vector{Dict{String,Vector{String}}},
    idx::IndexType,
    t_unit::Unitful.Units; 
    <keyword arguments>
)::DataFrame

Construct a dataframe with the path, the time stamp, the number, and the global index of each snapshot in a series of simulations.

Arguments

source::Vector{Dict{String,Vector{String}}}: Vector with the output of getSnapshotPaths for every simulation.

idx::IndexType: Indexing of the simulation, i.e. which snapshots will be read. It can be an integer (a single snapshot), a vector of integers (several snapshots), an UnitRange (e.g. 5:13) or : (every snapshot will be read).

t_unit::Unitful.Units: Target unit for the time-stamps.

sim_cosmo::Bool = false: If the simulation is cosmological,

false ⟶ Newtonian simulation (ComovingIntegrationOn = 0).

true ⟶ Cosmological simulation (ComovingIntegrationOn = 1).

Returns

A dataframe, with the following columns:

:id ⟶ global index of each snapshot (it should be the same across simulations)

:time_stamps ⟶ timestamp of each snapshot (it should be the same across simulations)

:numbers ⟶ numbers in the name of each snapshot (it should be the same across simulations)

All the rest are the full paths to the snapshots, one column per simulation