Data Analysis Functions

GadgetInspector.cmdf — Method

cmdf(
    data::Dict{Symbol,<:Any}; 
    <keyword arguments>
)::Union{NTuple{2,Vector{Float64}},Nothing}

Compute the cumulative metallicity distribution function (CMDF).

The CMDF is calculated separating the stellar metallicity in bins windows. For the $n$-th window the CMDF is

\[\sum_{i = 1}^n \dfrac{m_i}{M_T} \quad \mathrm{vs.} \quad \bar{Z}_n \, ,\]

or, for normalize_x = true,

\[\sum_{i = 1}^n \dfrac{m_i}{M_T} \quad \mathrm{vs.} \quad \dfrac{\bar{Z}_n}{\mathrm{max}(\bar{Z}_n)} \, ,\]

where $M_T$ is the total stellar mass, $m_i$ the stellar mass of the $i$-th window and $\bar{Z}_n$ the mean stellar metallicity of the $n$-th window.

Arguments

data::Dict{Symbol,<:Any}: The raw data returned by getSnapshotData. It has to contain the quantities "Z" and "MASS" for the particles of type :stars.

bins::Int64 = 20: Number of metallicity windows to be used for the CMDF.

normalize_x::Bool = false: If the x axis will be normalized to its maximum value or not (default).

solar::Bool = false: If the result will be in units of solar metallicity or not (default).

Returns

A Tuple with two elements:

A Vector with the mean metallicity of each window.

A Vector with the accumulated stellar mass fraction of each window.

It returns nothing in case of empty results or a mismatch in lengths.

GadgetInspector.correlation — Method

correlation(
    data::Dict{Symbol,<:Any},
    x_quantity::Tuple{Symbol,String},
    y_quantity::Tuple{Symbol,String}; 
    <keyword arguments>
)::Union{NTuple{2,Vector{<:RealOrQty}},Nothing}

Compute the result of applying a couple of functions to the output of getSnapshotData.

It is intended for plots of Y vs. X, where Y and X are vector quantities from QUANTITIES or are derived from them.

This method is for analysis functions that take a single quantity as input.

Arguments

data::Dict{Symbol,<:Any}: The raw data returned by getSnapshotData.

x_quantity::Tuple{Symbol,String}: The tuple: (particle type, quantity), where the posible particle types are given by ParticleType and the possible quantities by QUANTITIES.

y_quantity::Tuple{Symbol,String}: The tuple: (particle type, quantity), where the posible particle types are given by ParticleType and the possible quantities by QUANTITIES.

x_function::Function = identity: Function to be aplied to the x quantity. It has to have the signature:

foo(::typeof(data[x_quantity[1]][x_quantity[2]]))::Vector{<:RealOrQty}

y_function::Function = identity: Function to be aplied to the y quantity. It has to have the signature:

foo(::typeof(data[y_quantity[1]][y_quantity[2]]))::Vector{<:RealOrQty}

Returns

A Tuple with two elements:

A Vector with the final values for the x-axis.

A Vector with the final values for the y-axis.

It returns nothing in case of empty results or a mismatch in lengths.

GadgetInspector.correlation — Method

correlation(
    data::Dict{Symbol,<:Any},
    x_quantities::Vector{Tuple{Symbol,String}},
    y_quantities::Vector{Tuple{Symbol,String}}; 
    <keyword arguments>
)::Union{NTuple{2,Vector{<:RealOrQty}},Nothing}

Compute the result of applying a couple of functions to the output of getSnapshotData.

It is intended for plots of Y vs. X, where Y and X are vector quantities from QUANTITIES or are derived from them.

This method is for analysis functions that take several quantities as input.

Arguments

data::Dict{Symbol,<:Any}: The raw data returned by getSnapshotData.

x_quantities::Vector{Tuple{Symbol,String}}: Vector of tuples: (particle type, quantity), where the posible particle types are given by ParticleType and the possible quantities by QUANTITIES.

y_quantities::Vector{Tuple{Symbol,String}}: Vector of tuples: (particle type, quantity), where the posible particle types are given by ParticleType and the possible quantities by QUANTITIES.

x_function::Function = identity: Function to be aplied to the x quantities. It has to have the signature:

foo(
    ::typeof(data[x_quantities[1][1]][x_quantities[1][2]]), 
    ::typeof(data[x_quantities[2][1]][x_quantities[2][2]]),
    ...
)::Vector{<:RealOrQty}

y_function::Function = identity: Function to be applied to the y quantities. It has to have the signature:

foo(
    ::typeof(data[y_quantities[1][1]][y_quantities[1][2]]), 
    ::typeof(data[y_quantities[2][1]][y_quantities[2][2]]),
    ...
)::Vector{<:RealOrQty}

Returns

A Tuple with two elements:

A Vector with the final values for the x-axis.

A Vector with the final values for the y-axis.

It returns nothing in case of empty results or a mismatch in lengths.

GadgetInspector.gasDensityMap — Method

gasDensityMap(
    data::Dict{Symbol,<:Any},
    plane::String,
    box_size::Unitful.Length; 
    <keyword arguments>
)::Tuple{Vector{<:Unitful.Length},Vector{<:Unitful.Length},Matrix{Float64}}

Compute a density map of the gas, in the x-y, x-z, or y-z planes.

Arguments

data::Dict{Symbol,<:Any}: The raw data returned by getSnapshotData. It has to contain

the quantities "MASS", "RHO", "HSML" and "POS", for the particles of type :gas.

plane::String = "All": String indicating which plane will be considered,

"XY" ⟶ x-y plane.

"XZ" ⟶ x-z plane.

"YZ" ⟶ y-z plane.

box_size::Unitful.Quantity: Size of the region of interest if vacuum boundary conditions were used.

sim_cosmo::Bool = false: If the simulation is cosmological,

false ⟶ Newtonian simulation (ComovingIntegrationOn = 0).

true ⟶ Cosmological simulation (ComovingIntegrationOn = 1).

Returns

A Tuple with three elements:

An order list with the abscissa values of the plane.

An order list with the ordinate values of the plane.

A Matrix with the gas density at each point in the plane.

GadgetInspector.histogram — Method

histogram(
    data::Dict{Symbol,<:Any},
    type::Symbol,
    quantity::String;
    <keyword arguments>
)::Union{NTuple{1,Vector{<:RealOrQty}},Nothing}

Compute the result of applying func to the output of getSnapshotData.

This method is for analysis functions that take a single quantity as input.

Arguments

data::Dict{Symbol,<:Any}: The raw data returned by getSnapshotData.

type::Symbol: Particle type, where the possibilities are given by ParticleType.

quantity::String: GADGET quantity, where the possibilities are given by QUANTITIES.

func::Function = identity: Function to be applied to the quantity. It has to have the signature:

foo(::typeof(data[type][quantity]))::Vector{<:RealOrQty}

Returns

A Tuple with one element:

A Vector with the values for each bin.

It returns nothing in case of empty results.

GadgetInspector.histogram — Method

histogram(
    data::Dict{Symbol,<:Any},
    type::Symbol,
    quantities::Vector{String};
    <keyword arguments>
)::Union{NTuple{1,Vector{<:RealOrQty}},Nothing}

Compute the result of applying func to the output of getSnapshotData.

This method is for analysis functions that take several quantities as input.

Arguments

data::Dict{Symbol,<:Any}: The raw data returned by getSnapshotData.

type::Symbol: Particle type. The possibilities are given by ParticleType.

quantities::Vector{String}: GADGET quantities. The possibilities are given by QUANTITIES.

func::Function = identity: Function to be applied to the quantities. It has to have the signature:

foo(
    ::typeof(data[type][quantities[1]]), 
    ::typeof(data[type][quantities[2]]), 
    ...
)::Vector{<:RealOrQty}

Returns

A Tuple with one element:

A Vector with the values for each bin.

It returns nothing in case of empty results.

GadgetInspector.kennicuttSchmidt — Method

kennicuttSchmidt(
    data::Dict{Symbol,<:Any},
    temp_filter::Unitful.Temperature,
    age_filter::Unitful.Time,
    max_r::Unitful.Length; 
    <keyword arguments>
)::Union{NTuple{2,Vector{<:RealOrQty}},Nothing}

Compute the mass area density and the SFR area density for the Kennicutt-Schmidt law.

The area densities are calculated by projecting the positions of the stars and the gas particles to the x-y plane. Then the space is subdivided into bins concentric rings from 0 to a radius of max_r. Each ring is of equal width max_r / bins.

The $n$-th ring has an area of

\[A_n = π \, \mathrm{width}^2 \, (2 \, n - 1) \, .\]

So, the assigned SFR for that ring is

\[\Sigma_\mathrm{SFR}^n = \frac{M_*^n}{\mathrm{age\_filter}\,A_n} \, ,\]

where $M_*^n$ is the total mass of stars younger than age_filter within the ring. Equivalently, the mass area density of the gas is given by

\[\Sigma_\rho^n = \frac{M_\rho^n}{A_n} \, ,\]

where $M_\rho^n$ is the total mass of gas colder than temp_filter within the ring.

Arguments

data::Dict{Symbol,<:Any}: The raw data returned by getSnapshotData. It has to contain the quantities "Z", "MASS", "POS", "U" and "NE" for the particles of type :gas, and the quantities "MASS", "POS" and "AGE" for the particles of type :stars.

temp_filter::Unitful.Temperature: Maximum temperature allowed for the gas particles.

age_filter::Unitful.Time: Maximum age allowed for the stars.

max_r::Unitful.Length: Maximum distance up to which the parameters will be calculated.

bins::Int64 = 20: Number of subdivisions of [0, max_r].

x_unit::Unitful.Units = UnitfulAstro.Msun * UnitfulAstro.pc^-2: Unit for the surface density of gas.

y_unit::Unitful.Units = UnitfulAstro.Msun * UnitfulAstro.yr^-1 * UnitfulAstro.kpc^-2: Unit for the surface density of star formation rate.

x_log::Bool = true: If the x axis will be $\log(\Sigma_\mathrm{gas})$ (default) or just $\Sigma_\mathrm{gas}$.

y_log::Bool = true: If the y axis will be $\log(\Sigma_\mathrm{sfr})$ (default) or just $\Sigma_\mathrm{sfr}$.

Returns

A Tuple with two elements:

A Vector with the gas area density of each ring.

A Vector with the SFR area density of each ring.

It returns nothing in case the results would have less than five data points.

GadgetInspector.particleMap — Method

particleMap(
    data::Dict{Symbol,<:Any},
    plane::String,
    type::Symbol,
    box_size::Unitful.Length;
    <keyword arguments>
)::Tuple{Vector{<:Unitful.Length},Vector{<:Unitful.Length},Matrix{Float64}}

Compute a number density map, for particles of type type in the x-y, x-z, or y-z planes.

Arguments

data::Dict{Symbol,<:Any}: The raw data returned by getSnapshotData. It has to contain the quantity "POS" for the particles of type type.

plane::String = "All": String indicating which plane will be considered,

"XY" ⟶ x-y plane.

"XZ" ⟶ x-z plane.

"YZ" ⟶ y-z plane.

box_size::Unitful.Quantity: Size of the region of interest if vacuum boundary conditions were used.

sim_cosmo::Bool = false: If the simulation is cosmological,

false ⟶ Newtonian simulation (ComovingIntegrationOn = 0).

true ⟶ Cosmological simulation (ComovingIntegrationOn = 1).

Returns

A Tuple with three elements:

An order list with the abscissa values of the plane.

An order list with the ordinate values of the plane.

A Matrix with the number density at each point in the plane.

GadgetInspector.profile — Method

profile(
    data::Dict{Symbol,<:Any},
    quantity::Tuple{Symbol,String},
    max_radius::Unitful.Length,
    bins::Int64; 
    <keyword arguments>
)::Union{NTuple{2,Vector},Nothing}

Compute a 2D or 3D profile of quantity[2], or of something derived from it.

Arguments

data::Dict{Symbol,<:Any}: The raw data returned by getSnapshotData. It has to contain the quantity "POS" and quantity[2], for the particles of the type specified in quantity[1].

quantity::Tuple{Symbol,String}: The tuple: (particle type, quantity), where the possible particle types are given by ParticleType and the possible quantities by QUANTITIES.

max_radius::Unitful.Length: Maximum distance up to which the profile will be calculated.

bins::Int64: Number of subdivisions of [0, max_radius].

flat::Bool = true: If the profile will be 2D (default) or 3D.

func::Function = identity: Function to be applied to the quantity. It has to have the signature:

foo(::typeof(data[quantity[1]][quantity[2]]))::Vector{<:Unitful.Quantity}

cumulative::Bool = false: If the profile will be accumulated or not (default).

density::Bool = false: If the profile is of the area/volume density of quantity[2] or just of quantity[2] itself (default).

Returns

A Tuple with two elements:

A Vector with the distances to each ring or shell.

A Vector with the values of the profile at those distances.

It returns nothing in case of empty results.

GadgetInspector.qtyEvolution — Method

qtyEvolution(
    sim_data::SimData,
    x_qty::String,
    y_qty::String,
    type::Union{Symbol,Nothing};
    <keyword arguments>
)::NTuple{2,Vector}

Compute a time series of y_qty and x_qty.

Arguments

sim_data::SimData: Information about the simulation in a SimData object.

x_qty::String: Quantity for the x axis. The possibilities are given by the computeTimeSeries function:

"clock_time" ⟶ Phisical time associated with each snapshot (dimensions: time).

"scale_factor" ⟶ Scale factor associated with each snapshot (dimensionless).

"redshift" ⟶ Redshift associated with each snapshot (dimensionless).

"number" ⟶ Number of particles of a given type (dimensionless).

"mass" ⟶ Total mass of a given type of particle (dimensions: mass).

"sfr" ⟶ Star formation rate in each snapshot (dimensions: mass over time).

y_qty::String: Quantity for the y axis. The possibilities are given by the computeTimeSeries function:

"clock_time" ⟶ Phisical time associated with each snapshot (dimensions: time).

"scale_factor" ⟶ Scale factor associated with each snapshot (dimensionless).

"redshift" ⟶ Redshift associated with each snapshot (dimensionless).

"number" ⟶ Number of particles of a given type (dimensionless).

"mass" ⟶ Total mass of a given type of particle (dimensions: mass).

"sfr" ⟶ Star formation rate in each snapshot (dimensions: mass over time).

x_type::Union{Symbol,Nothing}: Particle type for the x axis. The posibilities are given by ParticleType in src/constants.jl. It is only relevant for the "number" and "mass" quantities.

y_type::Union{Symbol,Nothing}: Particle type for the y axis. The posibilities are given by ParticleType in src/constants.jl. It is only relevant for the "number" and "mass" quantities.

sim_cosmo::Bool = false: If the simulation is cosmological,

false ⟶ Newtonian simulation (ComovingIntegrationOn = 0).

true ⟶ Cosmological simulation (ComovingIntegrationOn = 1).

smooth::Union{Int64,Nothing} = nothing: The amount of bins to smooth out the output, if nothing no smoothing will be done, which is the default.

warnings::Bool = true: If a warning will be given when the data is not as expected, but the function can still run using sane defaults.

Returns

A Tuple with two elements:

A Vector with the time series of x_qty.

A Vector with the time series of y_qty.

GadgetInspector.zProfile — Method

zProfile(
    data::Dict{Symbol,<:Any},
    type::Symbol,
    max_radius::Unitful.Length,
    bins::Int64;
    <keyword arguments>
)::Union{NTuple{2,Vector},Nothing}

Compute a 2D or 3D metallicity profile.

Arguments

data::Dict{Symbol,<:Any}: The raw data returned by getSnapshotData. It has to contain the quantities "Z", "MASS" and "POS", for the particles of the type specified in type.

type::Symbol: Particle type. The possibilities are given by ParticleType.

max_radius::Unitful.Length: Maximum distance up to which the profile will be calculated.

bins::Int64: Number of subdivisions of [0, max_radius].

flat::Bool = true: If the profile will be 2D (default) or 3D.

cumulative::Bool = false: If the profile will be accumulated or not (default).

density::Bool = false: If the profile is of the metallicity area/volume density or just of the metallicity itself (default).

solar::Bool = false: If the result will be in units of solar metallicity or not (default).

Returns

A Tuple with two elements:

A Vector with the distances to each ring or shell.

A Vector with the values of the profile at those distances.

It returns nothing in case of empty results.