Pipelines

GadgetInspector.snapshotPlot — Method

snapshotPlot(
    source_paths::Vector{String},
    base_names::Vector{String},
    quantities::Dict{Symbol,<:Union{String,Vector{String}}},
    plot_function!::Function; 
    <keyword arguments>
)::Nothing

Generate one figure per snapshot, for one or more simulations.

Some of the features are:

It can produce scatter plots, line plots, histograms, and density maps. The last two are only for one simulation at a time.

It can produce a GIF or an MP4 animating the results.

It transparently manages units, you only have to indicate the final unit of each axis.

Many aspects of the plots can be configured, see the arguments below for details.

Arguments

source_paths::Vector{String}: Paths to the folders containing the snapshots of each simulation, set in the GADGET variable OutputDir.

base_names::Vector{String}: Base names of the snapshot files, set in the GADGET variable SnapshotFileBase.

quantities::Dict{Symbol,<:Union{String,Vector{String}}}: Dictionary where the keys are the types of particles (the possibilities are given by ParticleType in src/constants.jl), and the values are the data blocks to be read. Which data blocks are nedded depends on the provided function data_analysis.

plot_function!::Function: A plotting function from Makie.jl. The supported functions are:

hist! ⟶ Histograms.

scatter! ⟶ Scatter plots.

lines! ⟶ Line plots.

scatterlines! ⟶ Markers with lines between them.

heatmap! ⟶ Heatmaps.

This sets the type of plot for the output.

Data manipulation options

pf_kwargs::NamedTuple = (;): Keyword arguments for the plot_function! function.

sim_cosmo::Bool = false: If the simulation is cosmological.

false ⟶ Newtonian simulation (ComovingIntegrationOn = 0).

true ⟶ Cosmological simulation (ComovingIntegrationOn = 1).

idx::IndexType = :: Indexing of the simulation, i.e. indices of the snapshots that will be read. It can be an integer (a single snapshot), a vector of integers (several given snapshots), an UnitRange (e.g. 5:13) or : (every snapshot will be read).

filter_functions::Vector{<:Union{Function,Nothing}} = [nothing]: List of functions (one per simulation), all with the signature:

foo(file_path::String)::Vector{Int64}

Each function indicates which particles will be read in each simulation, taking the file path to a snapshot and returning the list of indices of the selected particles. If set to nothing, then every particle is read. See the GadgetIO.jl documentation for examples.

da_functions::Vector{<:Function} = [trivial]: Optional vector of data analysis functions. See the required signature and examples in src/data_analysis.jl.

da_args::Vector{<:Tuple} = [()]: Vector of positional arguments for the da_functions functions.

da_kwargs::Vector{<:NamedTuple} = [(;)]: Vector of keyword arguments for the da_functions functions.

post_processing!::Function = trivial: Optional post processing function. See the required signature and examples in src/post_processing.jl.

pp_args::Tuple = (): Positional arguments for the post_processing! function.

pp_kwargs::NamedTuple = (;): Keyword arguments for the post_processing! function.

Axes options

x_label::String = ":auto_label": Label for the x axis. It can contain the string :auto_label, which will be replaced by the automatic label that would otherwise be used. The default label has the shape: x_name / 10^x_factor x_unit.

y_label::String = ":auto_label": Label for the y axis. It can contain the string :auto_label, which will be replaced by the automatic label that would otherwise be used. The default label has the shape: y_name / 10^y_factor y_unit.

x_name::Union{String,LaTeXString} = "": Name of the variable for the x axis. It should not include units or scaling factors. If left empty, the automatic label will be an empty string.

y_name::Union{String,LaTeXString} = "": Name of the variable for the y axis. It should not include units or scaling factors. If left empty, the automatic label will be an empty string.

x_unit::Unitful.Units = Unitful.NoUnits: Target unit for the x axis. The values will be converted accordingly. Leave the default value of Unitful.NoUnits for dimensionless quantities.

y_unit::Unitful.Units = Unitful.NoUnits: Target unit for the y axis. The values will be converted accordingly. Leave the default value of Unitful.NoUnits for dimensionless quantities.

x_factor::Int64 = 0: Numerical exponent to scale the x axis, e.g. if x_factor = 10 the values will be scaled by $10^{10}$. The default is no scaling.

y_factor::Int64 = 0: Numerical exponent to scale the y axis, e.g. if y_factor = 10 the values will be scaled by $10^{10}$. The default is no scaling.

x_scale::Union{Function,Makie.Symlog10} = identity: Scaling function for the x axis. The options are the scaling functions accepted by Makie.jl: log10, log2, log, sqrt, Makie.logit, Makie.Symlog10, Makie.pseudolog10 and identity.

y_scale::Union{Function,Makie.Symlog10} = identity: Scaling function for the y axis. The options are the scaling functions accepted by Makie.jl: log10, log2, log, sqrt, Makie.logit, Makie.Symlog10, Makie.pseudolog10 and identity.

x_limits::Tuple{<:Union{Real,Nothing},<:Union{Real,Nothing}} = (nothing, nothing): Set it to a value diferent than nothing if you want to fix the limits of the x axis.

y_limits::Tuple{<:Union{Real,Nothing},<:Union{Real,Nothing}} = (nothing, nothing): Set it to a value diferent than nothing if you want to fix the limits of the y axis.

x_range::NTuple{2,<:Real} = (-Inf, Inf): The data will be trim down so its x coordinates fit within x_range.

y_range::NTuple{2,<:Real} = (-Inf, Inf): The data will be trim down so its y coordinates fit within y_range. Notice that this option does not affect histograms.

x_edges::Bool = false: Set it to true if you want to keep the borders of x_range.

y_edges::Bool = false: Set it to true if you want to keep the borders of y_range.

x_func::Function = identity: Function to be applied to the values of the x axis. It has to be a pure function with the signature:foo(x_data::Vector{Float64})::Vector{Float64}And the output vector has to have the same length as the input one. Notice that this function will be applied regardless of units and possible domain problems (use x_range to solve incompatibilities), and that it will not be reflected in the automatic label.

y_func::Function = identity: Function to be applied to the values of the y axis. It has to be a pure function with the signature:foo(y_data::Vector{Float64})::Vector{Float64}And the output vector has to have the same length as the input one. Notice that this function will be applied regardless of units and possible domain problems (use y_range to solve incompatibilities), and that it will not be reflected in the automatic label.

Plotting and animation options

output_path::String = "snapshot_plots": Path to the output folder. The figures will be stored in output_path/plots/ and will be named snapshot_XXXoutput_format where XXX is the snapshot number.

sim_labels::Union{Vector{String},Nothing} = basename.(source_paths): Labels for the plot legend, one per simulation. Set it to nothing if you don't want a legend.

title::Union{String,Symbol} = "": Title for the figures. If left empty, no title will be printed. If set to :time, the time stamp of the snapshot will be used as title.

t_unit::Unitful.Units = UnitfulAstro.Myr: Target unit for the time stamps of the snapshots (only relevant if the option :time was given for the argument title).

output_format::String = ".svg": File format of the output figures. All formats supported by Makie.jl can be used, namely .pdf, .svg and .png.

pt_per_unit::Float64 = 0.75: Factor to scale up or down the size of the figures, keeping their proportions. It only works for .pdf and .svg.

px_per_unit::Float64 = 1.0: Factor to scale up or down the size of the figures, keeping their proportions. It only works for .png.

resolution::NTuple{2,Int64} = (1000, 750): Resolution of the figures in points. By default, for PNGs: points = pixels (as given by px_per_unit = 1.0), and for PDFs and SVGs: points = 0.75 * pixels (as given by pt_per_unit = 0.75).

aspect::Union{DataAspect,AxisAspect,Nothing} = nothing: Aspect ratio of the figures. The options are:

nothing: Use the default given by Makie.jl.

AxisAspect(n): The aspect ratio will be given by the number n = width / height.

DataAspect(): The aspect ratio of the data is used.

hist_bins::Int64 = 20: Number of bins for the histograms.

series_colors::Union{Vector{<:ColorInput},Nothing} = nothing: Colors for the diferent series to be plotted (diferent simulations). If set to nothing, the colors will be assigned automatically. This is only relevant for scatter! and scatterlines! plots.

series_markers::Union{Vector{Symbol},Nothing} = nothing: Markers for the different series to be plotted (different simulations). If set to nothing, the markers will be assigned automatically. This is only relevant for scatter! and scatterlines! plots.

series_linetypes::Union{Vector{<:LineStyleInput},Nothing} = nothing: Line types for the different series to be plotted (different simulations). If set to nothing, the line types will be assigned automatically. This is only relevant for lines! and scatterlines! plots.

animation::Bool = true: If an animation of the results will be made and saved in output_path.

anim_file::String = "animation.gif": Filename for the animation, including its extension. All formats supported by Makie.jl can be used, namely .mkv, .mp4, .webm and .gif.

framerate::Int64 = 10: Frame rate of the animation.

warnings::Bool = true: If a warning will be raised when the data is not as expected, but the function can still run using sane defaults.

show_progress::Bool = true: If a progress bar will be shown.

GadgetInspector.snapshotTable — Method

snapshotTable(
    source_paths::Vector{String},
    base_names::Vector{String},
    quantities::Dict{Symbol,<:Union{String,Vector{String}}}; 
    <keyword arguments>
)::Nothing

Generate one table per snapshot, for one or more simulations.

It transparently manages units, you only have to indicate the final unit of each axis.

The output can be a human-readable TXT or a CSV. In both cases the columns in each snapshot table are:

x quantity for simulation 1

y quantity for simulation 1

x quantity for simulation 2

y quantity for simulation 2

x quantity for simulation 3

y quantity for simulation 3

etc.

Arguments

source_paths::Vector{String}: Paths to the folders containing the snapshots of each simulation, set in the GADGET variable OutputDir.

base_names::Vector{String}: Base names of the snapshot files, set in the GADGET variable SnapshotFileBase.

quantities::Dict{Symbol,<:Union{String,Vector{String}}}: Dictionary where the keys are the types of particles (the possibilities are given by ParticleType in src/constants.jl), and the values are the data blocks to be read. Which data blocks are nedded depends on the provided function data_analysis.

human_redable::Bool = true: If the results will be written in a human readable format. The output file in this case will be a .txt file.

machine_redable::Bool = true: If the results will be written in a machine readable format. The output file in this case will be a .csv file.

output_path::String = "snapshot_tables": Path to the output folder.

sim_labels::Vector{String}} = basename.(source_paths): Labels for the simulations.

title::String = "": Title to be written at the top of the file, only aplicable to the .txt file.

warnings::Bool = true: If a warning will be raised when the data is not as expected, but the function can still run using sane defaults.

show_progress::Bool = true: If a progress bar will be shown.

sim_cosmo::Bool = false: If the simulation is cosmological.

false ⟶ Newtonian simulation (ComovingIntegrationOn = 0).

true ⟶ Cosmological simulation (ComovingIntegrationOn = 1).

Data manipulation options

idx::IndexType = :: Indexing of the simulation, i.e. indices of the snapshots that will be read. It can be an integer (a single snapshot), a vector of integers (several given snapshots), an UnitRange (e.g. 5:13) or : (every snapshot will be read).

filter_functions::Vector{<:Union{Function,Nothing}} = [nothing]: List of functions (one per simulation), all with the signature:

foo(file_path::String)::Vector{Int64}

Each function indicates which particles will be read in each simulation, taking the file path to a snapshot and returning the list of indices of the selected particles. If set to nothing, then every particle is read. See the GadgetIO.jl documentation for examples.

da_functions::Vector{<:Function} = [trivial]: Optional vector of data analysis functions. See the required signature and examples in src/data_analysis.jl.

da_args::Vector{<:Tuple} = [()]: Vector of positional arguments for the da_functions functions.

da_kwargs::Vector{<:NamedTuple} = [(;)]: Vector of keyword arguments for the da_functions functions.

Axes options

x_name::Union{String,LaTeXString} = "X": Name of the variable for the x axis. It should not include units or scaling factors.

y_name::Union{String,LaTeXString} = "Y": Name of the variable for the y axis. It should not include units or scaling factors.

xunit_label::String = "[:auto_label]": Label indicating the unit of the x axis. It can contain the string ":auto_label", which will be replaced by the automatic unit label that would otherwise be used.

yunit_label::String = "[:auto_label]": Label indicating the unit of the x axis. It can contain the string ":auto_label", which will be replaced by the automatic unit label that would otherwise be used.

x_unit::Unitful.Units = Unitful.NoUnits: Target unit for the x axis. The values will be converted accordingly. Leave the default value of Unitful.NoUnits for dimensionless quantities.

y_unit::Unitful.Units = Unitful.NoUnits: Target unit for the y axis. The values will be converted accordingly. Leave the default value of Unitful.NoUnits for dimensionless quantities.

x_factor::Int64 = 0: Numerical exponent to scale the x axis, e.g. if x_factor = 10 the values will be scaled by $10^{10}$. The default is no scaling.

y_factor::Int64 = 0: Numerical exponent to scale the y axis, e.g. if y_factor = 10 the values will be scaled by $10^{10}$. The default is no scaling.

x_range::NTuple{2,<:Real} = (-Inf, Inf): The data will be trim down so its x coordinates fit within x_range.

y_range::NTuple{2,<:Real} = (-Inf, Inf): The data will be trim down so its y coordinates fit within y_range. Notice that this option does not affect histograms.

x_edges::Bool = false: Set it to true if you want to keep the borders of x_range.

y_edges::Bool = false: Set it to true if you want to keep the borders of y_range.

x_func::Function = identity: Function to be applied to the values of the x axis. It has to be a pure function with the signature:foo(x_data::Vector{Float64})::Vector{Float64}And the output vector has to have the same length as the input one. Notice that this function will be applied regardless of units and possible domain problems (use x_range to solve incompatibilities), and that it will not be reflected in the automatic label.

y_func::Function = identity: Function to be applied to the values of the y axis. It has to be a pure function with the signature:foo(y_data::Vector{Float64})::Vector{Float64}And the output vector has to have the same length as the input one. Notice that this function will be applied regardless of units and possible domain problems (use y_range to solve incompatibilities), and that it will not be reflected in the automatic label.

GadgetInspector.timeSeriesPlot — Method

timeSeriesPlot(
    source_paths::Vector{String},
    base_names::Vector{String},
    plot_function!::Function; 
    <keyword arguments>
)::Nothing

Generate a figure with the time evolution of some quantity, for one or more simulations.

Some of the features are:

It can produce a scatter plot, a line plot, or a combination of both.

It transparently manages units, you only have to indicate the final unit of each axis.

Many aspects of the plot can be configured, see the arguments below for details.

Arguments

source_paths::Vector{String}: Paths to the folders containing the snapshots of each simulation, set in the GADGET variable OutputDir.

base_names::Vector{String}: Base names of the snapshot files, set in the GADGET variable SnapshotFileBase.

plot_function!::Function: A plotting function from Makie.jl. The supported functions are:

scatter! ⟶ Scatter plots.

lines! ⟶ Line plots.

scatterlines! ⟶ Markers with lines between them.

This sets the type of plot for the output.

Data manipulation options

pf_kwargs::NamedTuple = (;): Keyword arguments for the plot_function! function.

idx::IndexType = :: Indexing of the simulation, i.e. indices of the snapshots that will be read. It can be an integer (a single snapshot), a vector of integers (several given snapshots), an UnitRange (e.g. 5:13) or : (every snapshot will be read).

filter_functions::Vector{<:Union{Function,Nothing}} = [nothing]: List of functions (one per simulation), all with the signature:

foo(file_path::String)::Vector{Int64}

Each function indicates which particles will be read in each simulation, taking the file path to a snapshot and returning the list of indices of the selected particles. If set to nothing, then every particle is read. See the GadgetIO.jl documentation for examples.

da_functions::Vector{<:Function} = [trivial]: Optional vector of data analysis functions. See the required signature and examples in src/data_analysis.jl.

da_args::Vector{<:Tuple} = [()]: Vector of positional arguments for the da_functions functions.

da_kwargs::Vector{<:NamedTuple} = [(;)]: Vector of keyword arguments for the da_functions functions.

post_processing!::Function = trivial: Optional post processing function. See the required signature and examples in src/post_processing.jl.

pp_args::Tuple = (): Positional arguments for the post_processing! function.

pp_kwargs::NamedTuple = (;): Keyword arguments for the post_processing! function.

Axes options

x_label::String = ":auto_label": Label for the x axis. It can contain the string :auto_label, which will be replaced by the automatic label that would otherwise be used. The default label has the shape: x_name / 10^x_factor x_unit.

y_label::String = ":auto_label": Label for the y axis. It can contain the string :auto_label, which will be replaced by the automatic label that would otherwise be used. The default label has the shape: y_name / 10^y_factor y_unit.

x_name::Union{String,LaTeXString} = "": Name of the variable for the x axis. It should not include units or scaling factors. If left empty, the automatic label will be an empty string.

y_name::Union{String,LaTeXString} = "": Name of the variable for the y axis. It should not include units or scaling factors. If left empty, the automatic label will be an empty string.

x_unit::Unitful.Units = Unitful.NoUnits: Target unit for the x axis. The values will be converted accordingly. Leave the default value of Unitful.NoUnits for dimensionless quantities.

y_unit::Unitful.Units = Unitful.NoUnits: Target unit for the y axis. The values will be converted accordingly. Leave the default value of Unitful.NoUnits for dimensionless quantities.

x_factor::Int64 = 0: Numerical exponent to scale the x axis, e.g. if x_factor = 10 the values will be scaled by $10^{10}$. The default is no scaling.

y_factor::Int64 = 0: Numerical exponent to scale the y axis, e.g. if y_factor = 10 the values will be scaled by $10^{10}$. The default is no scaling.

x_scale::Union{Function,Makie.Symlog10} = identity: Scaling function for the x axis. The options are the scaling functions accepted by Makie.jl: log10, log2, log, sqrt, Makie.logit, Makie.Symlog10, Makie.pseudolog10 and identity.

y_scale::Union{Function,Makie.Symlog10} = identity: Scaling function for the y axis. The options are the scaling functions accepted by Makie.jl: log10, log2, log, sqrt, Makie.logit, Makie.Symlog10, Makie.pseudolog10 and identity.

x_limits::Tuple{<:Union{Real,Nothing},<:Union{Real,Nothing}} = (nothing, nothing): Set it to a value diferent than nothing if you want to fix the limits of the x axis.

y_limits::Tuple{<:Union{Real,Nothing},<:Union{Real,Nothing}} = (nothing, nothing): Set it to a value diferent than nothing if you want to fix the limits of the y axis.

x_range::NTuple{2,<:Real} = (-Inf, Inf): The data will be trim down so its x coordinates fit within x_range.

y_range::NTuple{2,<:Real} = (-Inf, Inf): The data will be trim down so its y coordinates fit within y_range. Notice that this option does not affect histograms.

x_edges::Bool = false: Set it to true if you want to keep the borders of x_range.

y_edges::Bool = false: Set it to true if you want to keep the borders of y_range.

x_func::Function = identity: Function to be applied to the values of the x axis. It has to be a pure function with the signature:foo(x_data::Vector{Float64})::Vector{Float64}And the output vector has to have the same length as the input one. Notice that this function will be applied regardless of units and possible domain problems (use x_range to solve incompatibilities), and that it will not be reflected in the automatic label.

y_func::Function = identity: Function to be applied to the values of the y axis. It has to be a pure function with the signature:foo(y_data::Vector{Float64})::Vector{Float64}And the output vector has to have the same length as the input one. Notice that this function will be applied regardless of units and possible domain problems (use y_range to solve incompatibilities), and that it will not be reflected in the automatic label.

Plotting and animation options

output_path::String = "time_series_plots"": Path to the output folder.

file_name::String = "figure.svg": Filename for the output figure with its extension. All formats supported by Makie.jl can be used, namely .pdf, .svg and .png.

sim_labels::Union{Vector{String},Nothing} = basename.(source_paths): Labels for the plot legend, one per simulation. Set it to nothing if you don't want a legend.

title::String = "": Title for the figures. If left empty, no title will be printed.

pt_per_unit::Float64 = 0.75: Factor to scale up or down the size of the figures, keeping their proportions. It only works for .pdf and .svg.

px_per_unit::Float64 = 1.0: Factor to scale up or down the size of the figures, keeping their proportions. It only works for .png.

resolution::NTuple{2,Int64} = (1000, 750): Resolution of the figures in points. By default, for PNGs: points = pixels (as given by px_per_unit = 1.0), and for PDFs and SVGs: points = 0.75 * pixels (as given by pt_per_unit = 0.75).

aspect::Union{DataAspect,AxisAspect,Nothing} = nothing: Aspect ratio of the figures. The options are:

nothing: Use the default given by Makie.jl.

AxisAspect(n): The aspect ratio will be given by the number n = width / height.

DataAspect(): The aspect ratio of the data is used.

series_colors::Union{Vector{<:ColorInput},Nothing} = nothing: Colors for the diferent series to be plotted (diferent simulations). If set to nothing, the colors will be assigned automatically. This is only relevant for scatter! and scatterlines! plots.

series_markers::Union{Vector{Symbol},Nothing} = nothing: Markers for the different series to be plotted (different simulations). If set to nothing, the markers will be assigned automatically. This is only relevant for scatter! and scatterlines! plots.

series_linetypes::Union{Vector{<:LineStyleInput},Nothing} = nothing: Line types for the different series to be plotted (different simulations). If set to nothing, the line types will be assigned automatically. This is only relevant for lines! and scatterlines! plots.

GadgetInspector.timeSeriesTable — Method

timeSeriesTable(
    source_paths::Vector{String},
    base_names::Vector{String}; 
    <keyword arguments>
)::Nothing

Generate a table with the time evolution of some quantity, for one or more simulations.

It transparently manages units, you only have to indicate the final unit of each axis.

The output can be a human-readable TXT or a CSV. In both cases the columns are:

x quantity for simulation 1

y quantity for simulation 1

x quantity for simulation 2

y quantity for simulation 2

x quantity for simulation 3

y quantity for simulation 3

etc.

Arguments

source_paths::Vector{String}: Paths to the folders containing the snapshots of each simulation, set in the GADGET variable OutputDir.

base_names::Vector{String}: Base names of the snapshot files, set in the GADGET variable SnapshotFileBase.

human_redable::Bool = true: If the results will be written in a human readable format. The output file in this case will be a .txt file.

machine_redable::Bool = true: If the results will be written in a machine readable format. The output file in this case will be a .csv file.

output_path::String = "time_series_table": Path to the output folder.

sim_labels::Vector{String}} = basename.(source_paths): Labels for the simulations.

file_name::String = "table": Name of the output files, without extension.

title::String = "": Title to be written at the top of the file, only aplicable to the .txt file.

Data manipulation options

idx::IndexType = :: Indexing of the simulation, i.e. indices of the snapshots that will be read. It can be an integer (a single snapshot), a vector of integers (several given snapshots), an UnitRange (e.g. 5:13) or : (every snapshot will be read).

filter_functions::Vector{<:Union{Function,Nothing}} = [nothing]: List of functions (one per simulation), all with the signature: foo(file_path::String)::Vector{Int64}Each function indicates which particles will be read in each simulation, taking the file path to a snapshot and returning the list of indices of the selected particles. If set to nothing, then every particle is read. See the GadgetIO.jl documentation for examples.

da_functions::Vector{<:Function} = [trivial]: Optional vector of data analysis functions. See the required signature and examples in src/data_analysis.jl.

da_args::Vector{<:Tuple} = [()]: Vector of positional arguments for the da_functions functions.

da_kwargs::Vector{<:NamedTuple} = [(;)]: Vector of keyword arguments for the da_functions functions.

Axes options

x_name::Union{String,LaTeXString} = "X": Name of the variable for the x axis. It should not include units or scaling factors.

y_name::Union{String,LaTeXString} = "Y": Name of the variable for the y axis. It should not include units or scaling factors.

xunit_label::String = "[:auto_label]": Label indicating the unit of the x axis. It can contain the string :auto_label, which will be replaced by the automatic unit label that would otherwise be used.

yunit_label::String = "[:auto_label]": Label indicating the unit of the x axis. It can contain the string :auto_label, which will be replaced by the automatic unit label that would otherwise be used.

x_unit::Unitful.Units = Unitful.NoUnits: Target unit for the x axis. The values will be converted accordingly. Leave the default value of Unitful.NoUnits for dimensionless quantities.

y_unit::Unitful.Units = Unitful.NoUnits: Target unit for the y axis. The values will be converted accordingly. Leave the default value of Unitful.NoUnits for dimensionless quantities.

x_factor::Int64 = 0: Numerical exponent to scale the x axis, e.g. if x_factor = 10 the values will be scaled by $10^{10}$. The default is no scaling.

y_factor::Int64 = 0: Numerical exponent to scale the y axis, e.g. if y_factor = 10 the values will be scaled by $10^{10}$. The default is no scaling.

x_range::NTuple{2,<:Real} = (-Inf, Inf): The data will be trim down so its x coordinates fit within x_range.

y_range::NTuple{2,<:Real} = (-Inf, Inf): The data will be trim down so its y coordinates fit within y_range. Notice that this option does not affect histograms.

x_edges::Bool = false: Set it to true if you want to keep the borders of x_range.

y_edges::Bool = false: Set it to true if you want to keep the borders of y_range.

x_func::Function = identity: Function to be applied to the values of the x axis. It has to be a pure function with the signature:foo(x_data::Vector{Float64})::Vector{Float64}And the output vector has to have the same length as the input one. Notice that this function will be applied regardless of units and possible domain problems (use x_range to solve incompatibilities), and that it will not be reflected in the automatic label.

y_func::Function = identity: Function to be applied to the values of the y axis. It has to be a pure function with the signature:foo(y_data::Vector{Float64})::Vector{Float64}And the output vector has to have the same length as the input one. Notice that this function will be applied regardless of units and possible domain problems (use y_range to solve incompatibilities), and that it will not be reflected in the automatic label.