Post Processing Functions

GadgetInspector.compareExperiment! — Method

compareExperiment!(
    figure::Makie.Figure,
    quantity::String,
    x_unit::Unitful.Units,
    y_unit::Unitful.Units; 
    <keyword arguments>
)::Tuple{Vector{<:LegendElement},Vector{String}}

Draw the experimental data compiled by Mollá et al. (2015).

Arguments

figure::Makie.Figure: Figure to be modified.

quantity:String: The physical quantity for the y axis. The options are "ΣHI", "ΣHII", "logΣ*", "logΣsfr", "C/H", "N/H", "O/H". See the documentation on the function getMollá2015 in src/data_acquisition.jl for details. The x axis will be the radial distance.

x_unit::Unitful.Units: Unit of lenght for the x axis of figure.

y_unit::Unitful.Units: Unit of the quantity in the y axis of figure. If the quantity is inside a logarithm, you have to put the unit used to normalize it.

source_path::String = "./": Path to the directory containing the data.csv file.

color::ColorInput = :red: The color of the markers and error whiskers.

markersize::Int64 = 6: Size of the markers.

whiskerwidth::Int64 = 10: Width of the error whiskers.

Returns

Tuple with two elements:

MarkerElement to be used as markers in the legend.

Labels.

GadgetInspector.cross! — Method

cross!(
    figure::Makie.Figure; 
    <keyword arguments>
)::Nothing

Draw two lines, one horizontal and one vertical, crossing at cross.

Arguments

figure::Makie.Figure: Figure to be modified.

cross::NTuple{2,Float64} = (0.0, 0.0): Crossing point of the lines.

color::Symbol = :black: Color of the lines.

GadgetInspector.ksLine! — Method

ksLine!(
    figure::Makie.Figure; 
    <keyword arguments>
)::Tuple{Vector{<:LegendElement},Vector{String}}

Draw two line plots and three lines of text.

One line plot is the fit of the data in figure, while the other is the result of Kennicutt (1998). The lines of text will be drawn in the upper left corner of the plot, the first with the equation

\[\Sigma_\mathrm{SFR} = \mathrm{A}\,\Sigma_\mathrm{gas}^{\,\mathrm{N}} \, ,\]

and the second and third giving the values of $\mathrm{N}$ and $\mathrm{log}_{10}(\mathrm{A})$ respectively.

Note

x_unit, y_unit, x_log and y_log have to be set with the same values as they have been in kennicuttSchmidt, otherwise the results will not be correct.

Arguments

figure::Makie.Figure: Figure to be modified.

x_unit::Unitful.Units = UnitfulAstro.Msun * UnitfulAstro.pc^-2: Unit of the surface density of gas used in figure.

y_unit::Unitful.Units = UnitfulAstro.Msun * UnitfulAstro.yr^-1 * UnitfulAstro.kpc^-2: Unit of the surface density of star formation rate used in figure.

x_log::Bool = true: If the x axis will be $\log(\Sigma_\mathrm{gas})$ (default) or just $\Sigma_\mathrm{gas}$.

y_log::Bool = true: If the y axis will be $\log(\Sigma_\mathrm{sfr})$ (default) or just $\Sigma_\mathrm{sfr}$.

error_formating::String = "std_error": What to print as the error of the fit. The options are:

"std_error" ⟶ mean ± standard_error.

"conf_interval" ⟶ mean ± max(upper$_{95\%}$ - mean, mean - lower$_{95\%}$).

colors::NTuple{2,ColorInput} = (:red, :blue): The color of the two line plots. The first is for the result of Kennicutt (1998) and the second is for the linear fit.

linestyles::NTuple{2,LineStyleInput} = (nothing, :dash): The line styles of the two line plots. The first is for the result of Kennicutt (1998) and the second is for the linear fit.

Returns

Tuple with two elements:

LineElement to be used as markers in the legend.

Labels.

GadgetInspector.verticalFlags! — Method

verticalFlags!(
    figure::Makie.Figure,
    flags::Tuple{Vector{<:Real},Vector{<:AbstractString}}, 
)::Union{Tuple{Vector{<:LegendElement},Vector{String}},Nothing}

Draw vertical lines at specified positions.

Arguments

figure::Makie.Figure: Figure to be modified.

flags::Tuple{Vector{<:Real},Vector{<:AbstractString}}: The first vector has the positions of the vertical lines. The second has the corresponding labels.

Returns

Tuple with two vectors:

LineElement to be used as markers in the legend.

Labels.