Utilities

Base.isempty — Method

Extension of Base.isempty to check for empty LaTeXStrings.

GadgetInspector.absCoor — Method

absCoor(
    pl::Union{Makie.FigureAxisPlot,Makie.Axis,Makie.Figure},
    rx::Float64,
    ry::Float64,
)::NTuple{2,Float64}

Give the absolute x and y coordinates from the relative ones.

Arguments

pl::Union{Makie.FigureAxisPlot, Makie.Axis, Makie.Figure}: Plot, axis, or figure for which the absolute coordinates will be calculated. In the case of a figure, it will use the limits from the current axis object.

rx::Float64: relative x coordinate, rx ∈ [0, 1].

ry::Float64: relative y coordinate, ry ∈ [0, 1].

Returns

A tuple with the absolute coordinates: (x, y).

Examples

julia> GadgetInspector.absCoor(lines(rand(100)), 0.5, 0.5)
(50.49999666213989, 0.5031078979372978)

GadgetInspector.checkDataShape — Method

checkDataShape(
    ::Union{typeof(scatter!),typeof(scatterlines!),typeof(lines!)}, 
    data_length::Int64, 
    ::Int64,
)::Nothing

Throw an error if data_length is not compatible with the plot type (scatter/lines for this method).

Arguments

::Union{typeof(scatter!),typeof(scatterlines!),typeof(lines!)}: Plot function.

data_length::Int64: Length of the data tuple returned by data_analysis.

::Int64: Number of simulations (only for compatibility with the other methods).

GadgetInspector.checkDataShape — Method

checkDataShape(::typeof(hist!), data_length::Int64, sim_number::Int64)::Nothing

Throw an error if data_length is not compatible with the plot type (histogram for this method).

Arguments

::typeof(hist!): Plot function.

data_length::Int64: Length of the data tuple returned by data_analysis.

sim_number::Int64: Number of simulations. In this method, it is only for compatibility.

GadgetInspector.checkDataShape — Method

checkDataShape(::typeof(hist!), data_length::Int64, sim_number::Int64)::Nothing

Throw an error if data_length is not compatible with the plot type (heatmap for this method) or if there is more than one simulation to plot.

Arguments

::typeof(hist!): Plot function.

data_length::Int64: Length of the data tuple returned by data_analysis.

sim_number::Int64: Number of simulations.

GadgetInspector.cleanPlot! — Method

Default function to end the recursion if an unknown type is encountered.

GadgetInspector.cleanPlot! — Method

cleanPlot!(ax::Makie.Axis)::Bool

Empty an axis.

Arguments

ax::Makie.Axis: Axis to be emptied.

Returns

Flag to indicate that an axis has been emptied.

GadgetInspector.cleanPlot! — Method

cleanPlot!(figure::Makie.Figure)::Nothing

Delete all the legends of a figure and empty all its axes.

Arguments

figure::Makie.Figure: Figure to be cleaned.

GadgetInspector.cleanPlot! — Method

cleanPlot!(legend::Makie.Legend)::Bool

Delete a legend.

Arguments

legend::Makie.Legend: Legend to be deleted.

Returns

Flag to indicate that a legend has been deleted.

GadgetInspector.compare — Method

compare(x, y; <keyword arguments>)::Bool

Determine if two elements are equal, using the isequal function.

It's the default method, used when all other compare methods cannot be dispatched.

Arguments

x: First element to be compared.

y: The second element to be compared.

atol::Float64 = 1e-5: Absolute tolerance (just for compatibility with the other methods).

rtol::Float64 = 1e-5: Relative tolerance (just for compatibility with the other methods).

Returns

Returns isequal(x, y).

GadgetInspector.compare — Method

compare(x::Dict, y::Dict; <keyword arguments>)::Bool

Determine if two dataframes are equal (or approximately equal for numeric values).

Every pair of values is compared with the compare function.

Arguments

x::DataFrame: First dataframe to be compared.

y::DataFrame: Second dataframe to be compared.

atol::Float64 = 1e-5: Absolute tolerance (for numeric elements).

rtol::Float64 = 1e-5: Relative tolerance (for numeric elements).

Returns

Return true if every pair of elements within the dataframes pass the equality tests.

GadgetInspector.compare — Method

compare(x::Dict, y::Dict; <keyword arguments>)::Bool

Determine if two dictionaries are equal (or approximately equal for numeric values).

Every pair of values is compared with the compare function.

Arguments

x::Dict: First dictionary to be compared.

y::Dict: Second dictionary to be compared.

atol::Float64 = 1e-5: Absolute tolerance (for numeric elements).

rtol::Float64 = 1e-5: Relative tolerance (for numeric elements).

Returns

Return true if every pair of elements within the dictionaries pass the equality tests.

GadgetInspector.compare — Method

compare(
    x::GIO.SnapshotHeader, 
    y::GIO.SnapshotHeader; 
    <keyword arguments>
)::Bool

Determine if two snapshot headers are equal.

Arguments

x::GIO.SnapshotHeader: First header to be compared.

y::GIO.SnapshotHeader: Second header to be compared.

atol::Float64 = 1e-5: Absolute tolerance.

rtol::Float64 = 1e-5: Relative tolerance.

Returns

Returns true when the headers are equal, and false otherwise.

GadgetInspector.compare — Method

compare(
    x::RealOrQty, 
    y::RealOrQty; 
    <keyword arguments>
)::Bool

Determine if two numeric elements are approximately equal, using the isequal function for elements with NaN, and using isapprox for everything else.

Arguments

x::RealOrQty: First element to be compared.

y::RealOrQty: Second element to be compared.

atol::Float64 = 1e-5: Absolute tolerance.

rtol::Float64 = 1e-5: Relative tolerance.

Returns

Returns true when the values are approximately equal, and false otherwise.

GadgetInspector.compare — Method

compare(
    x::Union{AbstractVector,Tuple}, 
    y::Union{AbstractVector,Tuple};
    <keyword arguments>
)::Bool

Determine if two list-like elements are equal (or approximately equal for numeric values).

Every pair of values is compared with the compare function.

Arguments

x::Union{AbstractVector,Tuple}: First list to be compared.

y::Union{AbstractVector,Tuple}: Second list to be compared.

atol::Float64 = 1e-5: Absolute tolerance (for numeric elements).

rtol::Float64 = 1e-5: Relative tolerance (for numeric elements).

Returns

Returns true if the lists are equal, false if not.

GadgetInspector.computeCenterOfMass — Method

computeCenterOfMass(
    positions::Matrix{<:Unitful.Length},
    mass::Vector{<:Unitful.Mass},
)::Vector{<:Unitful.Length}

Compute the center of mass of a group of particles.

Arguments

positions::Matrix{<:Unitful.Length}: Coordinates of the particles, where each row is an axis, and each column is a particle.

mass::Vector{<:Unitful.Mass}: Masses of the particles.

Returns

The center of mass.

GadgetInspector.computeDistance — Method

computeDistance(
    positions::Matrix{<:Unitful.Length}; 
    <keyword arguments>
)::Vector{<:Unitful.Length}

Compute the distance of several particles from a center.

Arguments

positions::Matrix{<:Unitful.Length}: Coordinates of the particles, where each row is an axis, and each column is a particle.

center::Union{Vector{<:Unitful.Length},Nothing} = nothing: Origin used to calculate the distances.

Returns

The distance of every particle to the center.

GadgetInspector.computeElementFraction — Method

computeElementFraction(
    metals::Matrix{<:Unitful.Mass},
    element::String; 
    <keyword arguments>
)::Vector{Float64}

Compute the total fraction of element with respec to Hydrogen, and apply function to it, for every gas or stellar particle in a snapshot.

Arguments

metals::Matrix{<:Unitful.Mass}: Matrix with the mass content of every element. Each row must be an element and each column a particle.

Row 01: He (Helium)

Row 02: C (Carbon)

Row 03: Mg (Magnesium)

Row 04: 0 (Oxygen)

Row 05: Fe (Iron)

Row 06: Si (Silicon)

Row 07: H (Hydrogen)

Row 08: N (Nitrogen)

Row 09: Ne (Neon)

Row 10: S (Sulfur)

Row 11: Ca (Calcium)

Row 12: Zn (Zinc)

element::String: Name of the element which fraction will be calculated.

func::Function = identity: Function to be applied to the fraction.

Returns

function(X / H) where X is element.

GadgetInspector.computeMetallicity — Method

computeMetallicity(
    metals::Matrix{<:Unitful.Mass},
    mass::Vector{<:Unitful.Mass}; 
    <keyword arguments>
)::Vector{Float64}

Compute the metallicity as the dimensionless fraction of metal mass inside every gas or stellar particle in a snapshot.

Arguments

metals::Matrix{<:Unitful.Mass}: Matrix with the mass content of every element. Each row must be an element and each column a particle.

Row 01: He (Helium)

Row 02: C (Carbon)

Row 03: Mg (Magnesium)

Row 04: 0 (Oxygen)

Row 05: Fe (Iron)

Row 06: Si (Silicon)

Row 07: H (Hydrogen)

Row 08: N (Nitrogen)

Row 09: Ne (Neon)

Row 10: S (Sulfur)

Row 11: Ca (Calcium)

Row 12: Zn (Zinc)

mass::Vector{<:Unitful.Mass}: Mass of every gas or stellar particle.

solar::Bool = false: If the result will be in units of solar metallicity or not.

Returns

The metallicity of the particles.

GadgetInspector.computeProfile — Method

computeProfile(
    distances::Vector{<:Unitful.Length}, 
    quantity::Vector{<:RealOrQty},
    max_radius::Unitful.Length,
    bins::Int64;
    <keyword arguments>
)::NTuple{2,Vector}

Compute a profile of a given quantity, using rings or spherical shells of equal width.

Arguments

distances::Vector{<:Unitful.Length}: Distance to each particle.

quantity::Vector{<:RealOrQty}: Quantity for which the profile will be calculated. It should be defined as a single value per particle.

max_radius::Unitful.Length: Maximum distance up to which the profile will be calculated.

bins::Int64: Number of rings or shells.

flat::Bool = true: If the profile will be 2D, using rings (default), or 3D, using spherical shells.

cumulative::Bool = false: If the profile will be accumulated or not.

density::Bool = false: If the profile is of the area/volume density of quantity or just of quantity itself.

Returns

A Tuple with two elements:

A Vector with the distances to each ring or shell.

A Vector with the values of the profile at those distances.

GadgetInspector.computeRedshift — Method

computeRedshift(scale_factor::Vector{<:Real})::Vector{Float64}

Compute the redshift, from the scale factor.

Arguments

scale_factor::Vector{<:Real}: Scale factors to be converted.

Returns

The redshift.

GadgetInspector.computeSFR — Method

computeSFR(
    stellar_mass::Vector{<:Unitful.Mass}, 
    time::Vector{<:Unitful.Time},
)::Vector{<:Unitful.MassFlow}

Compute the total star formation rate of a series of snapshots, and replace negative values with 0.

Arguments

stellar_mass::Vector{<:Unitful.Mass}: Total mass of the stellar particles, one value per snapshot.

time::Vector{<:Unitful.Time}: Clock time of each snapshot.

Returns

The star formation rate of each snapshot.

GadgetInspector.computeStellarAge — Method

computeStellarAge(
    birth_a::Vector{<:Real},
    sim_data::SnapData, 
    snap_data::SnapData,
)::Vector{<:Unitful.Time}

Compute the age of the stars.

Arguments

birth_times::Vector{<:Real}: Scale factor at the moment of birth of the stars.

sim_data::SnapData: Basic information about the simulation.

snap_data::SnapData: Basic data of the snapshot containing the stars in question, including its timestamp.

Returns

The stellar ages.

GadgetInspector.computeStellarAge — Method

computeStellarAge(
    birth_times::Vector{<:Unitful.Time}, 
    ::SimData,
    snap_data::SnapData,
)::Vector{<:Unitful.Time}

Compute the age of the stars.

Arguments

birth_times::Vector{<:Unitful.Time}: Time of birth of the stars.

::SimData: For compatibility with the other method.

snap_data::SnapData: Basic data of the snapshot containing the stars in question, including its timestamp, header, etc.

Returns

The stellar ages.

GadgetInspector.computeSurfaceDensity — Method

computeSurfaceDensity(
    positions::Matrix{<:Unitful.Length},
    quantity::Vector{<:RealOrQty},
    radius::Unitful.Length,
    bins::Int64,
)::Matrix{<:Unitful.Quantity}

Compute the surface density of a given quantity, using square pixels of equal size.

Arguments

positions::Matrix{<:Unitful.Length}: Projected 2D position of each particle. Each row is an axis and each column is a particle.

quantity::Vector{<:RealOrQty}: Quantity for which the density will be calculated. It should be defined as a single value per particle.

radius::Unitful.Length: Maximum distance up to which the density will be calculated.

bins::Int64: Number of rows and columns used to divide the (-radius, radius] x (-radius, radius] region.

Returns

The density of quantity within each pixel.

GadgetInspector.computeSurfaceDensity — Method

computeSurfaceDensity(
    distances::Vector{<:Unitful.Length},
    quantity::Vector{<:RealOrQty},
    max_radius::Unitful.Length,
    bins::Int64,
)::Vector{<:Unitful.Quantity}

Compute the surface density of a given quantity, using concentric rings of equal width.

Arguments

distances::Vector{<:Unitful.Length}: Projected 2D distance to each particle.

quantity::Vector{<:RealOrQty}: Quantity for which the density will be calculated. It should be defined as a single value per particle.

max_radius::Unitful.Length: Maximum distance up to which the density will be calculated.

bins::Int64: Number of concentric rings.

Returns

The density of quantity within each ring.

GadgetInspector.computeTemperature — Method

computeTemperature(
    metals::Matrix{<:Unitful.Mass},
    mass::Vector{<:Unitful.Mass},
    internal_energy::Vector{<:SpecificEnergy},
    electron_fraction::Vector{<:Real},
)::Vector{<:Unitful.Temperature}

Compute the temperature of gas particles.

Arguments

metals::Matrix{<:Unitful.Mass}: Matrix with the mass content of every element. Each row must be an element and each column a particle.

Row 01: He (Helium)

Row 02: C (Carbon)

Row 03: Mg (Magnesium)

Row 04: 0 (Oxygen)

Row 05: Fe (Iron)

Row 06: Si (Silicon)

Row 07: H (Hydrogen)

Row 08: N (Nitrogen)

Row 09: Ne (Neon)

Row 10: S (Sulfur)

Row 11: Ca (Calcium)

Row 12: Zn (Zinc)

mass::Vector{<:Unitful.Mass}: Mass of every gas particle.

internal_energy::Vector{<:SpecificEnergy}: Specific internal energy of every gas particles.

electron_fraction::Vector{<:Real}: Number fraction of electrons in every gas particle.

Returns

The temperature of each gas particle.

GadgetInspector.computeTime — Method

computeTime(
    scale_factors::Vector{<:Real}, 
    header::GIO.SnapshotHeader; 
    <keyword arguments>
)::Vector{<:Unitful.Time}

Compute the clock time, from the scale factor.

To get the clock time from the scale factor one has to do the integral

\[t = H_0^{\,-1} \int_0^a \frac{\mathrm{d}a'}{a' \, \sqrt{\mathcal{E}(a')}} \, ,\]

where

\[\mathcal{E}(a) = \Omega_\Lambda + \Omega_m \, a^{-3} + \Omega_r \, a^{-4} + \Omega_K \, a^{-2} \, .\]

Arguments

scale_factors::Vector{<:Real}: Scale factors to be converted.

header::GIO.SnapshotHeader: Header of some snapshot of the simulation (it contains the cosmological parameters).

a0::Real = scale_factor[1]: Initial scale factor.

Returns

The clock time.

GadgetInspector.computeTimeSeries — Method

computeTimeSeries(
    snapshot_files::Vector{String},
    quantity::String,
    type::Union{Symbol,Nothing},
    sim_cosmo::Bool; 
    <keyword arguments>
)::Vector

Compute the time series of a given quantity.

Arguments

snapshot_files::Vector{String}: Paths to the snapshot files.

quantity::String: Some of the following quantities:

"clock_time" ⟶ Physical time of each snapshot (time units).

"scale_factor" ⟶ Scale factor of each snapshot (dimensionless).

"redshift" ⟶ Redshift of each snapshot (dimensionless).

"number" ⟶ Number of particles of a given type (dimensionless).

"mass" ⟶ Total mass of a given type of particle (mass units).

"sfr" ⟶ Star formation rate of each snapshot (mass over time units).

type::Union{Symbol,Nothing}: Particle type. The possibilities are given by ParticleType in src/constants.jl. It is only relevant for the "number" and "mass" quantities.

sim_cosmo::Bool: If the simulation is cosmological,

false ⟶ Newtonian simulation (ComovingIntegrationOn = 0).

true ⟶ Cosmological simulation (ComovingIntegrationOn = 1).

filter_function::Union{Function,Nothing} = nothing: A function with the signature:

foo(file_path::String)::Vector{Int64}

It indicates which particles will be read, taking the file path to a snapshot and returning the list of indices of the selected particles. If set to nothing, then no particles are filtered. See the GadgetIO.jl documentation for examples.

warnings::Bool = true: If a warning will be given when the data is not as expected, but the function can still run using sane defaults.

Returns

The time series.

GadgetInspector.computeZProfile — Method

computeZProfile(
    distances::Vector{<:Unitful.Length}, 
    metals::Matrix{<:Unitful.Mass},
    mass::Vector{<:Unitful.Mass},
    max_radius::Unitful.Length,
    bins::Int64;
    <keyword arguments>
)::NTuple{2,Vector}

Compute a metallicity profile.

Arguments

distances::Vector{<:Unitful.Length}: Distance to each particle.

metals::Matrix{<:Unitful.Mass}: Metallicity of the particles.

mass::Vector{<:Unitful.Mass}: Masses of the particles.

max_radius::Unitful.Length: Maximum distance up to which the profile will be calculated.

bins::Int64: Number of subdivisions of [0, max_radius].

flat::Bool = true: If the profile will be 2D, using rings (default), or 3D, using spherical shells.

cumulative::Bool = false: If the profile will be accumulated or not.

density::Bool = false: If the profile is of the area/volume density of quantity or just of quantity itself.

solar::Bool = false: If the result will be in units of solar metallicity or not.

Returns

A Tuple with two elements:

A Vector with the distances to each ring or shell.

A Vector with the values of the profile at those distances.

GadgetInspector.energyIntegrand — Method

energyIntegrand(header::GIO.SnapshotHeader, a::Float64)::Float64

The integrand of the integral that converts the scale factor to physical time,

\[\frac{1}{H\,\sqrt{\epsilon}} \, ,\]

where

\[\epsilon = \Omega_\lambda + \frac{1 - \Omega_\lambda - \Omega_0}{a^2} + \frac{\Omega_0}{a^3} \, , \]

\[H = H_0 \, a \, .\]

Arguments

header::GIO.SnapshotHeader: Header of the relevant snapshot file.

a::Float64: Dimensionless scale factor.

Returns

The integrand evaluated at value a. The result is in Gyr.

GadgetInspector.formatError — Method

formatError(q_mean::RealOrQty, q_error::RealOrQty)::NTuple{2,<:RealOrQty}

Format nicely a set of mean and error values.

It follows the traditional rules for error presentation, the error has only one significant digit, unless such digit is a one, in which case, two significant digits are used. The mean will have a number of digits such as to match the last significant position of the error. An error equal to 0.0 will leave the mean unmodified, and an error bigger than the mean may set the mean equal to 0.0.

Arguments

q_mean::RealOrQty: Mean value.

q_error::RealOrQty: Error value. It must be positive.

Returns

A Tuple with the formatted mean and error values.

Examples

julia> formatError(69.42069, 0.038796)
(69.42, 0.04)

julia> formatError(69.42069, 0.018796)
(69.421, 0.019)

julia> formatError(69.42069, 0.0)
(69.42069, 0.0)

julia> formatError(69.42069, 93.4)
(70.0, 90.0)

GadgetInspector.getLabel — Method

getLabel(
    label::Union{String,LaTeXString}, 
    factor::Int64, 
    unit::Unitful.Units; 
    <keyword arguments>
)::Union{String,LaTeXString}

Construct an axis label.

Arguments

`label::Union{String,LaTeXString}: Text part of the label.

factor::Int64: Exponential factor to scale the units. If different from 0, a term of the form "10^factor" will be added to the label.

unit::Unitful.Units: Unit of the axis.

latex::Bool = true: If the output will be a LaTeXString (default) or a plain String.

Returns

The LaTeXString or String: "label / 10^factor unit". If label is "", an empty string is returned. The factor term only appears if factor != 0, the unit term only appears if unit != Unitful.NoUnits, and the division only appears if there are a factor and/or a unit term.

GadgetInspector.getUnitLabel — Method

getUnitLabel(
    factor::Int64, 
    unit::Unitful.Units; 
    <keyword arguments>
)::Union{String,LaTeXString}

Construct the unit part of an axis label.

Arguments

factor::Int64: Exponential factor to scale the units. If different from 0, a term of the form "10^factor" will be added to the label.

unit::Unitful.Units: Unit of the axis.

latex::Bool = true: If the output will be a LaTeXString (default) or a plain String.

Returns

The LaTeXString or String: "10^factor unit". The factor term only appears if factor != 0, the unit term only appears if unit != Unitful.NoUnits.

GadgetInspector.inRange — Function

inRange(
    data::AbstractVector{<:RealOrQty},
    inferior::RealOrQty = -Inf,
    superior::RealOrQty = Inf;
    <keyword arguments>
)::Bool

If every element in data falls within the range (inferior, superior).

Arguments

data::AbstractVector{<:RealOrQty}: Elements to be checked.

inferior::RealOrQty = -Inf: Lower limit.

`superior::RealOrQty = Inf: Upper limit.

strict::Bool = true: If the the edges are excluded from the range or not.

Returns

If every element x in data satisfy inferior < x < superior, for strict = true, or inferior <= x <= superior, for strict = false.

GadgetInspector.internalUnits — Method

internalUnits(
    quantity::String, 
    header::GIO.SnapshotHeader; 
    <keyword arguments>
)::Union{Unitful.Quantity,Unitful.Units}

Get the factor to convert a plain number into a Unitful quantity, with the correct internal units used by GADGET.

Arguments

quantity::String: Key to the target quantity. The options are given by QUANTITIES in src/constants.jl.

header::GIO.SnapshotHeader: Header of the corresponding snapshot.

sim_cosmo::Bool = false: If the simulation is cosmological,

false ⟶ Newtonian simulation (ComovingIntegrationOn = 0).

true ⟶ Cosmological simulation (ComovingIntegrationOn = 1).

Returns

A Unitful quantity or unit.

GadgetInspector.longest — Method

longest(lists::AbstractVector{<:AbstractVector})::AbstractVector

Get the longest collection within lists.

Arguments

lists::AbstractVector{<:AbstractVector}: List of collections to be compared.

Returns

The longest array.

GadgetInspector.maxLength — Method

maxLength(data::Matrix{<:RealOrQty})::RealOrQty

The maximum value of the norms of the columns in data.

Arguments

data::Matrix{<:RealOrQty}: The values to be analyzed.

Returns

The maximum value of the norms of the columns in data.

GadgetInspector.metalMass — Method

Total mass of metals in a every particle of a GADGET simulation.

GadgetInspector.passAll — Method

passAll(file_path::String, type::String)::Vector{Int64}

Default filter function needed for the read_blocks_over_all_files function.

It does not filter out any particles, allowing the data acquisition functions to gather all the data.

Arguments

file_path::String: Path to the snapshot file to be filtered.

type::Symbol: Particle type. The possibilities are given by the constant ParticleType in src/constants.jl.

Returns

A Vector with the indices of the allowed particles.

GadgetInspector.passCritRho — Method

passCritRho(file_path::String; <keyword arguments>)::Vector{Int64}

Filters out particles with densities below crit_ρ.

Onme possible filter function needed for the read_blocks_over_all_files function.

Arguments

file_path::String: Path to the snapshot file to be filtered.

type::Symbol = :gas: Particle type. The possibilities are given by the constant ParticleType in src/constants.jl.

crit_ρ::Unitful.Quantity = CRITICAL_DENSITY: Critical density, with units.

sim_cosmo::Bool = false: If the simulation is cosmological,

false ⟶ Newtonian simulation (ComovingIntegrationOn = 0).

true ⟶ Cosmological simulation (ComovingIntegrationOn = 1).

Returns

A Vector with the indices of the allowed particles.

GadgetInspector.passMetallicity — Method

passMetallicity(
    file_path::String,
    z_min::Real, 
    z_max::Real; 
    <keyword arguments>
)::Vector{Int64}

Filters out gas particles with a metallicity outside a given range.

One possible filter function needed for the read_blocks_over_all_files function.

Arguments

file_path::String: Path to the snapshot file to be filtered.

z_min::Real: Minimum value of metallicity (as the dimensionless fraction of metal mass) allowed.

z_max::Real: Maximum value of metallicity (as the dimensionless fraction of metal mass) allowed.

sim_cosmo::Bool = false: If the simulation is cosmological,

false ⟶ Newtonian simulation (ComovingIntegrationOn = 0).

true ⟶ Cosmological simulation (ComovingIntegrationOn = 1).

warnings::Bool = true: If a warning will be given when the data is not as expected, but the function can still run using sane defaults.

Returns

A Vector with the indices of the allowed particles.

GadgetInspector.passPositiveQty — Method

passPositiveQty(file_path::String; <keyword arguments>)::Vector{Int64}

Filters out gas particles with negative values for a given quantity.

One possible filter function needed for the read_blocks_over_all_files function.

Arguments

file_path::String: Path to the snapshot file to be filtered.

qty::String = "FMOL": Quantity to be checked.

Returns

A Vector with the indices of the allowed particles.

GadgetInspector.pointData — Method

Extract the data points from a Makie.jl plot, axis, or figure, as a Vector{Point2D}. For the three cases the last series is taken. And in the case of a figure, it will only take the data from the current axis object. It may give non-sensical results or an error for anything other than scatter plots.

GadgetInspector.positiveCut! — Method

positiveCut!(
    m_data::AbstractVector{<:RealOrQty}, 
    s_data::AbstractVector;
    <keyword arguments> 
)::Int32

Delete every element in m_data that is negative. Every corresponding element in s_data (i.e. with the same index) will be deleted too.

Arguments

m_data::AbstractVector{<:RealOrQty}: Master dataset that will be pruned.

s_data::AbstractVector: Slave dataset that will be pruned according to which values of m_data are negative.

keep_edges::Bool = true: If zero will be kept.

min_left::Int64 = 0: Minimum number of values that need to be left in the master dataset after pruning to proceed with the deletion.

Returns

Error flag:

1 ⟶ There would be less than min_left values left in the master dataset after pruning, so the datasets will be kept untouched.

0 ⟶ There would be more than min_left - 1 values left in the master datasets after pruning, so the operation proceeds normally.

GadgetInspector.positiveCut! — Method

positiveCut!(
    data::AbstractVector{<:RealOrQty}; 
    <keyword arguments>
)::Int32

Delete every element in data that is negative.

Arguments

data::AbstractVector{<:RealOrQty}: Dataset that will be pruned.

keep_edges::Bool = true: If zero will be kept.

min_left::Int64 = 0: Minimum number of values that need to be left after pruning to proceed with the deletion.

Returns

Error flag:

1 ⟶ There would be less than min_left values left in the dataset after pruning, so the dataset will be kept untouched.

0 ⟶ There would be more than min_left - 1 values left in the datasets after pruning, so the operation proceeds normally.

GadgetInspector.rangeCut! — Method

rangeCut!(
    m_data::AbstractVector{<:RealOrQty}, 
    s_data::AbstractVector, 
    range::NTuple{2,<:Real};
    <keyword arguments>
)::Int32

Delete every element in m_data that is outside of a given range. Every corresponding element in s_data (i.e. with the same index) will be deleted too.

Arguments

m_data::AbstractVector{<:RealOrQty}: Master dataset that will be pruned outside the range.

s_data::AbstractVector: Slave dataset that will be pruned according to which values of m_data are outside the range.

range::NTuple{2,<:Real}: The range in question.

keep_edges::Bool = true: If the edges of the range will be kept.

min_left::Int64 = 0: Minimum number of values that need to be left in the master dataset after pruning to proceed with the deletion.

Returns

Error flag:

1 ⟶ There would be less than min_left values left in the master dataset after pruning, so the datasets will be kept untouched.

0 ⟶ There would be more than min_left - 1 values left in the master datasets after pruning, so the operation proceeds normally.

GadgetInspector.rangeCut! — Method

rangeCut!(
    data::AbstractVector{<:RealOrQty},  
    range::NTuple{2, <:Real};
    <keyword arguments>
)::Int32

Delete every element in data that is outside of a given range.

Arguments

data::AbstractVector{<:RealOrQty}: Dataset that will be pruned outside the range.

range::NTuple{2,<:Real}: The range in question.

keep_edges::Bool = true: If the edges of the range will be kept.

min_left::Int64 = 0: Minimum number of values that need to be left after pruning to proceed with the deletion.

Returns

Error flag:

1 ⟶ There would be less than min_left values left in the dataset after pruning, so the dataset will be kept untouched.

0 ⟶ There would be more than min_left - 1 values left in the datasets after pruning, so the operation proceeds normally.

GadgetInspector.read_csv — Method

Read the numeric data of a CSV file using DelimitedFiles. It skips the first line.

GadgetInspector.sanitizeData! — Method

sanitizeData!(
    x_data::AbstractVector{<:RealOrQty},
    y_data::AbstractVector{<:RealOrQty};
    <keyword arguments> 
)::NTuple{4,Int32}

Do three operations over x_data and y_data, in the following order:

Trim them to fit within the domain of func_domain[1] and func_domain[2] respectively.

Trim them to fit within range[1] and range[2] respectively.

Scale them by the factors 10^exp_factor[1] and 10^exp_factor[2] respectively.

By default, no transformation is done to any of the datasets.

Note

The dataset must have the same length, and any operation that deletes an element, will delete the corresponding element (i.e. with the same index) in the other dataset, so the dataset will be kept equal in length at the end.

Arguments

x_data::AbstractVector{<:RealOrQty}: First dataset to be sanitized.

y_data::AbstractVector{<:RealOrQty}: Second dataset to be sanitized.

func_domain::NTuple{2, Function} = (identity, identity): x_data will be trimmed to fit within the domain of the function func_domain[1], and y_data will be trimmed to fit within the domain of the function func_domain[2]. The possible options are the scaling functions accepted by Makie.jl: log10, log2, log, sqrt, Makie.logit, Makie.Symlog10, Makie.pseudolog10 and identity.

range::NTuple{2,NTuple{2,<:Real}} = ((-Inf, Inf), (-Inf, Inf)): Every element in x_data that falls outside of range[1] will be deleted, and every element in y_data that falls outside of range[2] will be deleted.

keep_edges::NTuple{2,Bool} = (true, true): If the edges of each corresponding range will be kept.

min_left::Int64 = 0: Minimum number of values that need to be left after prunning to proceed with the deletion.

exp_factor::NTuple{2,Int64} = (0, 0): Every element in x_data will be multiplied by 10^exp_factor[1], and every element in y_data will be multiplied by 10^exp_factor[2].

Returns

A tuple with four error flags:

After trying to fit x_data and y_data within the domain of func_domain, one flag for every dataset:

1 ⟶ There would be less than min_left values left in the dataset after pruning, so the dataset will be kept untouched.

0 ⟶ There would be more than min_left - 1 values left in the datasets after pruning, so the operation proceeds normally.

After trying to fit x_data and y_data in range, one flag for every dataset:

1 ⟶ There would be less than min_left values left in the dataset after pruning, so the dataset will be kept untouched.

0 ⟶ There would be more than min_left - 1 values left in the datasets after pruning, so the operation proceeds normally.

GadgetInspector.sanitizeData! — Method

sanitizeData!(
    data::AbstractVector{<:RealOrQty};
    <keyword arguments> 
)::NTuple{2,Int32}

Do three operations over data, in the following order:

Trim it to fit within the domain of func_domain.

Trim it to fit within range.

Scale it by a factor of 10^exp_factor.

By default, no transformation is done.

Arguments

data::AbstractVector{<:RealOrQty}: Dataset to be sanitized.

func_domain::Function = identity: data will be trimmed to fit within the domain of the function func_domain. The options are the scaling functions accepted by Makie.jl: log10, log2, log, sqrt, Makie.logit, Makie.Symlog10, Makie.pseudolog10 and identity.

range::NTuple{2,<:Real} = (-Inf, Inf): Every element in data that falls outside of range will be deleted.

keep_edges::Bool = true: If the edges of range will be kept.

min_left::Int64 = 0: Minimum number of values that need to be left after pruning to proceed with the deletion.

exp_factor::Int64 = 0: Every element in data will be multiplied by 10^exp_factor.

Returns

A tuple with two error flags:

After trying to fit data within the domain of func_domain:

1 ⟶ There would be less than min_left values left in the dataset after pruning, so the dataset will be kept untouched.

0 ⟶ There would be more than min_left - 1 values left in the datasets after pruning, so the operation proceeds normally.

After trying to fit the data in range:

1 ⟶ There would be less than min_left values left in the dataset after pruning, so the dataset will be kept untouched.

0 ⟶ There would be more than min_left - 1 values left in the datasets after pruning, so the operation proceeds normally.

GadgetInspector.scaledBins — Method

scaledBins(
    values::AbstractVector{<:RealOrQty}, 
    bins::Int64,
    scale::Function; 
    <keyword arguments>
)::Vector{Float64}

Compute the bin edges, for a given scaling function and set of values.

Arguments

values::AbstractVector{<:RealOrQty}: Values to be binned.

bins::Int64: Number of bins to be calculated.

scale::Function: Function to scale the values. The possibilities are the scaling functions accepted by Makie.jl: log10, log2, log, sqrt, Makie.logit, Makie.Symlog10, Makie.pseudolog10 and identity.

limits::Tuple{<:Union{Real,Nothing},<:Union{Real,Nothing}} = (nothing, nothing): Set it to a value different than nothing if you want to fix the limits of the axis.

warnings::Bool = true: If a warning will be given when the data is not as expected, but the function can still run using sane defaults.

Returns

A sorted list of bin edges.

GadgetInspector.smoothWindow — Method

smoothWindow(
    x_data::AbstractVector{<:RealOrQty},
    y_data::AbstractVector{<:RealOrQty},
    bins::Int64;
    <keyword arguments>
)::NTuple{2,Vector{<:RealOrQty}}

Separate the x_data values in bins contiguous windows, and replaces every x and y value within the window with the corresponding mean, to smooth out the data.

x_data should be within the domain of f_scale.

Arguments

x_data::AbstractVector{<:RealOrQty}: x-axis data.

y_data::AbstractVector{<:RealOrQty}: y-axis data.

bins::Int64: Number of bins to be used in the smoothing.

scale::Function = identity: Function to scale the x-axis. The bins will be computed accordingly. The options are the scaling functions accepted by Makie.jl: log10, log2, log, sqrt, Makie.logit, Makie.Symlog10, Makie.pseudolog10 and identity.

Returns

A tuple with two arrays containing the smoothed out x and y data.

GadgetInspector.trivial — Method

Function that always returns nothing, for any type and number of arguments.

GadgetInspector.xlimits — Method

Extract the limits of the axes from a Makie.jl plot, axis, or figure. In the case of a figure, it will take the limits from the current axis object.

GadgetInspector.xscale — Method

Extract the scales of the axes from a Makie.jl plot, axis, or figure. In the case of a figure, it will take the scales from the current axis object.