Computation of derived quantities

These functions are used internally and none are exported.

They may depend on the particulars of the simulation code (e.g. internal units).

GalaxyInspector.bigiel2008 — Method

bigiel2008(
    ΣH::Vector{<:SurfaceDensity};
    <keyword arguments>
)::Vector{<:Number}

Kennicutt-Schmidt law for the molecular or neutral gas, taken from a set of observations of nearby galaxies.

From Bigiel et al. (2008) (Section 3.1), we have

\[\Sigma_\mathrm{SFR} = a \left( \frac{\Sigma_\mathrm{HI, H_2, gas}}{10 \, \mathrm{M_\odot \, pc^{-2}}} \right)^{\!N} \, ,\]

where N is the power-law index, and $A = \log_{10}(a)$, where $a$ is $\Sigma_\mathrm{SFR}$ at the fiducial gas surface density of $10 \, \mathrm{M_\odot \, pc^{-2}}$.

Arguments

ΣH::Vector{<:SurfaceDensity}: Values of the molecular or neutral gas surface density, with units.

molecular::Bool=true: If the x axis will be the area mass density of molecular hydrogen, or, if set to false, the area mass density of neutral hydrogen.

log_output::Bool=true: If the output will the $\log_{10}$ of the star formation area density, or the star formation area density itself (with units). If log_output = true, the implied unit is $\mathrm{M_\odot \, yr^{-1} \, kpc^{-2}}$

Returns

The star formation area density.

References

F. Bigiel et al. (2008). THE STAR FORMATION LAW IN NEARBY GALAXIES ON SUB-KPC SCALES. The Astrophysical Journal, 136(6), 2846. doi:10.1088/0004-6256/136/6/2846

GalaxyInspector.computeClumpingFactor — Method

computeClumpingFactor(density::Vector{<:Number})::Float64

Compute the clumping factor,

\[C_\rho = \frac{\langle \rho^2 \rangle}{\langle \rho \rangle^2} \, .\]

Arguments

density::Vector{<:Number}: The density of the cells/particles.

Returns

The clumping factor.

GalaxyInspector.computeDepletionTime — Method

computeDepletionTime(
    mass::Vector{<:Unitful.Mass},
    sfr::Vector{<:Unitful.MassFlow},
)::Vector{<:Unitful.Time}

Compute the depletion time,

\[t_\mathrm{ff} = \frac{M_\mathrm{gas}}{\dot{M}_\star} \, .\]

Arguments

mass::Vector{<:Unitful.Mass}: The gas mass of the cells/particles.

sfr::Vector{<:Unitful.MassFlow}: The SFR associated to each cell/particle.

Returns

The depletion time.

GalaxyInspector.computeEfficiencyFF — Method

computeEfficiencyFF(
    density::Vector{<:Unitful.Density},
    mass::Vector{<:Unitful.Mass},
    sfr::Vector{<:Unitful.MassFlow},
)::Vector{Float64}

Compute the star formation efficiency per free-fall time, according to the definition in equation 1 of Krumholz (2012),

\[\epsilon_\mathrm{ff} = \frac{t_\mathrm{ff}}{t_\mathrm{dep}} \, .\]

where

\[t_\mathrm{ff} = \sqrt{\frac{3 \, \pi}{32 \, G \, \rho}} \, ,\]

is the free-fall time, and

\[t_\mathrm{dep} = \frac{M_\mathrm{H_2}}{\dot{M}_\star} \, ,\]

is the depletion time.

Arguments

density::Vector{<:Unitful.Density}: The molecular hydrogen ($\\mathrm{H_2}$) density of the cells/particles.

mass::Vector{<:Unitful.Mass}: The gas mass of the cells/particles.

sfr::Vector{<:Unitful.MassFlow}: The SFR associated to each cell/particle.

Returns

The star formation efficiency per free-fall time.

References

M. R. Krumholz et al. (2012). A UNIVERSAL, LOCAL STAR FORMATION LAW IN GALACTIC CLOUDS, NEARBY GALAXIES, HIGH-REDSHIFT DISKS, AND STARBURSTS. The Astrophysical Journal, 745(1), 69. doi:10.1088/0004-637X/745/1/69

GalaxyInspector.computeKineticEnergy — Method

computeKineticEnergy(
    masses::Vector{<:Unitful.Mass},
    velocities::Matrix{<:Unitful.Velocity},
)::Vector{<:Unitful.Energy}

Compute the kinetic energy.

Arguments

masses::Vector{<:Unitful.Mass}: Mass of every cell/particle.

velocities::Matrix{<:Unitful.Velocity}: Velocities of the cells/particles. Each column is a cell/particle and each row a dimension.

Returns

The kinetic energy of each cell/particle.

GalaxyInspector.computeParticleBandProfile — Method

computeParticleBandProfile(
    positions::Matrix{<:Unitful.Length},
    quantity::Vector{<:Number},
    grid::CircularGrid;
    <keyword arguments>
)::NTuple{2,Vector{<:Number}}

Compute a profile of the mean and standard deviation of quantity, using an 1D histogram

Empty bins have NaN as mean and standard deviation.

Arguments

positions::Matrix{<:Unitful.Length}: Positions of the cells/particles. Each column is a cell/particle and each row a dimension.

quantity::Vector{<:Number}: The profile will be of this quantity.

grid::CircularGrid: Circular grid.

flat::Bool=true: If the profile will be 2D, using rings, or 3D, using spherical shells.

Returns

A tuple with two elements:

A vector with the mean value for each bin.

A vector with the standard deviation for each bin.

GalaxyInspector.computeParticleProfile — Method

computeParticleProfile(
    positions::Matrix{<:Unitful.Length},
    quantity::Vector{<:Number},
    grid::CircularGrid;
    <keyword arguments>
)::Vector{<:Number}

Compute a profile, using an 1D histogram.

Arguments

positions::Matrix{<:Unitful.Length}: Positions of the cells/particles. Each column is a cell/particle and each row a dimension.

quantity::Vector{<:Number}: The profile will be of this quantity.

grid::CircularGrid: Circular grid.

norm_values::Vector{<:Number}=Number[]: Values to normalize quantity.

flat::Bool=true: If the profile will be 2D, using rings, or 3D, using spherical shells.

total::Bool=true: If the sum (default) or the mean of quantity will be computed for each bin.

cumulative::Bool=false: If the profile will be accumulated or not.

density::Bool=false: If the profile will be of the density of quantity.

empty_nan::Bool=true: If empty bins will be set to NaN, 0 is used otherwise. Be careful if empty_nan = true and cumulative = true, because every bin after the first NaN will be set to NaN.

Returns

Vector with the values of the profile.

GalaxyInspector.computePotentialEnergy — Method

computePotentialEnergy(
    potential::Vector{<:SpecificEnergy},
    masses::Vector{<:Unitful.Mass},
)::Vector{<:Unitful.Energy}

Compute the gravitational potencial energy.

Arguments

potential::Vector{<:SpecificEnergy}: Specific potential energy of every cell/particle.

masses::Vector{<:Unitful.Mass}: Mass of every cell/particle.

Returns

The gravitational potencial energy of each cell/particle.

GalaxyInspector.computeSFR — Method

computeSFR(
    data_dict::Dict;
    <keyword arguments>
)::Vector{<:Unitful.MassFlow}

Compute the star formation rate of each stellar particle.

For stellar particles younger than age_resol, the SFR is its mass divided by age_resol. It is defined as 0 for older particles.

Arguments

data_dict::Dict: A dictionary with the following shape:

:sim_data -> ::Simulation (see Simulation).

:snap_data -> ::Snapshot (see Snapshot).

:gc_data -> ::GroupCatalog (see GroupCatalog).

cell/particle type -> (block -> data of block, block -> data of block, ...).

cell/particle type -> (block -> data of block, block -> data of block, ...).

cell/particle type -> (block -> data of block, block -> data of block, ...).

...

groupcat type -> (block -> data of block, block -> data of block, ...).

groupcat type -> (block -> data of block, block -> data of block, ...).

groupcat type -> (block -> data of block, block -> data of block, ...).

...

age_resol::Unitful.Time=AGE_RESOLUTION: Age resolution for the SFR.

Returns

The star formation rate of each stellar particle.

GalaxyInspector.computeStellarAge — Method

computeStellarAge(data_dict::Dict)::Vector{<:Unitful.Time}

Compute the age of the stars.

Arguments

data_dict::Dict: A dictionary with the following shape:

:sim_data -> ::Simulation (see Simulation).

:snap_data -> ::Snapshot (see Snapshot).

:gc_data -> ::GroupCatalog (see GroupCatalog).

cell/particle type -> (block -> data of block, block -> data of block, ...).

cell/particle type -> (block -> data of block, block -> data of block, ...).

cell/particle type -> (block -> data of block, block -> data of block, ...).

...

groupcat type -> (block -> data of block, block -> data of block, ...).

groupcat type -> (block -> data of block, block -> data of block, ...).

groupcat type -> (block -> data of block, block -> data of block, ...).

...

Returns

The stellar ages.

GalaxyInspector.computeTemperature — Method

computeTemperature(
    internal_energy::Vector{<:SpecificEnergy},
    electron_fraction::Vector{Float32},
)::Vector{<:Unitful.Temperature}

Compute the gas temperature.

Arguments

internal_energy::Vector{<:SpecificEnergy}: Specific internal energy of every gas cell/particle.

electron_fraction::Vector{Float32}: Number fraction of electrons in every gas cell/particle.

Returns

The temperature of each gas cell/particle.

GalaxyInspector.computeTime — Method

computeTime(a::Real, header::SnapshotHeader; <keyword arguments>)::Unitful.Time

Compute the physical time corresponding to the scale factor a.

To get the physical time $t$ from the scale factor a, one does the integral:

\[t = \frac{1}{H_0} \int_0^a \frac{\mathrm{d}a'}{a' \, \sqrt{\mathcal{E}(a')}} \, ,\]

where

\[\mathcal{E}(a) = \Omega_\Lambda + \Omega_m \, a^{-3} + \Omega_r \, a^{-4} + \Omega_K \, a^{-2} \, .\]

Arguments

a::Real: Scale factor.

header::SnapshotHeader: A header of the simulation, containing the cosmological parameters.

a0::Float64=0.0: Initial scale factor.

Returns

The physical time.

GalaxyInspector.computeTime — Method

computeTime(
    scale_factors::Vector{<:Real},
    header::SnapshotHeader;
    <keyword arguments>
)::Vector{<:Unitful.Time}

Compute the physical time corresponding to each of the scale_factors.

To get the physical time $t$ from the scale factor a, one does the integral:

\[t = \frac{1}{H_0} \int_0^a \frac{\mathrm{d}a'}{a' \, \sqrt{\mathcal{E}(a')}} \, ,\]

where

\[\mathcal{E}(a) = \Omega_\Lambda + \Omega_m \, a^{-3} + \Omega_r \, a^{-4} + \Omega_K \, a^{-2} \, .\]

Arguments

scale_factors::Vector{<:Real}: Scale factors.

header::SnapshotHeader: A header of the simulation, containing the cosmological parameters.

a0::Float64=0.0: Initial scale factor.

Returns

A vector with the physical times.

GalaxyInspector.computeTimeTicks — Method

computeTimeTicks(
    paths::Vector{<:Union{Missing,String}},
)::Tuple{Vector{Float64},Vector{Float64},Vector{<:Unitful.Time},Vector{<:Unitful.Time}}

Compute the different times stamps associated with each snapshot in paths.

Arguments

paths::Vector{<:Union{Missing,String}}: Paths to the snapshots.

Returns

A tuple with four elements:

A vector with the scale factors.

A vector with the redshifts.

A vector with the physical times (physical time since the Big Bang).

A vector with the lookback times (physical time left to reach the last snapshot).

GalaxyInspector.computeTotalEnergy — Method

computeTotalEnergy(
    potential::Vector{<:SpecificEnergy},
    masses::Vector{<:Unitful.Mass},
    velocities::Matrix{<:Unitful.Velocity},
)::Vector{<:Unitful.Energy}

Compute the total energy (kinetic + potential).

Arguments

potential::Vector{<:SpecificEnergy}: Specific potential energy of every cell/particle.

masses::Vector{<:Unitful.Mass}: Mass of every cell/particle.

velocities::Matrix{<:Unitful.Velocity}: Velocities of the cells/particles. Each column is a cell/particle and each row a dimension.

Returns

The total energy of each cell/particle.

GalaxyInspector.density3DProjection — Method

density3DProjection(
    data_dict::Dict,
    grid::CubicGrid,
    quantity::Symbol,
    type::Symbol;
    <keyword arguments>
)::Array{Float64,3}

Sample the 3D density field of a given quantity using a cubic grid

Note

If the source of the field are particles, a simple 3D histogram is used. If they are Voronoi cells instead, the density of the cell that intersects each voxel is used.

Arguments

data_dict::Dict: A dictionary with the following shape:

:sim_data -> ::Simulation (see Simulation).

:snap_data -> ::Snapshot (see Snapshot).

:gc_data -> ::GroupCatalog (see GroupCatalog).

cell/particle type -> (block -> data of block, block -> data of block, ...).

cell/particle type -> (block -> data of block, block -> data of block, ...).

cell/particle type -> (block -> data of block, block -> data of block, ...).

...

groupcat type -> (block -> data of block, block -> data of block, ...).

groupcat type -> (block -> data of block, block -> data of block, ...).

groupcat type -> (block -> data of block, block -> data of block, ...).

...

grid::CubicGrid: Cubic grid.

quantity::Symbol: Which density will be calculated. The options are:

:stellar_mass -> Stellar density.

:gas_mass -> Gas density.

:hydrogen_mass -> Hydrogen density.

:dm_mass -> Dark matter density.

:bh_mass -> Black hole density.

:molecular_mass -> Molecular hydrogen ($\mathrm{H_2}$) density.

:br_molecular_mass -> Molecular hydrogen ($\mathrm{H_2}$) density, computed using the pressure relation in Blitz et al. (2006).

:atomic_mass -> Atomic hydrogen ($\mathrm{HI}$) density.

:ionized_mass -> Ionized hydrogen ($\mathrm{HII}$) density.

:neutral_mass -> Neutral hydrogen ($\mathrm{HI + H_2}$) density.

:stellar_gas_mass -> Stellar gas mass (according to our SF model).

:ode_metal_mass -> Metal mass (according to our SF model).

:dust_mass -> Dust mass.

type::Symbol: If the source of the field are :particles or Voronoi :cells.

m_unit::Unitful.Units=u"Msun": Mass unit.

l_unit::Unitful.Units=u"kpc": Length unit.

filter_function::Function=filterNothing: A function with the signature:

filter_function(data_dict) -> indices

where

data_dict::Dict: A dictionary with the following shape:

:sim_data -> ::Simulation (see Simulation).

:snap_data -> ::Snapshot (see Snapshot).

:gc_data -> ::GroupCatalog (see GroupCatalog).

cell/particle type -> (block -> data of block, block -> data of block, ...).

cell/particle type -> (block -> data of block, block -> data of block, ...).

cell/particle type -> (block -> data of block, block -> data of block, ...).

...

groupcat type -> (block -> data of block, block -> data of block, ...).

groupcat type -> (block -> data of block, block -> data of block, ...).

groupcat type -> (block -> data of block, block -> data of block, ...).

...

indices::Dict: A dictionary with the following shape:

cell/particle type -> idxs::IndexType

cell/particle type -> idxs::IndexType

cell/particle type -> idxs::IndexType

...

Returns

A 3D array with the density at each point of the 3D grid.

References

L. Blitz et al. (2006). The Role of Pressure in GMC Formation II: The H2-Pressure Relation. The Astrophysical Journal, 650(2), 933. doi:10.1086/505417

GalaxyInspector.energyIntegrand — Method

energyIntegrand(a::Real, header::SnapshotHeader)::Float64

The integrand of the integral that converts the scale factor into physical time:

\[\frac{1}{H\,\sqrt{\mathcal{E}}} \, ,\]

where

\[\mathcal{E} = \Omega_\Lambda + (1 - \Omega_\Lambda - \Omega_m) \, a^{-2} + \Omega_m \, a^{-3} \, ,\]

\[H = H_0 \, a \, .\]

Arguments

a::Real: Scale factor.

header::SnapshotHeader: Header of the relevant snapshot file.

Returns

The integrand evaluated at a, in $\mathrm{Gyr}$.

GalaxyInspector.invBigiel2008 — Method

invBigiel2008(
    Σsfr ::Vector{<:MassFlowDensity};
    <keyword arguments>
)::Vector{<:Number}

Inverse Kennicutt-Schmidt law for the molecular or neutral gas, taken from a set of observations of nearby galaxies.

From Bigiel et al. (2008) (Section 3.1, Eq. 2), we have

\[\Sigma_\mathrm{SFR} = a \left( \frac{\Sigma_\mathrm{HI, H_2, gas}}{10 \, \mathrm{M_\odot \, pc^{-2}}} \right)^{\!N} \, ,\]

where N is the power-law index, and $A = \log_{10}(a)$, where $a$ is $\Sigma_\mathrm{SFR}$ at the fiducial gas surface density of $10 \, \mathrm{M_\odot \, pc^{-2}}$.

Arguments

Σsfr ::Vector{<:MassFlowDensity}: Values of the star formation area density, with units.

molecular::Bool=true: If the output will be the area mass density of molecular hydrogen, or, if set to false, the area mass density of neutral hydrogen.

log_output::Bool=true: If the output will the $\log_{10}$ of the molecular or neutral gas surface density, or the molecular or neutral gas surface density itself (with units). If log_output = true, the implied unit is $10 \, \mathrm{M_\odot \, pc^{-2}}$

Returns

The molecular or neutral gas surface density.

References

F. Bigiel et al. (2008). THE STAR FORMATION LAW IN NEARBY GALAXIES ON SUB-KPC SCALES. The Astrophysical Journal, 136(6), 2846. doi:10.1088/0004-6256/136/6/2846

GalaxyInspector.invKennicutt1998 — Method

invKennicutt1998(Σsfr::Vector{<:MassFlowDensity}; <keyword arguments>)::Vector{<:Number}

Inverse Kennicutt-Schmidt law, taken from a set of observations of nearby galaxies.

From Kennicutt (1998) (Section 4, Eq. 4), we have

\[\Sigma_\mathrm{SFR} = a \left( \frac{\Sigma_\mathrm{gas}}{1 \, \mathrm{M_\odot \, pc^{-2}}} \right)^{\!N} \mathrm{M_\odot \, yr^{-1] \, kpc^{-2}} \, ,\]

where N is the power-law index and $a$ is $\Sigma_\mathrm{SFR}$ at the fiducial gas surface density of $1 \, \mathrm{M_\odot \, pc^{-2}}$.

Arguments

Σsfr::Vector{<:MassFlowDensity}: Values of the star formation area density, with units.

log_output::Bool=true: If the output will the $\log_{10}$ of the gas mass surface density, or the gas mass surface density itself (with units). If log_output = true, the implied unit is $\mathrm{M_\odot \, pc^{-2}}$

Returns

The gas mass surface density.

References

R. C. Kennicutt (1998). The Global Schmidt Law in Star-forming Galaxies. The Astrophysical Journal, 498(2), 541-552. doi:10.1086/305588

GalaxyInspector.kennicutt1998 — Method

kennicutt1998(Σgas::Vector{<:SurfaceDensity}; <keyword arguments>)::Vector{<:Number}

Kennicutt-Schmidt law, taken from a set of observations of nearby galaxies.

From Kennicutt (1998) (Section 4, Eq. 4), we have

\[\Sigma_\mathrm{SFR} = a \left( \frac{\Sigma_\mathrm{gas}}{1 \, \mathrm{M_\odot \, pc^{-2}}} \right)^{\!N} \mathrm{M_\odot \, yr^{-1] \, kpc^{-2}} \, ,\]

where N is the power-law index and $a$ is $\Sigma_\mathrm{SFR}$ at the fiducial gas surface density of $1 \, \mathrm{M_\odot \, pc^{-2}}$.

Arguments

Σgas::Vector{<:SurfaceDensity}: Values of the gas mass surface density, with units.

log_output::Bool=true: If the output will the $\log_{10}$ of the star formation area density, or the star formation area density itself (with units). If log_output = true, the implied unit is $\mathrm{M_\odot \, yr^{-1} \, kpc^{-2}}$

Returns

The star formation area density.

References

R. C. Kennicutt (1998). The Global Schmidt Law in Star-forming Galaxies. The Astrophysical Journal, 498(2), 541-552. doi:10.1086/305588