Filters

These functions are only used internally, and depend on the particulars of the simulation code (e.g. units).

Here you can find the different filters.

None of these functions are exported.

GalaxyInspector.filterCircularity — Method

filterCircularity(data_dict::Dict, l_ϵ::Float64, h_ϵ::Float64)::Dict{Symbol,IndexType}

Filter out stellar particles with circularity outside the range [l_ϵ, h_ϵ].

Arguments

data_dict::Dict: A dictionary with the following shape:

:sim_data -> ::Simulation (see Simulation).

:snap_data -> ::Snapshot (see Snapshot).

:gc_data -> ::GroupCatalog (see GroupCatalog).

cell/particle type -> (block -> data of block, block -> data of block, ...).

cell/particle type -> (block -> data of block, block -> data of block, ...).

cell/particle type -> (block -> data of block, block -> data of block, ...).

...

groupcat type -> (block -> data of block, block -> data of block, ...).

groupcat type -> (block -> data of block, block -> data of block, ...).

groupcat type -> (block -> data of block, block -> data of block, ...).

...

l_ϵ::Float64: Minimum circularity.

h_ϵ::Float64: Maximum circularity.

Returns

A dictionary with the following shape:

cell/particle type -> idxs::IndexType

cell/particle type -> idxs::IndexType

cell/particle type -> idxs::IndexType

...

GalaxyInspector.filterData! — Method

filterData!(data_dict::Dict; <keyword arguments>)::Nothing

Filter data_dict using the indices provided by filter_function.

Arguments

data_dict::Dict: A dictionary with the following shape:

:sim_data -> ::Simulation (see Simulation).

:snap_data -> ::Snapshot (see Snapshot).

:gc_data -> ::GroupCatalog (see GroupCatalog).

cell/particle type -> (block -> data of block, block -> data of block, ...).

cell/particle type -> (block -> data of block, block -> data of block, ...).

cell/particle type -> (block -> data of block, block -> data of block, ...).

...

groupcat type -> (block -> data of block, block -> data of block, ...).

groupcat type -> (block -> data of block, block -> data of block, ...).

groupcat type -> (block -> data of block, block -> data of block, ...).

...

filter_function::Function=filterNothing: A function with the signature:

filter_function(data_dict) -> indices

where

data_dict::Dict: A dictionary with the following shape:

:sim_data -> ::Simulation (see Simulation).

:snap_data -> ::Snapshot (see Snapshot).

:gc_data -> ::GroupCatalog (see GroupCatalog).

cell/particle type -> (block -> data of block, block -> data of block, ...).

cell/particle type -> (block -> data of block, block -> data of block, ...).

cell/particle type -> (block -> data of block, block -> data of block, ...).

...

groupcat type -> (block -> data of block, block -> data of block, ...).

groupcat type -> (block -> data of block, block -> data of block, ...).

groupcat type -> (block -> data of block, block -> data of block, ...).

...

indices::Dict: A dictionary with the following shape:

cell/particle type -> idxs::IndexType

cell/particle type -> idxs::IndexType

cell/particle type -> idxs::IndexType

...

GalaxyInspector.filterData — Method

filterData(data_dict::Dict; <keyword arguments>)::Dict

Returna filtered copy of data_dict using the indices provided by filter_function.

Arguments

data_dict::Dict: A dictionary with the following shape:

:sim_data -> ::Simulation (see Simulation).

:snap_data -> ::Snapshot (see Snapshot).

:gc_data -> ::GroupCatalog (see GroupCatalog).

cell/particle type -> (block -> data of block, block -> data of block, ...).

cell/particle type -> (block -> data of block, block -> data of block, ...).

cell/particle type -> (block -> data of block, block -> data of block, ...).

...

groupcat type -> (block -> data of block, block -> data of block, ...).

groupcat type -> (block -> data of block, block -> data of block, ...).

groupcat type -> (block -> data of block, block -> data of block, ...).

...

filter_function::Function=filterNothing: A function with the signature:

filter_function(data_dict) -> indices

where

data_dict::Dict: A dictionary with the following shape:

:sim_data -> ::Simulation (see Simulation).

:snap_data -> ::Snapshot (see Snapshot).

:gc_data -> ::GroupCatalog (see GroupCatalog).

cell/particle type -> (block -> data of block, block -> data of block, ...).

cell/particle type -> (block -> data of block, block -> data of block, ...).

cell/particle type -> (block -> data of block, block -> data of block, ...).

...

groupcat type -> (block -> data of block, block -> data of block, ...).

groupcat type -> (block -> data of block, block -> data of block, ...).

groupcat type -> (block -> data of block, block -> data of block, ...).

...

indices::Dict: A dictionary with the following shape:

cell/particle type -> idxs::IndexType

cell/particle type -> idxs::IndexType

cell/particle type -> idxs::IndexType

...

Returns

The filtered data.

GalaxyInspector.filterExsituStars — Method

filterExsituStars(
    data_dict::Dict;
    <keyword arguments>
)::Dict{Symbol,IndexType}

Filter out stars that where born inside the given halo and subhalo (insitu), leaving only the ones born outside (exsitu).

Arguments

data_dict::Dict: A dictionary with the following shape:

:sim_data -> ::Simulation (see Simulation).

:snap_data -> ::Snapshot (see Snapshot).

:gc_data -> ::GroupCatalog (see GroupCatalog).

cell/particle type -> (block -> data of block, block -> data of block, ...).

cell/particle type -> (block -> data of block, block -> data of block, ...).

cell/particle type -> (block -> data of block, block -> data of block, ...).

...

groupcat type -> (block -> data of block, block -> data of block, ...).

groupcat type -> (block -> data of block, block -> data of block, ...).

groupcat type -> (block -> data of block, block -> data of block, ...).

...

halo_idx::Int=1: Index of the target halo (FoF group). Starts at 1.

subhalo_rel_idx::Int=1: Index of the target subhalo (subfind), relative the target halo. Starts at 1. If set to 0, only stars born outside halo halo_idx are consider exsitu.

warnings::Bool=true: If a warning will be given when there is missing data.

Returns

A dictionary with the following shape:

cell/particle type -> idxs::IndexType

cell/particle type -> idxs::IndexType

cell/particle type -> idxs::IndexType

...

GalaxyInspector.filterGFM — Method

filterGFM(data_dict::Dict)::Dict{Symbol,IndexType}

Filter out gas cells that have not entered out routine.

Arguments

data_dict::Dict: A dictionary with the following shape:

:sim_data -> ::Simulation (see Simulation).

:snap_data -> ::Snapshot (see Snapshot).

:gc_data -> ::GroupCatalog (see GroupCatalog).

cell/particle type -> (block -> data of block, block -> data of block, ...).

cell/particle type -> (block -> data of block, block -> data of block, ...).

cell/particle type -> (block -> data of block, block -> data of block, ...).

...

groupcat type -> (block -> data of block, block -> data of block, ...).

groupcat type -> (block -> data of block, block -> data of block, ...).

groupcat type -> (block -> data of block, block -> data of block, ...).

...

Returns

A dictionary with the following shape:

cell/particle type -> idxs::IndexType

cell/particle type -> idxs::IndexType

cell/particle type -> idxs::IndexType

...

GalaxyInspector.filterGasDensity — Method

filterGasDensity(
    data_dict::Dict,
    min_ρ::Unitful.Density,
    max_ρ::Unitful.Density,
)::Dict{Symbol,IndexType}

Filter out gas that is outside the density range [min_ρ, max_ρ].

Arguments

data_dict::Dict: A dictionary with the following shape:

:sim_data -> ::Simulation (see Simulation).

:snap_data -> ::Snapshot (see Snapshot).

:gc_data -> ::GroupCatalog (see GroupCatalog).

cell/particle type -> (block -> data of block, block -> data of block, ...).

cell/particle type -> (block -> data of block, block -> data of block, ...).

cell/particle type -> (block -> data of block, block -> data of block, ...).

...

groupcat type -> (block -> data of block, block -> data of block, ...).

groupcat type -> (block -> data of block, block -> data of block, ...).

groupcat type -> (block -> data of block, block -> data of block, ...).

...

min_ρ::Unitful.Temperature: Minimum gas density.

max_ρ::Unitful.Temperature: Maximum gas density.

Returns

A dictionary with the following shape:

cell/particle type -> idxs::IndexType

cell/particle type -> idxs::IndexType

cell/particle type -> idxs::IndexType

...

GalaxyInspector.filterHotGas — Method

filterHotGas(data_dict::Dict, max_temp::Unitful.Temperature)::Dict{Symbol,IndexType}

Filter out gas cells hotter than max_temp.

Arguments

data_dict::Dict: A dictionary with the following shape:

:sim_data -> ::Simulation (see Simulation).

:snap_data -> ::Snapshot (see Snapshot).

:gc_data -> ::GroupCatalog (see GroupCatalog).

cell/particle type -> (block -> data of block, block -> data of block, ...).

cell/particle type -> (block -> data of block, block -> data of block, ...).

cell/particle type -> (block -> data of block, block -> data of block, ...).

...

groupcat type -> (block -> data of block, block -> data of block, ...).

groupcat type -> (block -> data of block, block -> data of block, ...).

groupcat type -> (block -> data of block, block -> data of block, ...).

...

max_temp::Unitful.Temperature: Maximum gas temperature.

Returns

A dictionary with the following shape:

cell/particle type -> idxs::IndexType

cell/particle type -> idxs::IndexType

cell/particle type -> idxs::IndexType

...

GalaxyInspector.filterInsituStars — Method

filterInsituStars(
    data_dict::Dict;
    <keyword arguments>
)::Dict{Symbol,IndexType}

Filter out stars that where born outside the given halo and subhalo (exsitu), leaving only the ones born inside the halo and subhalo (insitu).

Arguments

data_dict::Dict: A dictionary with the following shape:

:sim_data -> ::Simulation (see Simulation).

:snap_data -> ::Snapshot (see Snapshot).

:gc_data -> ::GroupCatalog (see GroupCatalog).

cell/particle type -> (block -> data of block, block -> data of block, ...).

cell/particle type -> (block -> data of block, block -> data of block, ...).

cell/particle type -> (block -> data of block, block -> data of block, ...).

...

groupcat type -> (block -> data of block, block -> data of block, ...).

groupcat type -> (block -> data of block, block -> data of block, ...).

groupcat type -> (block -> data of block, block -> data of block, ...).

...

halo_idx::Int=1: Index of the target halo (FoF group). Starts at 1.

subhalo_rel_idx::Int=1: Index of the target subhalo (subfind), relative the target halo. Starts at 1. If set to 0, all subhalos of the target halo are consider insitu.

warnings::Bool=true: If a warning will be given when there is missing data.

Returns

A dictionary with the following shape:

cell/particle type -> idxs::IndexType

cell/particle type -> idxs::IndexType

cell/particle type -> idxs::IndexType

...

GalaxyInspector.filterMetallicity — Method

filterMetallicity(data_dict::Dict, l_Z::Float64, h_Z::Float64)::Dict{Symbol,IndexType}

Filter out gas cells and stellar particles with metallicity outside the range [l_Z, h_Z].

Arguments

data_dict::Dict: A dictionary with the following shape:

:sim_data -> ::Simulation (see Simulation).

:snap_data -> ::Snapshot (see Snapshot).

:gc_data -> ::GroupCatalog (see GroupCatalog).

cell/particle type -> (block -> data of block, block -> data of block, ...).

cell/particle type -> (block -> data of block, block -> data of block, ...).

cell/particle type -> (block -> data of block, block -> data of block, ...).

...

groupcat type -> (block -> data of block, block -> data of block, ...).

groupcat type -> (block -> data of block, block -> data of block, ...).

groupcat type -> (block -> data of block, block -> data of block, ...).

...

l_Z::Float64: Minimum metallicity.

h_Z::Float64: Maximum metallicity.

Returns

A dictionary with the following shape:

cell/particle type -> idxs::IndexType

cell/particle type -> idxs::IndexType

cell/particle type -> idxs::IndexType

...

GalaxyInspector.filterNothing — Method

Default filter function that does not filter any cells/particles.

GalaxyInspector.filterStellarAge — Function

filterStellarAge(data_dict::Dict; <keyword arguments>)::Dict{Symbol,IndexType}

Filter out stars that are older than age.

Arguments

data_dict::Dict: A dictionary with the following shape:

:sim_data -> ::Simulation (see Simulation).

:snap_data -> ::Snapshot (see Snapshot).

:gc_data -> ::GroupCatalog (see GroupCatalog).

cell/particle type -> (block -> data of block, block -> data of block, ...).

cell/particle type -> (block -> data of block, block -> data of block, ...).

cell/particle type -> (block -> data of block, block -> data of block, ...).

...

groupcat type -> (block -> data of block, block -> data of block, ...).

groupcat type -> (block -> data of block, block -> data of block, ...).

groupcat type -> (block -> data of block, block -> data of block, ...).

...

age::Unitful.Time=AGE_RESOLUTION: Stars older than this value will be filtered out.

Returns

A dictionary with the following shape:

cell/particle type -> idxs::IndexType

cell/particle type -> idxs::IndexType

cell/particle type -> idxs::IndexType

...

GalaxyInspector.filterSubhalo — Method

filterSubhalo(data_dict::Dict, subhalo_abs_idx::Int)::Dict{Symbol,IndexType}

Filter out cells/particles that do not belong to a given subhalo.

Arguments

data_dict::Dict: A dictionary with the following shape:

:sim_data -> ::Simulation (see Simulation).

:snap_data -> ::Snapshot (see Snapshot).

:gc_data -> ::GroupCatalog (see GroupCatalog).

cell/particle type -> (block -> data of block, block -> data of block, ...).

cell/particle type -> (block -> data of block, block -> data of block, ...).

cell/particle type -> (block -> data of block, block -> data of block, ...).

...

groupcat type -> (block -> data of block, block -> data of block, ...).

groupcat type -> (block -> data of block, block -> data of block, ...).

groupcat type -> (block -> data of block, block -> data of block, ...).

...

subhalo_abs_idx::Int: Index of the target subhalo (subfind). Starts at 1.

Returns

A dictionary with the following shape:

cell/particle type -> idxs::IndexType

cell/particle type -> idxs::IndexType

cell/particle type -> idxs::IndexType

...

GalaxyInspector.filterSubhalo — Method

filterSubhalo(
    data_dict::Dict;
    <keyword arguments>
)::Dict{Symbol,IndexType}

Filter out cells/particles that do not belong to a given halo and subhalo.

Arguments

data_dict::Dict: A dictionary with the following shape:

:sim_data -> ::Simulation (see Simulation).

:snap_data -> ::Snapshot (see Snapshot).

:gc_data -> ::GroupCatalog (see GroupCatalog).

cell/particle type -> (block -> data of block, block -> data of block, ...).

cell/particle type -> (block -> data of block, block -> data of block, ...).

cell/particle type -> (block -> data of block, block -> data of block, ...).

...

groupcat type -> (block -> data of block, block -> data of block, ...).

groupcat type -> (block -> data of block, block -> data of block, ...).

groupcat type -> (block -> data of block, block -> data of block, ...).

...

halo_idx::Int: Index of the target halo (FoF group). Starts at 1.

subhalo_rel_idx::Int: Index of the target subhalo (subfind), relative the target halo. Starts at 1. If set to 0, all subhalos of the target halo are included.

Returns

A dictionary with the following shape:

cell/particle type -> idxs::IndexType

cell/particle type -> idxs::IndexType

cell/particle type -> idxs::IndexType

...

GalaxyInspector.filterWithinCylinder — Method

filterWithinCylinder(
    data_dict::Dict,
    max_r::Unitful.Length,
    max_z::Unitful.Length,
    origin...,
)::Dict{Symbol,IndexType}

Filter out the cell/particles outside a given cylinder.

Arguments

data::Dict: A dictionary with the following shape:

:sim_data -> ::Simulation (see Simulation).

:snap_data -> ::Snapshot (see Snapshot).

:gc_data -> ::GroupCatalog (see GroupCatalog).

cell/particle type -> (block -> data of block, block -> data of block, ...).

cell/particle type -> (block -> data of block, block -> data of block, ...).

cell/particle type -> (block -> data of block, block -> data of block, ...).

...

groupcat type -> (block -> data of block, block -> data of block, ...).

groupcat type -> (block -> data of block, block -> data of block, ...).

groupcat type -> (block -> data of block, block -> data of block, ...).

...

max_r::Unitful.Length: Radius of the cylinder.

max_z::Unitful.Length: Half height of the cylinder.

origin: It can be any number and type of argument compatible with the second to last arguments of a computeCenter method.

Returns

A dictionary with the following shape:

cell/particle type -> idxs::IndexType

cell/particle type -> idxs::IndexType

cell/particle type -> idxs::IndexType

...

GalaxyInspector.filterWithinSphere — Method

filterWithinSphere(
    data_dict::Dict,
    range::NTuple{2,<:Unitful.Length},
    origin...,
)::Dict{Symbol,IndexType}

Filter out the cell/particles outside a given spherical shell.

Arguments

data::Dict: A dictionary with the following shape:

:sim_data -> ::Simulation (see Simulation).

:snap_data -> ::Snapshot (see Snapshot).

:gc_data -> ::GroupCatalog (see GroupCatalog).

cell/particle type -> (block -> data of block, block -> data of block, ...).

cell/particle type -> (block -> data of block, block -> data of block, ...).

cell/particle type -> (block -> data of block, block -> data of block, ...).

...

groupcat type -> (block -> data of block, block -> data of block, ...).

groupcat type -> (block -> data of block, block -> data of block, ...).

groupcat type -> (block -> data of block, block -> data of block, ...).

...

range::NTuple{2,<:Unitful.Length}: Internal and external radius of the spherical shell.

origin: It can be any number and type of argument compatible with the second to last arguments of a computeCenter method.

Returns

A dictionary with the following shape:

cell/particle type -> idxs::IndexType

cell/particle type -> idxs::IndexType

cell/particle type -> idxs::IndexType

...

GalaxyInspector.filterYoungStars — Method

filterYoungStars(data_dict::Dict)::Dict{Symbol,IndexType}

Filter out stars that where born one or more snapshots ago.

Arguments

data_dict::Dict: A dictionary with the following shape:

:sim_data -> ::Simulation (see Simulation).

:snap_data -> ::Snapshot (see Snapshot).

:gc_data -> ::GroupCatalog (see GroupCatalog).

cell/particle type -> (block -> data of block, block -> data of block, ...).

cell/particle type -> (block -> data of block, block -> data of block, ...).

cell/particle type -> (block -> data of block, block -> data of block, ...).

...

groupcat type -> (block -> data of block, block -> data of block, ...).

groupcat type -> (block -> data of block, block -> data of block, ...).

groupcat type -> (block -> data of block, block -> data of block, ...).

...

Returns

A dictionary with the following shape:

cell/particle type -> idxs::IndexType

cell/particle type -> idxs::IndexType

cell/particle type -> idxs::IndexType

...

GalaxyInspector.selectFilter — Method

selectFilter(
    filter_mode::Dict{Symbol,Any},
    request::Dict{Symbol,Vector{String}},
)::Tuple{
    Function,
    Union{Symbol,NTuple{2,Int},Int},
    Union{Symbol,NTuple{2,Int},Int},
    Dict{Symbol,Vector{String}},
}

Select the filter function, translation, and rotation from filter_mode.

Creates a request dictionary, using request as a base, adding what is necessary for the filter function and corresponding transformations.

Arguments

filter_mode::Dict{Symbol,Any}: A dictionary with three entries:

:filter_function -> The filter function.

:translation -> Translation for the simulation box. The posibilities are:

:global_cm -> Selects the center of mass of the whole system as the new origin.

:{component} -> Sets the center of mass of the given component (e.g. :stars, :gas, :halo, etc) as the new origin. It can be any of the keys of PARTICLE_INDEX.

(halo_idx, subhalo_rel_idx) -> Sets the position of the potencial minimum for the subhalo_rel_idx::Int subhalo (of the halo_idx::Int halo) as the new origin.

(halo_idx, 0) -> Sets the center of mass of the halo_idx::Int halo as the new origin.

subhalo_abs_idx -> Sets the center of mass of the subhalo_abs_idx::Int as the new origin.

:rotation -> Rotation for the simulation box. The posibilities are:

:zero -> No rotation is appplied.

:global_am -> Sets the angular momentum of the whole system as the new z axis.

:stellar_am -> Sets the stellar angular momentum as the new z axis.

:stellar_pa -> Sets the stellar principal axis as the new coordinate system.

:stellar_subhalo_pa -> Sets the principal axis of the stars in the main subhalo as the new coordinate system.

(halo_idx, subhalo_rel_idx) -> Sets the principal axis of the stars in subhalo_rel_idx::Int subhalo (of the halo_idx::Int halo) as the new coordinate system.

(halo_idx, 0) -> Sets the principal axis of the stars in the halo_idx::Int halo as the new coordinate system.

subhalo_abs_idx -> Sets the principal axis of the stars in the subhalo_abs_idx::Int subhalo as the new coordinate system.

request::Dict{Symbol,Vector{String}}: Base request dictionary, nothing will be deleted from it.

Returns

A Tuple with four elements:

The filter function.

Translation for the simulation box. The posibilities are:

:global_cm -> Selects the center of mass of the whole system as the new origin.

:{component} -> Sets the center of mass of the given component (e.g. :stars, :gas, :halo, etc) as the new origin. It can be any of the keys of PARTICLE_INDEX.

(halo_idx, subhalo_rel_idx) -> Sets the position of the potencial minimum for the subhalo_rel_idx::Int subhalo (of the halo_idx::Int halo) as the new origin.

(halo_idx, 0) -> Sets the center of mass of the halo_idx::Int halo as the new origin.

subhalo_abs_idx -> Sets the center of mass of the subhalo_abs_idx::Int as the new origin.

Rotation for the simulation box. The posibilities are:

:zero -> No rotation is appplied.

:global_am -> Sets the angular momentum of the whole system as the new z axis.

:stellar_am -> Sets the stellar angular momentum as the new z axis.

:stellar_pa -> Sets the stellar principal axis as the new coordinate system.

:stellar_subhalo_pa -> Sets the principal axis of the stars in the main subhalo as the new coordinate system.

(halo_idx, subhalo_rel_idx) -> Sets the principal axis of the stars in subhalo_rel_idx::Int subhalo (of the halo_idx::Int halo) as the new coordinate system.

(halo_idx, 0) -> Sets the principal axis of the stars in the halo_idx::Int halo as the new coordinate system.

subhalo_abs_idx -> Sets the principal axis of the stars in the subhalo_abs_idx::Int subhalo as the new coordinate system.

New request dictionary.

GalaxyInspector.selectFilter — Method

selectFilter(
    filter_mode::Symbol,
    request::Dict{Symbol,Vector{String}},
)::Tuple{Function,Union{Symbol,NTuple{2,Int}},Symbol,Dict{Symbol,Vector{String}}}

Select a filter function, and the corresponding translation and rotation for the simulation box, from a list of premade ones.

Creates a request dictionary, using request as a base, adding what is necessary for the filter function and corresponding transformations.

Arguments

filter_mode::Symbol: Which cells/particles will be plotted, the options are:

:all -> Plot every cell/particle within the simulation box.

:halo -> Plot only the cells/particles that belong to the main halo.

:subhalo -> Plot only the cells/particles that belong to the main subhalo.

:sphere -> Plot only the cell/particle inside a sphere with radius DISK_R (see ./src/constants.jl).

:stellar_subhalo -> Plot only the cells/particles that belong to the main subhalo.

:all_subhalo -> Plot every cell/particle centered around the main subhalo.

request::Dict{Symbol,Vector{String}}: Base request dictionary, nothing will be deleted from it.

Returns

A Tuple with four elements:

The filter function.

Translation for the simulation box. The posibilities are:

:global_cm -> Selects the center of mass of the whole system as the new origin.

:{component} -> Sets the center of mass of the given component (e.g. :stars, :gas, :halo, etc) as the new origin. It can be any of the keys of PARTICLE_INDEX.

(halo_idx, subhalo_rel_idx) -> Sets the position of the potencial minimum for the subhalo_rel_idx::Int subhalo (of the halo_idx::Int halo), as the new origin.

(halo_idx, 0) -> Selects the center of mass of the halo_idx::Int halo, as the new origin.

Rotation for the simulation box. The posibilities are:

:global_am -> Sets the angular momentum of the whole system as the new z axis.

:stellar_am -> Sets the stellar angular momentum as the new z axis.

:stellar_pa -> Sets the stellar principal axis as the new coordinate system.

:stellar_subhalo_pa -> Sets the principal axis of the stars in the main subhalo as the new coordinate system.

New request dictionary.