Filter functions

These functions are used internally and none are exported.

They may depend on the particulars of the simulation code (e.g. internal units).

GalaxyInspector.filterAll — Method

Filter function that filters out all cells/particles.

GalaxyInspector.filterByBirthPlace — Method

filterByBirthPlace(
    data_dict::Dict,
    exclude::Symbol;
    <keyword arguments>
)::Dict{Symbol,IndexType}

Select stars that were born either inside the given halo and subhalo (exclude= :exsitu), or outside (exclude= :insitu).

Arguments

data_dict::Dict: Data dictionary (see makeDataDict for the canonical description).

exclude::Symbol: Which stars will be excluded, either the ones born outside the given halo and subhalo (:exsitu), or inside (:insitu).

halo_idx::Int=1: Index of the target halo (FoF group). Starts at 1.

subhalo_rel_idx::Int=1: Index of the target subhalo (subfind), relative to the target halo. Starts at 1. If it is set to 0, all subhalos of the target halo are considered in-situ.

Returns

A filter dictionary.

GalaxyInspector.filterByCylinder — Method

filterByCylinder(
    data_dict::Dict,
    max_r::Unitful.Length,
    max_z::Unitful.Length,
    origin...,
)::Dict{Symbol,IndexType}

Select the cells/particles within a given cylinder.

Arguments

data_dict::Dict: Data dictionary (see makeDataDict for the canonical description).

max_r::Unitful.Length: Radius of the cylinder.

max_z::Unitful.Length: Half height of the cylinder.

origin...: It can be any number and type of argument compatible with the second to last arguments of a computeCenter method.

Returns

A filter dictionary.

GalaxyInspector.filterByQuantity — Method

filterByQuantity(
    data_dict::Dict,
    quantity::Symbol,
    component::Symbol,
    min::Number,
    max::Number,
)::Dict{Symbol,IndexType}

Select particles/cells with a value of quantity within [min, max].

Arguments

data_dict::Dict: Data dictionary (see makeDataDict for the canonical description).

quantity::Symbol: Target quantity. For the possibilities see the documentation of scatterQty.

component::Symbol: Type of particle/cell. The possibilities are the keys of PARTICLE_INDEX.

min::Number: Minimum value of quantity.

max::Number: Maximum value of quantity.

Returns

A filter dictionary.

GalaxyInspector.filterBySFM — Method

filterBySFM(data_dict::Dict)::Dict{Symbol,IndexType}

Select the gas cells that have entered our star formation routine at least once.

Arguments

data_dict::Dict: Data dictionary (see makeDataDict for the canonical description).

Returns

A filter dictionary.

GalaxyInspector.filterBySphere — Method

filterBySphere(
    data_dict::Dict,
    min_r::Unitful.Length,
    max_r::Unitful.Length,
    origin...,
)::Dict{Symbol,IndexType}

Select the cells/particles within a given spherical shell.

Arguments

data_dict::Dict: Data dictionary (see makeDataDict for the canonical description).

min_r::Unitful.Length: Internal radius of the spherical shell.

max_r::Unitful.Length: External radius of the spherical shell.

origin...: It can be any number and type of argument compatible with the second to last arguments of a computeCenter method.

Returns

A filter dictionary.

GalaxyInspector.filterByStellarAge — Method

filterByStellarAge(
    data_dict::Dict;
    <keyword arguments>
)::Dict{Symbol,IndexType}

Select stars with an age within [min_age, max_age].

Arguments

data_dict::Dict: Data dictionary (see makeDataDict for the canonical description).

min_age::Unitful.Time=0.0u"Gyr": Minimum age.

max_age::Unitful.Time=AGE_RESOLUTION: Maximum age.

Returns

A filter dictionary.

GalaxyInspector.filterBySubhalo — Method

filterBySubhalo(data_dict::Dict, subhalo_abs_idx::Int)::Dict{Symbol,IndexType}

Select the cells/particles that belong to a given subhalo.

Arguments

data_dict::Dict: Data dictionary (see makeDataDict for the canonical description).

subhalo_abs_idx::Int: Index of the target subhalo (subfind). Starts at 1.

Returns

A filter dictionary.

GalaxyInspector.filterBySubhalo — Method

filterBySubhalo(
    data_dict::Dict;
    <keyword arguments>
)::Dict{Symbol,IndexType}

Select the cells/particles that belong to a given halo and subhalo.

Arguments

data_dict::Dict: Data dictionary (see makeDataDict for the canonical description).

halo_idx::Int=1: Index of the target halo (FoF group). Starts at 1.

subhalo_rel_idx::Int=1: Index of the target subhalo (subfind), relative to the target halo. Starts at 1. If it is set to 0, all subhalos of the target halo are included.

Returns

A filter dictionary.

GalaxyInspector.filterData! — Method

filterData!(data_dict::Dict; <keyword arguments>)::Nothing

Filter data_dict using the indices provided by filter_function.

Arguments

data_dict::Dict: Data dictionary (see makeDataDict for the canonical description).

filter_function::Function=filterNothing: Filter function. See the required signature and examples in ./src/analysis/filters.jl.

GalaxyInspector.filterData — Method

filterData(data_dict::Dict; <keyword arguments>)::Dict

Return a filtered copy of data_dict using the indices provided by filter_function.

Arguments

data_dict::Dict: Data dictionary (see makeDataDict for the canonical description).

filter_function::Function=filterNothing: Filter function. See the required signature and examples in ./src/analysis/filters.jl.

Returns

The filtered data.

GalaxyInspector.filterNothing — Method

Filter function that does not filter out any cell/particle.

GalaxyInspector.filterOldStars — Method

filterOldStars(data_dict::Dict)::Dict{Symbol,IndexType}

Select stars born after the previous snapshot.

Arguments

data_dict::Dict: Data dictionary (see makeDataDict for the canonical description).

Returns

A filter dictionary.

GalaxyInspector.intersectFilters — Method

intersectFilters(filters...)::Dict{Symbol,IndexType}

Generate the filter dictionary resulting from intersecting filters (using AND in boolean logic).

Arguments

filters: Filter dictionary.

Returns

The filter dictionary resulting from intersecting filters.

GalaxyInspector.intersectFilters — Method

intersectFilters(filters::Dict{Symbol,IndexType})::Dict{Symbol,IndexType}

Trivial method for the case of a single argument.

Arguments

filters::Dict{Symbol,IndexType}: Filter dictionary.

Returns

The filter dictionary given as argument.

GalaxyInspector.intersectFilters — Method

intersectFilters(filters::Function...)::Function

Generate the filter function resulting from intersecting filters (using AND in boolean logic).

Arguments

filters: Filter functions.

Returns

The filter function resulting from intersecting filters.

GalaxyInspector.invertFilter — Method

invertFilter(filter::Function)::Function

Invert a filter function

Arguments

filter::Function: Filter function to be inverted.

Returns

The inverted filter function.

GalaxyInspector.invertFilterDict — Method

invertFilterDict(filter::Dict{Symbol,IndexType})::Dict{Symbol,IndexType}

Invert a given filter dictionary.

Arguments

filter::Dict{Symbol,IndexType}: Filter dictionary to be inverted.

Returns

The inverted filter dictionary.

GalaxyInspector.selectFilter — Method

selectFilter(
    filter_mode::Symbol,
    base_request::Dict{Symbol,Vector{String}},
)::Tuple{Function,Dict{Symbol,Vector{String}}}

Select a filter function from a list of premade ones.

Creates a request dictionary, using base_request as a base, adding what is necessary for the filter function.

Arguments

filter_mode::Symbol: Which cells/particles will be selected, the options are:

:all -> Select every cell/particle within the simulation box.

:halo -> Select the cells/particles that belong to the main halo.

:subhalo -> Select the cells/particles that belong to the main subhalo.

:sphere -> Select the cells/particles inside a sphere centered at the origin with radius DISK_R (see ./src/constants/globals.jl).

:cylinder -> Select the cells/particles inside a cylinder centered at the origin with radius DISK_R and half height DISK_HEIGHT (see ./src/constants/globals.jl). The cylinder axis is aligned with the z axis.

base_request::Dict{Symbol,Vector{String}}: Base request dictionary.

Returns

A Tuple with two elements:

The filter function.

The new request dictionary.

GalaxyInspector.selectFilter — Method

selectFilter(
    filter_mode::Tuple{Function,Dict{Symbol,Vector{String}}},
    base_request::Dict{Symbol,Vector{String}},
)::Tuple{Function,Dict{Symbol,Vector{String}}}

Compatibility method to pass personalized filters to functions that only use the filter_mode argument.

Arguments

filter_mode::Tuple{Function,Dict{Symbol,Vector{String}}}: Filter function and the request that goes with it.

base_request::Dict{Symbol,Vector{String}}: Base request dictionary.

Returns

A Tuple with two elements:

The filter function.

The new request dictionary.