Tracer functions

These functions are used internally and none are exported.

They may depend on the particulars of the simulation code (e.g. internal units).

GalaxyInspector.findTracers — Method

findTracers(data_dict::Dict; <keyword arguments>)::Vector{UInt}

Find the tracers whose parents are allowed by filter_function.

Arguments

data_dict::Dict: Data dictionary (see makeDataDict for the canonical description).

filter_function::Function=filterNothing: Filter function. See the required signature and examples in ./src/analysis/filters.jl.

Returns

A vector with the IDs of the tracers.

GalaxyInspector.parentIDToIndex — Method

parentIDToIndex(data_dict::Dict, target_ids::Vector{<:Unsigned})::Dict{Symbol,Vector{Int}}

Find the indices of the cells/particles given their IDs.

Arguments

data_dict::Dict: Data dictionary (see makeDataDict for the canonical description).

target_ids::Vector{<:Unsigned}: List of parent IDs.

Returns

A dictionary with the following shape:

cell/particle type::Symbol -> indices::Int

cell/particle type::Symbol -> indices::Int

cell/particle type::Symbol -> indices::Int

...

GalaxyInspector.parentToTracerID — Method

parentToTracerID(data_dict::Dict, target_ids::Vector{<:Unsigned})::Vector{UInt}

Find the tracer IDs given a list of parent IDs.

Arguments

data_dict::Dict: Data dictionary (see makeDataDict for the canonical description).

target_ids::Vector{<:Unsigned}: List of parent IDs.

Returns

A vector with the IDs of the tracers.

GalaxyInspector.tracerToParentID — Method

tracerToParentID(data_dict::Dict, target_ids::Vector{<:Unsigned})::Vector{UInt}

Find the parent IDs given a list of tracer IDs.

Arguments

data_dict::Dict: Data dictionary (see makeDataDict for the canonical description).

target_ids::Vector{<:Unsigned}: List of tracer IDs.

Returns

A vector with the IDs of the parents.

GalaxyInspector.tracersToParentMass — Method

tracersToParentMass(
    data_dict::Dict,
    target_ids::Vector{<:Unsigned},
)::Dict{Symbol,Vector{<:Unitful.Mass}}

Find the masses of the parents given a list of tracer IDs.

Arguments

data_dict::Dict: Data dictionary (see makeDataDict for the canonical description).

target_ids::Vector{<:Unsigned}: List of tracer IDs.

Returns

A dictionary with the following shape:

cell/particle type -> masses::Vector{<:Unitful.Mass}

cell/particle type -> masses::Vector{<:Unitful.Mass}

cell/particle type -> masses::Vector{<:Unitful.Mass}

...

GalaxyInspector.tracersWithinDisc — Method

tracersWithinDisc(data_dict::Dict; <keyword arguments>)::Vector{UInt}

Find the tracers whose parents are within a given cylinder centered at the origin.

Arguments

data_dict::Dict: Data dictionary (see makeDataDict for the canonical description).

max_r::Unitful.Length=DISK_R: Radius of the cylinder.

max_z::Unitful.Length=DISK_HEIGHT: Half height of the cylinder.

Returns

A vector with the IDs of the tracers.

GalaxyInspector.tracersWithinR200 — Method

tracersWithinR200(data_dict::Dict; <keyword arguments>)::Vector{UInt}

Find the tracers whose parents are within the virial radius ($R_{200}$) of a given halo.

Note

The center of the target halo is its potential minimum in the simulation reference frame. This will not work if data_dict has been translated before calling this function.

Arguments

data_dict::Dict: Data dictionary (see makeDataDict for the canonical description).

halo_idx::Int=1: Index of the target halo (FoF group). Starts at 1.

Returns

A vector with the IDs of the tracers.