Coordinate transformation functions

These functions are used internally and none are exported.

They may depend on the particulars of the simulation code (e.g. internal units).

GalaxyInspector.computeRotation — Method

computeRotation(
    data_dict::Dict,
    z_axis::Symbol,
    component::Symbol,
    filter_function::Function,
)::Union{Matrix{Float64},UniformScaling{Bool}}

Compute the corresponding rotation matrix.

Arguments

data_dict::Dict: Data dictionary (see makeDataDict for the canonical description).

z_axis::Symbol: Target reference system axis. The options are:

:zero -> No rotation is applied.

:am -> The angular momentum will be used as the new z axis.

:pa -> The principal axes will be used as the new coordinate system.

component::Symbol: Which component will be considered to compute the angular momentum or the principal axes. The options are:

:all -> Every component present in data_dict.

:{component} -> Any of the keys of PARTICLE_INDEX, if present in data_dict.

filter_function::Function: A function with the signature:

filter_function(data_dict::Dict) -> filter_dict::Dict{Symbol,IndexType}

Determines which cell/particles will be considered to compute the angular momentum or the principal axes.

Returns

The rotation matrix.

GalaxyInspector.rotateData! — Method

rotateData!(
    data_dict::Dict,
    z_axis::Symbol,
    component::Symbol,
    filter_function::Function,
)::Nothing

Rotate the positions and velocities of the cells/particles in data_dict.

Arguments

data_dict::Dict: Data dictionary (see makeDataDict for the canonical description).

z_axis::Symbol: Target reference system axis. The options are:

:zero -> No rotation is applied.

:am -> The angular momentum will be used as the new z axis.

:pa -> The principal axes will be used as the new coordinate system.

component::Symbol: Which component will be considered to compute the angular momentum or the principal axes. The options are:

:all -> Every component present in data_dict.

:{component} -> Any of the keys of PARTICLE_INDEX, if present in data_dict.

filter_function::Function: A function with the signature:

filter_function(data_dict::Dict) -> filter_dict::Dict{Symbol,IndexType}

Determines which cell/particles will be considered to compute the angular momentum or the principal axes.

GalaxyInspector.rotateData! — Method

rotateData!(
    data_dict::Dict,
    rotation_matrix::Union{Matrix{Float64},UniformScaling{Bool}},
)::Nothing

Rotate the positions and velocities of the cells/particles in data_dict.

Arguments

data_dict::Dict: Data dictionary (see makeDataDict for the canonical description).

rotation_matrix::Union{Matrix{Float64},UniformScaling{Bool}}: Rotation matrix.

GalaxyInspector.selectTransformation — Method

selectTransformation(
    trans_mode::Symbol,
    base_request::Dict{Symbol,Vector{String}},
)::Tuple{TranslationType,RotationType,Dict{Symbol,Vector{String}}}

Select a translation and rotation from a list of premade ones.

Creates a request dictionary, using request as a base, adding what is necessary for the corresponding transformations.

Arguments

trans_mode::Symbol: How to translate and rotate the cells/particles, after filtering with filter_mode. The options are the elements of TRANSFORM_LIST which have the form :{component}_{group} selecting which cells/particle to consider for the center of mass (new origin for the translation) and principal axis (new reference system for the rotation), where component can be:

:all -> Every component present in data_dict

:{component} -> One of the keys of PARTICLE_INDEX

and group can be:

:box -> Whole simulation box

:halo -> Main halo

:subhalo -> Main subhalo

Alternatively, it can be :zero, in which case no translation nor rotation is applied.

base_request::Dict{Symbol,Vector{String}}: Base request dictionary.

Returns

A Tuple with tree elements:

The translation (see the posible arguments of translateData!).

The rotation (see the posible arguments of rotateData!).

The new request dictionary.

GalaxyInspector.selectTransformation — Method

selectTransformation(
    trans_mode::Tuple{TranslationType,RotationType,Dict{Symbol,Vector{String}}},
    base_request::Dict{Symbol,Vector{String}},
)::Tuple{TranslationType,RotationType,Dict{Symbol,Vector{String}}}

Compatibility method to pass personalized transformations to functions that only use the trans_mode argument.

Arguments

trans_mode::Tuple{TranslationType,RotationType,Dict{Symbol,Vector{String}}}: Translation, rotation, and the request that goes with them. See the posible arguments of translateData! and rotateData!.

base_request::Dict{Symbol,Vector{String}}: Base request dictionary.

Returns

A Tuple with tree elements:

The translation (see the posible arguments of translateData!).

The rotation (see the posible arguments of rotateData!).

The new request dictionary.

GalaxyInspector.translateData! — Method

translateData!(data_dict::Dict, translation::Union{Symbol,NTuple{2,Int},Int}...)::Nothing

Translate the positions and velocities of the cells/particles in data_dict.

Arguments

data_dict::Dict: Data dictionary (see makeDataDict for the canonical description).

translation::Union{Symbol,NTuple{2,Int},Int}: Target translation. The options are:

:zero -> No translation is applied.

:all -> Sets the center of mass of the whole system as the new origin.

:{component} -> Sets the center of mass of the given component (e.g. :stellar, :gas, :darkmatter, etc) as the new origin. It can be any of the keys of [`PARTICLEINDEX`](@ref).

(halo_idx, subhalo_rel_idx) -> Sets the position of the potential minimum for the subhalo_rel_idx::Int subhalo (of the halo_idx::Int halo) as the new origin.

(halo_idx, 0) -> Sets the center of mass of the halo_idx::Int halo as the new origin.

subhalo_abs_idx -> Sets the center of mass of the subhalo_abs_idx::Int as the new origin.

GalaxyInspector.translateData! — Method

translateData!(
    data_dict::Dict,
    origin::Vector{<:Unitful.Length},
    vcm::Vector{<:Unitful.Velocity},
)::Nothing

Translate the positions and velocities of the cells/particles in data_dict such that origin and vcm end up both being [0, 0, 0].

Arguments

data_dict::Dict: Data dictionary (see makeDataDict for the canonical description).

origin::Vector{<:Unitful.Length}: Target origin.

vcm::Vector{<:Unitful.Velocity}: Velocity of the center of mass.

GalaxyInspector.translatePoints! — Method

translatePoints!(points::Matrix{<:Number}, origin::Vector{<:Number})::Nothing

Translate a system of points, moving origin to [0, 0, 0].

Arguments

points::Matrix{<:Number}: Points to be translated. Each column is a point and each row a dimension.

origin::Vector{<:Number}: Target origin.