Histogram utilities

These functions are used internally and none are exported.

They may depend on the particulars of the simulation code (e.g. internal units).

GalaxyInspector.computeBandProfile — Method

computeBandProfile(
    positions::Matrix{<:Unitful.Length},
    quantity::Vector{<:Number},
    grid::LinearGrid;
    <keyword arguments>
)::NTuple{3,Vector{<:Number}}

Compute a profile of quantity, using an 1D histogram and three given aggregator functions.

Each aggregator functions with the signature:

`agg_func(::Vector{<:Number}) -> ::Number`

will be used to accumulate the values of quantity within each bin.

Arguments

positions::Matrix{<:Unitful.Length}: Positions of the cells/particles. Each column is a cell/particle and each row a dimension.

quantity::Vector{<:Number}: The profile will be of this quantity.

grid::LinearGrid: Linear grid.

flat::Bool=true: If the profile will be 2D, using rings, or 3D, using spherical shells.

density::Bool=false: If the profile will be of the density of quantity.

center_func::Function=x->quantile(x, 0.5): Aggregator function for the central value.

low_func::Function=x->quantile(x, 0.25): Aggregator function for the low value.

high_func::Function=x->quantile(x, 0.75): Aggregator function for the high value.

Returns

A tuple with three elements:

A vector with the central value for each bin.

A vector with the low value for each bin.

A vector with the high value for each bin.

GalaxyInspector.computeProfile — Method

computeProfile(
    positions::Matrix{<:Unitful.Length},
    quantity::Vector{<:Number},
    grid::LinearGrid;
    <keyword arguments>
)::Vector{<:Number}

Compute a profile of quantity, using an 1D histogram.

Arguments

positions::Matrix{<:Unitful.Length}: Positions of the cells/particles. Each column is a cell/particle and each row a dimension.

quantity::Vector{<:Number}: The profile will be of this quantity.

grid::LinearGrid: Linear grid.

norm::Vector{<:Number}=Number[]: The value of quantity in each bin will be divided by the corresponding value of norm.

flat::Bool=true: If the profile will be 2D (rings), or 3D (spherical shells).

total::Bool=true: If the sum (default) or the mean of quantity will be computed for each bin.

cumulative::Bool=false: If the profile will be accumulated.

density::Bool=false: If the profile will be of the density of quantity.

empty_nan::Bool=true: If empty bins will be set to NaN. 0 is used otherwise. Notice that if empty_nan = true and cumulative = true, every bin after the first NaN will be set to NaN.

Returns

Vector with the values of the profile.

GalaxyInspector.findIn2DGrid — Method

findIn2DGrid(
    positions::Matrix{<:Number},
    grid::SquareGrid;
    <keyword arguments>
)::VecOrMat{Int}

Compute the linear or cartesian index of the pixel where each of the positions is located.

Note

For positions outside the grid 0 is returned.

Arguments

positions::Matrix{<:Number}: Target coordinates to be located in the grid. Each column corresponds to a value and each row is a dimension.

grid::SquareGrid: A square grid.

cartesian::Bool=false: If the resulting indices will be cartesian or linear.

Returns

A vector with the linear indices or a matrix with cartesian indices.

GalaxyInspector.findIn3DGrid — Method

findIn3DGrid(
    positions::Matrix{<:Number},
    grid::CubicGrid;
    <keyword arguments>
)::VecOrMat{Int}

Compute the linear or cartesian index of the voxel where each of the positions is located.

Note

For positions outside the grid 0 is returned.

Arguments

positions::Matrix{<:Number}: Target coordinates to be located in the grid. Each column corresponds to a value and each row is a dimension.

grid::CubicGrid: A cubic grid.

cartesian::Bool=false: If the resulting indices will be cartesian or linear.

Returns

A vector with the linear indices or a matrix with cartesian indices.

GalaxyInspector.histogram1D — Method

histogram1D(
    positions::Vector{<:Number},
    grid::LinearGrid;
    <keyword arguments>
)::Vector{Float64}

Compute a 1D histogram of positions.

Arguments

positions::Vector{<:Number}: Values for which the histogram will be constructed.

grid::LinearGrid: A linear grid.

empty_nan::Bool=true: If NaN will be put into empty bins, 0 is used otherwise.

Returns

A vector with the counts.

GalaxyInspector.histogram1D — Method

histogram1D(
    positions::Vector{<:Number},
    values::Vector{<:Number},
    grid::LinearGrid;
    <keyword arguments>
)::Vector{<:Number}

Compute a 1D histogram of values.

Arguments

positions::Vector{<:Number}: Positions of the values within a 1D axis.

values::Vector{<:Number}: The values that will be added up to each bin, according to their positions.

grid::LinearGrid: A linear grid.

total::Bool=true: If the sum (total = true) or the mean (total = false) of values will be computed for each bin.

empty_nan::Bool=true: If NaN will be put into empty bins, 0 is used otherwise.

Returns

A vector with the histogram values.

GalaxyInspector.histogram1D — Method

histogram1D(
    positions::Vector{<:Number},
    values::Vector{<:Number},
    edges::Vector{<:Number};
    <keyword arguments>
)::Vector{<:Number}

Compute a 1D histogram of values.

Arguments

positions::Vector{<:Number}: Positions of the values within a 1D axis.

values::Vector{<:Number}: The values that will be added up in each bin, according to their positions.

edges::Vector{<:Number}: A list of bin edges.

total::Bool=true: If the sum (total = true) or the mean (total = false) of values will be computed for each bin.

empty_nan::Bool=true: If NaN will be put into empty bins, 0 is used otherwise.

Returns

A vector with the histogram values.

GalaxyInspector.histogram1D — Method

histogram1D(
    positions::Vector{<:Number},
    edges::Vector{<:Number};
    <keyword arguments>
)::Vector{Float64}

Compute a 1D histogram of positions.

Arguments

positions::Vector{<:Number}: Values for which the histogram will be constructed.

edges::Vector{<:Number}: A list of bin edges.

empty_nan::Bool=true: If NaN will be put into empty bins, 0 is used otherwise.

Returns

A vector with the counts.

GalaxyInspector.histogram2D — Method

histogram2D(
    positions::Matrix{<:Number},
    values::Vector{<:Number},
    grid::SquareGrid;
    <keyword arguments>
)::Matrix{<:Number}

Compute a 2D histogram of values.

Arguments

positions::Matrix{<:Number}: Positions of the values in the grid. Each column corresponds to a value and each row is a dimension.

values::Vector{<:Number}: The values that will be added up in each square bin, according to their positions.

grid::SquareGrid: A square grid.

total::Bool=true: If the sum (total = true) or the mean (total = false) of values will be computed for each bin.

empty_nan::Bool=true: If NaN will be put into empty bins, 0 is used otherwise.

Returns

A matrix with the histogram values.

GalaxyInspector.histogram2D — Method

histogram2D(
    positions::Matrix{<:Number},
    values::Vector{<:Number},
    x_edges::Vector{<:Number},
    y_edges::Vector{<:Number};
    <keyword arguments>
)::Matrix{<:Number}

Compute a 2D histogram of values.

Arguments

positions::Matrix{<:Number}: Positions of the values in the grid. Each column corresponds to a value and each row is a dimension.

values::Vector{<:Number}: The values that will be added up in each square bin, according to their positions.

x_edges::Vector{<:Number}: A list of bin edges for the x axis.

y_edges::Vector{<:Number}: A list of bin edges for the y axis.

total::Bool=true: If the sum (total = true) or the mean (total = false) of values will be computed for each bin.

empty_nan::Bool=true: If NaN will be put into empty bins, 0 is used otherwise.

Returns

A matrix with the histogram values.

GalaxyInspector.histogram2D — Method

histogram2D(
    positions::Matrix{<:Number},
    x_edges::Vector{<:Number},
    y_edges::Vector{<:Number};
    <keyword arguments>
)::Matrix{Float64}

Compute a 2D histogram of positions.

Arguments

positions::Matrix{<:Number}: Values for which the histogram will be constructed. Each column corresponds to a value and each row is a dimension.

x_edges::Vector{<:Number}: A list of bin edges for the x axis.

y_edges::Vector{<:Number}: A list of bin edges for the y axis.

empty_nan::Bool=true: If NaN will be put into empty bins, 0 is used otherwise.

Returns

A matrix with the counts.

GalaxyInspector.histogram3D — Method

histogram3D(
    positions::Matrix{<:Number},
    values::Vector{<:Number},
    grid::CubicGrid;
    <keyword arguments>
)::Array{<:Number,3}

Compute a 3D histogram of values.

Arguments

positions::Matrix{<:Number}: Positions of the values in the grid. Each column corresponds to a value and each row is a dimension.

values::Vector{<:Number}: The values that will be added up in each square bin, according to their positions.

grid::CubicGrid: A cubic grid.

total::Bool=true: If the sum (total = true) or the mean (total = false) of values will be computed for each bin.

empty_nan::Bool=true: If NaN will be put into empty bins, 0 is used otherwise.

Returns

A 3D tensor with the histogram values.

GalaxyInspector.listHistogram1D — Method

listHistogram1D(
    positions::Vector{<:Number},
    values::Vector{<:Number},
    grid::LinearGrid,
)::Vector{Vector{<:Number}}

Compute a 1D histogram of values, returning the full list of values within each bin.

Arguments

positions::Vector{<:Number}: The positions of values within a 1D axis.

values::Vector{<:Number}: The values that will be sorted, according to their positions.

grid::LinearGrid: A linear grid.

Returns

- A vector with the lists of `values` within each bin.

GalaxyInspector.listHistogram1D — Method

listHistogram1D(
    positions::Vector{<:Number},
    values::Vector{<:Number},
    edges::Vector{<:Number},
)::Vector{Vector{<:Number}}

Compute a 1D histogram of values, returning the full list of values within each bin.

Arguments

positions::Vector{<:Number}: The positions of values within a 1D axis.

values::Vector{<:Number}: The values that will be sorted, according to their positions.

edges::Vector{<:Number}: A list of bin edges.

Returns

- A vector with the lists of `values` within each bin.

GalaxyInspector.listHistogram3D — Method

listHistogram3D(positions::Matrix{<:Number}, grid::CubicGrid)::Array{Vector{Int},3}

Compute a 3D histogram of positions, returning the list of indices that fall within each bin.

Arguments

positions::Matrix{<:Number}: Positions of the points. Each column corresponds to a point and each row is a dimension.

grid::CubicGrid: A cubic grid.

Returns

- A tensor with the indices of the points within each bin.