Auxiliary functions

These functions are used internally and none are exported.

They may depend on the particulars of the simulation code (e.g. internal units).

Base.intersect — Method

New methods for Base.intersect to use with the Colon type.

Base.intersect — Method

New methods for Base.intersect to use with the Vector{Bool} type.

Base.isempty — Method

New method for Base.isempty to check for empty LaTeXStrings.

Base.iszero — Method

New method for Base.iszero to compare IndexType with 0.

Base.iterate — Method

New method for Base.iterate to manage symbols.

Base.round — Method

New method for Base.round to use with the Unitful.Quantity type.

Base.union — Method

New methods for Base.union to use with the Vector{Bool} type.

GalaxyInspector.area — Method

Area of a circle with radius r.

GalaxyInspector.evaluateNormal — Method

evaluateNormal(values::Vector{<:Number})::Vector{<:Number}

Evaluate a normal distribution at values.

The median and standard deviation for the distribution are the ones of values.

Arguments

values::Vector{<:Number}: Vector used to compute the mean and standard deviation of the normal distribution.

Returns

The normal distribution evaluated at values.

GalaxyInspector.format_seconds — Method

format_seconds(sec::Float64)::String

Format given number of seconds as "DDd-HHh:MM':SS''", where D is days, H is hours, M is minutes, and S is seconds.

Arguments

sec::Float64: Number of seconds.

Returns

String with the formatted time.

GalaxyInspector.getEmpty — Method

Always returns an empty vector, for any type and number of arguments.

GalaxyInspector.getNothing — Method

Always returns nothing, for any type and number of arguments.

GalaxyInspector.hvcatImages — Method

hvcatImages(
    blocks_per_row::Int,
    paths::Vector{String};
    <keyword arguments>
)::Nothing

Join several images vertically and horizontally.

The images in paths will fill the rows and columns starting at the top left, going from left to right and from top to bottom (row-major order).

Arguments

blocks_per_row::Int: Number of columns.

paths::Vector{String}: Paths to the images.

output_path::String="./joined_images.png": Path to the output image.

GalaxyInspector.isPositive — Method

Test for strict positivity.

GalaxyInspector.metaFormatter — Method

metaFormatter(level::LogLevel, _module, group, id, file, line)

Formatter for loggers.

See the documentation of ConsoleLogger

GalaxyInspector.parserWS — Method

parserWS(value::AbstractString)::Union{Float64,Missing}

Parse a string as a Float64, ignoring white spaces. If the string is empty return missing.

Arguments

value::AbstractString: String to be parsed.

Returns

Number in the string as a Float64.

GalaxyInspector.rangeCut! — Method

rangeCut!(
    raw_values::Vector{<:Number},
    range::Tuple{<:Number,<:Number};
    <keyword arguments>
)::Bool

Delete every element in raw_values that is outside the given range.

Arguments

raw_values::Vector{<:Number}: Dataset that will be pruned.

range::Tuple{<:Number,<:Number}: The range in question.

keep_edges::Bool=true: If the edges of the range will be kept.

min_left::Int=1: Minimum number of values that need to be left after pruning to proceed with the transformation.

Returns

If a transformation was performed.

GalaxyInspector.rangeCut! — Method

rangeCut!(
    m_data::Vector{<:Number},
    s_data::Vector,
    range::Tuple{<:Number,<:Number};
    <keyword arguments>
)::Bool

Delete every element in m_data that is outside the given range.

Every corresponding element in s_data (i.e. with the same index) will be deleted too.

Arguments

m_data::Vector{<:Number}: Master dataset that will be pruned.

s_data::Vector: Slave dataset that will be pruned according to which values of m_data are outside range.

range::Tuple{<:Number,<:Number}: The range in question.

keep_edges::Bool=true: If the edges of the range will be kept.

min_left::Int=1: Minimum number of values that need to be left in the master dataset after pruning to proceed with the transformation.

Returns

If a transformation was performed.

GalaxyInspector.ratio — Method

ratio(qty_1::Symbol, qty_2::Symbol)::Symbol

Returns the symbol for the ratio between two derived quantities.

Arguments

qty_1::Symbol: Symbol for the numerator derived quantity.

qty_2::Symbol: Symbol for the denominator derived quantity.

Returns

The symbol for the ration between qty_1 and qty_2.

GalaxyInspector.ring — Method

ring(vec::Vector, index::Integer)::Vector

Make the indexing operation vec[index] work with modular arithmetic for the indices.

Arguments

vec::Vector: Vector.

index::Integer: Index.

Returns

vec[mod1(index, length(vec))]

Examples

julia> ring([1, 2, 3], 11)
2

julia> ring([1, 2, 3], 3)
3

julia> ring([1, 2, 3], -5)
1

GalaxyInspector.safeSelect — Method

safeSelect(vec::Vector, index::IndexType)

Do the indexing operation vec[index] while ignoring indices that are out of bounds.

Arguments

vec::Vector: Vector.

index::IndexType: Indices. It can be an integer (a single element), a vector of integers (several elements), an UnitRange (e.g. 5:13), an StepRange (e.g. 5:2:13) or (:) (every element).

Returns

vec[index (minus out of bounds indices)]

Examples

julia> safeSelect([1, 2, 3], 11)
Int[]

julia> safeSelect([1, 2, 3], 1:5)
3-element Vector{Int}:
 1
 2
 3

julia> safeSelect([1, 2, 3], 1:3:10)
1

julia> safeSelect([1, 2, 3], [1, 2, 5, 9])
2-element Vector{Int}:
 1
 2

julia> safeSelect([1, 2, 3], (:))
3-element Vector{Int}:
 1
 2
 3

GalaxyInspector.sanitizeData! — Method

sanitizeData!(
    x_data::Vector{<:Number},
    y_data::Vector{<:Number};
    <keyword arguments>
)::NTuple{4,Bool}

Do the following transformations over x_data and y_data, in order:

Trim them to fit within the domain of the functions func_domain[1] and func_domain[2], respectively.

Trim them to fit within range[1] and range[2], respectively.

Scale them down by a factor 10^exp_factor[1] and 10^exp_factor[2], respectively.

By default, no transformation is done.

Note

The datasets must have the same length, and any operation that deletes an element, will delete the corresponding element (i.e. with the same index) in the other dataset, so that the datasets will remain of equal length.

Arguments

x_data::Vector{<:Number}: First dataset to be sanitized.

y_data::Vector{<:Number}: Second dataset to be sanitized.

func_domain::NTuple{2,Function}=(identity, identity): x_data will be trimmed to fit within the domain of the function func_domain[1], and y_data will be trimmed to fit within the domain of the function func_domain[2]. The options are the scaling functions accepted by Makie: log10, log2, log, sqrt, Makie.logit, Makie.Symlog10, Makie.pseudolog10, and identity.

range::Tuple{Tuple{<:Number,<:Number},Tuple{<:Number,<:Number}}=((-Inf, Inf), (-Inf, Inf)): Every element in x_data that falls outside of range[1] will be deleted, and every element in y_data that falls outside of range[2] will be deleted.

keep_edges::NTuple{2,Bool}=(true, true): If the edges of each corresponding range will be kept.

min_left::Int=1: Minimum number of values that need to be left in each dataset after any of the transformations to proceed with them.

exp_factor::NTuple{2,Int}=(0, 0): Every element in x_data will be divided by 10^exp_factor[1], and every element in y_data will be divided by 10^exp_factor[2].

Returns

A tuple with four flags:

If x_data was mutated to fit within the domain of func_domain[1].

If y_data was mutated to fit within the domain of func_domain[2].

If x_data was mutated to fit within range[1].

If y_data was mutated to fit within range[2].

GalaxyInspector.sanitizeData! — Method

sanitizeData!(
    raw_values::Vector{<:Number};
    <keyword arguments>
)::NTuple{2,Bool}

Do the following transformations over raw_values, in order:

Trim it to fit within the domain of the function func_domain.

Trim it to fit within range.

Scale it down by a factor of 10^exp_factor.

By default, no transformation is done.

Arguments

raw_values::Vector{<:Number}: Dataset to be sanitized.

func_domain::Function=identity: raw_values will be trimmed to fit within the domain of the function func_domain. The options are the scaling functions accepted by Makie: log10, log2, log, sqrt, Makie.logit, Makie.Symlog10, Makie.pseudolog10, and identity.

range::Tuple{<:Number,<:Number}=(-Inf, Inf): Every element in raw_values that falls outside of range will be deleted.

keep_edges::Bool=true: If the edges of range will be kept.

min_left::Int=1: Minimum number of values that need to be left after each transformation to proceed with it.

exp_factor::Int=0: Every element in raw_values will be divided by 10^exp_factor.

Returns

A tuple with two flags:

If raw_values was mutated to fit within the domain of func_domain.

If raw_values was mutated to fit within range.

GalaxyInspector.setLogging! — Method

setLogging!(log::Bool; <keyword arguments>)::Nothing

Set if logging messages will be printed out. By default no logs are printed.

Arguments

log::Bool: If logs will be printed out using the default logger.

stream::IO=stdout: Where to print the logs. It can be a file.

GalaxyInspector.setPositive — Method

Return a copy of list with every negative value set to 0.

GalaxyInspector.smoothWindow — Method

smoothWindow(
    x_data::Vector{<:Number},
    y_data::Vector{<:Number},
    n_bins::Int;
    <keyword arguments>
)::NTuple{2,Vector{<:Number}}

Separate the values of x_data in n_bins bins and compute the mean value of x_data and y_data within each one.

Arguments

x_data::Vector{<:Number}: x-axis data.

y_data::Vector{<:Number}: y-axis data.

n_bins::Int: Number of bins.

scaling::Function=identity: Scaling function for the x axis. The options are the scaling functions accepted by Makie: log10, log2, log, sqrt, Makie.logit, Makie.Symlog10, Makie.pseudolog10, and identity. All the values of x_data must be in the domain of scaling.

Returns

A tuple with two vectors, containing the smoothed-out x and y values.

GalaxyInspector.volume — Method

Volume of a sphere with radius r.