Plotting utilities

These functions are used internally and none are exported.

They may depend on the particulars of the simulation code (e.g. internal units).

GalaxyInspector.KSLawBigiel2008 — Method

KSLawBigiel2008(
    ΣH::Vector{<:SurfaceDensity};
    <keyword arguments>
)::Vector{<:Number}

Evaluate the Kennicutt-Schmidt law for the molecular or neutral gas, taken from Bigiel et al. (2008).

From Bigiel et al. (2008) (Section 3.1), we have

\[\Sigma_\mathrm{SFR} = a \left( \frac{\Sigma_\mathrm{HI, H_2, gas}}{10 \, \mathrm{M_\odot \, pc^{-2}}} \right)^{\!N} \, ,\]

where N is the power-law index, and $A = \log_{10}(a)$, where $a$ is $\Sigma_\mathrm{SFR}$ at the fiducial gas surface density of $10 \, \mathrm{M_\odot \, pc^{-2}}$.

Arguments

ΣH::Vector{<:SurfaceDensity}: Values of the molecular or neutral gas surface density.

molecular::Bool=true: If ΣH is the surface density of molecular hydrogen, or of neutral hydrogen.

log_output::Bool=true: If the output will the $\log_{10} \, \Sigma_\text{SFR}$, or $\Sigma_\text{SFR}$. If log_output = true, the implied units are $\mathrm{M_\odot \, yr^{-1} \, kpc^{-2}}$

Returns

The SFR surface density.

References

F. Bigiel et al. (2008). THE STAR FORMATION LAW IN NEARBY GALAXIES ON SUB-KPC SCALES. The Astrophysical Journal, 136(6), 2846. doi:10.1088/0004-6256/136/6/2846

GalaxyInspector.KSLawKennicutt1998 — Method

KSLawKennicutt1998(Σgas::Vector{<:SurfaceDensity}; <keyword arguments>)::Vector{<:Number}

Evaluate the Kennicutt-Schmidt law, taken from Kennicutt (1998).

From Kennicutt (1998) (Section 4, Eq. 4), we have

\[\Sigma_\mathrm{SFR} = a \left( \frac{\Sigma_\mathrm{gas}}{1 \, \mathrm{M_\odot \, pc^{-2}}} \right)^{\!N} \mathrm{M_\odot \, yr^{-1] \, kpc^{-2}} \, ,\]

where N is the power-law index and $a$ is $\Sigma_\mathrm{SFR}$ at the fiducial gas surface density of $1 \, \mathrm{M_\odot \, pc^{-2}}$.

Arguments

Σgas::Vector{<:SurfaceDensity}: Values of the gas mass surface density.

log_output::Bool=true: If the output will the $\log_{10} \, \Sigma_\text{SFR}$, or $\Sigma_\text{SFR}$. If log_output = true, the implied units are $\mathrm{M_\odot \, yr^{-1} \, kpc^{-2}}$

Returns

The SFR surface density.

References

R. C. Kennicutt (1998). The Global Schmidt Law in Star-forming Galaxies. The Astrophysical Journal, 498(2), 541-552. doi:10.1086/305588

GalaxyInspector.absCoor — Method

absCoor(
    plot::Union{Makie.FigureAxisPlot,Makie.Axis,Makie.Figure},
    r_x::Real,
    r_y::Real,
)::NTuple{2,Float64}

Compute the absolute x and y coordinates of a plot, from the relative ones.

Arguments

plot::Union{Makie.FigureAxisPlot,Makie.Axis,Makie.Figure}: Plot, axis, or figure for which the absolute coordinates will be calculated. In the case of a figure, it will use the limits from the current axis object.

r_x::Real: Relative x coordinate.

r_y::Real: Relative y coordinate.

Returns

A tuple with the absolute coordinates, (x, y).

Examples

julia> absCoor(lines(rand(100)), 0.5, 0.5)
(50.50000071525574, 0.48792968317866325)

GalaxyInspector.barPlotLabelFormater — Method

barPlotLabelFormater(x::LaTeXString)::LaTeXString

Format a number to be a barplot label.

Method for compatibility with the barplot! function of Makie.

Arguments

x::Number: Value to be formatted.

Returns

The bar label.

GalaxyInspector.barPlotLabelFormater — Method

barPlotLabelFormater(x::Number)::LaTeXString

Format a number to be a barplot label.

For values between 0 and 0.01 the label will be "< 0.01", otherwise it will be the value itself with 2 digits.

Arguments

x::Number: Value to be formatted.

Returns

The bar label.

GalaxyInspector.cleanPlot! — Method

Default function to end cleanPlot! recursion if an unknown type is encountered.

GalaxyInspector.cleanPlot! — Method

cleanPlot!(ax::Makie.Axis)::Bool

Empty an axis.

Arguments

ax::Makie.Axis: Axis to be emptied.

Returns

Flag to indicate that an axis has been emptied.

GalaxyInspector.cleanPlot! — Method

cleanPlot!(figure::Makie.Figure)::Nothing

Delete all the legends of a figure and empty all its axes.

Arguments

figure::Makie.Figure: Figure to be cleaned.

GalaxyInspector.cleanPlot! — Method

cleanPlot!(legend::Union{Makie.Legend,Makie.Colorbar})::Bool

Delete a legend or colorbar.

Arguments

legend::Union{Makie.Legend,Makie.Colorbar}: Legend or colorbar to be deleted.

Returns

Flag to indicate that a legend or colorbar has been deleted.

GalaxyInspector.derivedQtyPlotParams — Method

derivedQtyPlotParams(magnitude::Symbol, component::Symbol)::PlotParams

Return the plotting parameters for a given magnitude of component.

Arguments

magnitude::Symbol: One of the physical magnitudes in MAGNITUDES.

component::Symbol: One of the physical components in COMPONENTS.

Returns

A PlotParams object.

GalaxyInspector.derivedSFMQtyPlotParams — Method

derivedSFMQtyPlotParams(quantity::Symbol)::PlotParams

Return the plotting parameters for a given derived code quantity of our SF model.

Arguments

quantity::Symbol: One of the quantities in SFM_DERIVED_QTY.

Returns

A PlotParams object.

GalaxyInspector.formatError — Method

formatError(q_mean::Number, q_error::Number)::NTuple{2,<:Number}

Nicely format a magnitude with uncertainty.

It follows the traditional rules for error presentation: the error has only one significant digit, unless such digit is a one, in which case two significant digits are used. The mean will have as many digits as to match the last significant position of the error. An error equal to 0 will leave the mean unchanged.

Arguments

q_mean::Number: Mean value.

q_error::Number: Error value. It must be positive.

Returns

A tuple with the formatted mean and error values.

Examples

julia> formatError(69.42069, 0.038796)
(69.42, 0.04)

julia> formatError(69.42069, 0.018796)
(69.421, 0.019)

julia> formatError(15.42, 0.00004)
(15.42, 4.0e-5)

julia> formatError(69.42069, 0.0)
(69.42069, 0.0)

julia> formatError(69.42069, 93.4)
(70.0, 90.0)

julia> formatError(69.42069, 123.4)
(70.0, 120.0)

julia> formatError(15.42069, 16.4)
(15.0, 16.0)

GalaxyInspector.getLabel — Method

getLabel(
    label::AbstractString,
    factor::Int,
    unit::Unitful.Units;
    <keyword arguments>
)::AbstractString

Construct an axis label.

Arguments

label::AbstractString: Variable name.

factor::Int: Exponential factor to scale down the units. If different from 0, a term of the form 10^factor will be added to the label.

unit::Unitful.Units: Unit of the axis.

latex::Bool=true: If the output will be a LaTeXString or a plain String.

Returns

The LaTeXString or String: "label [10^factor unit]". If label is "", an empty string is returned. The factor term only appears if factor != 0, the unit term only appears if unit != Unitful.NoUnits, and the brackets only appears if there are a factor and/or a unit term.

GalaxyInspector.getUnitLabel — Method

getUnitLabel(factor::Int, unit::Unitful.Units; <keyword arguments>)::AbstractString

Construct the unit part of an axis label.

Arguments

factor::Int: Exponential factor to scale down the units. If different from 0, a term of the form 10^factor will be added to the label.

unit::Unitful.Units: Unit of the axis.

latex::Bool=true: If the output will be a LaTeXString or a plain String.

Returns

The LaTeXString or String: "10^factor unit". The factor term only appears if factor != 0, the unit term only appears if unit != Unitful.NoUnits.

GalaxyInspector.haloQtyPlotParams — Method

haloQtyPlotParams(quantity::Symbol)::PlotParams

Return the plotting parameters for a given halo magnitude.

Arguments

quantity::Symbol: Target halo and halo quantity, e.g., :halo_mass_12. The index of the target halo starts at 1 and the halo quantity has to be one of the keys of HALO_KEYS.

Returns

A PlotParams object.

GalaxyInspector.invKSLawBigiel2008 — Method

invKSLawBigiel2008(
    Σsfr ::Vector{<:MassFlowDensity};
    <keyword arguments>
)::Vector{<:Number}

Evaluate the inverse Kennicutt-Schmidt law for the molecular or neutral gas, taken from Bigiel et al. (2008).

From Bigiel et al. (2008) (Section 3.1, Eq. 2), we have

\[\Sigma_\mathrm{SFR} = a \left( \frac{\Sigma_\mathrm{HI, H_2, gas}}{10 \, \mathrm{M_\odot \, pc^{-2}}} \right)^{\!N} \, ,\]

where N is the power-law index, and $A = \log_{10}(a)$, where $a$ is $\Sigma_\mathrm{SFR}$ at the fiducial gas surface density of $10 \, \mathrm{M_\odot \, pc^{-2}}$.

Arguments

Σsfr ::Vector{<:MassFlowDensity}: Values of the SFR surface density.

molecular::Bool=true: If the output will be the surface density of molecular hydrogen, or of neutral hydrogen.

log_output::Bool=true: If the output will the $\log_{10} \, \Sigma_\text{H}$, or $\Sigma_\text{H}$. If log_output = true, the implied units are $10 \, \mathrm{M_\odot \, pc^{-2}}$

Returns

The molecular or neutral gas surface density.

References

F. Bigiel et al. (2008). THE STAR FORMATION LAW IN NEARBY GALAXIES ON SUB-KPC SCALES. The Astrophysical Journal, 136(6), 2846. doi:10.1088/0004-6256/136/6/2846

GalaxyInspector.invKSLawKennicutt1998 — Method

invKSLawKennicutt1998(
    Σsfr::Vector{<:MassFlowDensity};
    <keyword arguments>
)::Vector{<:Number}

Evaluate the inverse Kennicutt-Schmidt law, taken from Kennicutt (1998).

From Kennicutt (1998) (Section 4, Eq. 4), we have

\[\Sigma_\mathrm{SFR} = a \left( \frac{\Sigma_\mathrm{gas}}{1 \, \mathrm{M_\odot \, pc^{-2}}} \right)^{\!N} \mathrm{M_\odot \, yr^{-1] \, kpc^{-2}} \, ,\]

where N is the power-law index and $a$ is $\Sigma_\mathrm{SFR}$ at the fiducial gas surface density of $1 \, \mathrm{M_\odot \, pc^{-2}}$.

Arguments

Σsfr::Vector{<:MassFlowDensity}: Values of the SFR surface density..

log_output::Bool=true: If the output will the $\log_{10} \, \Sigma_\text{gas}$, or $\Sigma_\text{gas}$. If log_output = true, the implied units are $\mathrm{M_\odot \, pc^{-2}}$

Returns

The gas mass surface density.

References

R. C. Kennicutt (1998). The Global Schmidt Law in Star-forming Galaxies. The Astrophysical Journal, 498(2), 541-552. doi:10.1086/305588

GalaxyInspector.plotParams — Method

plotParams(quantity::Symbol)::PlotParams

Return the plotting parameters for a given quantity.

Arguments

quantity::Symbol: Target quantity. See ./src/constants/quantities.jl for options.

Returns

A PlotParams object.

GalaxyInspector.pointData — Method

Extract all the data points in a Makie plot, axis, or figure. In the case of a figure, it will only take the data from the current axis object. It only works for scatter, line, and scatterline plots.

GalaxyInspector.reduceTicks — Method

reduceTicks(hr_ticks::Vector{<:Number}, factor::Int)::Vector{<:Number}

Reduce the length of a given list of axis ticks, while keeping the total length of the axis the same.

Arguments

hr_ticks::Vector{<:Number}: Original "high resolution" list of ticks. It has to be regularly spaced

factor::Int: Factor by which the number of values will be reduced. It has to divide the size of hr_ticks exactly.

Returns

The new shorter list of ticks.

GalaxyInspector.sfmQtyPlotParams — Method

sfmQtyPlotParams(quantity::Symbol)::PlotParams

Return the plotting parameters for a given base code quantity of our SF model.

Arguments

quantity::Symbol: One of the quantities in SFM_QTY.

Returns

A PlotParams object.

GalaxyInspector.validatePlotData — Method

validatePlotData(
    plot_function::Function,
    plot_data::VecOrMat{<:Number}...;
    <keyword arguments>
)::Tuple{Vector{<:VecOrMat{<:Number}},Bool,Bool}

Adapt plot_data to be plotted with plot_function.

This function strips the units of every argument in plot_data, and it may trims down the first two arguments in plot_data.

Arguments

plot_function::Function: Target plotting function from Makie. The supported functions are:

scatter! -> Scatter plot.

lines! -> Line plot.

scatterlines! -> Scatter plot with lines between the markers.

hist! -> Histogram.

heatmap! -> Heatmap.

arrows2d! -> 2D vector field.

barplot! -> Bar plots.

band! -> Band plots.

errorbars! -> Error bars.

plot_data::VecOrMat{<:Number}: Raw plot data.

x_unit::Unitful.Units=Unitful.NoUnits: Target unit for the x axis data (first arguments in plot_data). The values will be converted accordingly. Use the default value of Unitful.NoUnits for dimensionless quantities.

y_unit::Unitful.Units=Unitful.NoUnits: Target unit for the y axis data (second arguments in plot_data). The values will be converted accordingly. Use the default value of Unitful.NoUnits for dimensionless quantities.

x_exp_factor::Int=0: Numerical exponent to scale down the x axis data (first arguments in plot_data), e.g. if x_exp_factor = 10 the values will be divided by $10^{10}$. The default is no scaling.

y_exp_factor::Int=0: Numerical exponent to scale down the y axis data (second arguments in plot_data), e.g. if y_exp_factor = 10 the values will be divided by $10^{10}$. The default is no scaling.

x_trim::NTuple{2,<:Real}=(-Inf, Inf): The data (first arguments in plot_data) will be trim down so the x coordinates fit within x_trim, in the units given by x_unit.

y_trim::NTuple{2,<:Real}=(-Inf, Inf): The data (second arguments in plot_data) will be trim down so the y coordinates fit within y_trim, in the units given by y_unit. This option does not affect histograms.

x_edges::Bool=false: Set it to true if you want to keep the borders of x_trim.

y_edges::Bool=false: Set it to true if you want to keep the borders of y_trim.

x_scale_func::Function=identity: Scaling function for the x axis (first arguments in plot_data). The options are the scaling functions accepted by Makie: log10, log2, log, sqrt, Makie.logit, Makie.Symlog10, Makie.pseudolog10, and identity. The data will be trimmed down to fit within the domain of x_scale_func.

y_scale_func::Function=identity: Scaling function for the y axis (second arguments in plot_data). The options are the scaling functions accepted by Makie: log10, log2, log, sqrt, Makie.logit, Makie.Symlog10, Makie.pseudolog10, and identity. The data will be trimmed down to fit within the domain of y_scale_func.

Returns

A tuple with three elements:

A vector with the elements of plot_data ready to be plotted by plot_function.

If plot_data[1] was mutated to fit within the domain of x_scale_func.

If plot_data[2] was mutated to fit within the domain of y_scale_func (only relevant if length(plot_data) > 1).

GalaxyInspector.xlimits! — Method

Extract the limits of the x axis, from a Makie plot, axis, or figure. In the case of a figure, it will take the limits from the current axis object.

GalaxyInspector.xscale — Method

Extract the scale function of the x axis, from a Makie plot, axis, or figure. In the case of a figure, it will take the scale from the current axis object.

GalaxyInspector.ylimits! — Method

Extract the limits of the y axis, from a Makie plot, axis, or figure. In the case of a figure, it will take the limits from the current axis object.

GalaxyInspector.yscale — Method

Extract the scale function of the y axis, from a Makie plot, axis, or figure. In the case of a figure, it will take the scale from the current axis object.