Plotting utilities

These functions are used internally and none are exported.

They may depend on the particulars of the simulation code (e.g. internal units).

GalaxyInspector.absCoor — Method

absCoor(
    plot::Union{Makie.FigureAxisPlot,Makie.Axis,Makie.Figure},
    r_x::Real,
    r_y::Real,
)::NTuple{2,Float64}

Compute the absolute x and y coordinates of a plot, from the relative ones.

Arguments

plot::Union{Makie.FigureAxisPlot,Makie.Axis,Makie.Figure}: Plot, axis, or figure for which the absolute coordinates will be calculated. In the case of a figure, it will use the limits from the current axis object.

r_x::Real: Relative x coordinate.

r_y::Real: Relative y coordinate.

Returns

A tuple with the absolute coordinates, (x, y).

Examples

julia> absCoor(lines(rand(100)), 0.5, 0.5)
(50.50000071525574, 0.48792968317866325)

GalaxyInspector.barPlotLabelFormater — Method

barPlotLabelFormater(x::LaTeXString)::LaTeXString

Format a number to be a barplot label.

Method for compatibility with the barplot! function of Makie.

Arguments

x::Number: Value to be formatted.

Returns

The bar label.

GalaxyInspector.barPlotLabelFormater — Method

barPlotLabelFormater(x::Number)::LaTeXString

Format a number to be a barplot label.

For values between 0 and 0.01 the label will be "< 0.01", otherwise it will be the value itself with 2 digits.

Arguments

x::Number: Value to be formatted.

Returns

The bar label.

GalaxyInspector.cleanPlot! — Method

Default function to end cleanPlot! recursion if an unknown type is encountered.

GalaxyInspector.cleanPlot! — Method

cleanPlot!(ax::Makie.Axis)::Bool

Empty an axis.

Arguments

ax::Makie.Axis: Axis to be emptied.

Returns

Flag to indicate that an axis has been emptied.

GalaxyInspector.cleanPlot! — Method

cleanPlot!(figure::Makie.Figure)::Nothing

Delete all the legends of a figure and empty all its axes.

Arguments

figure::Makie.Figure: Figure to be cleaned.

GalaxyInspector.cleanPlot! — Method

cleanPlot!(legend::Union{Makie.Legend,Makie.Colorbar})::Bool

Delete a legend or colorbar.

Arguments

legend::Union{Makie.Legend,Makie.Colorbar}: Legend or colorbar to be deleted.

Returns

Flag to indicate that a legend or colorbar has been deleted.

GalaxyInspector.formatError — Method

formatError(q_mean::Number, q_error::Number)::NTuple{2,<:Number}

Nicely format a magnitude with uncertainty.

It follows the traditional rules for error presentation: the error has only one significant digit, unless such digit is a one, in which case two significant digits are used. The mean will have as many digits as to match the last significant position of the error. An error equal to 0 will leave the mean unchanged.

Arguments

q_mean::Number: Mean value.

q_error::Number: Error value. It must be positive.

Returns

A tuple with the formatted mean and error values.

Examples

julia> formatError(69.42069, 0.038796)
(69.42, 0.04)

julia> formatError(69.42069, 0.018796)
(69.421, 0.019)

julia> formatError(15.42, 0.00004)
(15.42, 4.0e-5)

julia> formatError(69.42069, 0.0)
(69.42069, 0.0)

julia> formatError(69.42069, 93.4)
(70.0, 90.0)

julia> formatError(69.42069, 123.4)
(70.0, 120.0)

julia> formatError(15.42069, 16.4)
(15.0, 16.0)

GalaxyInspector.getLabel — Method

getLabel(
    label::AbstractString,
    factor::Int,
    unit::Unitful.Units;
    <keyword arguments>
)::AbstractString

Construct an axis label.

Arguments

label::AbstractString: Variable name.

factor::Int: Exponential factor to scale down the units. If different from 0, a term of the form 10^factor will be added to the label.

unit::Unitful.Units: Unit of the axis.

latex::Bool=true: If the output will be a LaTeXString or a plain String.

Returns

The LaTeXString or String: "label [10^factor unit]". If label is "", an empty string is returned. The factor term only appears if factor != 0, the unit term only appears if unit != Unitful.NoUnits, and the brackets only appears if there are a factor and/or a unit term.

GalaxyInspector.getUnitLabel — Method

getUnitLabel(factor::Int, unit::Unitful.Units; <keyword arguments>)::AbstractString

Construct the unit part of an axis label.

Arguments

factor::Int: Exponential factor to scale down the units. If different from 0, a term of the form 10^factor will be added to the label.

unit::Unitful.Units: Unit of the axis.

latex::Bool=true: If the output will be a LaTeXString or a plain String.

Returns

The LaTeXString or String: "10^factor unit". The factor term only appears if factor != 0, the unit term only appears if unit != Unitful.NoUnits.

GalaxyInspector.plotParams — Method

plotParams(quantity::Symbol)::PlotParams

Return the plotting parameters for a given quantity.

Arguments

quantity::Symbol: The options are:

:stellar_mass -> Stellar mass.

:gas_mass -> Gas mass.

:hydrogen_mass -> Hydrogen mass.

:dm_mass -> Dark matter mass.

:bh_mass -> Black hole mass.

:molecular_mass -> Molecular hydrogen ($\mathrm{H_2}$) mass.

:br_molecular_mass -> Molecular hydrogen ($\mathrm{H_2}$) mass, computed using the pressure relation in Blitz et al. (2006).

:atomic_mass -> Atomic hydrogen ($\mathrm{HI}$) mass.

:ionized_mass -> Ionized hydrogen ($\mathrm{HII}$) mass.

:neutral_mass -> Neutral hydrogen ($\mathrm{HI + H_2}$) mass.

:stellar_gas_mass -> Stellar gas mass (according to our SF model).

:ode_metal_mass -> Metal mass (according to our SF model).

:ode_metallicity -> Metallicity (according to our SF model).

:dust_mass -> Dust mass.

:generic_mass -> Parameters for plots with several different masses.

:stellar_number -> Number of stellar particles.

:gas_number -> Number of gas cells.

:dm_number -> Number of dark matter particles.

:bh_number -> Number of black hole particles.

:molecular_fraction -> Gas mass fraction of molecular hydrogen.

:br_molecular_fraction -> Gas mass fraction of molecular hydrogen, computed using the pressure relation in Blitz et al. (2006).

:atomic_fraction -> Gas mass fraction of atomic hydrogen.

:ionized_fraction -> Gas mass fraction of ionized hydrogen.

:neutral_fraction -> Gas mass fraction of neutral hydrogen.

:molecular_neutral_fraction -> Fraction of molecular hydrogen in the neutral gas.

:ionized_neutral_fraction -> Fraction of ionized gas to neutral gas.

:stellar_gas_fraction -> Stellar gas fraction (according to our SF model).

:metal_gas_fraction -> Metallicity (according to our SF model).

:dust_fraction -> Dust mass fraction.

:generic_fraction -> Parameters for plots with several different fraction.

:gas_mass_density -> Gas mass density.

:hydrogen_mass_density -> Hydrogen mass density.

:gas_number_density -> Gas number density.

:molecular_number_density -> Molecular hydrogen number density.

:br_molecular_number_density -> Molecular hydrogen number density, computed using the pressure relation in Blitz et al. (2006).

:atomic_number_density -> Atomic hydrogen number density.

:ionized_number_density -> Ionized hydrogen number density.

:neutral_number_density -> Neutral hydrogen number density.

:stellar_area_density -> Stellar area mass density, for a radius of DISK_R.

:gas_area_density -> Gas mass surface density, for a radius of DISK_R.

:molecular_area_density -> Molecular mass surface density, for a radius of DISK_R.

:br_molecular_area_density -> Molecular mass surface density, for a radius of DISK_R, computed using the pressure relation in Blitz et al. (2006).

:atomic_area_density -> Atomic hydrogen area mass density, for a radius of DISK_R.

:ionized_area_density -> Ionized hydrogen area mass density, for a radius of DISK_R.

:neutral_area_density -> Neutral mass surface density, for a radius of DISK_R.

:sfr_area_density -> Star formation rate area density, for the last AGE_RESOLUTION and a radius of DISK_R.

:generic_area_density -> Parameters for plots with several different area densities.

:gas_td -> Total gas depletion time.

:molecular_td -> Molecular hydrogen ($\mathrm{H_2}$) depletion time.

:br_molecular_td -> Molecular hydrogen ($\mathrm{H_2}$) depletion time, computed using the pressure relation in Blitz et al. (2006).

:atomic_td -> Atomic hydrogen ($\mathrm{HI}$) depletion time.

:ionized_td -> Ionized hydrogen ($\mathrm{HII}$) depletion time.

:neutral_td -> Neutral hydrogen ($\mathrm{HI + H_2}$) depletion time.

:gas_metallicity -> Mass fraction of all elements above He in the gas (solar units).

:stellar_metallicity -> Mass fraction of all elements above He in the stars (solar units).

:X_gas_abundance -> Gas abundance of element $\mathrm{X}$, as $12 + \log_{10}(\mathrm{X \, / \, H})$. The possibilities are the keys of ELEMENT_INDEX.

:X_stellar_abundance -> Stellar abundance of element $\mathrm{X}$, as $12 + \log_{10}(\mathrm{X \, / \, H})$. The possibilities are the keys of ELEMENT_INDEX.

:stellar_radial_distance -> Distance of every stellar particle to the origin.

:gas_radial_distance -> Distance of every gas cell to the origin.

:dm_radial_distance -> Distance of every dark matter particle to the origin.

:stellar_xy_distance -> Projected distance of every stellar particle to the origin.

:gas_xy_distance -> Projected distance of every gas cell to the origin.

:dm_xy_distance -> Projected distance of every dark matter particle to the origin.

:gas_sfr -> SFR associated to each gas particle/cell within the code.

:mass_accretion -> Gas accretion rate. Positive values mean gas infall into the virial radius $R_{200}$, and negative values mean outflow.

:stellar_specific_am -> Norm of the stellar specific angular momentum.

:gas_specific_am -> Norm of the gas specific angular momentum.

:dm_specific_am -> Norm of the dark matter specific angular momentum.

:stellar_circularity -> Stellar circularity.

:stellar_vcirc -> Stellar circular velocity.

:stellar_vradial -> Stellar radial speed.

:stellar_vtangential -> Stellar tangential speed.

:stellar_vzstar -> Stellar speed in the z direction, computed as $v_z \, \mathrm{sign}(z)$.

:stellar_age -> Stellar age.

:sfr -> Star formation rate.

:ssfr -> The specific star formation rate.

:observational_sfr -> Star formation rate of the last AGE_RESOLUTION.

:observational_ssfr -> The specific star formation rate of the last AGE_RESOLUTION.

:stellar_eff -> Star formation efficiency per free-fall time for the gas that has turn into stars.

:gas_eff -> Star formation efficiency per free-fall time for the gas.

:molecular_eff -> Star formation efficiency per free-fall time for the molecular hydrogen ($\mathrm{H_2}$) gas.

:br_molecular_eff -> Star formation efficiency per free-fall time for the molecular hydrogen ($\mathrm{H_2}$) gas, computed using the pressure relation in Blitz et al. (2006).

:atomic_eff -> Star formation efficiency per free-fall time for the atomic hydrogen ($\mathrm{HI}$) gas.

:ionized_eff -> Star formation efficiency per free-fall time for the ionized hydrogen ($\mathrm{HII}$) gas.

:neutral_eff -> Star formation efficiency per free-fall time for the neutral hydrogen ($\mathrm{HI + H_2}$) gas.

:temperature -> Gas temperature, as $\log_{10}(T \, / \, \mathrm{K})$.

:pressure -> Gas pressure.

:scale_factor -> Scale factor.

:redshift -> Redshift.

:physical_time -> Physical time since the Big Bang.

:lookback_time -> Physical time left to reach the last snapshot.

:ode_gas_it -> Integration time.

:ode_gas_tau_s -> Star formation time scale, $\tau_\mathrm{S}$.

:ode_gas_eta_d -> Photodissociation efficiency, $\eta_\mathrm{diss}$.

:ode_gas_eta_i -> Photoionization efficiency, $\eta_\mathrm{ion}$.

:ode_gas_r -> Mass recycling parameter, $R$.

:ode_gas_cold_mf -> Cold gas mass fraction.

:ode_stellar_it -> Integration time, for the gas that form the stars.

:ode_stellar_tau_s -> Star formation time scale, $\tau_\mathrm{S}$, for the gas that form the stars.

:ode_stellar_eta_d -> Photodissociation efficiency, $\eta_\mathrm{diss}$, for the gas that form the stars.

:ode_stellar_eta_i -> Photoionization efficiency, $\eta_\mathrm{ion}$, for the gas that form the stars.

:ode_stellar_r -> Mass recycling parameter, $R$, for the gas that form the stars.

:ode_stellar_cold_mf -> Cold gas mass fraction, for the gas that form the stars.

:ode_stellar_gas_rho -> Gas mass density, for the gas that form the stars.

:ode_stellar_gas_Z -> Gas metallicity, for the gas that form the stars (solar units).

:ode_stellar_gas_mass -> Cell mass, for the gas that form the stars.

:ode_stellar_gas_sfr -> SFR associated to the gas particles/cells within the code, for the gas that form the stars.

:ode_stellar_gas_P -> Gas pressure, for the gas that form the stars.

Returns

A PlotParams object, with entries:

request::Dict{Symbol,Vector{String}} -> Data request for readSnapshot.

var_name::AbstractString -> Name of the quantity for the plot axis.

exp_factor::Int -> Numerical exponent to scale down the axis.

unit::Unitful.Units -> Target unit for the axis.

axis_label::AbstractString -> Label for the axis.

References

L. Blitz et al. (2006). The Role of Pressure in GMC Formation II: The H2-Pressure Relation. The Astrophysical Journal, 650(2), 933. doi:10.1086/505417

GalaxyInspector.pointData — Method

Extract all the data points in a Makie plot, axis, or figure. In the case of a figure, it will only take the data from the current axis object. It only works for scatter, line, and scatterline plots.

GalaxyInspector.reduceTicks — Method

reduceTicks(hr_ticks::Vector{<:Number}, factor::Int)::Vector{<:Number}

Reduce the length of a given list of axis ticks, while keeping the total length of the axis the same.

Arguments

hr_ticks::Vector{<:Number}: Original "high resolution" list of ticks. It has to be regularly spaced

factor::Int: Factor by which the number of values will be reduced. It has to divide the size of hr_ticks exactly.

Returns

The new shorter list of ticks.

GalaxyInspector.xlimits! — Method

Extract the limits of the x axis, from a Makie plot, axis, or figure. In the case of a figure, it will take the limits from the current axis object.

GalaxyInspector.xscale — Method

Extract the scale function of the x axis, from a Makie plot, axis, or figure. In the case of a figure, it will take the scale from the current axis object.

GalaxyInspector.ylimits! — Method

Extract the limits of the y axis, from a Makie plot, axis, or figure. In the case of a figure, it will take the limits from the current axis object.

GalaxyInspector.yscale — Method

Extract the scale function of the y axis, from a Makie plot, axis, or figure. In the case of a figure, it will take the scale from the current axis object.