Arepo constants

This is a list of constant and types used in GalaxyInspector, specific for the Arepo codebase.

None of these are exported.

GalaxyInspector.COSMO_THRESHOLD_DENSITY — Constant

Cosmological threshold density above which the gas cells/particles can turn into stars.

This value corresponds to CritOverDensity $= 57.7 \, [\mathrm{cm^{-3}}]$ in the param.txt file (used only in cosmological simulations). Which is converted to internal units within the code using OverDensThresh = CritOverDensity * OmegaBaryon * 3 * Hubble * Hubble / (8 * M_PI * G). Then, to go to physical units again one has to do:OverDensThresh*UnitDensityincgs*cf_a3inv*HubbleParam*HubbleParam`.

Using the unit factors,

UnitLength_in_cm = $3.085678 \times 10^{24}$

UnitMass_in_g = $1.989 \times 10^{43}$

UnitVelocity_in_cm_per_s = $100000$

The derived units,

UnitTime_in_s = UnitLength_in_cm * UnitVelocity_in_cm_per_s^-1 = $3.08568 \times 10^{19}$

UnitDensity_in_cgs = UnitMass_in_g * UnitLength_in_cm^-3 = $6.76991 \times 10^{-31}$

The parameters,

OmegaBaryon = $0.048$

HubbleParam = $0.6777$

PROTONMASS = $1.67262178 \times 10^{-24}$

HYDROGEN_MASSFRAC = $0.76$

GRAVITY = $6.6738 \times 10^{-8}$

HUBBLE = $3.2407789 \times 10^{-18}$

And the derived parameters,

Hubble = HUBBLE * UnitTime_in_s = $100$

G = GRAVITY * UnitLength_in_cm^-3 * UnitMass_in_g * UnitTime_in_s^2 = $43.0187$

One gets,

OverDensThresh = 76.8495 [internal units of density]

And, for a cosmological simulation at redshift 0 (cf_a3inv = 1), this result in a physical density threshold of $1.42857 \times 10^{-5} \, [\mathrm{cm^{-3}}]$, or, adding the proton mass, a value of

$\log_{10} \rho \ [\mathrm{M_\odot \, kpc^{-3}}] = 2.548$

GalaxyInspector.CPU_REL_PATH — Constant

Relative path, within the simulation directory, of cpu.txt.

GalaxyInspector.Cxd — Constant

Star formation model parameters

GalaxyInspector.DEFAULT_L_UNIT — Constant

Default internal unit of length.

GalaxyInspector.DEFAULT_M_UNIT — Constant

Default internal unit of mass.

GalaxyInspector.DEFAULT_V_UNIT — Constant

Default internal unit of velocity.

GalaxyInspector.ELEMENT_INDEX — Constant

Code index for each tracked element.

GalaxyInspector.GC_BASENAME — Constant

Base name of the group catalog files.

GalaxyInspector.HUBBLE_CONSTANT — Constant

Hubble constant in $\mathrm{Gyr^{-1}}$.

This value corresponds to $H_0 = 0.102201 \, \mathrm{Gyr}^{-1} = 100 \, \mathrm{km} \, \mathrm{s}^{-1} \, \mathrm{Mpc}^{-1}$.

GalaxyInspector.HYDROGEN_MASSFRAC — Constant

Mass fraction of hydrogen.

GalaxyInspector.ISOLATED_PARTICLE_NAMES — Constant

Human readable name for each type of cell/particle.

GalaxyInspector.LONG_PARTICLE_INDEX — Constant

Code index for each type of cell/particle.

References

See for example Gadget2 User's Guide, or Gadget4 documentation.

GalaxyInspector.LONG_PARTICLE_NAMES — Constant

Human readable name for each type of cell/particle.

GalaxyInspector.METAL_LIST — Constant

List of element indices above helium.

GalaxyInspector.MORPHOLOGICAL_COMPONENTS — Constant

Human readable name for each morphological component.

GalaxyInspector.PARTICLE_INDEX — Constant

Current cell/particle index in use.

GalaxyInspector.PARTICLE_NAMES — Constant

Current human readable name of each cell/particle type in use.

GalaxyInspector.SFM_IDX — Constant

ODE index corresponding to each gas phase in our star formation model.

Ionized gas fraction: fi(t) = Mi(t) / MC –> data_dict[:gas]["FRAC"][1, :]

Atomic gas fraction: fa(t) = Ma(t) / MC –> data_dict[:gas]["FRAC"][2, :]

Molecular gas fraction: fm(t) = Mm(t) / MC –> data_dict[:gas]["FRAC"][3, :]

Stellar fraction: fs(t) = Ms(t) / MC –> data_dict[:gas]["FRAC"][4, :]

Metal fraction: fZ(t) = MZ(t) / MC –> data_dict[:gas]["FRAC"][5, :]

Dust fraction: fd(t) = Md(t) / MC –> data_dict[:gas]["FRAC"][6, :]

GalaxyInspector.SFR_REL_PATH — Constant

Relative path, within the simulation directory, of sfr.txt.

GalaxyInspector.SNAP_BASENAME — Constant

Base name of the snapshot files, set in the code variable SnapshotFileBase.

GalaxyInspector.SOLAR_METALLICITY — Constant

Solar metallicity, as used in Arepo.

References

M. Asplund et al. (2006). The new solar abundances - Part I: the observations. Communications in Asteroseismology, 147. doi:10.1553/cia147s76

GalaxyInspector.THRESHOLD_DENSITY — Constant

Threshold density above which the gas cells/particles can turn into stars.

This value corresponds to CritPhysDensity $= 0.318 \, [\mathrm{cm^{-3}}]$ in the param.txt file (used in cosmological and non-cosmological simulations). Which is converted to internal units within the code using PhysDensThresh = CritPhysDensity * PROTONMASS / HYDROGEN_MASSFRAC / UnitDensity_in_cgs. Then, to go to physical units again one has to do: PhysDensThresh * UnitDensity_in_cgs * cf_a3inv * HubbleParam * HubbleParam.

PhysDensThresh = $1.03378 \times 10^{6}$ [internal units of density]

For a cosmological simulation at redshift 0 (cf_a3inv = 1), this result in a physical density threshold of $0.192 \, [\mathrm{cm^{-3}}]$, or, adding the proton mass, a value of

$\log_{10} \rho \, [\mathrm{M_\odot \, kpc^{-3}}] = 6.677$

GalaxyInspector.TRACER_MASS — Constant

Mass of the tracers in internal code units. Its value comes from All.TargetGasMass = All.TargetGasMassFactor * All.ReferenceGasPartMass in the code.

It is only printed in the output files (stdout_n), as All.TargetGasMass=3.65456e-06