Arepo constants

This is a list of constant and types used in GalaxyInspector, specific for the Arepo codebase.

None of these constants and types are exported.

GalaxyInspector.ELEMENT_INDEX — Constant

Code index for each tracked element.

GalaxyInspector.GC_BASENAME — Constant

Base name of the group catalog files.

GalaxyInspector.HYDROGEN_MASSFRAC — Constant

Mass fraction of hydrogen.

GalaxyInspector.METAL_LIST — Constant

List of element indices above helium.

GalaxyInspector.PARTICLE_INDEX — Constant

Current cell/particle index in use.

GalaxyInspector.PARTICLE_NAMES — Constant

Current human readable name of each cell/particle type in use.

GalaxyInspector.QUANTITIES — Constant

Dictionary of dimensional properties for the quantities in the code.

GalaxyInspector.SNAP_BASENAME — Constant

Base name of the snapshot files, set in the code variable SnapshotFileBase.

GalaxyInspector.SOLAR_METALLICITY — Constant

Solar metallicity, as used in Arepo.

References

M. Asplund et al. (2006). The new solar abundances - Part I: the observations. Communications in Asteroseismology, 147. doi:10.1553/cia147s76

GalaxyInspector.TRACER_MASS — Constant

Mass of the tracers in internal code units. Its value comes from All.TargetGasMass = All.TargetGasMassFactor * All.ReferenceGasPartMass in the code.