Quantities

This is a list of constant related to code and derived quantities used in GalaxyInspector.

None of these are exported.

GalaxyInspector.ABUNDANCE_QTY — Constant

List of symbols for all the abundance quantities.

GalaxyInspector.COMPONENTS — Constant

List of symbols for the physical components of a simulation.

GalaxyInspector.DERIVED_QTY — Constant

List of symbols for the derived quantities (component + magnitude).

GalaxyInspector.GAS_ABUNDANCE — Constant

List of symbols for the gas abundance quantities.

GalaxyInspector.GAS_ABUNDANCE_SPLITS — Constant

Dictionary mapping each gas abundance quantity to its element.

GalaxyInspector.HALO_KEYS — Constant

Dictionary mapping each halo quantity with its QUANTITIES key.

GalaxyInspector.MAGNITUDES — Constant

List of symbols for physical magnitudes.

GalaxyInspector.QTY_GLOBAL_LIST — Constant

List of symbols for all the quantities in GalaxyInspector.

GalaxyInspector.QTY_SINGLE_LIST — Constant

List of symbols for all the quantities in GalaxyInspector, except ratios.

GalaxyInspector.QUANTITIES — Constant

Dimensional properties for the quantities in the code.

GalaxyInspector.QUANTITY_SPLITS — Constant

Dictionary mapping each derived quantity to its component and magnitude.

GalaxyInspector.RATIO_QTY — Constant

List of symbols for the ratios between derived quantities.

GalaxyInspector.RATIO_SPLITS — Constant

Dictionary mapping each ratio to its two derived quantities.

GalaxyInspector.SFM_DERIVED_MAGNITUDES — Constant

List of symbols for every derived code magnitude in our SF model.

GalaxyInspector.SFM_DERIVED_QTY — Constant

List of symbols for every derived code quantity in our SF model.

GalaxyInspector.SFM_DERIVED_QTY_SPLITS — Constant

Dictionary mapping each derived code quantity to its component and magnitude.

GalaxyInspector.SFM_GAS_DERIVED_QTY — Constant

List of symbols for every derived code quantity associated with a gas cell in our SF model.

GalaxyInspector.SFM_GAS_QTY — Constant

List of symbols for every code quantity associated with a gas cell in our SF model.

GalaxyInspector.SFM_KEYS — Constant

Dictionary mapping each code quantity in our SF model with its QUANTITIES key.

GalaxyInspector.SFM_QTY — Constant

List of symbols for every code quantity in our SF model.

GalaxyInspector.SFM_QTY_SPLITS — Constant

Dictionary mapping each code quantity in our SF model to its component and magnitude.

GalaxyInspector.SFM_STELLAR_DERIVED_QTY — Constant

List of symbols for every derived code quantity associated with a stellar particle in our SF model.

This values come from the state of the parent gas cell (at the stellar birth time).

GalaxyInspector.SFM_STELLAR_QTY — Constant

List of symbols for every code quantity associated with a stellar particle in our SF model.

This values come from the state of the parent gas cell (at the stellar birth time).

GalaxyInspector.STELLAR_ABUNDANCE — Constant

List of symbols for the stellar abundance quantities.

GalaxyInspector.STELLAR_ABUNDANCE_SPLITS — Constant

Dictionary mapping each stellar abundance quantity to its element.

GalaxyInspector.parseHaloQuantity — Method

parseHaloQuantity(s::Symbol)::Tuple{Symbol,Int}

Takes a symbol for an halo (FoF group) quantity in the format :quantity_haloidx and splits it into the text part as a Symbol and the number part as an Int.

Arguments

s::Symbol: Target halo and halo quantity, e.g., :halo_mass_12. The index of the target halo starts at 1 and the halo quantity has to be one of the keys of HALO_KEYS.

Returns

A Tuple with two elements:

A Symbol for the target quantity.

The index of the target halo.

Examples

julia> parseHaloQuantity(:halo_mass_12)
(:halo_mass, 12)

julia> parseHaloQuantity(:halo_M200_1)
(:halo_M200, 1)