Data analysis functions

These functions read the data generated by the data acquisition functions, and produce the values that will be plotted.

None of these functions are exported.

Signature for the snapshotPlot function.

A data analysis functions for snapshotPlot must take a dictionary with the following shape:

:sim_data -> ::Simulation (see Simulation).

:snap_data -> ::Snapshot (see Snapshot).

:gc_data -> ::GroupCatalog (see GroupCatalog).

cell/particle type -> (block -> data of block, block -> data of block, ...).

cell/particle type -> (block -> data of block, block -> data of block, ...).

cell/particle type -> (block -> data of block, block -> data of block, ...).

...

groupcat type -> (block -> data of block, block -> data of block, ...).

groupcat type -> (block -> data of block, block -> data of block, ...).

groupcat type -> (block -> data of block, block -> data of block, ...).

...

and return one or more vectors or matrices with the processed data. It should return nothing if the input data has some problem that prevents computation (e.g. is empty).

Expected signature:

  da_function(data_dict, args...; kwargs...) -> (processed_data, ...)

where:

data_dict::Dict

processed_data::Union{VecOrMat{<:Number},Nothing}

Signature for the timeSeriesPlot function.

A data analysis functions for timeSeriesPlot must take a Simulation struct, and return two vectors. It should return nothing if the input data has some problem that prevents computation (e.g. is empty).

Expected signature:

  da_function(sim_data, args...; kw_args...) -> (processed_data_x, processed_data_y)

where:

sim_data::Simulation, see Simulation

processed_data_x::Vector{<:Number}

processed_data_y::Vector{<:Number}

GalaxyInspector.daCPUtxt — Method

daCPUtxt(
    sim_data::Simulation,
    target::String,
    x_quantity::Symbol,
    y_quantity::Symbol;
    <keyword arguments>
)::NTuple{2,Vector{<:Number}}

Compute the evolution of a measured quantity in the cpu.txt file, for a given target process.

Arguments

sim_data::Simulation: Information about the simulation in a Simulation object.

target::String: Target process.

x_quantity::Symbol: Quantity for the x axis. The possibilities are:

:time_step -> Time step.

:physical_time -> Physical time since the Big Bang.

:clock_time_s -> Clock time duration of the time step in seconds.

:clock_time_percent -> Clock time duration of the time step as a percentage.

:tot_clock_time_s -> Total clock time in seconds.

:tot_clock_time_percent -> Total clock time as a percentage.

y_quantity::Symbol: Quantity for the y axis. The possibilities are:

:time_step -> Time step.

:physical_time -> Physical time since the Big Bang.

:clock_time_s -> Clock time duration of the time step in seconds.

:clock_time_percent -> Clock time duration of the time step as a percentage.

:tot_clock_time_s -> Total clock time in seconds.

:tot_clock_time_percent -> Total clock time as a percentage.

smooth::Int=0: The result will be smooth out using smooth bins. Set it to 0 if you want no smoothing.

warnings::Bool=true: If a warning will be given when the target process is missing.

Returns

A Tuple with two elements:

A Vector with the time series of x_quantity.

A Vector with the time series of y_quantity.

GalaxyInspector.daCircularityHistogram — Method

daCircularityHistogram(
    data_dict::Dict,
    grid::LinearGrid,
)::Union{NTuple{2,Vector{Float64}},Nothing}

Compute a histogram of the stellar circularity, normalized to the maximum number of counts.

Arguments

data_dict::Dict: A dictionary with the following shape:

:sim_data -> ::Simulation (see Simulation).

:snap_data -> ::Snapshot (see Snapshot).

:gc_data -> ::GroupCatalog (see GroupCatalog).

cell/particle type -> (block -> data of block, block -> data of block, ...).

cell/particle type -> (block -> data of block, block -> data of block, ...).

cell/particle type -> (block -> data of block, block -> data of block, ...).

...

groupcat type -> (block -> data of block, block -> data of block, ...).

groupcat type -> (block -> data of block, block -> data of block, ...).

groupcat type -> (block -> data of block, block -> data of block, ...).

...

grid::LinearGrid: Linear grid.

Returns

A tuple with two elements:

A vector with the circularity of each bin.

A vector with the counts, normalized to the maximum value.

GalaxyInspector.daDensity2DHistogram — Method

daDensity2DHistogram(
    data_dict::Dict,
    grid::SquareGrid,
    quantity::Symbol;
    <keyword arguments>
)::Tuple{Vector{<:Unitful.Length},Vector{<:Unitful.Length},Matrix{Float64}}

Compute a 2D density histogram.

Note

By default, $\mathrm{M_\odot \, kpc^{-2}}$ is used as unit of density, so the output will be $\log_{10}(\rho \, / \, \mathrm{M_\odot \, kpc^{-2}})$.

Arguments

data_dict::Dict: A dictionary with the following shape:

:sim_data -> ::Simulation (see Simulation).

:snap_data -> ::Snapshot (see Snapshot).

:gc_data -> ::GroupCatalog (see GroupCatalog).

cell/particle type -> (block -> data of block, block -> data of block, ...).

cell/particle type -> (block -> data of block, block -> data of block, ...).

cell/particle type -> (block -> data of block, block -> data of block, ...).

...

groupcat type -> (block -> data of block, block -> data of block, ...).

groupcat type -> (block -> data of block, block -> data of block, ...).

groupcat type -> (block -> data of block, block -> data of block, ...).

...

grid::SquareGrid: Square grid.

quantity::Symbol: For which quantity the density will be calculated. The possibilities are:

:stellar_mass -> Stellar mass.

:gas_mass -> Gas mass.

:dm_mass -> Dark matter mass.

:bh_mass -> Black hole mass.

:molecular_mass -> Molecular hydrogen ($\mathrm{H_2}$) mass.

:atomic_mass -> Atomic hydrogen ($\mathrm{HI}$) mass.

:ionized_mass -> Ionized hydrogen ($\mathrm{HII}$) mass.

:neutral_mass -> Neutral hydrogen ($\mathrm{HI + H_2}$) mass.

projection_plane::Symbol=:xy: To which plane the cells/particles will be projected. The options are :xy, :xz, and :yz.

smooth::Bool=false: If the results will be smooth out using the cubicSplineKernel kernel.

neighbors::Int=18: Number of neighbors for the 2D smoothing (only relevant if smooth = true). The default value comes form Price2010: $N_{2D} = \pi \, (\zeta \, \eta)^2$, where we use $\zeta = 2$ and $\eta = 1.2$.

smoothing_length::Union{Unitful.Length,Nothing}=nothing: Smoothing length. If set to nothing, the mean value of the "SOFT" block will be used. If the "SOFT" block is no available, the mean of the cell characteristic size will be used.

print_range::Bool=false: Print an info block detailing the logarithmic density range.

Returns

A tuple with three elements:

A vector with the x coordinates of the grid.

A vector with the y coordinates of the grid.

A matrix with the values of density in each bin.

GalaxyInspector.daEvolution — Method

daEvolution(
    sim_data::Simulation,
    x_quantity::Symbol,
    y_quantity::Symbol;
    <keyword arguments>
)::NTuple{2,Vector{<:Number}}

Compute the time series of two quantities.

Arguments

sim_data::Simulation: Information about the simulation in a Simulation object.

x_quantity::Symbol: Quantity for the x axis. The possibilities are:

:stellar_mass -> Stellar mass.

:gas_mass -> Gas mass.

:dm_mass -> Dark matter mass.

:bh_mass -> Black hole mass.

:stellar_number -> Number of stellar particles.

:gas_number -> Number of gas cells.

:dm_number -> Number of dark matter particles.

:bh_number -> Number of black hole particles.

:molecular_mass -> Molecular hydrogen ($\mathrm{H_2}$) mass.

:atomic_mass -> Atomic hydrogen ($\mathrm{HI}$) mass.

:ionized_mass -> Ionized hydrogen ($\mathrm{HII}$) mass.

:neutral_mass -> Neutral hydrogen ($\mathrm{HI + H_2}$) mass.

:molecular_fraction -> Gas mass fraction of molecular hydrogen.

:atomic_fraction -> Gas mass fraction of atomic hydrogen.

:ionized_fraction -> Gas mass fraction of ionized hydrogen.

:neutral_fraction -> Gas mass fraction of neutral hydrogen.

:stellar_area_density -> Stellar area mass density, for a radius of FILTER_R.

:gas_area_density -> Gas area mass density, for a radius of FILTER_R.

:molecular_area_density -> Molecular hydrogen area mass density, for a radius of FILTER_R.

:atomic_area_density -> Atomic hydrogen area mass density, for a radius of FILTER_R.

:ionized_area_density -> Ionized hydrogen area mass density, for a radius of FILTER_R.

:neutral_area_density -> Neutral hydrogen area mass density, for a radius of FILTER_R.

:sfr_area_density -> Star formation rate area density, for the last AGE_RESOLUTION_ρ and a radius of FILTER_R.

:gas_metallicity -> Mass fraction of all elements above He in the gas (solar units).

:stellar_metallicity -> Mass fraction of all elements above He in the stars (solar units).

:X_gas_abundance -> Gas abundance of element $\mathrm{X}$, as $12 + \log_{10}(\mathrm{X \, / \, H})$. The possibilities are the keys of ELEMENT_INDEX.

:X_stellar_abundance -> Stellar abundance of element $\mathrm{X}$, as $12 + \log_{10}(\mathrm{X \, / \, H})$. The possibilities are the keys of ELEMENT_INDEX.

:stellar_specific_am -> Norm of the stellar specific angular momentum.

:gas_specific_am -> Norm of the gas specific angular momentum.

:dm_specific_am -> Norm of the dark matter specific angular momentum.

:sfr -> The star formation rate of the last AGE_RESOLUTION.

:ssfr -> The specific star formation rate of the last AGE_RESOLUTION.

:scale_factor -> Scale factor.

:redshift -> Redshift.

:physical_time -> Physical time since the Big Bang.

:lookback_time -> Physical time left to reach the last snapshot.

y_quantity::Symbol: Quantity for the y axis. The possibilities are:

:stellar_mass -> Stellar mass.

:gas_mass -> Gas mass.

:dm_mass -> Dark matter mass.

:bh_mass -> Black hole mass.

:stellar_number -> Number of stellar particles.

:gas_number -> Number of gas cells.

:dm_number -> Number of dark matter particles.

:bh_number -> Number of black hole particles.

:molecular_mass -> Molecular hydrogen ($\mathrm{H_2}$) mass.

:atomic_mass -> Atomic hydrogen ($\mathrm{HI}$) mass.

:ionized_mass -> Ionized hydrogen ($\mathrm{HII}$) mass.

:neutral_mass -> Neutral hydrogen ($\mathrm{HI + H_2}$) mass.

:molecular_fraction -> Gas mass fraction of molecular hydrogen.

:atomic_fraction -> Gas mass fraction of atomic hydrogen.

:ionized_fraction -> Gas mass fraction of ionized hydrogen.

:neutral_fraction -> Gas mass fraction of neutral hydrogen.

:stellar_area_density -> Stellar area mass density, for a radius of FILTER_R.

:gas_area_density -> Gas area mass density, for a radius of FILTER_R.

:molecular_area_density -> Molecular hydrogen area mass density, for a radius of FILTER_R.

:atomic_area_density -> Atomic hydrogen area mass density, for a radius of FILTER_R.

:ionized_area_density -> Ionized hydrogen area mass density, for a radius of FILTER_R.

:neutral_area_density -> Neutral hydrogen area mass density, for a radius of FILTER_R.

:sfr_area_density -> Star formation rate area density, for the last AGE_RESOLUTION_ρ and a radius of FILTER_R.

:gas_metallicity -> Mass fraction of all elements above He in the gas (solar units).

:stellar_metallicity -> Mass fraction of all elements above He in the stars (solar units).

:X_gas_abundance -> Gas abundance of element $\mathrm{X}$, as $12 + \log_{10}(\mathrm{X \, / \, H})$. The possibilities are the keys of ELEMENT_INDEX.

:X_stellar_abundance -> Stellar abundance of element $\mathrm{X}$, as $12 + \log_{10}(\mathrm{X \, / \, H})$. The possibilities are the keys of ELEMENT_INDEX.

:stellar_specific_am -> Norm of the stellar specific angular momentum.

:gas_specific_am -> Norm of the gas specific angular momentum.

:dm_specific_am -> Norm of the dark matter specific angular momentum.

:sfr -> The star formation rate of the last AGE_RESOLUTION.

:ssfr -> The specific star formation rate of the last AGE_RESOLUTION.

:scale_factor -> Scale factor.

:redshift -> Redshift.

:physical_time -> Physical time since the Big Bang.

:lookback_time -> Physical time left to reach the last snapshot.

filter_mode::Symbol=:all: Which cells/particles will be plotted, the options are:

:all -> Plot every cell/particle within the simulation box.

:halo -> Plot only the cells/particles that belong to the main halo.

:subhalo -> Plot only the cells/particles that belong to the main subhalo.

:sphere -> Plot only the cell/particle inside a sphere with radius FILTER_R (see ./src/constants.jl).

:stellar_subhalo -> Plot only the cells/particles that belong to the main subhalo.

:all_subhalo -> Plot every cell/particle centered around the main subhalo.

smooth::Int=0: The result of integrateQty will be smooth out using smooth bins. Set it to 0 if you want no smoothing.

scaling::Function=identity: Function to scale the x-axis (only relevant if smooth != 0). The bins will be computed accordingly. The options are the scaling functions accepted by Makie.jl: log10, log2, log, sqrt, Makie.logit, Makie.Symlog10, Makie.pseudolog10, and identity.

warnings::Bool=true: If a warning will be given when there is missing data.

Returns

A Tuple with two elements:

A Vector with the time series of x_quantity.

A Vector with the time series of y_quantity.

GalaxyInspector.daIntegrateGalaxy — Method

daIntegrateGalaxy(
    data_dict::Dict,
    x_quantity::Symbol,
    y_quantity::Symbol,
)::NTuple{2,Vector{<:Number}}

Compute two global quantities of the simulation.

Arguments

data_dict::Dict: A dictionary with the following shape:

:sim_data -> ::Simulation (see Simulation).

:snap_data -> ::Snapshot (see Snapshot).

:gc_data -> ::GroupCatalog (see GroupCatalog).

cell/particle type -> (block -> data of block, block -> data of block, ...).

cell/particle type -> (block -> data of block, block -> data of block, ...).

cell/particle type -> (block -> data of block, block -> data of block, ...).

...

groupcat type -> (block -> data of block, block -> data of block, ...).

groupcat type -> (block -> data of block, block -> data of block, ...).

groupcat type -> (block -> data of block, block -> data of block, ...).

...

x_quantity::Symbol: Quantity for the x axis. The possibilities are:

:stellar_mass -> Stellar mass.

:gas_mass -> Gas mass.

:dm_mass -> Dark matter mass.

:bh_mass -> Black hole mass.

:stellar_number -> Number of stellar particles.

:gas_number -> Number of gas cells.

:dm_number -> Number of dark matter particles.

:bh_number -> Number of black hole particles.

:molecular_mass -> Molecular hydrogen ($\mathrm{H_2}$) mass.

:atomic_mass -> Atomic hydrogen ($\mathrm{HI}$) mass.

:ionized_mass -> Ionized hydrogen ($\mathrm{HII}$) mass.

:neutral_mass -> Neutral hydrogen ($\mathrm{HI + H_2}$) mass.

:molecular_fraction -> Gas mass fraction of molecular hydrogen.

:atomic_fraction -> Gas mass fraction of atomic hydrogen.

:ionized_fraction -> Gas mass fraction of ionized hydrogen.

:neutral_fraction -> Gas mass fraction of neutral hydrogen.

:stellar_area_density -> Stellar area mass density, for a radius of FILTER_R.

:gas_area_density -> Gas area mass density, for a radius of FILTER_R.

:molecular_area_density -> Molecular hydrogen area mass density, for a radius of FILTER_R.

:atomic_area_density -> Atomic hydrogen area mass density, for a radius of FILTER_R.

:ionized_area_density -> Ionized hydrogen area mass density, for a radius of FILTER_R.

:neutral_area_density -> Neutral hydrogen area mass density, for a radius of FILTER_R.

:sfr_area_density -> Star formation rate area density, for the last AGE_RESOLUTION_ρ and a radius of FILTER_R.

:gas_metallicity -> Mass fraction of all elements above He in the gas (solar units).

:stellar_metallicity -> Mass fraction of all elements above He in the stars (solar units).

:X_gas_abundance -> Gas abundance of element $\mathrm{X}$, as $12 + \log_{10}(\mathrm{X \, / \, H})$. The possibilities are the keys of ELEMENT_INDEX.

:X_stellar_abundance -> Stellar abundance of element $\mathrm{X}$, as $12 + \log_{10}(\mathrm{X \, / \, H})$. The possibilities are the keys of ELEMENT_INDEX.

:stellar_specific_am -> Norm of the stellar specific angular momentum.

:gas_specific_am -> Norm of the gas specific angular momentum.

:dm_specific_am -> Norm of the dark matter specific angular momentum.

:sfr -> The star formation rate of the last AGE_RESOLUTION.

:ssfr -> The specific star formation rate of the last AGE_RESOLUTION.

:scale_factor -> Scale factor.

:redshift -> Redshift.

:physical_time -> Physical time since the Big Bang.

:lookback_time -> Physical time left to reach the last snapshot.

y_quantity::Symbol: Quantity for the y axis. The possibilities are:

:stellar_mass -> Stellar mass.

:gas_mass -> Gas mass.

:dm_mass -> Dark matter mass.

:bh_mass -> Black hole mass.

:stellar_number -> Number of stellar particles.

:gas_number -> Number of gas cells.

:dm_number -> Number of dark matter particles.

:bh_number -> Number of black hole particles.

:molecular_mass -> Molecular hydrogen ($\mathrm{H_2}$) mass.

:atomic_mass -> Atomic hydrogen ($\mathrm{HI}$) mass.

:ionized_mass -> Ionized hydrogen ($\mathrm{HII}$) mass.

:neutral_mass -> Neutral hydrogen ($\mathrm{HI + H_2}$) mass.

:molecular_fraction -> Gas mass fraction of molecular hydrogen.

:atomic_fraction -> Gas mass fraction of atomic hydrogen.

:ionized_fraction -> Gas mass fraction of ionized hydrogen.

:neutral_fraction -> Gas mass fraction of neutral hydrogen.

:stellar_area_density -> Stellar area mass density, for a radius of FILTER_R.

:gas_area_density -> Gas area mass density, for a radius of FILTER_R.

:molecular_area_density -> Molecular hydrogen area mass density, for a radius of FILTER_R.

:atomic_area_density -> Atomic hydrogen area mass density, for a radius of FILTER_R.

:ionized_area_density -> Ionized hydrogen area mass density, for a radius of FILTER_R.

:neutral_area_density -> Neutral hydrogen area mass density, for a radius of FILTER_R.

:sfr_area_density -> Star formation rate area density, for the last AGE_RESOLUTION_ρ and a radius of FILTER_R.

:gas_metallicity -> Mass fraction of all elements above He in the gas (solar units).

:stellar_metallicity -> Mass fraction of all elements above He in the stars (solar units).

:X_gas_abundance -> Gas abundance of element $\mathrm{X}$, as $12 + \log_{10}(\mathrm{X \, / \, H})$. The possibilities are the keys of ELEMENT_INDEX.

:X_stellar_abundance -> Stellar abundance of element $\mathrm{X}$, as $12 + \log_{10}(\mathrm{X \, / \, H})$. The possibilities are the keys of ELEMENT_INDEX.

:stellar_specific_am -> Norm of the stellar specific angular momentum.

:gas_specific_am -> Norm of the gas specific angular momentum.

:dm_specific_am -> Norm of the dark matter specific angular momentum.

:sfr -> The star formation rate of the last AGE_RESOLUTION.

:ssfr -> The specific star formation rate of the last AGE_RESOLUTION.

:scale_factor -> Scale factor.

:redshift -> Redshift.

:physical_time -> Physical time since the Big Bang.

:lookback_time -> Physical time left to reach the last snapshot.

Returns

A tuple with two elements:

A single element vector with the value of x_quantity.

A single element vector with the value of y_quantity.

GalaxyInspector.daKennicuttSchmidt — Method

daKennicuttSchmidt(
    data_dict::Dict,
    grid::CircularGrid,
)::Union{Tuple{Vector{<:SurfaceDensity},Vector{<:MassFlowDensity}},Nothing}

Compute the gas mass and SFR surface densities, used in the Kennicutt-Schmidt law.

Arguments

data_dict::Dict: A dictionary with the following shape:

:sim_data -> ::Simulation (see Simulation).

:snap_data -> ::Snapshot (see Snapshot).

:gc_data -> ::GroupCatalog (see GroupCatalog).

cell/particle type -> (block -> data of block, block -> data of block, ...).

cell/particle type -> (block -> data of block, block -> data of block, ...).

cell/particle type -> (block -> data of block, block -> data of block, ...).

...

groupcat type -> (block -> data of block, block -> data of block, ...).

groupcat type -> (block -> data of block, block -> data of block, ...).

groupcat type -> (block -> data of block, block -> data of block, ...).

...

grid::CircularGrid: Circular grid.

Returns

A tuple with two elements:

A vector with the gas mass surface density of each ring.

A vector with the SFR surface density of each ring.

It returns nothing if any of the necessary quantities are missing.

References

R. C. Kennicutt (1998). The Global Schmidt Law in Star-forming Galaxies. The Astrophysical Journal, 498(2), 541-552. doi:10.1086/305588

GalaxyInspector.daKennicuttSchmidtLaw — Method

function daKennicuttSchmidtLaw(
    data_dict::Dict,
    grid::CircularGrid,
    quantity::Symbol,
)::Union{Tuple{Vector{<:SurfaceDensity},Vector{<:MassFlowDensity}},Nothing}

Compute the (molecular or neutral) gas mass and SFR surface densities, used in the Kennicutt-Schmidt law.

Arguments

data_dict::Dict: A dictionary with the following shape:

:sim_data -> ::Simulation (see Simulation).

:snap_data -> ::Snapshot (see Snapshot).

:gc_data -> ::GroupCatalog (see GroupCatalog).

cell/particle type -> (block -> data of block, block -> data of block, ...).

cell/particle type -> (block -> data of block, block -> data of block, ...).

cell/particle type -> (block -> data of block, block -> data of block, ...).

...

groupcat type -> (block -> data of block, block -> data of block, ...).

groupcat type -> (block -> data of block, block -> data of block, ...).

groupcat type -> (block -> data of block, block -> data of block, ...).

...

grid::CircularGrid: Circular grid.

quantity::Symbol: Quantity for the x axis. The possibilities are:

:molecular_area_density -> Molecular hydrogen area mass density.

:neutral_area_density -> Neutral hydrogen area mass density.

Returns

A tuple with two elements:

A vector with the gas mass surface density of each ring.

A vector with the SFR surface density of each ring.

It returns nothing if any of the necessary quantities are missing.

GalaxyInspector.daMolla2015 — Method

daMolla2015(
    data_dict::Dict,
    grid::CircularGrid,
    quantity::Symbol,
)::Union{
    Tuple{
        Vector{<:Unitful.Length},
        <:Union{Vector{<:SurfaceDensity},Vector{<:MassFlowDensity},Vector{Float64}}
    },
    Nothing,
}

Compute a profile for the Milky Way, compatible with the experimental data in Mollá et al. (2015).

Arguments

data_dict::Dict: A dictionary with the following shape:

:sim_data -> ::Simulation (see Simulation).

:snap_data -> ::Snapshot (see Snapshot).

:gc_data -> ::GroupCatalog (see GroupCatalog).

cell/particle type -> (block -> data of block, block -> data of block, ...).

cell/particle type -> (block -> data of block, block -> data of block, ...).

cell/particle type -> (block -> data of block, block -> data of block, ...).

...

groupcat type -> (block -> data of block, block -> data of block, ...).

groupcat type -> (block -> data of block, block -> data of block, ...).

groupcat type -> (block -> data of block, block -> data of block, ...).

...

grid::CircularGrid: Circular grid.

quantity::Symbol: Quantity for the y axis. The options are:

:stellar_area_density -> Stellar area mass density.

:molecular_area_density -> Molecular hydrogen area mass density.

:atomic_area_density -> Atomic hydrogen area mass density.

:sfr_area_density -> Star formation rate area density, for the last AGE_RESOLUTION_ρ.

:O_stellar_abundance -> Stellar abundance of oxygen, as $12 + \log_{10}(\mathrm{O \, / \, H})$.

:N_stellar_abundance -> Stellar abundance of nitrogen, as $12 + \log_{10}(\mathrm{N \, / \, H})$.

:C_stellar_abundance -> Stellar abundance of carbon, as $12 + \log_{10}(\mathrm{C \, / \, H})$.

Returns

A tuple with two elements:

A vector with the position of each ring.

A vector with the quantity area density of each ring.

It returns nothing if any of the necessary quantities are missing.

References

M. Mollá et al. (2015). Galactic chemical evolution: stellar yields and the initial mass function. Monthly Notices of the Royal Astronomical Society 451(4), 3693–3708. doi:10.1093/mnras/stv1102

GalaxyInspector.daProfile — Method

daProfile(
    data_dict::Dict,
    grid::CircularGrid,
    quantity::Symbol;
    <keyword arguments>
)::Union{Tuple{Vector{<:Unitful.Length},Vector{<:Number}},Nothing}

Compute a profile.

Arguments

data_dict::Dict: A dictionary with the following shape:

:sim_data -> ::Simulation (see Simulation).

:snap_data -> ::Snapshot (see Snapshot).

:gc_data -> ::GroupCatalog (see GroupCatalog).

cell/particle type -> (block -> data of block, block -> data of block, ...).

cell/particle type -> (block -> data of block, block -> data of block, ...).

cell/particle type -> (block -> data of block, block -> data of block, ...).

...

groupcat type -> (block -> data of block, block -> data of block, ...).

groupcat type -> (block -> data of block, block -> data of block, ...).

groupcat type -> (block -> data of block, block -> data of block, ...).

...

grid::CircularGrid: Circular grid.

quantity::Symbol: Quantity for the y axis. The options are:

:stellar_area_density -> Stellar area mass density, for a radius of FILTER_R.

:gas_area_density -> Gas area mass density, for a radius of FILTER_R.

:molecular_area_density -> Molecular hydrogen area mass density, for a radius of FILTER_R.

:atomic_area_density -> Atomic hydrogen area mass density, for a radius of FILTER_R.

:ionized_area_density -> Ionized hydrogen area mass density, for a radius of FILTER_R.

:neutral_area_density -> Neutral hydrogen area mass density, for a radius of FILTER_R.

:sfr_area_density -> Star formation rate area density, for the last AGE_RESOLUTION_ρ and a radius of FILTER_R.

:stellar_vcirc -> Stellar circular velocity, upto a radius of FILTER_R.

:stellar_vradial -> Stellar radial speed, upto a radius of FILTER_R.

:stellar_vtangential -> Stellar tangential speed, upto a radius of FILTER_R.

flat::Bool=true: If the profile will be 2D, using rings, or 3D, using spherical shells.

total::Bool=true: If the sum (default) or the mean of quantity will be computed for each bin.

cumulative::Bool=false: If the profile will be accumulated or not.

density::Bool=true: If the profile will be of the density of quantity.

Returns

A tuple with two elements:

A vector with the position of each ring.

A vector with the quantity area density of each ring.

It returns nothing if any of the necessary quantities are missing.

GalaxyInspector.daRotationCurve — Method

daRotationCurve(
    data_dict::Dict,
    R::Unitful.Length,
)::Tuple{Vector{<:Unitful.Length},Vector{<:Unitful.Velocity}}

Compute a rotation curve.

Arguments

data_dict::Dict: A dictionary with the following shape:

:sim_data -> ::Simulation (see Simulation).

:snap_data -> ::Snapshot (see Snapshot).

:gc_data -> ::GroupCatalog (see GroupCatalog).

cell/particle type -> (block -> data of block, block -> data of block, ...).

cell/particle type -> (block -> data of block, block -> data of block, ...).

cell/particle type -> (block -> data of block, block -> data of block, ...).

...

groupcat type -> (block -> data of block, block -> data of block, ...).

groupcat type -> (block -> data of block, block -> data of block, ...).

groupcat type -> (block -> data of block, block -> data of block, ...).

...

R::Unitful.Length: Radius for the profile.

Returns

A tuple with two elements:

A vector with the distances to each star.

A vector with the circular velocity of each star.

GalaxyInspector.daSFRtxt — Method

daSFRtxt(
    sim_data::Simulation,
    x_quantity::Symbol,
    y_quantity::Symbol;
    <keyword arguments>
)::NTuple{2,Vector{<:Number}}

Compute the stellar mass or SFR evolution using the data in the sfr.txt file.

Arguments

sim_data::Simulation: Information about the simulation in a Simulation object.

x_quantity::Symbol: Quantity for the x axis. The possibilities are:

:scale_factor -> Scale factor.

:redshift -> Redshift.

:physical_time -> Physical time since the Big Bang.

:lookback_time -> Physical time left to reach the last snapshot.

y_quantity::Symbol: Quantity for the y axis. The possibilities are:

:stellar_mass -> Stellar mass.

:sfr -> The star formation rate.

smooth::Int=0: The result will be smooth out using smooth bins. Set it to 0 if you want no smoothing.

warnings::Bool=true: If a warning will be given when trying to use the scale factor or the redshift in the x axis for a non-cosmological simulation.

Returns

A Tuple with two elements:

A Vector with the time series of x_quantity.

A Vector with the time series of y_quantity.

GalaxyInspector.daScatterDensity — Method

daScatterDensity(
    data_dict::Dict,
    x_quantity::Symbol,
    y_quantity::Symbol;
    <keyword arguments>
)::Tuple{Vector{<:Number},Vector{<:Number},Matrix{Float64}}

Compute a 2D histogram.

Arguments

data_dict::Dict: A dictionary with the following shape:

:sim_data -> ::Simulation (see Simulation).

:snap_data -> ::Snapshot (see Snapshot).

:gc_data -> ::GroupCatalog (see GroupCatalog).

cell/particle type -> (block -> data of block, block -> data of block, ...).

cell/particle type -> (block -> data of block, block -> data of block, ...).

cell/particle type -> (block -> data of block, block -> data of block, ...).

...

groupcat type -> (block -> data of block, block -> data of block, ...).

groupcat type -> (block -> data of block, block -> data of block, ...).

groupcat type -> (block -> data of block, block -> data of block, ...).

...

x_quantity::Symbol: Quantity for the x axis. The possibilities are:

:stellar_mass -> Stellar mass.

:gas_mass -> Gas mass.

:dm_mass -> Dark matter mass.

:bh_mass -> Black hole mass.

:molecular_mass -> Molecular hydrogen ($\mathrm{H_2}$) mass.

:atomic_mass -> Atomic hydrogen ($\mathrm{HI}$) mass.

:ionized_mass -> Ionized hydrogen ($\mathrm{HII}$) mass.

:neutral_mass -> Neutral hydrogen ($\mathrm{HI + H_2}$) mass.

:molecular_fraction -> Gas mass fraction of molecular hydrogen.

:atomic_fraction -> Gas mass fraction of atomic hydrogen.

:ionized_fraction -> Gas mass fraction of ionized hydrogen.

:neutral_fraction -> Gas mass fraction of neutral hydrogen.

:molecular_neutral_fraction -> Fraction of molecular hydrogen in the neutral gas.

:gas_mass_density -> Gas mass density.

:gas_number_density -> Gas number density.

:molecular_number_density -> Molecular hydrogen number density.

:atomic_number_density -> Atomic hydrogen number density.

:ionized_number_density -> Ionized hydrogen number density.

:neutral_number_density -> Neutral hydrogen number density.

:gas_metallicity -> Mass fraction of all elements above He in the gas (solar units).

:stellar_metallicity -> Mass fraction of all elements above He in the stars (solar units).

:X_gas_abundance -> Gas abundance of element $\mathrm{X}$, as $12 + \log_{10}(\mathrm{X \, / \, H})$. The possibilities are the keys of ELEMENT_INDEX.

:X_stellar_abundance -> Stellar abundance of element $\mathrm{X}$, as $12 + \log_{10}(\mathrm{X \, / \, H})$. The possibilities are the keys of ELEMENT_INDEX.

:stellar_radial_distance -> Distance of every stellar particle to the origin.

:gas_radial_distance -> Distance of every gas cell to the origin.

:dm_radial_distance -> Distance of every dark matter particle to the origin.

:stellar_xy_distance -> Projected distance of every stellar particle to the origin.

:gas_xy_distance -> Projected distance of every gas cell to the origin.

:dm_xy_distance -> Projected distance of every dark matter particle to the origin.

:stellar_circularity -> Stellar circularity.

:stellar_vcirc -> Stellar circular velocity.

:stellar_vradial -> Stellar radial speed.

:stellar_vtangential -> Stellar tangential speed.

:stellar_vzstar -> Stellar speed in the z direction, computed as $v_z \, \sign(z)$.

:stellar_age -> Stellar age.

:sfr -> The star formation rate of the last AGE_RESOLUTION.

:ssfr -> The specific star formation rate of the last AGE_RESOLUTION.

:temperature -> Gas temperature, as $\log_{10}(T \, / \, \mathrm{K})$.

y_quantity::Symbol: Quantity for the y axis. The possibilities are:

:stellar_mass -> Stellar mass.

:gas_mass -> Gas mass.

:dm_mass -> Dark matter mass.

:bh_mass -> Black hole mass.

:molecular_mass -> Molecular hydrogen ($\mathrm{H_2}$) mass.

:atomic_mass -> Atomic hydrogen ($\mathrm{HI}$) mass.

:ionized_mass -> Ionized hydrogen ($\mathrm{HII}$) mass.

:neutral_mass -> Neutral hydrogen ($\mathrm{HI + H_2}$) mass.

:molecular_fraction -> Gas mass fraction of molecular hydrogen.

:atomic_fraction -> Gas mass fraction of atomic hydrogen.

:ionized_fraction -> Gas mass fraction of ionized hydrogen.

:neutral_fraction -> Gas mass fraction of neutral hydrogen.

:molecular_neutral_fraction -> Fraction of molecular hydrogen in the neutral gas.

:gas_mass_density -> Gas mass density.

:gas_number_density -> Gas number density.

:molecular_number_density -> Molecular hydrogen number density.

:atomic_number_density -> Atomic hydrogen number density.

:ionized_number_density -> Ionized hydrogen number density.

:neutral_number_density -> Neutral hydrogen number density.

:gas_metallicity -> Mass fraction of all elements above He in the gas (solar units).

:stellar_metallicity -> Mass fraction of all elements above He in the stars (solar units).

:X_gas_abundance -> Gas abundance of element $\mathrm{X}$, as $12 + \log_{10}(\mathrm{X \, / \, H})$. The possibilities are the keys of ELEMENT_INDEX.

:X_stellar_abundance -> Stellar abundance of element $\mathrm{X}$, as $12 + \log_{10}(\mathrm{X \, / \, H})$. The possibilities are the keys of ELEMENT_INDEX.

:stellar_radial_distance -> Distance of every stellar particle to the origin.

:gas_radial_distance -> Distance of every gas cell to the origin.

:dm_radial_distance -> Distance of every dark matter particle to the origin.

:stellar_xy_distance -> Projected distance of every stellar particle to the origin.

:gas_xy_distance -> Projected distance of every gas cell to the origin.

:dm_xy_distance -> Projected distance of every dark matter particle to the origin.

:stellar_circularity -> Stellar circularity.

:stellar_vcirc -> Stellar circular velocity.

:stellar_vradial -> Stellar radial speed.

:stellar_vtangential -> Stellar tangential speed.

:stellar_vzstar -> Stellar speed in the z direction, computed as $v_z \, \sign(z)$.

:stellar_age -> Stellar age.

:sfr -> The star formation rate of the last AGE_RESOLUTION.

:ssfr -> The specific star formation rate of the last AGE_RESOLUTION.

:temperature -> Gas temperature, as $\log_{10}(T \, / \, \mathrm{K})$.

x_range::Union{NTuple{2,<:Number},Nothing}=nothing: x axis range for the histogram grid. If set to nothing, the extrema of the values will be used.

y_range::Union{NTuple{2,<:Number},Nothing}=nothing: y axis range for the histogram grid. If set to nothing, the extrema of the values will be used.

x_log::Union{Unitful.Units,Nothing}=nothing: Set to the desired unit of x_quantity, if you want to use log10(x_quantity) for the x axis.

y_log::Union{Unitful.Units,Nothing}=nothing: Set to the desired unit of y_quantity, if you want to use log10(y_quantity) for the y axis.

n_bins::Int=100: Number of bins per side of the grid.

Returns

A tuple with three elements:

A vector with the x coordinates of the grid.

A vector with the y coordinates of the grid.

A matrix with the counts.

GalaxyInspector.daScatterGalaxy — Method

daScatterGalaxy(
    data_dict::Dict,
    x_quantity::Symbol,
    y_quantity::Symbol;
    <keyword arguments>
)::NTuple{2,Vector{<:Number}}

Compute two quantities for every cell/particle in the simulation.

Arguments

data_dict::Dict: A dictionary with the following shape:

:sim_data -> ::Simulation (see Simulation).

:snap_data -> ::Snapshot (see Snapshot).

:gc_data -> ::GroupCatalog (see GroupCatalog).

cell/particle type -> (block -> data of block, block -> data of block, ...).

cell/particle type -> (block -> data of block, block -> data of block, ...).

cell/particle type -> (block -> data of block, block -> data of block, ...).

...

groupcat type -> (block -> data of block, block -> data of block, ...).

groupcat type -> (block -> data of block, block -> data of block, ...).

groupcat type -> (block -> data of block, block -> data of block, ...).

...

x_quantity::Symbol: Quantity for the x axis. The possibilities are:

:stellar_mass -> Stellar mass.

:gas_mass -> Gas mass.

:dm_mass -> Dark matter mass.

:bh_mass -> Black hole mass.

:molecular_mass -> Molecular hydrogen ($\mathrm{H_2}$) mass.

:atomic_mass -> Atomic hydrogen ($\mathrm{HI}$) mass.

:ionized_mass -> Ionized hydrogen ($\mathrm{HII}$) mass.

:neutral_mass -> Neutral hydrogen ($\mathrm{HI + H_2}$) mass.

:molecular_fraction -> Gas mass fraction of molecular hydrogen.

:atomic_fraction -> Gas mass fraction of atomic hydrogen.

:ionized_fraction -> Gas mass fraction of ionized hydrogen.

:neutral_fraction -> Gas mass fraction of neutral hydrogen.

:molecular_neutral_fraction -> Fraction of molecular hydrogen in the neutral gas.

:gas_mass_density -> Gas mass density.

:gas_number_density -> Gas number density.

:molecular_number_density -> Molecular hydrogen number density.

:atomic_number_density -> Atomic hydrogen number density.

:ionized_number_density -> Ionized hydrogen number density.

:neutral_number_density -> Neutral hydrogen number density.

:gas_metallicity -> Mass fraction of all elements above He in the gas (solar units).

:stellar_metallicity -> Mass fraction of all elements above He in the stars (solar units).

:X_gas_abundance -> Gas abundance of element $\mathrm{X}$, as $12 + \log_{10}(\mathrm{X \, / \, H})$. The possibilities are the keys of ELEMENT_INDEX.

:X_stellar_abundance -> Stellar abundance of element $\mathrm{X}$, as $12 + \log_{10}(\mathrm{X \, / \, H})$. The possibilities are the keys of ELEMENT_INDEX.

:stellar_radial_distance -> Distance of every stellar particle to the origin.

:gas_radial_distance -> Distance of every gas cell to the origin.

:dm_radial_distance -> Distance of every dark matter particle to the origin.

:stellar_xy_distance -> Projected distance of every stellar particle to the origin.

:gas_xy_distance -> Projected distance of every gas cell to the origin.

:dm_xy_distance -> Projected distance of every dark matter particle to the origin.

:stellar_circularity -> Stellar circularity.

:stellar_vcirc -> Stellar circular velocity.

:stellar_vradial -> Stellar radial speed.

:stellar_vtangential -> Stellar tangential speed.

:stellar_vzstar -> Stellar speed in the z direction, computed as $v_z \, \sign(z)$.

:stellar_age -> Stellar age.

:sfr -> The star formation rate of the last AGE_RESOLUTION.

:ssfr -> The specific star formation rate of the last AGE_RESOLUTION.

:temperature -> Gas temperature, as $\log_{10}(T \, / \, \mathrm{K})$.

y_quantity::Symbol: Quantity for the y axis. The possibilities are:

:stellar_mass -> Stellar mass.

:gas_mass -> Gas mass.

:dm_mass -> Dark matter mass.

:bh_mass -> Black hole mass.

:molecular_mass -> Molecular hydrogen ($\mathrm{H_2}$) mass.

:atomic_mass -> Atomic hydrogen ($\mathrm{HI}$) mass.

:ionized_mass -> Ionized hydrogen ($\mathrm{HII}$) mass.

:neutral_mass -> Neutral hydrogen ($\mathrm{HI + H_2}$) mass.

:molecular_fraction -> Gas mass fraction of molecular hydrogen.

:atomic_fraction -> Gas mass fraction of atomic hydrogen.

:ionized_fraction -> Gas mass fraction of ionized hydrogen.

:neutral_fraction -> Gas mass fraction of neutral hydrogen.

:molecular_neutral_fraction -> Fraction of molecular hydrogen in the neutral gas.

:gas_mass_density -> Gas mass density.

:gas_number_density -> Gas number density.

:molecular_number_density -> Molecular hydrogen number density.

:atomic_number_density -> Atomic hydrogen number density.

:ionized_number_density -> Ionized hydrogen number density.

:neutral_number_density -> Neutral hydrogen number density.

:gas_metallicity -> Mass fraction of all elements above He in the gas (solar units).

:stellar_metallicity -> Mass fraction of all elements above He in the stars (solar units).

:X_gas_abundance -> Gas abundance of element $\mathrm{X}$, as $12 + \log_{10}(\mathrm{X \, / \, H})$. The possibilities are the keys of ELEMENT_INDEX.

:X_stellar_abundance -> Stellar abundance of element $\mathrm{X}$, as $12 + \log_{10}(\mathrm{X \, / \, H})$. The possibilities are the keys of ELEMENT_INDEX.

:stellar_radial_distance -> Distance of every stellar particle to the origin.

:gas_radial_distance -> Distance of every gas cell to the origin.

:dm_radial_distance -> Distance of every dark matter particle to the origin.

:stellar_xy_distance -> Projected distance of every stellar particle to the origin.

:gas_xy_distance -> Projected distance of every gas cell to the origin.

:dm_xy_distance -> Projected distance of every dark matter particle to the origin.

:stellar_circularity -> Stellar circularity.

:stellar_vcirc -> Stellar circular velocity.

:stellar_vradial -> Stellar radial speed.

:stellar_vtangential -> Stellar tangential speed.

:stellar_vzstar -> Stellar speed in the z direction, computed as $v_z \, \sign(z)$.

:stellar_age -> Stellar age.

:sfr -> The star formation rate of the last AGE_RESOLUTION.

:ssfr -> The specific star formation rate of the last AGE_RESOLUTION.

:temperature -> Gas temperature, as $\log_{10}(T \, / \, \mathrm{K})$.

filter_function::Function=filterNothing: A functions with the signature:

filter_function(data_dict) -> indices

where

+ `data_dict::Dict`: A dictionary with the following shape:

    * `:sim_data`          -> ::Simulation (see [`Simulation`](@ref)).
    * `:snap_data`         -> ::Snapshot (see [`Snapshot`](@ref)).
    * `:gc_data`           -> ::GroupCatalog (see [`GroupCatalog`](@ref)).
    * `cell/particle type` -> (`block` -> data of `block`, `block` -> data of `block`, ...).
    * `cell/particle type` -> (`block` -> data of `block`, `block` -> data of `block`, ...).
    * `cell/particle type` -> (`block` -> data of `block`, `block` -> data of `block`, ...).
    * ...
    * `groupcat type`      -> (`block` -> data of `block`, `block` -> data of `block`, ...).
    * `groupcat type`      -> (`block` -> data of `block`, `block` -> data of `block`, ...).
    * `groupcat type`      -> (`block` -> data of `block`, `block` -> data of `block`, ...).
    * ...

+ `indices::Dict`: A dictionary with the following shape:

    * `cell/particle type` -> idxs::IndexType
    * `cell/particle type` -> idxs::IndexType
    * `cell/particle type` -> idxs::IndexType
    * ...

Returns

A tuple with two elements:

A vector with the values of x_quantity.

A vector with the values of y_quantity.

GalaxyInspector.daStellarHistory — Method

daStellarHistory(
    data_dict::Dict;
    <keyword arguments>
)::Union{Tuple{Vector{<:Unitful.Time},Vector{<:Number}},Nothing}

Compute the evolution of a given stellar quantity using the stellar ages at a given instant in time.

Arguments

data_dict::Dict: A dictionary with the following shape:

:sim_data -> ::Simulation (see Simulation).

:snap_data -> ::Snapshot (see Snapshot).

:gc_data -> ::GroupCatalog (see GroupCatalog).

cell/particle type -> (block -> data of block, block -> data of block, ...).

cell/particle type -> (block -> data of block, block -> data of block, ...).

cell/particle type -> (block -> data of block, block -> data of block, ...).

...

groupcat type -> (block -> data of block, block -> data of block, ...).

groupcat type -> (block -> data of block, block -> data of block, ...).

groupcat type -> (block -> data of block, block -> data of block, ...).

...

quantity::Symbol=:sfr: Which quantity will be calculated. The options are:

:sfr -> The star formation rate.

:ssfr -> The specific star formation rate.

:stellar_mass -> Stellar mass.

n_bins::Int=50: Number of bins (time intervals).

Returns

A tuple with two elements:

A vector with the physical time of each bin.

A vector with the values of quantity for each bin.

GalaxyInspector.daVelocityField — Method

daVelocityField(
    data_dict::Dict,
    grid::SquareGrid,
    type_symbol::Symbol;
    <keyword arguments>
)::Tuple{
    Vector{<:Unitful.Length},
    Vector{<:Unitful.Length},
    Matrix{<:Unitful.Velocity},
    Matrix{<:Unitful.Velocity},
}

Compute a 2D mean velocity field.

Note

If the stellar masses and velocities can be found in data_dict, the velocity field is boosted with respect to the stellar center of mass.

Arguments

data_dict::Dict: A dictionary with the following shape:

:sim_data -> ::Simulation (see Simulation).

:snap_data -> ::Snapshot (see Snapshot).

:gc_data -> ::GroupCatalog (see GroupCatalog).

cell/particle type -> (block -> data of block, block -> data of block, ...).

cell/particle type -> (block -> data of block, block -> data of block, ...).

cell/particle type -> (block -> data of block, block -> data of block, ...).

...

groupcat type -> (block -> data of block, block -> data of block, ...).

groupcat type -> (block -> data of block, block -> data of block, ...).

groupcat type -> (block -> data of block, block -> data of block, ...).

...

grid::SquareGrid: Square grid.

type_symbol::Symbol: For which cell/particle type the velocity field will be computed. The possibilities are the keys of PARTICLE_INDEX.

projection_plane::Symbol=:xy: To which plane the cells/particles will be projected. The options are :xy, :xz, and :yz.

Returns

A tuple with four elements:

A vector with the x coordinates of the grid.

A vector with the y coordinates of the grid.

A matrix with the mean velocity in the x direction for each bin.

A matrix with the mean velocity in the y direction for each bin.