Post processing functions

These functions are used internally and none are exported.

A post-processing function must take a Makie figure, add something to it, and return how to label the additions (or nothing when no new labels should be drawn).

Expected signature:

  post_processing(figure, args...; kwargs...) -> ([marker, ...], [label, ...])

where:

figure::Makie.Figure

marker::LegendElement

label::String

GalaxyInspector.ppABFlags! — Method

ppABFlags!(
    figure::Makie.Figure,
    slopes::Vector{<:Real}
    intercepts::Vector{<:Real};
    <keyword arguments>
)::Nothing

Draw lines defined by f(x) = slope * x + intercept.

Arguments

figure::Makie.Figure: Makie figure to be drawn over.

slopes::Vector{<:Real}: List of slopes.

intercepts::Vector{<:Real}: List of intercepts.

colors::Vector{<:ColorType}=[:red]: Colors of the lines.

line_styles::Vector{<:LineStyleType}=[:solid]: Styles of the lines.

GalaxyInspector.ppAgertz2021! — Method

ppAgertz2021!(
    figure::Makie.Figure;
    <keyword arguments>
)::Tuple{Vector{<:LegendElement},Vector{AbstractString}}

Draw the stellar density profiles from McMillan (2011) and Leroy et al. (2008), show in Agertz et al. (2021).

Arguments

figure::Makie.Figure: Makie figure.

galaxies::Vector=["MW"]: Target galaxies. The options are:

One of the 23 galaxies in Leroy et al. (2008), e.g. "DDO154", "HOI", "HOII", "IC2574", "NGC0628", "NGC0925", etc. For a full list see Agertz et al. (2021).

:all: All 23 galaxies in the dataset of Leroy et al. (2008) (this options will plot the data as an scatter plot with transparency instead of a line plot).

"MW": The Milky Way fits from McMillan (2011).

x_unit::Unitful.Units=u"kpc": Unit for the x axis in figure.

y_unit::Unitful.Units=u"Msun * kpc^-2": Unit for the y axis in figure.

x_log::Bool=false: If the x axis will be plotted as the $\log_{10}$ of the galactocentric radius.

y_log::Bool=true: If the y axis will be plotted as the $\log_{10}$ of the stellar surface density.

error_band::Bool=true: If the error band will be plotted.

colors::Vector{<:ColorType}=[WONG_RED]: Colors for stellar profiles.

linestyle::LineStyleType=:solid: Style of the line.

linewidth::Int=3: Width of the lines.

Returns

A tuple with the elements for the legend:

A LineElement to be used as the marker.

The label.

References

O. Agertz et al. (2021). VINTERGATAN – I. The origins of chemically, kinematically, and structurally distinct discs in a simulated Milky Way-mass galaxy. Monthly Notices of the Royal Astronomical Society. **503(4), 5826–5845. doi:10.1093/mnras/stab322

A. K. Leroy et al. (2008). THE STAR FORMATION EFFICIENCY IN NEARBY GALAXIES: MEASURING WHERE GAS FORMS STARS EFFECTIVELY. The Astronomical Journal 136(6), 2782–2845. doi:10.1088/0004-6256/136/6/2782

P. J. McMillan (2011). Mass models of the Milky Way. Monthly Notices of the Royal Astronomical Society 414(3), 2446–2457. doi:10.1111/j.1365-2966.2011.18564.x

GalaxyInspector.ppAnnotation! — Method

ppAnnotation!(figure::Makie.Figure, text::AbstractString; <keyword arguments>)::Nothing

Add an annotation to the plot.

Arguments

figure::Makie.Figure: Makie figure.

text::AbstractString: Text to be written.

position::Tuple{<:Real,<:Real}=(0.04, 0.98): Relative position of the top left corner of the text box.

color=:black: Text color.

fontsize::Int=35: Font size.

GalaxyInspector.ppArrows! — Method

ppArrows!(
    figure::Makie.Figure,
    positions::Vector{NTuple{4,Float64}};
    <keyword arguments>
)::Nothing

Draw arrows.

Arguments

figure::Makie.Figure: Makie figure.

positions::Vector{NTuple{4,Float64}}: The x, y, u, and v coordinates of the arrows. x and y indicate the position of the tails, and u and v the position of the point of the arrows.

colors::Vector{<:ColorType}=[WONG_RED]: Colors of the arrows.

GalaxyInspector.ppBarPlotLabels — Method

ppBarPlotLabels(
    ::Makie.Figure,
    components::Vector{Symbol};
    <keyword arguments>
)::Tuple{Vector{<:LegendElement},Vector{AbstractString}}

Return the legend elements for the plot made by gasBarPlot.

Arguments

::Makie.Figure: Makie figure.

components::Vector{Symbol}: List of gas components to be considered. See COMPONENTS for options.

colors=Makie.wong_colors(): Colors of the bars.

Returns

A tuple with the elements for the legend:

PolyElements to be used in the legend.

The label strings.

GalaxyInspector.ppBigiel2008! — Method

ppBigiel2008!(
    figure::Makie.Figure,
    molecular::Bool;
    <keyword arguments>
)::Union{Tuple{Vector{<:LegendElement},Vector{AbstractString}},Nothing}

Draw a line plot with the fit for the Kennicutt-Schmidt relation from Bigiel et al. (2008).

Note

The resolution used in Bigiel et al. (2008) is 750 pc (see Section 1).

Arguments

figure::Makie.Figure: Makie figure.

molecular::Bool: If the x axis will be the area mass density of molecular hydrogen, or, if set to false, the area mass density of neutral hydrogen.

x_unit::Unitful.Units=u"Msun * pc^-2": Unit for the area density of gas used in figure.

y_unit::Unitful.Units=u"Msun * yr^-1 * kpc^-2": Unit for the area density of star formation rate used in figure.

x_log::Bool=true: If the x axis is $\log_{10}(\Sigma_\mathrm{H})$ (x_log = true) or just $\Sigma_\mathrm{H}$ (x_log = false).

y_log::Bool=true: If the y axis is $\log_{10}(\Sigma_\mathrm{SFR})$ (y_log = true) or just $\Sigma_\mathrm{SFR}$ (y_log = false).

extend::Float64=0.0: By default the y axis limits will be the vertical range of the points in the plot. This can be extended by the multiplicative factor extend of the vertical range.

color::ColorType=WONG_RED: Color for the line.

linestyles::Vector{<:LineStyleType}=[:solid, :dash],: Styles for the lines. The first style will indicate the range for which there are experimental data, and the second one will be for the extrapolation.

linewidth::Int=3: Line width.

Returns

A tuple with the elements for the legend:

A LineElement to be used as the marker.

The label.

References

F. Bigiel et al. (2008). THE STAR FORMATION LAW IN NEARBY GALAXIES ON SUB-KPC SCALES. The Astrophysical Journal, 136(6), 2846. doi:10.1088/0004-6256/136/6/2846

GalaxyInspector.ppBigiel2010! — Method

ppBigiel2010!(
    figure::Makie.Figure;
    <keyword arguments>
)::Tuple{Vector{<:LegendElement},Vector{AbstractString}}

Draw a scatter plot of the SFR surface density vs gas surface density (Kennicutt-Schmidt law) for a given galaxy, using the data from Bigiel et al. (2010).

Note

The resolution used in Bigiel et al. (2010) is 600 pc (see Section 2).

Arguments

figure::Makie.Figure: Makie figure.

galaxy::Union{String,Symbol}="NGC 628": Target galaxy. The options are:

With molecular and and atomic data (Table 2): "NGC 628", "NGC 3184", "NGC 3521", "NGC 4736", "NGC 5055", "NGC 5194", "NGC 6946".

With only atomic data (Table 3): "NGC 925", "NGC 2403", "NGC 2841", "NGC 2903", "NGC 3198", "NGC 3351", "NGC 3621", "NGC 3627", "NGC 5236", "NGC 5457", "NGC 7331", "NGC 7793".

:all: Every galaxy that is available for the given quantity. For more information on each galaxy see Bigiel et al. (2010).

quantity::Symbol=:molecular: Gas quantity for the x axis. The options are:

:molecular -> Surface density of molecular gas.

:neutral -> Surface density of neutral gas.

:atomic -> Surface density of atomic gas.

x_unit::Unitful.Units=u"Msun * pc^-2": Unit for the area density of gas used in figure.

y_unit::Unitful.Units=u"Msun * yr^-1 * kpc^-2": Unit for the area density of star formation rate used in figure.

x_log::Bool=true: If the x axis will be plotted as the $\log_{10}$ of the gas surface density.

y_log::Bool=true: If the y axis will be plotted as the $\log_{10}$ of the SFR surface density.

color::ColorType=WONG_RED: Color of the markers.

Returns

A tuple with the elements for the legend:

A MarkerElement to be used as the marker.

The label.

References

F. Bigiel et al. (2010). EXTREMELY INEFFICIENT STAR FORMATION IN THE OUTER DISKS OF NEARBY GALAXIES. The Astrophysical Journal, 140(5), 1194. doi:10.1088/0004-6256/140/5/1194

GalaxyInspector.ppCross! — Method

ppCross!(
    figure::Makie.Figure,
    cross_point::Tuple{<:Real,<:Real};
    <keyword arguments>
)::Union{Tuple{Vector{<:LegendElement},Vector{AbstractString}},Nothing}

Draw two lines, one horizontal and one vertical.

Arguments

figure::Makie.Figure: Makie figure.

cross_point::Tuple{<:Real,<:Real}: Crossing point of the lines.

color::ColorType=WONG_RED: Color of the lines.

linestyle::LineStyleType=:solid: Style of the lines.

label::Union{AbstractString,Nothing}=nothing: Label for the lines. If set to nothing no label is printed.

Returns

A tuple with the elements for the legend:

A LineElement to be used as the marker.

The label.

GalaxyInspector.ppFeldmann2020! — Method

ppFeldmann2020!(
    figure::Makie.Figure,
    x_quantity::Symbol,
    y_quantity::Symbol;
    <keyword arguments>
)::Tuple{Vector{<:LegendElement},Vector{AbstractString}}

Draw the experimental data from xGASS and xCOLD GASS, compiled by Feldmann (2020), as a line or scatter plot.

Arguments

figure::Makie.Figure: Makie figure.

x_quantity::Symbol: Quantity for the x axis. The options are:

:stellar -> Stellar mass.

:molecular -> Molecular mass.

:atomic -> Atomic mass.

:sfr -> Star formation rate of the last AGE_RESOLUTION.

y_quantity::Symbol: Quantity for the y axis. The options are:

:stellar -> Stellar mass.

:molecular -> Molecular mass.

:atomic -> Atomic mass.

:sfr -> Star formation rate of the last AGE_RESOLUTION.

scatter::Bool=false: If the data will be presented as a line plot with error bands (default), or alternatively, a scatter plot.

xlog::Bool=true: If the x axis will be plotted as the $\log_{10}$ of the x quantity.

ylog::Bool=true: If the y axis will be plotted as the $\log_{10}$ of the y quantity.

Returns

A tuple with the elements for the legend:

A MarkerElement to be used as the marker.

The label.

References

R. Feldmann (2020). The link between star formation and gas in nearby galaxies. Communications Physics 3(226). doi:10.1038/s42005-020-00493-0

GalaxyInspector.ppFillBelowLine! — Method

ppFillBelowLine!(
    figure::Makie.Figure;
    <keyword arguments>
)::Union{Tuple{Vector{<:LegendElement},Vector{AbstractString}},Nothing}

Fill the space below a line plot with a solid color down to a lower_limit.

Arguments

figure::Makie.Figure: Makie figure.

lower_limit::Number=0.0: Lower bound.

color::ColorType=WONG_RED: Color.

alpha::Float64=0.3: Level of transparency. 1.0 is completely opaque and 0.0 is completely transparent.

label::Union{AbstractString,Nothing}=nothing: Label for the shaded region. If set to nothing no label is printed.

Returns

A tuple with the elements for the legend:

A MarkerElement to be used as the marker.

The label.

GalaxyInspector.ppFitLine! — Method

ppFitLine!(
    figure::Makie.Figure;
    <keyword arguments>
)::Union{Tuple{Vector{<:LegendElement},Vector{AbstractString}},Nothing}

Draw a linear fit for the data in figure.

An annotation with the equation $y = a \, x + b$, and the fitted values for $a$ and $b$, will be positioned in the upper right corner of the plot.

Arguments

figure::Makie.Figure: Makie figure.

text_position::Tuple{<:Real,<:Real}=(0.04, 0.98): Relative position of the top left corner of the text box.

text_align::NTuple{2,Symbol}=(:left, :top): Corner of the legend box whose positions is set with text_position.

weights::Union{Vector{Float64},Nothing}=nothing: Weights for the fits. Set to nothing for a non-weighted fit.

error_formating::Symbol=:std_error: Error format for the annotation. The options are:

:std_error -> mean ± standard_error.

:conf_interval -> mean ± max(upper$_{95\%}$ - mean, mean - lower$_{95\%}$).

interval::Symbol=:prediction: Type of prediction band for the linear fit. The options are:

:confidence: Confidence interval.

:prediction: Prediction interval.

color::ColorType=WONG_RED: Color of the line.

linestyle::LineStyleType=:solid: Style of the line.

linewidth::Int=3: Line width.

Returns

A tuple with the elements for the legend:

A LineElement to be used as the marker.

The label.

GalaxyInspector.ppHorizontalFlags! — Method

ppHorizontalFlags!(
    figure::Makie.Figure,
    positions::Vector{<:Real};
    <keyword arguments>
)::Union{Tuple{Vector{<:LegendElement},Vector{AbstractString}},Nothing}

Draw horizontal lines.

Arguments

figure::Makie.Figure: Makie figure.

positions::Vector{<:Real}: The y coordinates of the lines.

colors::Vector{<:ColorType}=[WONG_RED]: Colors of the lines.

line_styles::Vector{<:LineStyleType}=[:solid]: Styles of the lines.

labels::Union{Vector{AbstractString},Nothing}=nothing: Labels for the lines. If set to nothing no label is printed.

Returns

A tuple with the elements for the legend:

LineElements to be used as the marker.

The labels.

GalaxyInspector.ppKennicutt1998! — Method

ppKennicutt1998!(
    figure::Makie.Figure;
    <keyword arguments>
)::Union{Tuple{Vector{<:LegendElement},Vector{AbstractString}},Nothing}

Draw a line plot with the fit for the Kennicutt-Schmidt relation from Kennicutt (1998).

Arguments

figure::Makie.Figure: Makie figure.

x_unit::Unitful.Units=u"Msun * pc^-2": Unit for the area density of gas used in figure.

y_unit::Unitful.Units=u"Msun * yr^-1 * kpc^-2": Unit for the area density of star formation rate used in figure.

x_log::Bool=true: If the x axis is $\log_{10}(\Sigma_\mathrm{HI + H_2})$ (x_log = true) or just $\Sigma_\mathrm{HI + H_2}$ (x_log = false).

y_log::Bool=true: If the y axis is $\log_{10}(\Sigma_\mathrm{SFR})$ (y_log = true) or just $\Sigma_\mathrm{SFR}$ (y_log = false).

extend::Float64=0.0: By default the y axis limits of the line will be the vertical range of points in the plot. This can be extended by the fraction extend of the vertical range.

color::ColorType=WONG_RED: Color for the line.

linestyles::Vector{<:LineStyleType}=[:solid, :dash],: Styles for the lines. The first style will indicate the range for which there are experimental data, and the second one will be for the extrapolation.

linewidth::Int=3: Line width.

Returns

A tuple with the elements for the legend:

A LineElement to be used as the marker.

The label.

References

R. C. Kennicutt (1998). The Global Schmidt Law in Star-forming Galaxies. The Astrophysical Journal, 498(2), 541-552. doi:10.1086/305588

GalaxyInspector.ppLeroy2008! — Method

ppLeroy2008!(
    figure::Makie.Figure;
    <keyword arguments>
)::Tuple{Vector{<:LegendElement},Vector{AbstractString}}

Draw a scatter plot of the SFR surface density vs gas surface density (Kennicutt-Schmidt law) for a given galaxy, using the data from Leroy et al. (2008).

Note

The resolution used in Leroy et al. (2008) is 800 pc for spirals and 400 pc for dwarf galaxies (Section 3.1).

Arguments

figure::Makie.Figure: Makie figure.

quantity::Symbol=:molecular: Gas quantity for the x axis. The options are:

:molecular -> Surface density of molecular gas.

:neutral -> Surface density of neutral gas.

:atomic -> Surface density of atomic gas.

galaxies::Vector=["MW"]: Target galaxies. The options are:

One of the 23 galaxies in Leroy et al. (2008), e.g. "DDO154", "HOI", "HOII", "IC2574", "NGC0628", "NGC0925", etc. For a full list see Leroy et al. (2008).

:all: All 23 galaxies in Leroy et al. (2008).

x_unit::Unitful.Units=u"Msun * pc^-2": Unit for the area density of gas used in figure.

y_unit::Unitful.Units=u"Msun * yr^-1 * kpc^-2": Unit for the area density of star formation rate used in figure.

x_log::Bool=true: If the x axis will be plotted as the $\log_{10}$ of the gas surface density.

y_log::Bool=true: If the y axis will be plotted as the $\log_{10}$ of the SFR surface density.

error_bars::Bool=false: If error bars will be plotted.

color::ColorType=WONG_RED,: Color for the markers.

Returns

A tuple with the elements for the legend:

A MarkerElement to be used as the marker.

The label.

References

A. K. Leroy et al. (2008). THE STAR FORMATION EFFICIENCY IN NEARBY GALAXIES: MEASURING WHERE GAS FORMS STARS EFFECTIVELY. The Astronomical Journal 136(6), 2782–2845. doi:10.1088/0004-6256/136/6/2782

GalaxyInspector.ppMolla2015! — Method

ppMolla2015!(
    figure::Makie.Figure,
    quantity::Symbol,
    <keyword arguments>
)::Tuple{Vector{<:LegendElement},Vector{AbstractString}}

Draw a profile for the Milky Way using the data compiled by Mollá et al. (2015).

Arguments

figure::Makie.Figure: Makie figure.

quantity::Symbol: Quantity for the y axis. See compareMolla2015 for options.

y_unit::Unitful.Units=Unitful.NoUnits: Target unit for quantity.

color::ColorType=WONG_RED: Color of the line.

linestyle::LineStyleType=:solid: Style of the line.

error_bars::Bool=true: If the error bars will be plotted.

Returns

A tuple with the elements for the legend:

A MarkerElement to be used as the marker.

The label.

References

M. Mollá et al. (2015). Galactic chemical evolution: stellar yields and the initial mass function. Monthly Notices of the Royal Astronomical Society 451(4), 3693–3708. doi:10.1093/mnras/stv1102

GalaxyInspector.ppSun2023! — Method

ppSun2023!(
    figure::Makie.Figure;
    <keyword arguments>
)::Tuple{Vector{<:LegendElement},Vector{AbstractString}}

Draw a scatter plot of the SFR surface density vs molecular surface density (molecular Kennicutt-Schmidt law) for a given galaxy, using the data from Sun et al. (2023).

Note

The resolution used in Sun et al. (2023) is 1.5 kpc (see Section 2).

Arguments

figure::Makie.Figure: Makie figure.

galaxy::Union{String,Symbol}=:main: Target galaxy. The options are:

One of the 80 galaxies in the dataset, e.g. "ESO097-013", "IC1954", "IC5273", "NGC1546", "NGC1559", "NGC1566", etc. For a full list see the reference below.

:main: Every galaxy with $-2.0 < \log_{10}(t_\mathrm{dep} \, / \, \mathrm{Gyr}) < 2.0$, where $t_\mathrm{dep} = \Sigma_\mathrm{H_2} / \Sigma_\mathrm{SFR}$ is the depletion time.

:all: All 80 galaxies in the dataset. For more information on each galaxy see Sun et al. (2023).

sfr_calibration::Symbol=:Halpha: SFR calibration for combining UV, optical, and/or IR data. The options are: :Halpha, :FUV, and :AV_corrected_Halpha. For an explanation of each one see section 2 of Sun et al. (2023).

h2_prescription::Symbol=:S20: Prescription for the CO-to-H₂ conversion factor. The options are: :S20, :Mw, :B13, and :G20. For an explanation of each one see section 2 of Sun et al. (2023).

x_unit::Unitful.Units=u"Msun * pc^-2": Unit for the area density of gas used in figure.

y_unit::Unitful.Units=u"Msun * yr^-1 * kpc^-2": Unit for the area density of star formation rate used in figure.

x_log::Bool=true: If the x axis will be plotted as the $\\log_{10}$ of the gas surface density.

y_log::Bool=true: If the y axis will be plotted as the $\\log_{10}$ of the SFR surface density.

color::ColorType=WONG_RED: Color of the markers.

Returns

A tuple with the elements for the legend:

A MarkerElement to be used as the marker.

The label.

References

J. Sun et al. (2023). Star Formation Laws and Efficiencies across 80 Nearby Galaxies. The Astrophysical Journal Letters, 945(2), L19. doi:10.3847/2041-8213/acbd9c

GalaxyInspector.ppVerticalFlags! — Method

ppVerticalFlags!(
    figure::Makie.Figure,
    positions::Vector{<:Real};
    <keyword arguments>
)::Union{Tuple{Vector{<:LegendElement},Vector{AbstractString}},Nothing}

Draw vertical lines.

Arguments

figure::Makie.Figure: Makie figure.

positions::Vector{<:Real}: The x coordinates of the lines.

colors::Vector{<:ColorType}=[WONG_RED]: Colors of the lines.

line_styles::Vector{<:LineStyleType}=[:solid]: Styles of the lines.

labels::Union{Vector{AbstractString},Nothing}=nothing: Labels for the lines. If set to nothing no label is printed.

Returns

A tuple with the elements for the legend:

LineElements to be used as the marker.

The labels.